% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_create_connection}
\alias{apprunner_create_connection}
\title{Create an App Runner connection resource}
\usage{
apprunner_create_connection(ConnectionName, ProviderType, Tags = NULL)
}
\arguments{
\item{ConnectionName}{[required] A name for the new connection. It must be unique across all App Runner
connections for the Amazon Web Services account in the Amazon Web
Services Region.}

\item{ProviderType}{[required] The source repository provider.}

\item{Tags}{A list of metadata items that you can associate with your connection
resource. A tag is a key-value pair.}
}
\description{
Create an App Runner connection resource. App Runner requires a connection resource when you create App Runner services that access private repositories from certain third-party providers. You can share a connection across multiple services.

See \url{https://www.paws-r-sdk.com/docs/apprunner_create_connection/} for full documentation.
}
\keyword{internal}
