% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_create_vpc_ingress_connection}
\alias{apprunner_create_vpc_ingress_connection}
\title{Create an App Runner VPC Ingress Connection resource}
\usage{
apprunner_create_vpc_ingress_connection(
  ServiceArn,
  VpcIngressConnectionName,
  IngressVpcConfiguration,
  Tags = NULL
)
}
\arguments{
\item{ServiceArn}{[required] The Amazon Resource Name (ARN) for this App Runner service that is used
to create the VPC Ingress Connection resource.}

\item{VpcIngressConnectionName}{[required] A name for the VPC Ingress Connection resource. It must be unique across
all the active VPC Ingress Connections in your Amazon Web Services
account in the Amazon Web Services Region.}

\item{IngressVpcConfiguration}{[required] Specifications for the customer’s Amazon VPC and the related Amazon Web
Services PrivateLink VPC endpoint that are used to create the VPC
Ingress Connection resource.}

\item{Tags}{An optional list of metadata items that you can associate with the VPC
Ingress Connection resource. A tag is a key-value pair.}
}
\description{
Create an App Runner VPC Ingress Connection resource. App Runner requires this resource when you want to associate your App Runner service with an Amazon VPC endpoint.

See \url{https://www.paws-r-sdk.com/docs/apprunner_create_vpc_ingress_connection/} for full documentation.
}
\keyword{internal}
