% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_describe_observability_configuration}
\alias{apprunner_describe_observability_configuration}
\title{Return a full description of an App Runner observability configuration
resource}
\usage{
apprunner_describe_observability_configuration(ObservabilityConfigurationArn)
}
\arguments{
\item{ObservabilityConfigurationArn}{[required] The Amazon Resource Name (ARN) of the App Runner observability
configuration that you want a description for.

The ARN can be a full observability configuration ARN, or a partial ARN
ending with either \code{.../name } or \code{.../name/revision }. If a revision
isn't specified, the latest active revision is described.}
}
\description{
Return a full description of an App Runner observability configuration resource.

See \url{https://www.paws-r-sdk.com/docs/apprunner_describe_observability_configuration/} for full documentation.
}
\keyword{internal}
