% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_describe_service}
\alias{apprunner_describe_service}
\title{Return a full description of an App Runner service}
\usage{
apprunner_describe_service(ServiceArn)
}
\arguments{
\item{ServiceArn}{[required] The Amazon Resource Name (ARN) of the App Runner service that you want a
description for.}
}
\description{
Return a full description of an App Runner service.

See \url{https://www.paws-r-sdk.com/docs/apprunner_describe_service/} for full documentation.
}
\keyword{internal}
