% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_list_services}
\alias{apprunner_list_services}
\title{Returns a list of running App Runner services in your Amazon Web
Services account}
\usage{
apprunner_list_services(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A token from a previous result page. Used for a paginated request. The
request retrieves the next result page. All other parameter values must
be identical to the ones specified in the initial request.

If you don't specify \code{NextToken}, the request retrieves the first result
page.}

\item{MaxResults}{The maximum number of results to include in each response (result page).
It's used for a paginated request.

If you don't specify \code{MaxResults}, the request retrieves all available
results in a single response.}
}
\description{
Returns a list of running App Runner services in your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/apprunner_list_services/} for full documentation.
}
\keyword{internal}
