% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_list_tags_for_resource}
\alias{apprunner_list_tags_for_resource}
\title{List tags that are associated with for an App Runner resource}
\usage{
apprunner_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that a tag list is
requested for.

It must be the ARN of an App Runner resource.}
}
\description{
List tags that are associated with for an App Runner resource. The response contains a list of tag key-value pairs.

See \url{https://www.paws-r-sdk.com/docs/apprunner_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
