% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_pause_service}
\alias{apprunner_pause_service}
\title{Pause an active App Runner service}
\usage{
apprunner_pause_service(ServiceArn)
}
\arguments{
\item{ServiceArn}{[required] The Amazon Resource Name (ARN) of the App Runner service that you want
to pause.}
}
\description{
Pause an active App Runner service. App Runner reduces compute capacity for the service to zero and loses state (for example, ephemeral storage is removed).

See \url{https://www.paws-r-sdk.com/docs/apprunner_pause_service/} for full documentation.
}
\keyword{internal}
