% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_resume_service}
\alias{apprunner_resume_service}
\title{Resume an active App Runner service}
\usage{
apprunner_resume_service(ServiceArn)
}
\arguments{
\item{ServiceArn}{[required] The Amazon Resource Name (ARN) of the App Runner service that you want
to resume.}
}
\description{
Resume an active App Runner service. App Runner provisions compute capacity for the service.

See \url{https://www.paws-r-sdk.com/docs/apprunner_resume_service/} for full documentation.
}
\keyword{internal}
