% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_start_deployment}
\alias{apprunner_start_deployment}
\title{Initiate a manual deployment of the latest commit in a source code
repository or the latest image in a source image repository to an App
Runner service}
\usage{
apprunner_start_deployment(ServiceArn)
}
\arguments{
\item{ServiceArn}{[required] The Amazon Resource Name (ARN) of the App Runner service that you want
to manually deploy to.}
}
\description{
Initiate a manual deployment of the latest commit in a source code repository or the latest image in a source image repository to an App Runner service.

See \url{https://www.paws-r-sdk.com/docs/apprunner_start_deployment/} for full documentation.
}
\keyword{internal}
