% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_update_vpc_ingress_connection}
\alias{apprunner_update_vpc_ingress_connection}
\title{Update an existing App Runner VPC Ingress Connection resource}
\usage{
apprunner_update_vpc_ingress_connection(
  VpcIngressConnectionArn,
  IngressVpcConfiguration
)
}
\arguments{
\item{VpcIngressConnectionArn}{[required] The Amazon Resource Name (Arn) for the App Runner VPC Ingress Connection
resource that you want to update.}

\item{IngressVpcConfiguration}{[required] Specifications for the customer’s Amazon VPC and the related Amazon Web
Services PrivateLink VPC endpoint that are used to update the VPC
Ingress Connection resource.}
}
\description{
Update an existing App Runner VPC Ingress Connection resource. The VPC Ingress Connection must be in one of the following states to be updated:

See \url{https://www.paws-r-sdk.com/docs/apprunner_update_vpc_ingress_connection/} for full documentation.
}
\keyword{internal}
