% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_service.R
\name{batch}
\alias{batch}
\title{AWS Batch}
\usage{
batch(config = list(), credentials = list(), endpoint = NULL, region = NULL)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
\item{\strong{use_dual_stack}: Set this to \code{true} to use the dualstack (IPv4 and IPv6) endpoint for a service, where available, falling back to the regular endpoint when it isn't. Defaults to the \code{AWS_USE_DUALSTACK_ENDPOINT} environment variable when unset.}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Batch

Using Batch, you can run batch computing workloads on the Amazon Web Services Cloud. Batch computing is a common means for developers, scientists, and engineers to access large amounts of compute resources. Batch uses the advantages of the batch computing to remove the undifferentiated heavy lifting of configuring and managing required infrastructure. At the same time, it also adopts a familiar batch computing software approach. You can use Batch to efficiently provision resources, and work toward eliminating capacity constraints, reducing your overall compute costs, and delivering results more quickly.

As a fully managed service, Batch can run batch computing workloads of any scale. Batch automatically provisions compute resources and optimizes workload distribution based on the quantity and scale of your specific workloads. With Batch, there's no need to install or manage batch computing software. This means that you can focus on analyzing results and solving your specific problems instead.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- batch(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string",
    use_dual_stack = "logical"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=batch_cancel_job]{cancel_job} \tab Cancels a job in an Batch job queue\cr
\link[=batch_cancel_jobs]{cancel_jobs} \tab Cancels up to 50 jobs in an Batch job queue\cr
\link[=batch_create_compute_environment]{create_compute_environment} \tab Creates an Batch compute environment\cr
\link[=batch_create_consumable_resource]{create_consumable_resource} \tab Creates an Batch consumable resource\cr
\link[=batch_create_job_queue]{create_job_queue} \tab Creates an Batch job queue\cr
\link[=batch_create_quota_share]{create_quota_share} \tab Creates an Batch quota share\cr
\link[=batch_create_scheduling_policy]{create_scheduling_policy} \tab Creates an Batch scheduling policy\cr
\link[=batch_create_service_environment]{create_service_environment} \tab Creates a service environment for running service jobs\cr
\link[=batch_delete_compute_environment]{delete_compute_environment} \tab Deletes an Batch compute environment\cr
\link[=batch_delete_consumable_resource]{delete_consumable_resource} \tab Deletes the specified consumable resource\cr
\link[=batch_delete_job_queue]{delete_job_queue} \tab Deletes the specified job queue\cr
\link[=batch_delete_quota_share]{delete_quota_share} \tab Deletes the specified quota share\cr
\link[=batch_delete_scheduling_policy]{delete_scheduling_policy} \tab Deletes the specified scheduling policy\cr
\link[=batch_delete_service_environment]{delete_service_environment} \tab Deletes a Service environment\cr
\link[=batch_deregister_job_definition]{deregister_job_definition} \tab Deregisters an Batch job definition\cr
\link[=batch_describe_compute_environments]{describe_compute_environments} \tab Describes one or more of your compute environments\cr
\link[=batch_describe_consumable_resource]{describe_consumable_resource} \tab Returns a description of the specified consumable resource\cr
\link[=batch_describe_job_definitions]{describe_job_definitions} \tab Describes a list of job definitions\cr
\link[=batch_describe_job_queues]{describe_job_queues} \tab Describes one or more of your job queues\cr
\link[=batch_describe_jobs]{describe_jobs} \tab Describes a list of Batch jobs\cr
\link[=batch_describe_quota_share]{describe_quota_share} \tab Returns a description of the specified quota share\cr
\link[=batch_describe_scheduling_policies]{describe_scheduling_policies} \tab Describes one or more of your scheduling policies\cr
\link[=batch_describe_service_environments]{describe_service_environments} \tab Describes one or more of your service environments\cr
\link[=batch_describe_service_job]{describe_service_job} \tab The details of a service job\cr
\link[=batch_get_job_queue_snapshot]{get_job_queue_snapshot} \tab Provides a snapshot of job queue state, including ordering of RUNNABLE jobs, as well as capacity utilization for already dispatched jobs\cr
\link[=batch_list_consumable_resources]{list_consumable_resources} \tab Returns a list of Batch consumable resources\cr
\link[=batch_list_jobs]{list_jobs} \tab Returns a list of Batch jobs\cr
\link[=batch_list_jobs_by_consumable_resource]{list_jobs_by_consumable_resource} \tab Returns a list of Batch jobs that require a specific consumable resource\cr
\link[=batch_list_quota_shares]{list_quota_shares} \tab Returns a list of Batch quota shares associated with a job queue\cr
\link[=batch_list_scheduling_policies]{list_scheduling_policies} \tab Returns a list of Batch scheduling policies\cr
\link[=batch_list_service_jobs]{list_service_jobs} \tab Returns a list of service jobs for a specified job queue\cr
\link[=batch_list_tags_for_resource]{list_tags_for_resource} \tab Lists the tags for an Batch resource\cr
\link[=batch_register_job_definition]{register_job_definition} \tab Registers an Batch job definition\cr
\link[=batch_submit_job]{submit_job} \tab Submits an Batch job from a job definition\cr
\link[=batch_submit_service_job]{submit_service_job} \tab Submits a service job to a specified job queue to run on SageMaker AI\cr
\link[=batch_tag_resource]{tag_resource} \tab Associates the specified tags to a resource with the specified resourceArn\cr
\link[=batch_terminate_job]{terminate_job} \tab Terminates a job in a job queue\cr
\link[=batch_terminate_jobs]{terminate_jobs} \tab Terminates up to 50 jobs in a job queue\cr
\link[=batch_terminate_service_job]{terminate_service_job} \tab Terminates a service job in a job queue\cr
\link[=batch_terminate_service_jobs]{terminate_service_jobs} \tab Terminates up to 50 service jobs in a job queue\cr
\link[=batch_untag_resource]{untag_resource} \tab Deletes specified tags from an Batch resource\cr
\link[=batch_update_compute_environment]{update_compute_environment} \tab Updates an Batch compute environment\cr
\link[=batch_update_consumable_resource]{update_consumable_resource} \tab Updates a consumable resource\cr
\link[=batch_update_job_queue]{update_job_queue} \tab Updates a job queue\cr
\link[=batch_update_quota_share]{update_quota_share} \tab Updates a quota share\cr
\link[=batch_update_scheduling_policy]{update_scheduling_policy} \tab Updates a scheduling policy\cr
\link[=batch_update_service_environment]{update_service_environment} \tab Updates a service environment\cr
\link[=batch_update_service_job]{update_service_job} \tab Updates the priority of a specified service job in an Batch job queue
}
}

\examples{
\dontrun{
svc <- batch()
# This example cancels a job with the specified job ID.
svc$cancel_job(
  jobId = "1d828f65-7a4d-42e8-996d-3b900ed59dc4",
  reason = "Cancelling job."
)
}

}
