% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_cancel_jobs}
\alias{batch_cancel_jobs}
\title{Cancels up to 50 jobs in an Batch job queue}
\usage{
batch_cancel_jobs(jobs, reason)
}
\arguments{
\item{jobs}{[required] An array of up to 50 Batch job IDs of the jobs to cancel.}

\item{reason}{[required] A message to attach to the job that explains the reason for cancelling it. This message is returned by future \code{\link[=batch_describe_jobs]{describe_jobs}} operations on the job. It is also recorded in the Batch activity logs.

This parameter has a limit of 1024 characters.}
}
\description{
Cancels up to 50 jobs in an Batch job queue. This is a bulk version of \code{\link[=batch_cancel_job]{cancel_job}}. Jobs that are in a \code{SUBMITTED}, \code{PENDING}, or \code{RUNNABLE} state are cancelled and the job status is updated to \code{FAILED}.

See \url{https://www.paws-r-sdk.com/docs/batch_cancel_jobs/} for full documentation.
}
\keyword{internal}
