% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_create_consumable_resource}
\alias{batch_create_consumable_resource}
\title{Creates an Batch consumable resource}
\usage{
batch_create_consumable_resource(
  consumableResourceName,
  totalQuantity = NULL,
  resourceType = NULL,
  tags = NULL
)
}
\arguments{
\item{consumableResourceName}{[required] The name of the consumable resource. Must be unique.}

\item{totalQuantity}{The total amount of the consumable resource that is available. Must be
non-negative.}

\item{resourceType}{Indicates whether the resource is available to be re-used after a job
completes. Can be one of:
\itemize{
\item \code{REPLENISHABLE} (default)
\item \code{NON_REPLENISHABLE}
}}

\item{tags}{The tags that you apply to the consumable resource to help you
categorize and organize your resources. Each tag consists of a key and
an optional value. For more information, see \href{https://docs.aws.amazon.com/batch/latest/userguide/using-tags.html}{Tagging your Batch resources}.}
}
\description{
Creates an Batch consumable resource.

See \url{https://www.paws-r-sdk.com/docs/batch_create_consumable_resource/} for full documentation.
}
\keyword{internal}
