% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_delete_compute_environment}
\alias{batch_delete_compute_environment}
\title{Deletes an AWS Batch compute environment}
\usage{
batch_delete_compute_environment(computeEnvironment)
}
\arguments{
\item{computeEnvironment}{[required] The name or Amazon Resource Name (ARN) of the compute environment to
delete.}
}
\value{
An empty list.
}
\description{
Deletes an AWS Batch compute environment.

Before you can delete a compute environment, you must set its state to
\code{DISABLED} with the
\code{\link[=batch_update_compute_environment]{update_compute_environment}} API
operation and disassociate it from any job queues with the
\code{\link[=batch_update_job_queue]{update_job_queue}} API operation. Compute
environments that use AWS Fargate resources must terminate all active
jobs on that compute environment before deleting the compute
environment. If this isn't done, the compute environment enters an
invalid state.
}
\section{Request syntax}{
\preformatted{svc$delete_compute_environment(
  computeEnvironment = "string"
)
}
}

\examples{
\dontrun{
# This example deletes the P2OnDemand compute environment.
svc$delete_compute_environment(
  computeEnvironment = "P2OnDemand"
)
}

}
\keyword{internal}
