% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_delete_consumable_resource}
\alias{batch_delete_consumable_resource}
\title{Deletes the specified consumable resource}
\usage{
batch_delete_consumable_resource(consumableResource)
}
\arguments{
\item{consumableResource}{[required] The name or ARN of the consumable resource that will be deleted.}
}
\description{
Deletes the specified consumable resource.

See \url{https://www.paws-r-sdk.com/docs/batch_delete_consumable_resource/} for full documentation.
}
\keyword{internal}
