% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_delete_job_queue}
\alias{batch_delete_job_queue}
\title{Deletes the specified job queue}
\usage{
batch_delete_job_queue(jobQueue)
}
\arguments{
\item{jobQueue}{[required] The short name or full Amazon Resource Name (ARN) of the queue to
delete.}
}
\value{
An empty list.
}
\description{
Deletes the specified job queue. You must first disable submissions for
a queue with the \code{\link[=batch_update_job_queue]{update_job_queue}} operation.
All jobs in the queue are eventually terminated when you delete a job
queue. The jobs are terminated at a rate of about 16 jobs each second.

It's not necessary to disassociate compute environments from a queue
before submitting a \code{\link[=batch_delete_job_queue]{delete_job_queue}}
request.
}
\section{Request syntax}{
\preformatted{svc$delete_job_queue(
  jobQueue = "string"
)
}
}

\examples{
\dontrun{
# This example deletes the GPGPU job queue.
svc$delete_job_queue(
  jobQueue = "GPGPU"
)
}

}
\keyword{internal}
