% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_delete_quota_share}
\alias{batch_delete_quota_share}
\title{Deletes the specified quota share}
\usage{
batch_delete_quota_share(quotaShareArn)
}
\arguments{
\item{quotaShareArn}{[required] The Amazon Resource Name (ARN) of the quota share.}
}
\description{
Deletes the specified quota share. You must first disable submissions for the share by updating the state to \code{DISABLED} using the \code{\link[=batch_update_quota_share]{update_quota_share}} operation. All jobs in the share are eventually terminated when you delete a quota share.

See \url{https://www.paws-r-sdk.com/docs/batch_delete_quota_share/} for full documentation.
}
\keyword{internal}
