% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_delete_scheduling_policy}
\alias{batch_delete_scheduling_policy}
\title{Deletes the specified scheduling policy}
\usage{
batch_delete_scheduling_policy(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the scheduling policy to delete.}
}
\description{
Deletes the specified scheduling policy.

See \url{https://www.paws-r-sdk.com/docs/batch_delete_scheduling_policy/} for full documentation.
}
\keyword{internal}
