% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_delete_service_environment}
\alias{batch_delete_service_environment}
\title{Deletes a Service environment}
\usage{
batch_delete_service_environment(serviceEnvironment)
}
\arguments{
\item{serviceEnvironment}{[required] The name or ARN of the service environment to delete.}
}
\description{
Deletes a Service environment. Before you can delete a service environment, you must first set its state to \code{DISABLED} with the \code{\link[=batch_update_service_environment]{update_service_environment}} API operation and disassociate it from any job queues with the \code{\link[=batch_update_job_queue]{update_job_queue}} API operation.

See \url{https://www.paws-r-sdk.com/docs/batch_delete_service_environment/} for full documentation.
}
\keyword{internal}
