% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_deregister_job_definition}
\alias{batch_deregister_job_definition}
\title{Deregisters an Batch job definition}
\usage{
batch_deregister_job_definition(jobDefinition)
}
\arguments{
\item{jobDefinition}{[required] The name and revision (\code{name:revision}) or full Amazon Resource Name
(ARN) of the job definition to deregister.}
}
\description{
Deregisters an Batch job definition. Job definitions are permanently deleted after 180 days.

See \url{https://www.paws-r-sdk.com/docs/batch_deregister_job_definition/} for full documentation.
}
\keyword{internal}
