% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_describe_consumable_resource}
\alias{batch_describe_consumable_resource}
\title{Returns a description of the specified consumable resource}
\usage{
batch_describe_consumable_resource(consumableResource)
}
\arguments{
\item{consumableResource}{[required] The name or ARN of the consumable resource whose description will be
returned.}
}
\description{
Returns a description of the specified consumable resource.

See \url{https://www.paws-r-sdk.com/docs/batch_describe_consumable_resource/} for full documentation.
}
\keyword{internal}
