% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_describe_jobs}
\alias{batch_describe_jobs}
\title{Describes a list of Batch jobs}
\usage{
batch_describe_jobs(jobs)
}
\arguments{
\item{jobs}{[required] A list of up to 100 job IDs.}
}
\description{
Describes a list of Batch jobs.

See \url{https://paws-r-sdk.com/docs/describe_jobs/} for full documentation.
}
\keyword{internal}
