% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_describe_quota_share}
\alias{batch_describe_quota_share}
\title{Returns a description of the specified quota share}
\usage{
batch_describe_quota_share(quotaShareArn)
}
\arguments{
\item{quotaShareArn}{[required] The Amazon Resource Name (ARN) of the quota share.}
}
\description{
Returns a description of the specified quota share.

See \url{https://www.paws-r-sdk.com/docs/batch_describe_quota_share/} for full documentation.
}
\keyword{internal}
