% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_describe_scheduling_policies}
\alias{batch_describe_scheduling_policies}
\title{Describes one or more of your scheduling policies}
\usage{
batch_describe_scheduling_policies(arns)
}
\arguments{
\item{arns}{[required] A list of up to 100 scheduling policy Amazon Resource Name (ARN)
entries.}
}
\description{
Describes one or more of your scheduling policies.

See \url{https://www.paws-r-sdk.com/docs/batch_describe_scheduling_policies/} for full documentation.
}
\keyword{internal}
