% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_describe_service_job}
\alias{batch_describe_service_job}
\title{The details of a service job}
\usage{
batch_describe_service_job(jobId)
}
\arguments{
\item{jobId}{[required] The job ID for the service job to describe.}
}
\description{
The details of a service job.

See \url{https://www.paws-r-sdk.com/docs/batch_describe_service_job/} for full documentation.
}
\keyword{internal}
