% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_list_consumable_resources}
\alias{batch_list_consumable_resources}
\title{Returns a list of Batch consumable resources}
\usage{
batch_list_consumable_resources(
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filters}{The filters to apply to the consumable resource list query. If used,
only those consumable resources that match the filter are listed. Filter
names and values can be:
\itemize{
\item name: \code{CONSUMABLE_RESOURCE_NAME }

values: case-insensitive matches for the consumable resource name.
If a filter value ends with an asterisk (\emph{), it matches any
consumable resource name that begins with the string before the
'}'.
}}

\item{maxResults}{The maximum number of results returned by
\code{\link[=batch_list_consumable_resources]{list_consumable_resources}} in
paginated output. When this parameter is used,
\code{\link[=batch_list_consumable_resources]{list_consumable_resources}} only
returns \code{maxResults} results in a single page and a \code{nextToken} response
element. The remaining results of the initial request can be seen by
sending another
\code{\link[=batch_list_consumable_resources]{list_consumable_resources}} request
with the returned \code{nextToken} value. This value can be between 1 and
100. If this parameter isn't used, then
\code{\link[=batch_list_consumable_resources]{list_consumable_resources}} returns
up to 100 results and a \code{nextToken} value if applicable.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{\link[=batch_list_consumable_resources]{list_consumable_resources}} request
where \code{maxResults} was used and the results exceeded the value of that
parameter. Pagination continues from the end of the previous results
that returned the \code{nextToken} value. This value is \code{null} when there are
no more results to return.

Treat this token as an opaque identifier that's only used to retrieve
the next items in a list and not for other programmatic purposes.}
}
\description{
Returns a list of Batch consumable resources.

See \url{https://www.paws-r-sdk.com/docs/batch_list_consumable_resources/} for full documentation.
}
\keyword{internal}
