% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_list_jobs_by_consumable_resource}
\alias{batch_list_jobs_by_consumable_resource}
\title{Returns a list of Batch jobs that require a specific consumable resource}
\usage{
batch_list_jobs_by_consumable_resource(
  consumableResource,
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{consumableResource}{[required] The name or ARN of the consumable resource.}

\item{filters}{The filters to apply to the job list query. If used, only those jobs
requiring the specified consumable resource (\code{consumableResource}) and
that match the value of the filters are listed. The filter names and
values can be:
\itemize{
\item name: \code{JOB_STATUS}

values:
\code{SUBMITTED | PENDING | RUNNABLE | STARTING | RUNNING | SUCCEEDED | FAILED}
\item name: \code{JOB_NAME }

The values are case-insensitive matches for the job name. If a
filter value ends with an asterisk (\emph{), it matches any job name
that begins with the string before the '}'.
}}

\item{maxResults}{The maximum number of results returned by
\code{\link[=batch_list_jobs_by_consumable_resource]{list_jobs_by_consumable_resource}}
in paginated output. When this parameter is used,
\code{\link[=batch_list_jobs_by_consumable_resource]{list_jobs_by_consumable_resource}}
only returns \code{maxResults} results in a single page and a \code{nextToken}
response element. The remaining results of the initial request can be
seen by sending another
\code{\link[=batch_list_jobs_by_consumable_resource]{list_jobs_by_consumable_resource}}
request with the returned \code{nextToken} value. This value can be between 1
and 100. If this parameter isn't used, then
\code{\link[=batch_list_jobs_by_consumable_resource]{list_jobs_by_consumable_resource}}
returns up to 100 results and a \code{nextToken} value if applicable.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{\link[=batch_list_jobs_by_consumable_resource]{list_jobs_by_consumable_resource}}
request where \code{maxResults} was used and the results exceeded the value
of that parameter. Pagination continues from the end of the previous
results that returned the \code{nextToken} value. This value is \code{null} when
there are no more results to return.

Treat this token as an opaque identifier that's only used to retrieve
the next items in a list and not for other programmatic purposes.}
}
\description{
Returns a list of Batch jobs that require a specific consumable resource.

See \url{https://www.paws-r-sdk.com/docs/batch_list_jobs_by_consumable_resource/} for full documentation.
}
\keyword{internal}
