% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_list_service_jobs}
\alias{batch_list_service_jobs}
\title{Returns a list of service jobs for a specified job queue}
\usage{
batch_list_service_jobs(
  jobQueue = NULL,
  jobStatus = NULL,
  maxResults = NULL,
  nextToken = NULL,
  filters = NULL
)
}
\arguments{
\item{jobQueue}{The name or ARN of the job queue with which to list service jobs.}

\item{jobStatus}{The job status used to filter service jobs in the specified queue. If
the \code{filters} parameter is specified, the \code{jobStatus} parameter is
ignored and jobs with any status are returned. The exceptions are the
\code{SHARE_IDENTIFIER} filter and \code{QUOTA_SHARE_NAME} filter, which can be
used with \code{jobStatus}. If you don't specify a status, only \code{RUNNING}
jobs are returned.

The \code{SHARE_IDENTIFIER} filter or \code{QUOTA_SHARE_NAME} filter can be used
with the \code{jobStatus} field to filter results.}

\item{maxResults}{The maximum number of results returned by
\code{\link[=batch_list_service_jobs]{list_service_jobs}} in paginated output. When
this parameter is used, \code{\link[=batch_list_service_jobs]{list_service_jobs}}
only returns \code{maxResults} results in a single page and a \code{nextToken}
response element. The remaining results of the initial request can be
seen by sending another \code{\link[=batch_list_service_jobs]{list_service_jobs}}
request with the returned \code{nextToken} value. This value can be between 1
and 100. If this parameter isn't used, then
\code{\link[=batch_list_service_jobs]{list_service_jobs}} returns up to 100 results
and a \code{nextToken} value if applicable.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{\link[=batch_list_service_jobs]{list_service_jobs}} request where
\code{maxResults} was used and the results exceeded the value of that
parameter. Pagination continues from the end of the previous results
that returned the \code{nextToken} value. This value is \code{null} when there are
no more results to return.

Treat this token as an opaque identifier that's only used to retrieve
the next items in a list and not for other programmatic purposes.}

\item{filters}{The filter to apply to the query. Only one filter can be used at a time.
When the filter is used, \code{jobStatus} is ignored with the exception that
\code{SHARE_IDENTIFIER} or \code{QUOTA_SHARE_NAME} and \code{jobStatus} can be used
together. The results are sorted by the \code{createdAt} field, with the most
recent jobs being first.

The \code{SHARE_IDENTIFIER} or \code{QUOTA_SHARE_NAME} filter and the \code{jobStatus}
field can be used together to filter results.

\strong{JOB_NAME}

The value of the filter is a case-insensitive match for the job name. If
the value ends with an asterisk (\emph{), the filter matches any job name
that begins with the string before the '}'. This corresponds to the
\code{jobName} value. For example, \code{test1} matches both \code{Test1} and \code{test1},
and \verb{test1*} matches both \code{test1} and \code{Test10}. When the \code{JOB_NAME}
filter is used, the results are grouped by the job name and version.

\strong{BEFORE_CREATED_AT}

The value for the filter is the time that's before the job was created.
This corresponds to the \code{createdAt} value. The value is a string
representation of the number of milliseconds since 00:00:00 UTC
(midnight) on January 1, 1970.

\strong{AFTER_CREATED_AT}

The value for the filter is the time that's after the job was created.
This corresponds to the \code{createdAt} value. The value is a string
representation of the number of milliseconds since 00:00:00 UTC
(midnight) on January 1, 1970.

\strong{SHARE_IDENTIFIER}

The value for the filter is the fairshare scheduling share identifier.

\strong{QUOTA_SHARE_NAME}

The value for the filter is the quota management share name.}
}
\description{
Returns a list of service jobs for a specified job queue.

See \url{https://www.paws-r-sdk.com/docs/batch_list_service_jobs/} for full documentation.
}
\keyword{internal}
