% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_submit_service_job}
\alias{batch_submit_service_job}
\title{Submits a service job to a specified job queue to run on SageMaker AI}
\usage{
batch_submit_service_job(
  jobName,
  jobQueue,
  retryStrategy = NULL,
  schedulingPriority = NULL,
  serviceRequestPayload,
  serviceJobType,
  shareIdentifier = NULL,
  timeoutConfig = NULL,
  tags = NULL,
  clientToken = NULL
)
}
\arguments{
\item{jobName}{[required] The name of the service job. It can be up to 128 characters long. It can
contain uppercase and lowercase letters, numbers, hyphens (-), and
underscores (_).}

\item{jobQueue}{[required] The job queue into which the service job is submitted. You can specify
either the name or the ARN of the queue. The job queue must have the
type \code{SAGEMAKER_TRAINING}.}

\item{retryStrategy}{The retry strategy to use for failed service jobs that are submitted
with this service job request.}

\item{schedulingPriority}{The scheduling priority of the service job. Valid values are integers
between 0 and 9999.}

\item{serviceRequestPayload}{[required] The request, in JSON, for the service that the SubmitServiceJob
operation is queueing.}

\item{serviceJobType}{[required] The type of service job. For SageMaker Training jobs, specify
\code{SAGEMAKER_TRAINING}.}

\item{shareIdentifier}{The share identifier for the service job. Don't specify this parameter
if the job queue doesn't have a fair-share scheduling policy. If the job
queue has a fair-share scheduling policy, then this parameter must be
specified.}

\item{timeoutConfig}{The timeout configuration for the service job. If none is specified,
Batch defers to the default timeout of the underlying service handling
the job.}

\item{tags}{The tags that you apply to the service job request. Each tag consists of
a key and an optional value. For more information, see \href{https://docs.aws.amazon.com/batch/latest/userguide/using-tags.html}{Tagging your Batch resources}.}

\item{clientToken}{A unique identifier for the request. This token is used to ensure
idempotency of requests. If this parameter is specified and two submit
requests with identical payloads and \code{clientToken}s are received, these
requests are considered the same request and the second request is
rejected.}
}
\description{
Submits a service job to a specified job queue to run on SageMaker AI. A service job is a unit of work that you submit to Batch for execution on SageMaker AI.

See \url{https://www.paws-r-sdk.com/docs/batch_submit_service_job/} for full documentation.
}
\keyword{internal}
