% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_terminate_jobs}
\alias{batch_terminate_jobs}
\title{Terminates up to 50 jobs in a job queue}
\usage{
batch_terminate_jobs(jobs, reason)
}
\arguments{
\item{jobs}{[required] An array of up to 50 Batch job IDs of the jobs to terminate.}

\item{reason}{[required] A message to attach to the job that explains the reason for terminating it. This message is returned by future \code{\link[=batch_describe_jobs]{describe_jobs}} operations on the job. It is also recorded in the Batch activity logs.

This parameter has a limit of 1024 characters.}
}
\description{
Terminates up to 50 jobs in a job queue. This is a bulk version of \code{\link[=batch_terminate_job]{terminate_job}}. Jobs that are in the \code{STARTING} or \code{RUNNING} state are terminated, which causes them to transition to \code{FAILED}. Jobs that have not progressed to the \code{STARTING} state are cancelled.

See \url{https://www.paws-r-sdk.com/docs/batch_terminate_jobs/} for full documentation.
}
\keyword{internal}
