% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_terminate_service_job}
\alias{batch_terminate_service_job}
\title{Terminates a service job in a job queue}
\usage{
batch_terminate_service_job(jobId, reason)
}
\arguments{
\item{jobId}{[required] The service job ID of the service job to terminate.}

\item{reason}{[required] A message to attach to the service job that explains the reason for terminating it. This message is returned by \code{\link[=batch_describe_service_job]{describe_service_job}} operations on the service job.}
}
\description{
Terminates a service job in a job queue.

See \url{https://www.paws-r-sdk.com/docs/batch_terminate_service_job/} for full documentation.
}
\keyword{internal}
