% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_terminate_service_jobs}
\alias{batch_terminate_service_jobs}
\title{Terminates up to 50 service jobs in a job queue}
\usage{
batch_terminate_service_jobs(jobs, reason)
}
\arguments{
\item{jobs}{[required] An array of up to 50 service job IDs of the service jobs to terminate.}

\item{reason}{[required] A message to attach to the service job that explains the reason for terminating it. This message is returned by \code{\link[=batch_describe_service_job]{describe_service_job}} operations on the service job.}
}
\description{
Terminates up to 50 service jobs in a job queue. This is a bulk version of \code{\link[=batch_terminate_service_job]{terminate_service_job}}.

See \url{https://www.paws-r-sdk.com/docs/batch_terminate_service_jobs/} for full documentation.
}
\keyword{internal}
