% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_update_quota_share}
\alias{batch_update_quota_share}
\title{Updates a quota share}
\usage{
batch_update_quota_share(
  quotaShareArn,
  capacityLimits = NULL,
  resourceSharingConfiguration = NULL,
  preemptionConfiguration = NULL,
  state = NULL
)
}
\arguments{
\item{quotaShareArn}{[required] The Amazon Resource Name (ARN) of the quota share to update.}

\item{capacityLimits}{A list that specifies the quantity and type of compute capacity
allocated to the quota share.}

\item{resourceSharingConfiguration}{Specifies whether a quota share reserves, lends, or both lends and
borrows idle compute capacity.}

\item{preemptionConfiguration}{Specifies the preemption behavior for jobs in a quota share.}

\item{state}{The state of the quota share. If the quota share is \code{ENABLED}, it is
able to accept jobs. If the quota share is \code{DISABLED}, new jobs won't be
accepted but jobs already submitted can finish.}
}
\description{
Updates a quota share.

See \url{https://www.paws-r-sdk.com/docs/batch_update_quota_share/} for full documentation.
}
\keyword{internal}
