% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_update_service_environment}
\alias{batch_update_service_environment}
\title{Updates a service environment}
\usage{
batch_update_service_environment(
  serviceEnvironment,
  state = NULL,
  capacityLimits = NULL
)
}
\arguments{
\item{serviceEnvironment}{[required] The name or ARN of the service environment to update.}

\item{state}{The state of the service environment.}

\item{capacityLimits}{The capacity limits for the service environment. This defines the
maximum resources that can be used by service jobs in this environment.}
}
\description{
Updates a service environment. You can update the state of a service environment from \code{ENABLED} to \code{DISABLED} to prevent new service jobs from being placed in the service environment.

See \url{https://www.paws-r-sdk.com/docs/batch_update_service_environment/} for full documentation.
}
\keyword{internal}
