% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_cancel_quantum_task}
\alias{braket_cancel_quantum_task}
\title{Cancels the specified task}
\usage{
braket_cancel_quantum_task(clientToken, quantumTaskArn)
}
\arguments{
\item{clientToken}{[required] The client token associated with the request.}

\item{quantumTaskArn}{[required] The ARN of the task to cancel.}
}
\description{
Cancels the specified task.

See \url{https://paws-r.github.io/docs/braket/cancel_quantum_task.html} for full documentation.
}
\keyword{internal}
