% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_create_spending_limit}
\alias{braket_create_spending_limit}
\title{Creates a spending limit for a specified quantum device}
\usage{
braket_create_spending_limit(
  clientToken,
  deviceArn,
  spendingLimit,
  timePeriod = NULL,
  tags = NULL
)
}
\arguments{
\item{clientToken}{[required] A unique, case-sensitive identifier to ensure that the operation
completes no more than one time. If this token matches a previous
request, Amazon Braket ignores the request, but does not return an
error.}

\item{deviceArn}{[required] The Amazon Resource Name (ARN) of the quantum device to apply the
spending limit to.}

\item{spendingLimit}{[required] The maximum amount that can be spent on the specified device, in USD.}

\item{timePeriod}{The time period during which the spending limit is active, including
start and end dates.}

\item{tags}{The tags to apply to the spending limit. Each tag consists of a key and
an optional value.}
}
\description{
Creates a spending limit for a specified quantum device. Spending limits help you control costs by setting maximum amounts that can be spent on quantum computing tasks within a specified time period. Simulators do not support spending limits.

See \url{https://www.paws-r-sdk.com/docs/braket_create_spending_limit/} for full documentation.
}
\keyword{internal}
