% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_get_job}
\alias{braket_get_job}
\title{Retrieves the specified Amazon Braket job}
\usage{
braket_get_job(additionalAttributeNames = NULL, jobArn)
}
\arguments{
\item{additionalAttributeNames}{A list of attributes to return information for.}

\item{jobArn}{[required] The ARN of the job to retrieve.}
}
\description{
Retrieves the specified Amazon Braket job.

See \url{https://www.paws-r-sdk.com/docs/braket_get_job/} for full documentation.
}
\keyword{internal}
