% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_get_quantum_task}
\alias{braket_get_quantum_task}
\title{Retrieves the specified quantum task}
\usage{
braket_get_quantum_task(additionalAttributeNames = NULL, quantumTaskArn)
}
\arguments{
\item{additionalAttributeNames}{A list of attributes to return information for.}

\item{quantumTaskArn}{[required] The ARN of the task to retrieve.}
}
\description{
Retrieves the specified quantum task.

See \url{https://www.paws-r-sdk.com/docs/braket_get_quantum_task/} for full documentation.
}
\keyword{internal}
