% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_list_tags_for_resource}
\alias{braket_list_tags_for_resource}
\title{Shows the tags associated with this resource}
\usage{
braket_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] Specify the \code{resourceArn} for the resource whose tags to display.}
}
\description{
Shows the tags associated with this resource.

See \url{https://www.paws-r-sdk.com/docs/braket_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
