% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_search_devices}
\alias{braket_search_devices}
\title{Searches for devices using the specified filters}
\usage{
braket_search_devices(filters, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{filters}{[required] The filter values to use to search for a device.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token used for pagination of results returned in the response. Use the
token returned from the previous request continue results where the
previous request ended.}
}
\description{
Searches for devices using the specified filters.

See \url{https://www.paws-r-sdk.com/docs/braket_search_devices/} for full documentation.
}
\keyword{internal}
