% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_search_quantum_tasks}
\alias{braket_search_quantum_tasks}
\title{Searches for tasks that match the specified filter values}
\usage{
braket_search_quantum_tasks(filters, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{filters}{[required] Array of \code{SearchQuantumTasksFilter} objects.}

\item{maxResults}{Maximum number of results to return in the response.}

\item{nextToken}{A token used for pagination of results returned in the response. Use the
token returned from the previous request continue results where the
previous request ended.}
}
\description{
Searches for tasks that match the specified filter values.

See \url{https://www.paws-r-sdk.com/docs/braket_search_quantum_tasks/} for full documentation.
}
\keyword{internal}
