% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_search_spending_limits}
\alias{braket_search_spending_limits}
\title{Searches and lists spending limits based on specified filters}
\usage{
braket_search_spending_limits(
  nextToken = NULL,
  maxResults = NULL,
  filters = NULL
)
}
\arguments{
\item{nextToken}{The token to retrieve the next page of results. This value is returned
from a previous call to SearchSpendingLimits when there are more results
available.}

\item{maxResults}{The maximum number of results to return in a single call. Minimum value
of 1, maximum value of 100. Default is 20.}

\item{filters}{The filters to apply when searching for spending limits. Use filters to
narrow down the results based on specific criteria.}
}
\description{
Searches and lists spending limits based on specified filters. This operation supports pagination and allows filtering by various criteria to find specific spending limits. We recommend using pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/braket_search_spending_limits/} for full documentation.
}
\keyword{internal}
