% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_tag_resource}
\alias{braket_tag_resource}
\title{Add a tag to the specified resource}
\usage{
braket_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] Specify the \code{resourceArn} of the resource to which a tag will be added.}

\item{tags}{[required] Specify the tags to add to the resource. Tags can be specified as a
key-value map.}
}
\description{
Add a tag to the specified resource.

See \url{https://www.paws-r-sdk.com/docs/braket_tag_resource/} for full documentation.
}
\keyword{internal}
