% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_untag_resource}
\alias{braket_untag_resource}
\title{Remove tags from a resource}
\usage{
braket_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] Specify the \code{resourceArn} for the resource from which to remove the
tags.}

\item{tagKeys}{[required] Specify the keys for the tags to remove from the resource.}
}
\description{
Remove tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/braket_untag_resource/} for full documentation.
}
\keyword{internal}
