% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_ecs_service_recommendation_projected_metrics}
\alias{computeoptimizer_get_ecs_service_recommendation_projected_metrics}
\title{Returns the projected metrics of Amazon ECS service recommendations}
\usage{
computeoptimizer_get_ecs_service_recommendation_projected_metrics(
  serviceArn,
  stat,
  period,
  startTime,
  endTime
)
}
\arguments{
\item{serviceArn}{[required] The ARN that identifies the Amazon ECS service.

The following is the format of the ARN:

\code{arn:aws:ecs:region:aws_account_id:service/cluster-name/service-name}}

\item{stat}{[required] The statistic of the projected metrics.}

\item{period}{[required] The granularity, in seconds, of the projected metrics data points.}

\item{startTime}{[required] The timestamp of the first projected metrics data point to return.}

\item{endTime}{[required] The timestamp of the last projected metrics data point to return.}
}
\description{
Returns the projected metrics of Amazon ECS service recommendations.

See \url{https://www.paws-r-sdk.com/docs/computeoptimizer_get_ecs_service_recommendation_projected_metrics/} for full documentation.
}
\keyword{internal}
