% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_enrollment_status}
\alias{computeoptimizer_get_enrollment_status}
\title{Returns the enrollment (opt in) status of an account to the Compute
Optimizer service}
\usage{
computeoptimizer_get_enrollment_status()
}
\description{
Returns the enrollment (opt in) status of an account to the Compute Optimizer service.

See \url{https://www.paws-r-sdk.com/docs/computeoptimizer_get_enrollment_status/} for full documentation.
}
\keyword{internal}
