% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_rds_database_recommendations}
\alias{computeoptimizer_get_rds_database_recommendations}
\title{Returns Amazon Aurora and RDS database recommendations}
\usage{
computeoptimizer_get_rds_database_recommendations(
  resourceArns = NULL,
  nextToken = NULL,
  maxResults = NULL,
  filters = NULL,
  accountIds = NULL,
  recommendationPreferences = NULL
)
}
\arguments{
\item{resourceArns}{The ARN that identifies the Amazon Aurora or RDS database.

The following is the format of the ARN:

\code{arn:aws:rds:{region}:{accountId}:db:{resourceName}}

The following is the format of a DB Cluster ARN:

\code{arn:aws:rds:{region}:{accountId}:cluster:{resourceName}}}

\item{nextToken}{The token to advance to the next page of Amazon Aurora and RDS database recommendations.}

\item{maxResults}{The maximum number of Amazon Aurora and RDS database recommendations to return with a single request.

To retrieve the remaining results, make another request with the returned \code{nextToken} value.}

\item{filters}{An array of objects to specify a filter that returns a more specific list of Amazon Aurora and RDS database recommendations.}

\item{accountIds}{Return the Amazon Aurora and RDS database recommendations to the specified Amazon Web Services account IDs.

If your account is the management account or the delegated administrator of an organization, use this parameter to return the Amazon Aurora and RDS database recommendations to specific member accounts.

You can only specify one account ID per request.}

\item{recommendationPreferences}{Describes the recommendation preferences to return in the response of a \code{\link[=computeoptimizer_get_auto_scaling_group_recommendations]{get_auto_scaling_group_recommendations}}, \code{\link[=computeoptimizer_get_ec2_instance_recommendations]{get_ec2_instance_recommendations}}, \code{\link[=computeoptimizer_get_ec2_recommendation_projected_metrics]{get_ec2_recommendation_projected_metrics}}, \code{\link[=computeoptimizer_get_rds_database_recommendations]{get_rds_database_recommendations}}, and \code{\link[=computeoptimizer_get_rds_database_recommendation_projected_metrics]{get_rds_database_recommendation_projected_metrics}} request.}
}
\description{
Returns Amazon Aurora and RDS database recommendations.

See \url{https://www.paws-r-sdk.com/docs/computeoptimizer_get_rds_database_recommendations/} for full documentation.
}
\keyword{internal}
