% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_recommendation_summaries}
\alias{computeoptimizer_get_recommendation_summaries}
\title{Returns the optimization findings for an account}
\usage{
computeoptimizer_get_recommendation_summaries(
  accountIds = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{accountIds}{The ID of the Amazon Web Services account for which to return
recommendation summaries.

If your account is the management account of an organization, use this
parameter to specify the member account for which you want to return
recommendation summaries.

Only one account ID can be specified per request.}

\item{nextToken}{The token to advance to the next page of recommendation summaries.}

\item{maxResults}{The maximum number of recommendation summaries to return with a single
request.

To retrieve the remaining results, make another request with the
returned \code{nextToken} value.}
}
\description{
Returns the optimization findings for an account.

See \url{https://www.paws-r-sdk.com/docs/computeoptimizer_get_recommendation_summaries/} for full documentation.
}
\keyword{internal}
