% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_accept_transit_gateway_client_vpn_attachment}
\alias{ec2_accept_transit_gateway_client_vpn_attachment}
\title{Accepts a Transit Gateway attachment request for a Client VPN endpoint}
\usage{
ec2_accept_transit_gateway_client_vpn_attachment(
  TransitGatewayAttachmentId,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayAttachmentId}{[required] The ID of the Transit Gateway attachment.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Accepts a Transit Gateway attachment request for a Client VPN endpoint. The Transit Gateway owner must accept the attachment request before the Client VPN endpoint can route traffic through the Transit Gateway.

See \url{https://www.paws-r-sdk.com/docs/ec2_accept_transit_gateway_client_vpn_attachment/} for full documentation.
}
\keyword{internal}
