% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_advertise_byoip_cidr}
\alias{ec2_advertise_byoip_cidr}
\title{Advertises an IPv4 or IPv6 address range that is provisioned for use
with your Amazon Web Services resources through bring your own IP
addresses (BYOIP)}
\usage{
ec2_advertise_byoip_cidr(Cidr, DryRun = NULL)
}
\arguments{
\item{Cidr}{[required] The address range, in CIDR notation. This must be the exact range that
you provisioned. You can't advertise only a portion of the provisioned
range.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Advertises an IPv4 or IPv6 address range that is provisioned for use with your Amazon Web Services resources through bring your own IP addresses (BYOIP).

See \url{https://paws-r.github.io/docs/ec2/advertise_byoip_cidr.html} for full documentation.
}
\keyword{internal}
