% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_associate_application_status_check}
\alias{ec2_associate_application_status_check}
\title{Associates an application status check with instances or tags}
\usage{
ec2_associate_application_status_check(
  ApplicationStatusCheckId,
  TargetTagAssociations = NULL,
  InstanceIds = NULL,
  ClientToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{ApplicationStatusCheckId}{[required] The ID of the application status check to associate.}

\item{TargetTagAssociations}{The \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_Tags.html}{tags} to associate the application status check with. Each tag is a key-value pair. When you associate tags, the application status check automatically monitors all instances that have the specified tags.}

\item{InstanceIds}{The IDs of the instances to associate with the application status check.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure that the operation completes no more than one time. If you retry a request with the same token, the service ignores the request but does not return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Associates an application status check with instances or \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_Tags.html}{tags}. Once you create an association, health monitoring automatically begins for the specified instances or for instances that match the specified tags. The following rules apply:

See \url{https://www.paws-r-sdk.com/docs/ec2_associate_application_status_check/} for full documentation.
}
\keyword{internal}
