% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_associate_instance_event_window}
\alias{ec2_associate_instance_event_window}
\title{Associates one or more targets with an event window}
\usage{
ec2_associate_instance_event_window(
  DryRun = NULL,
  InstanceEventWindowId,
  AssociationTarget
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{InstanceEventWindowId}{[required] The ID of the event window.}

\item{AssociationTarget}{[required] One or more targets associated with the specified event window.}
}
\description{
Associates one or more targets with an event window. Only one type of target (instance IDs, Dedicated Host IDs, or tags) can be specified with an event window.

See \url{https://www.paws-r-sdk.com/docs/ec2_associate_instance_event_window/} for full documentation.
}
\keyword{internal}
