% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_associate_route_server}
\alias{ec2_associate_route_server}
\title{Associates a route server with a VPC to enable dynamic route updates}
\usage{
ec2_associate_route_server(RouteServerId, VpcId, DryRun = NULL)
}
\arguments{
\item{RouteServerId}{[required] The unique identifier for the route server to be associated.}

\item{VpcId}{[required] The ID of the VPC to associate with the route server.}

\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Associates a route server with a VPC to enable dynamic route updates.

See \url{https://www.paws-r-sdk.com/docs/ec2_associate_route_server/} for full documentation.
}
\keyword{internal}
