% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_associate_security_group_vpc}
\alias{ec2_associate_security_group_vpc}
\title{Associates a security group with another VPC in the same Region}
\usage{
ec2_associate_security_group_vpc(GroupId, VpcId, DryRun = NULL)
}
\arguments{
\item{GroupId}{[required] A security group ID.}

\item{VpcId}{[required] A VPC ID.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Associates a security group with another VPC in the same Region. This enables you to use the same security group with network interfaces and instances in the specified VPC.

See \url{https://www.paws-r-sdk.com/docs/ec2_associate_security_group_vpc/} for full documentation.
}
\keyword{internal}
