% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_associate_subnet_cidr_block}
\alias{ec2_associate_subnet_cidr_block}
\title{Associates a CIDR block with your subnet}
\usage{
ec2_associate_subnet_cidr_block(Ipv6CidrBlock, SubnetId)
}
\arguments{
\item{Ipv6CidrBlock}{[required] The IPv6 CIDR block for your subnet. The subnet must have a /64 prefix
length.}

\item{SubnetId}{[required] The ID of your subnet.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Ipv6CidrBlockAssociation = list(
    AssociationId = "string",
    Ipv6CidrBlock = "string",
    Ipv6CidrBlockState = list(
      State = "associating"|"associated"|"disassociating"|"disassociated"|"failing"|"failed",
      StatusMessage = "string"
    )
  ),
  SubnetId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Associates a CIDR block with your subnet. You can only associate a
single IPv6 CIDR block with your subnet. An IPv6 CIDR block must have a
prefix length of /64.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_subnet_cidr_block(
  Ipv6CidrBlock = "string",
  SubnetId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
