% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_associate_transit_gateway_policy_table}
\alias{ec2_associate_transit_gateway_policy_table}
\title{Associates the specified transit gateway attachment with a transit
gateway policy table}
\usage{
ec2_associate_transit_gateway_policy_table(
  TransitGatewayPolicyTableId,
  TransitGatewayAttachmentId,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayPolicyTableId}{[required] The ID of the transit gateway policy table to associate with the transit
gateway attachment.}

\item{TransitGatewayAttachmentId}{[required] The ID of the transit gateway attachment to associate with the policy
table.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Associates the specified transit gateway attachment with a transit gateway policy table.

See \url{https://www.paws-r-sdk.com/docs/ec2_associate_transit_gateway_policy_table/} for full documentation.
}
\keyword{internal}
