% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_associate_trunk_interface}
\alias{ec2_associate_trunk_interface}
\title{Associates a branch network interface with a trunk network interface}
\usage{
ec2_associate_trunk_interface(
  BranchInterfaceId,
  TrunkInterfaceId,
  VlanId = NULL,
  GreKey = NULL,
  ClientToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{BranchInterfaceId}{[required] The ID of the branch network interface.}

\item{TrunkInterfaceId}{[required] The ID of the trunk network interface.}

\item{VlanId}{The ID of the VLAN. This applies to the VLAN protocol.}

\item{GreKey}{The application key. This applies to the GRE protocol.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{How to Ensure Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Associates a branch network interface with a trunk network interface.

See \url{https://www.paws-r-sdk.com/docs/ec2_associate_trunk_interface/} for full documentation.
}
\keyword{internal}
