% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_attach_image_watermark}
\alias{ec2_attach_image_watermark}
\title{Attaches a watermark to a non-public AMI}
\usage{
ec2_attach_image_watermark(ImageId, WatermarkName, DryRun = NULL)
}
\arguments{
\item{ImageId}{[required] The ID of the AMI.}

\item{WatermarkName}{[required] The name for the watermark. Combined with the caller's account ID to form the \code{WatermarkKey} (\code{accountId:watermarkName}).

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Attaches a watermark to a non-public AMI. The watermark is a structured identifier that automatically propagates to all derivative images created through \code{\link[=ec2_create_image]{create_image}}, and \code{\link[=ec2_copy_image]{copy_image}}.

See \url{https://www.paws-r-sdk.com/docs/ec2_attach_image_watermark/} for full documentation.
}
\keyword{internal}
