% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_attach_internet_gateway}
\alias{ec2_attach_internet_gateway}
\title{Attaches an internet gateway or a virtual private gateway to a VPC,
enabling connectivity between the internet and the VPC}
\usage{
ec2_attach_internet_gateway(DryRun = NULL, InternetGatewayId, VpcId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{InternetGatewayId}{[required] The ID of the internet gateway.}

\item{VpcId}{[required] The ID of the VPC.}
}
\description{
Attaches an internet gateway or a virtual private gateway to a VPC, enabling connectivity between the internet and the VPC. For more information, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/VPC_Internet_Gateway.html}{Internet gateways} in the \emph{Amazon VPC User Guide}.

See \url{https://www.paws-r-sdk.com/docs/ec2_attach_internet_gateway/} for full documentation.
}
\keyword{internal}
