% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_attach_verified_access_trust_provider}
\alias{ec2_attach_verified_access_trust_provider}
\title{Attaches the specified Amazon Web Services Verified Access trust
provider to the specified Amazon Web Services Verified Access instance}
\usage{
ec2_attach_verified_access_trust_provider(
  VerifiedAccessInstanceId,
  VerifiedAccessTrustProviderId,
  ClientToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{VerifiedAccessInstanceId}{[required] The ID of the Verified Access instance.}

\item{VerifiedAccessTrustProviderId}{[required] The ID of the Verified Access trust provider.}

\item{ClientToken}{A unique, case-sensitive token that you provide to ensure idempotency of
your modification request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Attaches the specified Amazon Web Services Verified Access trust provider to the specified Amazon Web Services Verified Access instance.

See \url{https://www.paws-r-sdk.com/docs/ec2_attach_verified_access_trust_provider/} for full documentation.
}
\keyword{internal}
