% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_batch_modify_ipam_routing_policy_registrations}
\alias{ec2_batch_modify_ipam_routing_policy_registrations}
\title{Modifies multiple routing policy registrations in a single operation}
\usage{
ec2_batch_modify_ipam_routing_policy_registrations(
  DryRun = NULL,
  IpamInternetRegistryAssociationId,
  DeltaJson,
  Force = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamInternetRegistryAssociationId}{[required] The ID of the IPAM internet registry association.}

\item{DeltaJson}{[required] The batch modifications to apply, in JSON format.}

\item{Force}{Forces the batch modification even if individual changes conflict with announced routes. Default: \code{false}.}

\item{ClientToken}{A unique, case-sensitive identifier to ensure that the operation completes no more than one time. If this token matches a previous request, the operation ignores the request, but does not return an error.}
}
\description{
Modifies multiple routing policy registrations in a single operation. You can create, update, or delete Route Origin Authorizations (ROAs) in batch.

See \url{https://www.paws-r-sdk.com/docs/ec2_batch_modify_ipam_routing_policy_registrations/} for full documentation.
}
\keyword{internal}
