% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_cancel_capacity_reservation_fleets}
\alias{ec2_cancel_capacity_reservation_fleets}
\title{Cancels one or more Capacity Reservation Fleets}
\usage{
ec2_cancel_capacity_reservation_fleets(
  DryRun = NULL,
  CapacityReservationFleetIds
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{CapacityReservationFleetIds}{[required] The IDs of the Capacity Reservation Fleets to cancel.}
}
\description{
Cancels one or more Capacity Reservation Fleets. When you cancel a Capacity Reservation Fleet, the following happens:

See \url{https://www.paws-r-sdk.com/docs/ec2_cancel_capacity_reservation_fleets/} for full documentation.
}
\keyword{internal}
