% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_cancel_declarative_policies_report}
\alias{ec2_cancel_declarative_policies_report}
\title{Cancels the generation of an account status report}
\usage{
ec2_cancel_declarative_policies_report(DryRun = NULL, ReportId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ReportId}{[required] The ID of the report.}
}
\description{
Cancels the generation of an account status report.

See \url{https://www.paws-r-sdk.com/docs/ec2_cancel_declarative_policies_report/} for full documentation.
}
\keyword{internal}
