% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_cancel_spot_fleet_requests}
\alias{ec2_cancel_spot_fleet_requests}
\title{Cancels the specified Spot Fleet requests}
\usage{
ec2_cancel_spot_fleet_requests(
  DryRun = NULL,
  SpotFleetRequestIds,
  TerminateInstances
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{SpotFleetRequestIds}{[required] The IDs of the Spot Fleet requests.}

\item{TerminateInstances}{[required] Indicates whether to terminate instances for a Spot Fleet request if it
is canceled successfully.}
}
\description{
Cancels the specified Spot Fleet requests.

See \url{https://paws-r.github.io/docs/ec2/cancel_spot_fleet_requests.html} for full documentation.
}
\keyword{internal}
