% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_copy_volumes}
\alias{ec2_copy_volumes}
\title{Creates a crash-consistent, point-in-time copy of an existing Amazon EBS
volume within the same Availability Zone}
\usage{
ec2_copy_volumes(
  SourceVolumeId,
  Iops = NULL,
  Size = NULL,
  VolumeType = NULL,
  DryRun = NULL,
  TagSpecifications = NULL,
  MultiAttachEnabled = NULL,
  Throughput = NULL,
  ClientToken = NULL,
  Encrypted = NULL,
  KmsKeyId = NULL
)
}
\arguments{
\item{SourceVolumeId}{[required] The ID of the source EBS volume to copy.}

\item{Iops}{The number of I/O operations per second (IOPS) to provision for the volume copy. Required for \code{io1} and \code{io2} volumes. Optional for \code{gp3} volumes. Omit for all other volume types. Full provisioned IOPS performance can be achieved only once the volume copy is fully initialized.

Valid ranges:
\itemize{
\item gp3: \verb{3,000 }(\emph{default})\verb{ - 80,000} IOPS
\item io1: \verb{100 - 64,000} IOPS
\item io2: \verb{100 - 256,000} IOPS
}

\href{https://docs.aws.amazon.com/ec2/latest/instancetypes/ec2-nitro-instances.html}{Instances built on the Nitro System} can support up to 256,000 IOPS. Other instances can support up to 32,000 IOPS.}

\item{Size}{The size of the volume copy, in GiBs. The size must be equal to or greater than the size of the source volume. If not specified, the size defaults to the size of the source volume.

Maximum supported sizes:
\itemize{
\item gp2: \verb{16,384} GiB
\item gp3: \verb{65,536} GiB
\item io1: \verb{16,384} GiB
\item io2: \verb{65,536} GiB
\item st1 and sc1: \verb{16,384} GiB
\item standard: \code{1024} GiB
}}

\item{VolumeType}{The volume type for the volume copy. If not specified, the volume type defaults to \code{gp2}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{TagSpecifications}{The tags to apply to the volume copy during creation.}

\item{MultiAttachEnabled}{Indicates whether to enable Amazon EBS Multi-Attach for the volume copy. If you enable Multi-Attach, you can attach the volume to up to 16 Nitro instances in the same Availability Zone simultaneously. Supported with \code{io1} and \code{io2} volumes only. For more information, see \href{https://docs.aws.amazon.com/ebs/latest/userguide/ebs-volumes-multi.html}{Amazon EBS Multi-Attach}.}

\item{Throughput}{The throughput to provision for the volume copy, in MiB/s. Supported for \code{gp3} volumes only. Omit for all other volume types. Full provisioned throughput performance can be achieved only once the volume copy is fully initialized.

Valid Range: \code{125 - 2000} MiB/s}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensure Idempotency}.}

\item{Encrypted}{Indicates whether to encrypt the volume copy. If the source volume is encrypted, the service always encrypts the copy regardless of this value. Set to \code{true} to encrypt a copy of an unencrypted source volume during the copy operation. If you set \code{Encrypted} to \code{true} but do not specify \code{KmsKeyId}, the service uses the default KMS key for EBS encryption in your account.}

\item{KmsKeyId}{The identifier of the KMS key to use for encryption of the volume copy. Specify a symmetric encryption KMS key. You can specify a KMS key using the key ID, key ARN, alias name, or alias ARN. If you set \code{Encrypted} to \code{true} but do not specify this parameter, the service uses the default KMS key for EBS encryption in your account. For cross-account volume copies, this must be a KMS key in the calling account.}
}
\description{
Creates a crash-consistent, point-in-time copy of an existing Amazon EBS volume within the same Availability Zone. The volume copy can be attached to an Amazon EC2 instance once it reaches the \code{available} state. For more information, see \href{https://docs.aws.amazon.com/ebs/latest/userguide/ebs-copying-volume.html}{Copy an Amazon EBS volume}.

See \url{https://www.paws-r-sdk.com/docs/ec2_copy_volumes/} for full documentation.
}
\keyword{internal}
