% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_capacity_reservation_cancellation_quote}
\alias{ec2_create_capacity_reservation_cancellation_quote}
\title{Generates a cancellation quote for a future-dated Capacity Reservation
that is within its commitment duration}
\usage{
ec2_create_capacity_reservation_cancellation_quote(
  CapacityReservationId,
  ClientToken = NULL,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{CapacityReservationId}{[required] The ID of the Capacity Reservation.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensure Idempotency}.}

\item{TagSpecifications}{The tags to apply to the cancellation quote.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Generates a cancellation quote for a future-dated Capacity Reservation that is within its commitment duration. The quote includes the cancellation terms and a quote ID that you can pass to the \code{\link[=ec2_cancel_capacity_reservation]{cancel_capacity_reservation}} action. Cancellation quotes are valid for 24 hours.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_capacity_reservation_cancellation_quote/} for full documentation.
}
\keyword{internal}
