% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_capacity_reservation_date_change_quote}
\alias{ec2_create_capacity_reservation_date_change_quote}
\title{Generates a quote for changing the start date of a future-dated Capacity
Reservation that has not yet been delivered}
\usage{
ec2_create_capacity_reservation_date_change_quote(
  CapacityReservationId,
  NewStartDate,
  ClientToken = NULL,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{CapacityReservationId}{[required] The ID of the Capacity Reservation.}

\item{NewStartDate}{[required] The requested new start date for the Capacity Reservation, in the ISO8601 format in the UTC time zone (\code{YYYY-MM-DDThh:mm:ss.sssZ}). The new start date must be later than the current start date and within the cumulative 30-day pushout limit.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensure Idempotency}.}

\item{TagSpecifications}{The tags to apply to the date change quote.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Generates a quote for changing the start date of a future-dated Capacity Reservation that has not yet been delivered. The quote includes the new start date, the resulting commitment end date, and a quote ID. Pass the quote ID to \code{\link[=ec2_modify_capacity_reservation]{modify_capacity_reservation}} to apply the change.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_capacity_reservation_date_change_quote/} for full documentation.
}
\keyword{internal}
