% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_coip_cidr}
\alias{ec2_create_coip_cidr}
\title{Creates a range of customer-owned IP addresses}
\usage{
ec2_create_coip_cidr(Cidr, CoipPoolId, DryRun = NULL)
}
\arguments{
\item{Cidr}{[required] A customer-owned IP address range to create.}

\item{CoipPoolId}{[required] The ID of the address pool.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a range of customer-owned IP addresses.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_coip_cidr/} for full documentation.
}
\keyword{internal}
