% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_coip_pool}
\alias{ec2_create_coip_pool}
\title{Creates a pool of customer-owned IP (CoIP) addresses}
\usage{
ec2_create_coip_pool(
  LocalGatewayRouteTableId,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{LocalGatewayRouteTableId}{[required] The ID of the local gateway route table.}

\item{TagSpecifications}{The tags to assign to the CoIP address pool.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a pool of customer-owned IP (CoIP) addresses.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_coip_pool/} for full documentation.
}
\keyword{internal}
