% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_delegate_mac_volume_ownership_task}
\alias{ec2_create_delegate_mac_volume_ownership_task}
\title{Delegates ownership of the Amazon EBS root volume for an Apple silicon
Mac instance to an administrative user}
\usage{
ec2_create_delegate_mac_volume_ownership_task(
  ClientToken = NULL,
  DryRun = NULL,
  InstanceId,
  MacCredentials,
  TagSpecifications = NULL
)
}
\arguments{
\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{InstanceId}{[required] The ID of the Amazon EC2 Mac instance.}

\item{MacCredentials}{[required] Specifies the following credentials:
\itemize{
\item \strong{Internal disk administrative user}
\itemize{
\item \strong{Username} - Only the default administrative user
(\code{aws-managed-user}) is supported and it is used by default. You
can't specify a different administrative user.
\item \strong{Password} - If you did not change the default password for
\code{aws-managed-user}, specify the default password, which is
\emph{blank}. Otherwise, specify your password.
}
\item \strong{Amazon EBS root volume administrative user}
\itemize{
\item \strong{Username} - If you did not change the default administrative
user, specify \code{ec2-user}. Otherwise, specify the username for
your administrative user.
\item \strong{Password} - Specify the password for the administrative user.
}
}

The credentials must be specified in the following JSON format:

\verb{\{ "internalDiskPassword":"internal-disk-admin_password", "rootVolumeUsername":"root-volume-admin_username", "rootVolumepassword":"root-volume-admin_password" \}}}

\item{TagSpecifications}{The tags to assign to the volume ownership delegation task.}
}
\description{
Delegates ownership of the Amazon EBS root volume for an Apple silicon Mac instance to an administrative user.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_delegate_mac_volume_ownership_task/} for full documentation.
}
\keyword{internal}
