% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_image_usage_report}
\alias{ec2_create_image_usage_report}
\title{Creates a report that shows how your image is used across other Amazon
Web Services accounts}
\usage{
ec2_create_image_usage_report(
  ImageId,
  DryRun = NULL,
  ResourceTypes,
  AccountIds = NULL,
  ClientToken = NULL,
  TagSpecifications = NULL
)
}
\arguments{
\item{ImageId}{[required] The ID of the image to report on.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ResourceTypes}{[required] The resource types to include in the report.}

\item{AccountIds}{The Amazon Web Services account IDs to include in the report. To include
all accounts, omit this parameter.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure
idempotency of the request.}

\item{TagSpecifications}{The tags to apply to the report on creation. The \code{ResourceType} must be
set to \code{image-usage-report}; any other value will cause the report
creation to fail.

To tag a report after it has been created, see
\code{\link[=ec2_create_tags]{create_tags}}.}
}
\description{
Creates a report that shows how your image is used across other Amazon Web Services accounts. The report provides visibility into which accounts are using the specified image, and how many resources (EC2 instances or launch templates) are referencing it.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_image_usage_report/} for full documentation.
}
\keyword{internal}
