% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_interruptible_capacity_reservation_allocation}
\alias{ec2_create_interruptible_capacity_reservation_allocation}
\title{Creates an interruptible Capacity Reservation by specifying the number
of unused instances you want to allocate from your source reservation}
\usage{
ec2_create_interruptible_capacity_reservation_allocation(
  CapacityReservationId,
  InstanceCount,
  ClientToken = NULL,
  DryRun = NULL,
  TagSpecifications = NULL,
  ZeroSizePreference = NULL
)
}
\arguments{
\item{CapacityReservationId}{[required] The ID of the source Capacity Reservation from which to create the interruptible Capacity Reservation. Your Capacity Reservation must be in active state with no end date set and have available capacity for allocation.}

\item{InstanceCount}{[required] The number of instances to allocate from your source reservation. You can only allocate available instances (also called unused capacity).}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response.}

\item{TagSpecifications}{The tags to apply to the interruptible Capacity Reservation during creation.}

\item{ZeroSizePreference}{Specifies the behavior for the interruptible Capacity Reservation when you reduce its allocation to zero instances. Specify \code{retain} to keep the interruptible Capacity Reservation active at zero capacity so that you can allocate instances to it again later. Specify \code{default} to cancel the interruptible Capacity Reservation and return the capacity to your source Capacity Reservation. The default value is \code{default}.}
}
\description{
Creates an interruptible Capacity Reservation by specifying the number of unused instances you want to allocate from your source reservation. This helps you make unused capacity available for other workloads within your account while maintaining control to reclaim it.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_interruptible_capacity_reservation_allocation/} for full documentation.
}
\keyword{internal}
