% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_ipam_internet_registry_association}
\alias{ec2_create_ipam_internet_registry_association}
\title{Creates an association between an IPAM and a Regional Internet Registry
(RIR) for Resource Public Key Infrastructure (RPKI) management}
\usage{
ec2_create_ipam_internet_registry_association(
  DryRun = NULL,
  IpamId,
  Rir,
  OrganizationHandle,
  Description = NULL,
  TagSpecifications = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamId}{[required] The ID of the IPAM to associate with the internet registry.}

\item{Rir}{[required] The Regional Internet Registry to associate with. Possible values:
\itemize{
\item \code{ripe} - RIPE NCC (Europe, the Middle East, and Central Asia).
\item \code{apnic} - APNIC (Asia Pacific).
\item \code{arin} - ARIN (North America).
\item \code{lacnic} - LACNIC (Latin America and the Caribbean).
}}

\item{OrganizationHandle}{[required] The organization handle at the internet registry (for example, a RIPE NCC organization ID or ARIN Org ID).}

\item{Description}{A description for the internet registry association.}

\item{TagSpecifications}{The tags to assign to the internet registry association.}

\item{ClientToken}{A unique, case-sensitive identifier to ensure that the operation completes no more than one time. If this token matches a previous request, the operation ignores the request, but does not return an error.}
}
\description{
Creates an association between an IPAM and a Regional Internet Registry (RIR) for Resource Public Key Infrastructure (RPKI) management. You can use this association to create Route Origin Authorizations (ROAs) for IP address prefixes registered with the internet registry. Your IPAM must be in the Advanced tier to use this feature.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_ipam_internet_registry_association/} for full documentation.
}
\keyword{internal}
