% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_ipam_policy}
\alias{ec2_create_ipam_policy}
\title{Creates an IPAM policy}
\usage{
ec2_create_ipam_policy(
  DryRun = NULL,
  TagSpecifications = NULL,
  ClientToken = NULL,
  IpamId
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{TagSpecifications}{The tags to assign to the IPAM policy.}

\item{ClientToken}{A unique, case-sensitive identifier to ensure the idempotency of the
request.}

\item{IpamId}{[required] The ID of the IPAM for which you're creating the policy.}
}
\description{
Creates an IPAM policy.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_ipam_policy/} for full documentation.
}
\keyword{internal}
