% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_ipam_prefix_list_resolver}
\alias{ec2_create_ipam_prefix_list_resolver}
\title{Creates an IPAM prefix list resolver}
\usage{
ec2_create_ipam_prefix_list_resolver(
  DryRun = NULL,
  IpamId,
  Description = NULL,
  AddressFamily,
  Rules = NULL,
  TagSpecifications = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamId}{[required] The ID of the IPAM that will serve as the source of the IP address
database for CIDR selection. The IPAM must be in the Advanced tier to
use this feature.}

\item{Description}{A description for the IPAM prefix list resolver to help you identify its
purpose and configuration.}

\item{AddressFamily}{[required] The address family for the IPAM prefix list resolver. Valid values are
\code{ipv4} and \code{ipv6}. You must create separate resolvers for IPv4 and IPv6
CIDRs as they cannot be mixed in the same resolver.}

\item{Rules}{The CIDR selection rules for the resolver.

CIDR selection rules define the business logic for selecting CIDRs from
IPAM. If a CIDR matches any of the rules, it will be included. If a rule
has multiple conditions, the CIDR has to match every condition of that
rule. You can create a prefix list resolver without any CIDR selection
rules, but it will generate empty versions (containing no CIDRs) until
you add rules.}

\item{TagSpecifications}{The tags to apply to the IPAM prefix list resolver during creation. Tags
help you organize and manage your Amazon Web Services resources.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}
}
\description{
Creates an IPAM prefix list resolver.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_ipam_prefix_list_resolver/} for full documentation.
}
\keyword{internal}
