% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_local_gateway_route_table}
\alias{ec2_create_local_gateway_route_table}
\title{Creates a local gateway route table}
\usage{
ec2_create_local_gateway_route_table(
  LocalGatewayId,
  Mode = NULL,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{LocalGatewayId}{[required] The ID of the local gateway.}

\item{Mode}{The mode of the local gateway route table.}

\item{TagSpecifications}{The tags assigned to the local gateway route table.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a local gateway route table.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_local_gateway_route_table/} for full documentation.
}
\keyword{internal}
