% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_local_gateway_route_table_vpc_association}
\alias{ec2_create_local_gateway_route_table_vpc_association}
\title{Associates the specified VPC with the specified local gateway route
table}
\usage{
ec2_create_local_gateway_route_table_vpc_association(
  LocalGatewayRouteTableId, VpcId, DryRun)
}
\arguments{
\item{LocalGatewayRouteTableId}{[required] The ID of the local gateway route table.}

\item{VpcId}{[required] The ID of the VPC.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Associates the specified VPC with the specified local gateway route
table.
}
\section{Request syntax}{
\preformatted{svc$create_local_gateway_route_table_vpc_association(
  LocalGatewayRouteTableId = "string",
  VpcId = "string",
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
