% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_local_gateway_virtual_interface_group}
\alias{ec2_create_local_gateway_virtual_interface_group}
\title{Create a local gateway virtual interface group}
\usage{
ec2_create_local_gateway_virtual_interface_group(
  LocalGatewayId,
  LocalBgpAsn = NULL,
  LocalBgpAsnExtended = NULL,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{LocalGatewayId}{[required] The ID of the local gateway.}

\item{LocalBgpAsn}{The Autonomous System Number(ASN) for the local Border Gateway Protocol
(BGP).}

\item{LocalBgpAsnExtended}{The extended 32-bit ASN for the local BGP configuration.}

\item{TagSpecifications}{The tags to apply to the local gateway virtual interface group when the
resource is being created.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Create a local gateway virtual interface group.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_local_gateway_virtual_interface_group/} for full documentation.
}
\keyword{internal}
