% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_mac_system_integrity_protection_modification_task}
\alias{ec2_create_mac_system_integrity_protection_modification_task}
\title{Creates a System Integrity Protection (SIP) modification task to
configure the SIP settings for an x86 Mac instance or Apple silicon Mac
instance}
\usage{
ec2_create_mac_system_integrity_protection_modification_task(
  ClientToken = NULL,
  DryRun = NULL,
  InstanceId,
  MacCredentials = NULL,
  MacSystemIntegrityProtectionConfiguration = NULL,
  MacSystemIntegrityProtectionStatus,
  TagSpecifications = NULL
)
}
\arguments{
\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{InstanceId}{[required] The ID of the Amazon EC2 Mac instance.}

\item{MacCredentials}{\strong{[Apple silicon Mac instances only]} Specifies the following
credentials:
\itemize{
\item \strong{Internal disk administrative user}
\itemize{
\item \strong{Username} - Only the default administrative user
(\code{aws-managed-user}) is supported and it is used by default. You
can't specify a different administrative user.
\item \strong{Password} - If you did not change the default password for
\code{aws-managed-user}, specify the default password, which is
\emph{blank}. Otherwise, specify your password.
}
\item \strong{Amazon EBS root volume administrative user}
\itemize{
\item \strong{Username} - If you did not change the default administrative
user, specify \code{ec2-user}. Otherwise, specify the username for
your administrative user.
\item \strong{Password} - Specify the password for the administrative user.
}
}

The credentials must be specified in the following JSON format:

\verb{\{ "internalDiskPassword":"internal-disk-admin_password", "rootVolumeUsername":"root-volume-admin_username", "rootVolumepassword":"root-volume-admin_password" \}}}

\item{MacSystemIntegrityProtectionConfiguration}{Specifies the overrides to selectively enable or disable individual SIP
settings. The individual settings you specify here override the overall
SIP status you specify for \strong{MacSystemIntegrityProtectionStatus}.}

\item{MacSystemIntegrityProtectionStatus}{[required] Specifies the overall SIP status for the instance. To enable all SIP
settings, specify \code{enabled}. To disable all SIP settings, specify
\code{disabled}.}

\item{TagSpecifications}{Specifies tags to apply to the SIP modification task.}
}
\description{
Creates a System Integrity Protection (SIP) modification task to configure the SIP settings for an x86 Mac instance or Apple silicon Mac instance. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/mac-sip-settings.html#mac-sip-configure}{Configure SIP for Amazon EC2 instances} in the \emph{Amazon EC2 User Guide}.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_mac_system_integrity_protection_modification_task/} for full documentation.
}
\keyword{internal}
