% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_network_acl}
\alias{ec2_create_network_acl}
\title{Creates a network ACL in a VPC}
\usage{
ec2_create_network_acl(DryRun, VpcId, TagSpecifications)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{VpcId}{[required] The ID of the VPC.}

\item{TagSpecifications}{The tags to assign to the network ACL.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NetworkAcl = list(
    Associations = list(
      list(
        NetworkAclAssociationId = "string",
        NetworkAclId = "string",
        SubnetId = "string"
      )
    ),
    Entries = list(
      list(
        CidrBlock = "string",
        Egress = TRUE|FALSE,
        IcmpTypeCode = list(
          Code = 123,
          Type = 123
        ),
        Ipv6CidrBlock = "string",
        PortRange = list(
          From = 123,
          To = 123
        ),
        Protocol = "string",
        RuleAction = "allow"|"deny",
        RuleNumber = 123
      )
    ),
    IsDefault = TRUE|FALSE,
    NetworkAclId = "string",
    Tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    VpcId = "string",
    OwnerId = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a network ACL in a VPC. Network ACLs provide an optional layer
of security (in addition to security groups) for the instances in your
VPC.

For more information, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/}{Network ACLs} in the \emph{Amazon
Virtual Private Cloud User Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_network_acl(
  DryRun = TRUE|FALSE,
  VpcId = "string",
  TagSpecifications = list(
    list(
      ResourceType = "client-vpn-endpoint"|"customer-gateway"|"dedicated-host"|"dhcp-options"|"egress-only-internet-gateway"|"elastic-ip"|"elastic-gpu"|"export-image-task"|"export-instance-task"|"fleet"|"fpga-image"|"host-reservation"|"image"|"import-image-task"|"import-snapshot-task"|"instance"|"internet-gateway"|"key-pair"|"launch-template"|"local-gateway-route-table-vpc-association"|"natgateway"|"network-acl"|"network-interface"|"network-insights-analysis"|"network-insights-path"|"placement-group"|"reserved-instances"|"route-table"|"security-group"|"snapshot"|"spot-fleet-request"|"spot-instances-request"|"subnet"|"traffic-mirror-filter"|"traffic-mirror-session"|"traffic-mirror-target"|"transit-gateway"|"transit-gateway-attachment"|"transit-gateway-connect-peer"|"transit-gateway-multicast-domain"|"transit-gateway-route-table"|"volume"|"vpc"|"vpc-peering-connection"|"vpn-connection"|"vpn-gateway"|"vpc-flow-log",
      Tags = list(
        list(
          Key = "string",
          Value = "string"
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# This example creates a network ACL for the specified VPC.
svc$create_network_acl(
  VpcId = "vpc-a01106c2"
)
}

}
\keyword{internal}
