% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_network_insights_access_scope}
\alias{ec2_create_network_insights_access_scope}
\title{Creates a Network Access Scope}
\usage{
ec2_create_network_insights_access_scope(
  MatchPaths = NULL,
  ExcludePaths = NULL,
  ClientToken,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{MatchPaths}{The paths to match.}

\item{ExcludePaths}{The paths to exclude.}

\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{How to ensure idempotency}.}

\item{TagSpecifications}{The tags to apply.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a Network Access Scope.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_network_insights_access_scope/} for full documentation.
}
\keyword{internal}
