% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_restore_image_task}
\alias{ec2_create_restore_image_task}
\title{Starts a task that restores an AMI from an S3 object that was previously
created by using CreateStoreImageTask}
\usage{
ec2_create_restore_image_task(Bucket, ObjectKey, Name,
  TagSpecifications, DryRun)
}
\arguments{
\item{Bucket}{[required] The name of the S3 bucket that contains the stored AMI object.}

\item{ObjectKey}{[required] The name of the stored AMI object in the bucket.}

\item{Name}{The name for the restored AMI. The name must be unique for AMIs in the
Region for this account. If you do not provide a name, the new AMI gets
the same name as the original AMI.}

\item{TagSpecifications}{The tags to apply to the AMI and snapshots on restoration. You can tag
the AMI, the snapshots, or both.
\itemize{
\item To tag the AMI, the value for \code{ResourceType} must be \code{image}.
\item To tag the snapshots, the value for \code{ResourceType} must be
\code{snapshot}. The same tag is applied to all of the snapshots that are
created.
}}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:\preformatted{list(
  ImageId = "string"
)
}
}
\description{
Starts a task that restores an AMI from an S3 object that was previously
created by using
\code{\link[=ec2_create_store_image_task]{create_store_image_task}}.

To use this API, you must have the required permissions. For more
information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ami-store-restore.html#ami-s3-permissions}{Permissions for storing and restoring AMIs using S3}
in the \emph{Amazon Elastic Compute Cloud User Guide}.

For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ami-store-restore.html}{Store and restore an AMI using S3}
in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_restore_image_task(
  Bucket = "string",
  ObjectKey = "string",
  Name = "string",
  TagSpecifications = list(
    list(
      ResourceType = "client-vpn-endpoint"|"customer-gateway"|"dedicated-host"|"dhcp-options"|"egress-only-internet-gateway"|"elastic-ip"|"elastic-gpu"|"export-image-task"|"export-instance-task"|"fleet"|"fpga-image"|"host-reservation"|"image"|"import-image-task"|"import-snapshot-task"|"instance"|"internet-gateway"|"key-pair"|"launch-template"|"local-gateway-route-table-vpc-association"|"natgateway"|"network-acl"|"network-interface"|"network-insights-analysis"|"network-insights-path"|"placement-group"|"reserved-instances"|"route-table"|"security-group"|"snapshot"|"spot-fleet-request"|"spot-instances-request"|"subnet"|"traffic-mirror-filter"|"traffic-mirror-session"|"traffic-mirror-target"|"transit-gateway"|"transit-gateway-attachment"|"transit-gateway-connect-peer"|"transit-gateway-multicast-domain"|"transit-gateway-route-table"|"volume"|"vpc"|"vpc-peering-connection"|"vpn-connection"|"vpn-gateway"|"vpc-flow-log",
      Tags = list(
        list(
          Key = "string",
          Value = "string"
        )
      )
    )
  ),
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
