% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_route_server}
\alias{ec2_create_route_server}
\title{Creates a new route server to manage dynamic routing in a VPC}
\usage{
ec2_create_route_server(
  AmazonSideAsn,
  ClientToken = NULL,
  DryRun = NULL,
  PersistRoutes = NULL,
  PersistRoutesDuration = NULL,
  SnsNotificationsEnabled = NULL,
  TagSpecifications = NULL
)
}
\arguments{
\item{AmazonSideAsn}{[required] The private Autonomous System Number (ASN) for the Amazon side of the
BGP session. Valid values are from 1 to 4294967295. We recommend using a
private ASN in the 64512–65534 (16-bit ASN) or 4200000000–4294967294
(32-bit ASN) range.}

\item{ClientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}

\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{PersistRoutes}{Indicates whether routes should be persisted after all BGP sessions are
terminated.}

\item{PersistRoutesDuration}{The number of minutes a route server will wait after BGP is
re-established to unpersist the routes in the FIB and RIB. Value must be
in the range of 1-5. Required if PersistRoutes is \code{enabled}.

If you set the duration to 1 minute, then when your network appliance
re-establishes BGP with route server, it has 1 minute to relearn it's
adjacent network and advertise those routes to route server before route
server resumes normal functionality. In most cases, 1 minute is probably
sufficient. If, however, you have concerns that your BGP network may not
be capable of fully re-establishing and re-learning everything in 1
minute, you can increase the duration up to 5 minutes.}

\item{SnsNotificationsEnabled}{Indicates whether SNS notifications should be enabled for route server
events. Enabling SNS notifications persists BGP status changes to an SNS
topic provisioned by Amazon Web Services.}

\item{TagSpecifications}{The tags to apply to the route server during creation.}
}
\description{
Creates a new route server to manage dynamic routing in a VPC.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_route_server/} for full documentation.
}
\keyword{internal}
