% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_route_server_endpoint}
\alias{ec2_create_route_server_endpoint}
\title{Creates a new endpoint for a route server in a specified subnet}
\usage{
ec2_create_route_server_endpoint(
  RouteServerId,
  SubnetId,
  ClientToken = NULL,
  DryRun = NULL,
  TagSpecifications = NULL
)
}
\arguments{
\item{RouteServerId}{[required] The ID of the route server for which to create an endpoint.}

\item{SubnetId}{[required] The ID of the subnet in which to create the route server endpoint.}

\item{ClientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}

\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{TagSpecifications}{The tags to apply to the route server endpoint during creation.}
}
\description{
Creates a new endpoint for a route server in a specified subnet.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_route_server_endpoint/} for full documentation.
}
\keyword{internal}
