% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_route_server_peer}
\alias{ec2_create_route_server_peer}
\title{Creates a new BGP peer for a specified route server endpoint}
\usage{
ec2_create_route_server_peer(
  RouteServerEndpointId,
  PeerAddress,
  BgpOptions,
  DryRun = NULL,
  TagSpecifications = NULL
)
}
\arguments{
\item{RouteServerEndpointId}{[required] The ID of the route server endpoint for which to create a peer.}

\item{PeerAddress}{[required] The IPv4 address of the peer device.}

\item{BgpOptions}{[required] The BGP options for the peer, including ASN (Autonomous System Number)
and BFD (Bidrectional Forwarding Detection) settings.}

\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{TagSpecifications}{The tags to apply to the route server peer during creation.}
}
\description{
Creates a new BGP peer for a specified route server endpoint.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_route_server_peer/} for full documentation.
}
\keyword{internal}
