% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_secondary_network}
\alias{ec2_create_secondary_network}
\title{Creates a secondary network}
\usage{
ec2_create_secondary_network(
  ClientToken = NULL,
  DryRun = NULL,
  Ipv4CidrBlock,
  NetworkType,
  TagSpecifications = NULL
)
}
\arguments{
\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensure Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Ipv4CidrBlock}{[required] The IPv4 CIDR block for the secondary network. The CIDR block size must
be between /12 and /28.}

\item{NetworkType}{[required] The type of secondary network.}

\item{TagSpecifications}{The tags to assign to the secondary network.}
}
\description{
Creates a secondary network.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_secondary_network/} for full documentation.
}
\keyword{internal}
