% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_secondary_subnet}
\alias{ec2_create_secondary_subnet}
\title{Creates a secondary subnet in a secondary network}
\usage{
ec2_create_secondary_subnet(
  ClientToken = NULL,
  AvailabilityZone = NULL,
  AvailabilityZoneId = NULL,
  DryRun = NULL,
  Ipv4CidrBlock,
  SecondaryNetworkId,
  TagSpecifications = NULL
)
}
\arguments{
\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensure Idempotency}.}

\item{AvailabilityZone}{The Availability Zone for the secondary subnet. You cannot specify both
\code{AvailabilityZone} and \code{AvailabilityZoneId} in the same request.}

\item{AvailabilityZoneId}{The ID of the Availability Zone for the secondary subnet. This option is
preferred over \code{AvailabilityZone} as it provides a consistent identifier
across Amazon Web Services accounts. You cannot specify both
\code{AvailabilityZone} and \code{AvailabilityZoneId} in the same request.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Ipv4CidrBlock}{[required] The IPv4 CIDR block for the secondary subnet. The CIDR block size must
be between /12 and /28.}

\item{SecondaryNetworkId}{[required] The ID of the secondary network in which to create the secondary subnet.}

\item{TagSpecifications}{The tags to assign to the secondary subnet.}
}
\description{
Creates a secondary subnet in a secondary network.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_secondary_subnet/} for full documentation.
}
\keyword{internal}
