% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_spot_datafeed_subscription}
\alias{ec2_create_spot_datafeed_subscription}
\title{Creates a data feed for Spot Instances, enabling you to view Spot
Instance usage logs}
\usage{
ec2_create_spot_datafeed_subscription(Bucket, DryRun, Prefix)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket in which to store the Spot Instance
data feed. For more information about bucket names, see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/BucketRestrictions.html#bucketnamingrules}{Rules for bucket naming}
in the \emph{Amazon S3 Developer Guide}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Prefix}{The prefix for the data feed file names.}
}
\value{
A list with the following syntax:\preformatted{list(
  SpotDatafeedSubscription = list(
    Bucket = "string",
    Fault = list(
      Code = "string",
      Message = "string"
    ),
    OwnerId = "string",
    Prefix = "string",
    State = "Active"|"Inactive"
  )
)
}
}
\description{
Creates a data feed for Spot Instances, enabling you to view Spot
Instance usage logs. You can create one data feed per AWS account. For
more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/spot-data-feeds.html}{Spot Instance data feed}
in the \emph{Amazon EC2 User Guide for Linux Instances}.
}
\section{Request syntax}{
\preformatted{svc$create_spot_datafeed_subscription(
  Bucket = "string",
  DryRun = TRUE|FALSE,
  Prefix = "string"
)
}
}

\examples{
\dontrun{
# This example creates a Spot Instance data feed for your AWS account.
svc$create_spot_datafeed_subscription(
  Bucket = "my-s3-bucket",
  Prefix = "spotdata"
)
}

}
\keyword{internal}
