% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_store_image_task}
\alias{ec2_create_store_image_task}
\title{Stores an AMI as a single object in an S3 bucket}
\usage{
ec2_create_store_image_task(ImageId, Bucket, S3ObjectTags, DryRun)
}
\arguments{
\item{ImageId}{[required] The ID of the AMI.}

\item{Bucket}{[required] The name of the S3 bucket in which the AMI object will be stored. The
bucket must be in the Region in which the request is being made. The AMI
object appears in the bucket only after the upload task has completed.}

\item{S3ObjectTags}{The tags to apply to the AMI object that will be stored in the S3
bucket.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:\preformatted{list(
  ObjectKey = "string"
)
}
}
\description{
Stores an AMI as a single object in an S3 bucket.

To use this API, you must have the required permissions. For more
information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ami-store-restore.html#ami-s3-permissions}{Permissions for storing and restoring AMIs using S3}
in the \emph{Amazon Elastic Compute Cloud User Guide}.

For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ami-store-restore.html}{Store and restore an AMI using S3}
in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_store_image_task(
  ImageId = "string",
  Bucket = "string",
  S3ObjectTags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
