% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_transit_gateway_metering_policy}
\alias{ec2_create_transit_gateway_metering_policy}
\title{Creates a metering policy for a transit gateway to track and measure
network traffic}
\usage{
ec2_create_transit_gateway_metering_policy(
  TransitGatewayId,
  MiddleboxAttachmentIds = NULL,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayId}{[required] The ID of the transit gateway for which to create the metering policy.}

\item{MiddleboxAttachmentIds}{The IDs of the middlebox attachments to include in the metering policy.}

\item{TagSpecifications}{The tags to assign to the metering policy.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a metering policy for a transit gateway to track and measure network traffic.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_transit_gateway_metering_policy/} for full documentation.
}
\keyword{internal}
