% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_transit_gateway_metering_policy_entry}
\alias{ec2_create_transit_gateway_metering_policy_entry}
\title{Creates an entry in a transit gateway metering policy to define traffic
measurement rules}
\usage{
ec2_create_transit_gateway_metering_policy_entry(
  TransitGatewayMeteringPolicyId,
  PolicyRuleNumber,
  SourceTransitGatewayAttachmentId = NULL,
  SourceTransitGatewayAttachmentType = NULL,
  SourceCidrBlock = NULL,
  SourcePortRange = NULL,
  DestinationTransitGatewayAttachmentId = NULL,
  DestinationTransitGatewayAttachmentType = NULL,
  DestinationCidrBlock = NULL,
  DestinationPortRange = NULL,
  Protocol = NULL,
  MeteredAccount,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayMeteringPolicyId}{[required] The ID of the transit gateway metering policy to add the entry to.}

\item{PolicyRuleNumber}{[required] The rule number for the metering policy entry. Rules are processed in
order from lowest to highest number.}

\item{SourceTransitGatewayAttachmentId}{The ID of the source transit gateway attachment for traffic matching.}

\item{SourceTransitGatewayAttachmentType}{The type of the source transit gateway attachment for traffic matching.
Note that the \code{tgw-peering} resource type has been deprecated. To
configure metering policies for Connect, use the transport attachment
type.}

\item{SourceCidrBlock}{The source CIDR block for traffic matching.}

\item{SourcePortRange}{The source port range for traffic matching.}

\item{DestinationTransitGatewayAttachmentId}{The ID of the destination transit gateway attachment for traffic
matching.}

\item{DestinationTransitGatewayAttachmentType}{The type of the destination transit gateway attachment for traffic
matching. Note that the \code{tgw-peering} resource type has been deprecated.
To configure metering policies for Connect, use the transport attachment
type.}

\item{DestinationCidrBlock}{The destination CIDR block for traffic matching.}

\item{DestinationPortRange}{The destination port range for traffic matching.}

\item{Protocol}{The protocol for traffic matching (1, 6, 17, etc.).}

\item{MeteredAccount}{[required] The Amazon Web Services account ID to which the metered traffic should
be attributed.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates an entry in a transit gateway metering policy to define traffic measurement rules.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_transit_gateway_metering_policy_entry/} for full documentation.
}
\keyword{internal}
