% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_transit_gateway_policy_table_entry}
\alias{ec2_create_transit_gateway_policy_table_entry}
\title{Creates an entry in a transit gateway policy table to route matching
traffic to a specified route table}
\usage{
ec2_create_transit_gateway_policy_table_entry(
  TransitGatewayPolicyTableId,
  PolicyRuleNumber,
  PolicyRule = NULL,
  TargetRouteTableId,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayPolicyTableId}{[required] The ID of the transit gateway policy table.}

\item{PolicyRuleNumber}{[required] The rule number for the policy table entry. Lower rule numbers are evaluated first and take precedence.}

\item{PolicyRule}{The matching criteria for the policy table entry.}

\item{TargetRouteTableId}{[required] The ID of the transit gateway route table to use for traffic matching this rule.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates an entry in a transit gateway policy table to route matching traffic to a specified route table.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_transit_gateway_policy_table_entry/} for full documentation.
}
\keyword{internal}
