% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_transit_gateway_route_table_announcement}
\alias{ec2_create_transit_gateway_route_table_announcement}
\title{Advertises a new transit gateway route table}
\usage{
ec2_create_transit_gateway_route_table_announcement(
  TransitGatewayRouteTableId,
  PeeringAttachmentId,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayRouteTableId}{[required] The ID of the transit gateway route table.}

\item{PeeringAttachmentId}{[required] The ID of the peering attachment.}

\item{TagSpecifications}{The tags specifications applied to the transit gateway route table
announcement.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Advertises a new transit gateway route table.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_transit_gateway_route_table_announcement/} for full documentation.
}
\keyword{internal}
