% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_vpc_encryption_control}
\alias{ec2_create_vpc_encryption_control}
\title{Creates a VPC Encryption Control configuration for a specified VPC}
\usage{
ec2_create_vpc_encryption_control(
  DryRun = NULL,
  VpcId,
  TagSpecifications = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{VpcId}{[required] The ID of the VPC for which to create the encryption control
configuration.}

\item{TagSpecifications}{The tags to apply to the VPC Encryption Control resource.}
}
\description{
Creates a VPC Encryption Control configuration for a specified VPC. VPC Encryption Control enables you to enforce encryption for all data in transit within and between VPCs to meet compliance requirements for standards like HIPAA, FedRAMP, and PCI DSS.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_vpc_encryption_control/} for full documentation.
}
\keyword{internal}
