% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_vpn_concentrator}
\alias{ec2_create_vpn_concentrator}
\title{Creates a VPN concentrator that aggregates multiple VPN connections to a
transit gateway}
\usage{
ec2_create_vpn_concentrator(
  Type,
  TransitGatewayId = NULL,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{Type}{[required] The type of VPN concentrator to create.}

\item{TransitGatewayId}{The ID of the transit gateway to attach the VPN concentrator to.}

\item{TagSpecifications}{The tags to apply to the VPN concentrator during creation.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a VPN concentrator that aggregates multiple VPN connections to a transit gateway.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_vpn_concentrator/} for full documentation.
}
\keyword{internal}
