% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_application_status_check}
\alias{ec2_delete_application_status_check}
\title{Deletes an application status check}
\usage{
ec2_delete_application_status_check(
  ApplicationStatusCheckId,
  ClientToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{ApplicationStatusCheckId}{[required] The ID of the application status check to delete.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure that the operation completes no more than one time. If you retry a request with the same token, the service ignores the request but does not return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes an application status check. The following rules apply:

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_application_status_check/} for full documentation.
}
\keyword{internal}
