% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_capacity_manager_data_export}
\alias{ec2_delete_capacity_manager_data_export}
\title{Deletes an existing Capacity Manager data export configuration}
\usage{
ec2_delete_capacity_manager_data_export(
  CapacityManagerDataExportId,
  DryRun = NULL
)
}
\arguments{
\item{CapacityManagerDataExportId}{[required] The unique identifier of the data export configuration to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes an existing Capacity Manager data export configuration. This stops future scheduled exports but does not delete previously exported files from S3.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_capacity_manager_data_export/} for full documentation.
}
\keyword{internal}
