% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_client_vpn_endpoint}
\alias{ec2_delete_client_vpn_endpoint}
\title{Deletes the specified Client VPN endpoint}
\usage{
ec2_delete_client_vpn_endpoint(ClientVpnEndpointId, DryRun)
}
\arguments{
\item{ClientVpnEndpointId}{[required] The ID of the Client VPN to be deleted.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Status = list(
    Code = "pending-associate"|"available"|"deleting"|"deleted",
    Message = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified Client VPN endpoint. You must disassociate all
target networks before you can delete a Client VPN endpoint.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_client_vpn_endpoint(
  ClientVpnEndpointId = "string",
  DryRun = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
