% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_coip_pool}
\alias{ec2_delete_coip_pool}
\title{Deletes a pool of customer-owned IP (CoIP) addresses}
\usage{
ec2_delete_coip_pool(CoipPoolId, DryRun = NULL)
}
\arguments{
\item{CoipPoolId}{[required] The ID of the CoIP pool that you want to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes a pool of customer-owned IP (CoIP) addresses.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_coip_pool/} for full documentation.
}
\keyword{internal}
