% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_fpga_image}
\alias{ec2_delete_fpga_image}
\title{Deletes the specified Amazon FPGA Image (AFI)}
\usage{
ec2_delete_fpga_image(DryRun, FpgaImageId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{FpgaImageId}{[required] The ID of the AFI.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Return = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified Amazon FPGA Image (AFI).
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_fpga_image(
  DryRun = TRUE|FALSE,
  FpgaImageId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
