% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_image_usage_report}
\alias{ec2_delete_image_usage_report}
\title{Deletes the specified image usage report}
\usage{
ec2_delete_image_usage_report(ReportId, DryRun = NULL)
}
\arguments{
\item{ReportId}{[required] The ID of the report to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified image usage report.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_image_usage_report/} for full documentation.
}
\keyword{internal}
