% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_ipam_external_resource_verification_token}
\alias{ec2_delete_ipam_external_resource_verification_token}
\title{Delete a verification token}
\usage{
ec2_delete_ipam_external_resource_verification_token(
  DryRun = NULL,
  IpamExternalResourceVerificationTokenId
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamExternalResourceVerificationTokenId}{[required] The token ID.}
}
\description{
Delete a verification token. A verification token is an Amazon Web Services-generated random value that you can use to prove ownership of an external resource. For example, you can use a verification token to validate that you control a public IP address range when you bring an IP address range to Amazon Web Services (BYOIP).

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_ipam_external_resource_verification_token/} for full documentation.
}
\keyword{internal}
