% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_ipam_policy}
\alias{ec2_delete_ipam_policy}
\title{Deletes an IPAM policy}
\usage{
ec2_delete_ipam_policy(DryRun = NULL, IpamPolicyId)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPolicyId}{[required] The ID of the IPAM policy to delete.}
}
\description{
Deletes an IPAM policy.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_ipam_policy/} for full documentation.
}
\keyword{internal}
