% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_ipam_prefix_list_resolver}
\alias{ec2_delete_ipam_prefix_list_resolver}
\title{Deletes an IPAM prefix list resolver}
\usage{
ec2_delete_ipam_prefix_list_resolver(DryRun = NULL, IpamPrefixListResolverId)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPrefixListResolverId}{[required] The ID of the IPAM prefix list resolver to delete.}
}
\description{
Deletes an IPAM prefix list resolver. Before deleting a resolver, you must first delete all resolver targets associated with it.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_ipam_prefix_list_resolver/} for full documentation.
}
\keyword{internal}
