% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_ipam_prefix_list_resolver_target}
\alias{ec2_delete_ipam_prefix_list_resolver_target}
\title{Deletes an IPAM prefix list resolver target}
\usage{
ec2_delete_ipam_prefix_list_resolver_target(
  DryRun = NULL,
  IpamPrefixListResolverTargetId
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPrefixListResolverTargetId}{[required] The ID of the IPAM prefix list resolver target to delete.}
}
\description{
Deletes an IPAM prefix list resolver target. This removes the association between the resolver and the managed prefix list, stopping automatic CIDR synchronization.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_ipam_prefix_list_resolver_target/} for full documentation.
}
\keyword{internal}
