% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_ipam_resource_discovery}
\alias{ec2_delete_ipam_resource_discovery}
\title{Deletes an IPAM resource discovery}
\usage{
ec2_delete_ipam_resource_discovery(DryRun = NULL, IpamResourceDiscoveryId)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamResourceDiscoveryId}{[required] The IPAM resource discovery ID.}
}
\description{
Deletes an IPAM resource discovery. A resource discovery is an IPAM component that enables IPAM to manage and monitor resources that belong to the owning account.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_ipam_resource_discovery/} for full documentation.
}
\keyword{internal}
