% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_ipam_scope}
\alias{ec2_delete_ipam_scope}
\title{Delete the scope for an IPAM}
\usage{
ec2_delete_ipam_scope(DryRun = NULL, IpamScopeId)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamScopeId}{[required] The ID of the scope to delete.}
}
\description{
Delete the scope for an IPAM. You cannot delete the default scopes.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_ipam_scope/} for full documentation.
}
\keyword{internal}
