% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_key_pair}
\alias{ec2_delete_key_pair}
\title{Deletes the specified key pair, by removing the public key from Amazon
EC2}
\usage{
ec2_delete_key_pair(KeyName = NULL, KeyPairId = NULL, DryRun = NULL)
}
\arguments{
\item{KeyName}{The name of the key pair.}

\item{KeyPairId}{The ID of the key pair.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified key pair, by removing the public key from Amazon EC2.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_key_pair/} for full documentation.
}
\keyword{internal}
