% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_local_gateway_route}
\alias{ec2_delete_local_gateway_route}
\title{Deletes the specified route from the specified local gateway route table}
\usage{
ec2_delete_local_gateway_route(DestinationCidrBlock,
  LocalGatewayRouteTableId, DryRun)
}
\arguments{
\item{DestinationCidrBlock}{[required] The CIDR range for the route. This must match the CIDR for the route
exactly.}

\item{LocalGatewayRouteTableId}{[required] The ID of the local gateway route table.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:\preformatted{list(
  Route = list(
    DestinationCidrBlock = "string",
    LocalGatewayVirtualInterfaceGroupId = "string",
    Type = "static"|"propagated",
    State = "pending"|"active"|"blackhole"|"deleting"|"deleted",
    LocalGatewayRouteTableId = "string",
    LocalGatewayRouteTableArn = "string",
    OwnerId = "string"
  )
)
}
}
\description{
Deletes the specified route from the specified local gateway route
table.
}
\section{Request syntax}{
\preformatted{svc$delete_local_gateway_route(
  DestinationCidrBlock = "string",
  LocalGatewayRouteTableId = "string",
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
