% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_local_gateway_virtual_interface}
\alias{ec2_delete_local_gateway_virtual_interface}
\title{Deletes the specified local gateway virtual interface}
\usage{
ec2_delete_local_gateway_virtual_interface(
  LocalGatewayVirtualInterfaceId,
  DryRun = NULL
)
}
\arguments{
\item{LocalGatewayVirtualInterfaceId}{[required] The ID of the local virtual interface to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified local gateway virtual interface.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_local_gateway_virtual_interface/} for full documentation.
}
\keyword{internal}
