% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_local_gateway_virtual_interface_group}
\alias{ec2_delete_local_gateway_virtual_interface_group}
\title{Delete the specified local gateway interface group}
\usage{
ec2_delete_local_gateway_virtual_interface_group(
  LocalGatewayVirtualInterfaceGroupId,
  DryRun = NULL
)
}
\arguments{
\item{LocalGatewayVirtualInterfaceGroupId}{[required] The ID of the local gateway virtual interface group to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Delete the specified local gateway interface group.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_local_gateway_virtual_interface_group/} for full documentation.
}
\keyword{internal}
