% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_managed_prefix_list}
\alias{ec2_delete_managed_prefix_list}
\title{Deletes the specified managed prefix list}
\usage{
ec2_delete_managed_prefix_list(DryRun, PrefixListId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{PrefixListId}{[required] The ID of the prefix list.}
}
\value{
A list with the following syntax:\preformatted{list(
  PrefixList = list(
    PrefixListId = "string",
    AddressFamily = "string",
    State = "create-in-progress"|"create-complete"|"create-failed"|"modify-in-progress"|"modify-complete"|"modify-failed"|"restore-in-progress"|"restore-complete"|"restore-failed"|"delete-in-progress"|"delete-complete"|"delete-failed",
    StateMessage = "string",
    PrefixListArn = "string",
    PrefixListName = "string",
    MaxEntries = 123,
    Version = 123,
    Tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    OwnerId = "string"
  )
)
}
}
\description{
Deletes the specified managed prefix list. You must first remove all
references to the prefix list in your resources.
}
\section{Request syntax}{
\preformatted{svc$delete_managed_prefix_list(
  DryRun = TRUE|FALSE,
  PrefixListId = "string"
)
}
}

\keyword{internal}
