% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_network_insights_access_scope}
\alias{ec2_delete_network_insights_access_scope}
\title{Deletes the specified Network Access Scope}
\usage{
ec2_delete_network_insights_access_scope(
  DryRun = NULL,
  NetworkInsightsAccessScopeId
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{NetworkInsightsAccessScopeId}{[required] The ID of the Network Access Scope.}
}
\description{
Deletes the specified Network Access Scope.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_network_insights_access_scope/} for full documentation.
}
\keyword{internal}
