% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_network_insights_access_scope_analysis}
\alias{ec2_delete_network_insights_access_scope_analysis}
\title{Deletes the specified Network Access Scope analysis}
\usage{
ec2_delete_network_insights_access_scope_analysis(
  NetworkInsightsAccessScopeAnalysisId,
  DryRun = NULL
)
}
\arguments{
\item{NetworkInsightsAccessScopeAnalysisId}{[required] The ID of the Network Access Scope analysis.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified Network Access Scope analysis.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_network_insights_access_scope_analysis/} for full documentation.
}
\keyword{internal}
