% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_network_insights_path}
\alias{ec2_delete_network_insights_path}
\title{Deletes the specified path}
\usage{
ec2_delete_network_insights_path(DryRun, NetworkInsightsPathId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{NetworkInsightsPathId}{[required] The ID of the path.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NetworkInsightsPathId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified path.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_network_insights_path(
  DryRun = TRUE|FALSE,
  NetworkInsightsPathId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
