% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_route_server}
\alias{ec2_delete_route_server}
\title{Deletes the specified route server}
\usage{
ec2_delete_route_server(RouteServerId, DryRun = NULL)
}
\arguments{
\item{RouteServerId}{[required] The ID of the route server to delete.}

\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified route server.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_route_server/} for full documentation.
}
\keyword{internal}
