% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_route_server_endpoint}
\alias{ec2_delete_route_server_endpoint}
\title{Deletes the specified route server endpoint}
\usage{
ec2_delete_route_server_endpoint(RouteServerEndpointId, DryRun = NULL)
}
\arguments{
\item{RouteServerEndpointId}{[required] The ID of the route server endpoint to delete.}

\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified route server endpoint.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_route_server_endpoint/} for full documentation.
}
\keyword{internal}
