% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_route_table}
\alias{ec2_delete_route_table}
\title{Deletes the specified route table}
\usage{
ec2_delete_route_table(DryRun = NULL, RouteTableId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{RouteTableId}{[required] The ID of the route table.}
}
\description{
Deletes the specified route table. You must disassociate the route table from any subnets before you can delete it. You can't delete the main route table.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_route_table/} for full documentation.
}
\keyword{internal}
