% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_secondary_network}
\alias{ec2_delete_secondary_network}
\title{Deletes a secondary network}
\usage{
ec2_delete_secondary_network(
  ClientToken = NULL,
  DryRun = NULL,
  SecondaryNetworkId
)
}
\arguments{
\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensure Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{SecondaryNetworkId}{[required] The ID of the secondary network.}
}
\description{
Deletes a secondary network. You must delete all secondary subnets in the secondary network before you can delete the secondary network.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_secondary_network/} for full documentation.
}
\keyword{internal}
