% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_secondary_subnet}
\alias{ec2_delete_secondary_subnet}
\title{Deletes a secondary subnet}
\usage{
ec2_delete_secondary_subnet(
  ClientToken = NULL,
  DryRun = NULL,
  SecondarySubnetId
)
}
\arguments{
\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensure Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{SecondarySubnetId}{[required] The ID of the secondary subnet to delete.}
}
\description{
Deletes a secondary subnet. A secondary subnet must not contain any secondary interfaces prior to deletion.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_secondary_subnet/} for full documentation.
}
\keyword{internal}
