% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_connect}
\alias{ec2_delete_transit_gateway_connect}
\title{Deletes the specified Connect attachment}
\usage{
ec2_delete_transit_gateway_connect(TransitGatewayAttachmentId, DryRun)
}
\arguments{
\item{TransitGatewayAttachmentId}{[required] The ID of the Connect attachment.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TransitGatewayConnect = list(
    TransitGatewayAttachmentId = "string",
    TransportTransitGatewayAttachmentId = "string",
    TransitGatewayId = "string",
    State = "initiating"|"initiatingRequest"|"pendingAcceptance"|"rollingBack"|"pending"|"available"|"modifying"|"deleting"|"deleted"|"failed"|"rejected"|"rejecting"|"failing",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    Options = list(
      Protocol = "gre"
    ),
    Tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified Connect attachment. You must first delete any
Connect peers for the attachment.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_transit_gateway_connect(
  TransitGatewayAttachmentId = "string",
  DryRun = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
