% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_metering_policy}
\alias{ec2_delete_transit_gateway_metering_policy}
\title{Deletes a transit gateway metering policy}
\usage{
ec2_delete_transit_gateway_metering_policy(
  TransitGatewayMeteringPolicyId,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayMeteringPolicyId}{[required] The ID of the transit gateway metering policy to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes a transit gateway metering policy.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_transit_gateway_metering_policy/} for full documentation.
}
\keyword{internal}
