% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_metering_policy_entry}
\alias{ec2_delete_transit_gateway_metering_policy_entry}
\title{Deletes an entry from a transit gateway metering policy}
\usage{
ec2_delete_transit_gateway_metering_policy_entry(
  TransitGatewayMeteringPolicyId,
  PolicyRuleNumber,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayMeteringPolicyId}{[required] The ID of the transit gateway metering policy containing the entry to
delete.}

\item{PolicyRuleNumber}{[required] The rule number of the metering policy entry to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes an entry from a transit gateway metering policy.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_transit_gateway_metering_policy_entry/} for full documentation.
}
\keyword{internal}
