% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_multicast_domain}
\alias{ec2_delete_transit_gateway_multicast_domain}
\title{Deletes the specified transit gateway multicast domain}
\usage{
ec2_delete_transit_gateway_multicast_domain(
  TransitGatewayMulticastDomainId, DryRun)
}
\arguments{
\item{TransitGatewayMulticastDomainId}{[required] The ID of the transit gateway multicast domain.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TransitGatewayMulticastDomain = list(
    TransitGatewayMulticastDomainId = "string",
    TransitGatewayId = "string",
    TransitGatewayMulticastDomainArn = "string",
    OwnerId = "string",
    Options = list(
      Igmpv2Support = "enable"|"disable",
      StaticSourcesSupport = "enable"|"disable",
      AutoAcceptSharedAssociations = "enable"|"disable"
    ),
    State = "pending"|"available"|"deleting"|"deleted",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    Tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified transit gateway multicast domain.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_transit_gateway_multicast_domain(
  TransitGatewayMulticastDomainId = "string",
  DryRun = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
