% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_policy_table}
\alias{ec2_delete_transit_gateway_policy_table}
\title{Deletes the specified transit gateway policy table}
\usage{
ec2_delete_transit_gateway_policy_table(
  TransitGatewayPolicyTableId,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayPolicyTableId}{[required] The transit gateway policy table to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified transit gateway policy table.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_transit_gateway_policy_table/} for full documentation.
}
\keyword{internal}
