% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_policy_table_entry}
\alias{ec2_delete_transit_gateway_policy_table_entry}
\title{Deletes the specified transit gateway policy table entry}
\usage{
ec2_delete_transit_gateway_policy_table_entry(
  TransitGatewayPolicyTableId,
  PolicyRuleNumber,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayPolicyTableId}{[required] The ID of the transit gateway policy table.}

\item{PolicyRuleNumber}{[required] The rule number of the policy table entry to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified transit gateway policy table entry.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_transit_gateway_policy_table_entry/} for full documentation.
}
\keyword{internal}
