% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_prefix_list_reference}
\alias{ec2_delete_transit_gateway_prefix_list_reference}
\title{Deletes a reference (route) to a prefix list in a specified transit
gateway route table}
\usage{
ec2_delete_transit_gateway_prefix_list_reference(
  TransitGatewayRouteTableId, PrefixListId, DryRun)
}
\arguments{
\item{TransitGatewayRouteTableId}{[required] The ID of the route table.}

\item{PrefixListId}{[required] The ID of the prefix list.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:\preformatted{list(
  TransitGatewayPrefixListReference = list(
    TransitGatewayRouteTableId = "string",
    PrefixListId = "string",
    PrefixListOwnerId = "string",
    State = "pending"|"available"|"modifying"|"deleting",
    Blackhole = TRUE|FALSE,
    TransitGatewayAttachment = list(
      TransitGatewayAttachmentId = "string",
      ResourceType = "vpc"|"vpn"|"direct-connect-gateway"|"connect"|"peering"|"tgw-peering",
      ResourceId = "string"
    )
  )
)
}
}
\description{
Deletes a reference (route) to a prefix list in a specified transit
gateway route table.
}
\section{Request syntax}{
\preformatted{svc$delete_transit_gateway_prefix_list_reference(
  TransitGatewayRouteTableId = "string",
  PrefixListId = "string",
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
