% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_route_table_announcement}
\alias{ec2_delete_transit_gateway_route_table_announcement}
\title{Advertises to the transit gateway that a transit gateway route table is
deleted}
\usage{
ec2_delete_transit_gateway_route_table_announcement(
  TransitGatewayRouteTableAnnouncementId,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayRouteTableAnnouncementId}{[required] The transit gateway route table ID that's being deleted.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Advertises to the transit gateway that a transit gateway route table is deleted.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_transit_gateway_route_table_announcement/} for full documentation.
}
\keyword{internal}
