% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_vpc_encryption_control}
\alias{ec2_delete_vpc_encryption_control}
\title{Deletes a VPC Encryption Control configuration}
\usage{
ec2_delete_vpc_encryption_control(DryRun = NULL, VpcEncryptionControlId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{VpcEncryptionControlId}{[required] The ID of the VPC Encryption Control resource to delete.}
}
\description{
Deletes a VPC Encryption Control configuration. This removes the encryption policy enforcement from the specified VPC.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_vpc_encryption_control/} for full documentation.
}
\keyword{internal}
