% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_vpc_endpoint_connection_notifications}
\alias{ec2_delete_vpc_endpoint_connection_notifications}
\title{Deletes one or more VPC endpoint connection notifications}
\usage{
ec2_delete_vpc_endpoint_connection_notifications(DryRun,
  ConnectionNotificationIds)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ConnectionNotificationIds}{[required] One or more notification IDs.}
}
\value{
A list with the following syntax:\preformatted{list(
  Unsuccessful = list(
    list(
      Error = list(
        Code = "string",
        Message = "string"
      ),
      ResourceId = "string"
    )
  )
)
}
}
\description{
Deletes one or more VPC endpoint connection notifications.
}
\section{Request syntax}{
\preformatted{svc$delete_vpc_endpoint_connection_notifications(
  DryRun = TRUE|FALSE,
  ConnectionNotificationIds = list(
    "string"
  )
)
}
}

\keyword{internal}
