% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_vpc_endpoints}
\alias{ec2_delete_vpc_endpoints}
\title{Deletes one or more specified VPC endpoints}
\usage{
ec2_delete_vpc_endpoints(DryRun, VpcEndpointIds)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{VpcEndpointIds}{[required] One or more VPC endpoint IDs.}
}
\value{
A list with the following syntax:\preformatted{list(
  Unsuccessful = list(
    list(
      Error = list(
        Code = "string",
        Message = "string"
      ),
      ResourceId = "string"
    )
  )
)
}
}
\description{
Deletes one or more specified VPC endpoints. You can delete any of the
following types of VPC endpoints.
\itemize{
\item Gateway endpoint,
\item Gateway Load Balancer endpoint,
\item Interface endpoint
}

The following rules apply when you delete a VPC endpoint:
\itemize{
\item When you delete a gateway endpoint, we delete the endpoint routes in
the route tables that are associated with the endpoint.
\item When you delete a Gateway Load Balancer endpoint, we delete the
endpoint network interfaces.

You can only delete Gateway Load Balancer endpoints when the routes
that are associated with the endpoint are deleted.
\item When you delete an interface endpoint, we delete the endpoint
network interfaces.
}
}
\section{Request syntax}{
\preformatted{svc$delete_vpc_endpoints(
  DryRun = TRUE|FALSE,
  VpcEndpointIds = list(
    "string"
  )
)
}
}

\keyword{internal}
