% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_vpn_concentrator}
\alias{ec2_delete_vpn_concentrator}
\title{Deletes the specified VPN concentrator}
\usage{
ec2_delete_vpn_concentrator(VpnConcentratorId, DryRun = NULL)
}
\arguments{
\item{VpnConcentratorId}{[required] The ID of the VPN concentrator to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified VPN concentrator.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_vpn_concentrator/} for full documentation.
}
\keyword{internal}
