% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_deregister_image}
\alias{ec2_deregister_image}
\title{Deregisters the specified AMI}
\usage{
ec2_deregister_image(ImageId, DryRun = NULL)
}
\arguments{
\item{ImageId}{[required] The ID of the AMI.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deregisters the specified AMI. After you deregister an AMI, it can't be used to launch new instances.

See \url{https://paws-r.github.io/docs/ec2/deregister_image.html} for full documentation.
}
\keyword{internal}
