% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_account_vpc_encryption_control}
\alias{ec2_describe_account_vpc_encryption_control}
\title{Describes the account-level VPC Encryption Control configuration for
your account}
\usage{
ec2_describe_account_vpc_encryption_control(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the account-level VPC Encryption Control configuration for your account. VPC Encryption Control enables you to enforce encryption for all data in transit within and between VPCs to meet compliance requirements.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_account_vpc_encryption_control/} for full documentation.
}
\keyword{internal}
