% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_application_status}
\alias{ec2_describe_application_status}
\title{Describes the aggregated application health status for the specified
instances}
\usage{
ec2_describe_application_status(
  InstanceIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{InstanceIds}{The IDs of the instances for which to describe application status.}

\item{Filters}{The filters.
\itemize{
\item \code{availability-zone-id} – The ID of the Availability Zone.
\item \code{status} – The instance-level application status. For valid values and their meanings, see \code{ApplicationStatus}.
}}

\item{MaxResults}{The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.}

\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the aggregated application health status for the specified instances. The following rules apply:

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_application_status/} for full documentation.
}
\keyword{internal}
