% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_application_status_check_associations}
\alias{ec2_describe_application_status_check_associations}
\title{Describes the associations for one or more application status checks}
\usage{
ec2_describe_application_status_check_associations(
  ApplicationStatusCheckIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{ApplicationStatusCheckIds}{The IDs of the application status checks for which to describe associations.}

\item{Filters}{The filters to use to limit the results.
\itemize{
\item \code{association-type} – The type of association. Valid values: \code{tag} and \code{instance-id}.
}}

\item{MaxResults}{The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.}

\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the associations for one or more application status checks. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/}{Application status checks}. To avoid timeouts and retrieve complete results, use the pagination parameters.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_application_status_check_associations/} for full documentation.
}
\keyword{internal}
