% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_application_status_checks}
\alias{ec2_describe_application_status_checks}
\title{Describes application status checks, including configuration details
such as protocol, port, path, thresholds, and associations}
\usage{
ec2_describe_application_status_checks(
  ApplicationStatusCheckIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  IncludeAll = NULL,
  DryRun = NULL
)
}
\arguments{
\item{ApplicationStatusCheckIds}{The IDs of the application status checks to describe.}

\item{Filters}{The filters.
\itemize{
\item \code{aggregation} – The aggregation setting. Valid values: \code{included} and \code{excluded}.
}}

\item{MaxResults}{The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.}

\item{IncludeAll}{Specifies whether to include recently deleted application status checks that remain available during the deletion grace period. If you omit this parameter or set it to \code{false}, the response includes only active checks.}

\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes application status checks, including configuration details such as protocol, port, path, thresholds, and associations. Results are paginated. Use the \code{NextToken} parameter to retrieve additional results. The following rules apply:

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_application_status_checks/} for full documentation.
}
\keyword{internal}
