% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_capacity_blocks}
\alias{ec2_describe_capacity_blocks}
\title{Describes details about Capacity Blocks in the Amazon Web Services
Region that you're currently using}
\usage{
ec2_describe_capacity_blocks(
  CapacityBlockIds = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Filters = NULL,
  DryRun = NULL
)
}
\arguments{
\item{CapacityBlockIds}{The IDs of the Capacity Blocks.}

\item{NextToken}{The token to use to retrieve the next page of results.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}

\item{Filters}{One or more filters.
\itemize{
\item \code{capacity-block-id} - The ID of the Capacity Block.
\item \code{ultraserver-type} - The Capacity Block type. The type can be
\code{instances} or \code{ultraservers}.
\item \code{availability-zone} - The Availability Zone of the Capacity Block.
\item \code{start-date} - The date and time at which the Capacity Block was
started.
\item \code{end-date} - The date and time at which the Capacity Block expires.
When a Capacity Block expires, all instances in the Capacity Block
are terminated.
\item \code{create-date} - The date and time at which the Capacity Block was
created.
\item \code{state} - The state of the Capacity Block (\code{active} | \code{expired} |
\code{unavailable} | \code{cancelled} | \code{failed} | \code{scheduled} |
\code{payment-pending} | \code{payment-failed}).
\item \code{tags} - The tags assigned to the Capacity Block.
}}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes details about Capacity Blocks in the Amazon Web Services Region that you're currently using.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_capacity_blocks/} for full documentation.
}
\keyword{internal}
