% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_capacity_reservation_cancellation_quotes}
\alias{ec2_describe_capacity_reservation_cancellation_quotes}
\title{Describes one or more Capacity Reservation cancellation quotes}
\usage{
ec2_describe_capacity_reservation_cancellation_quotes(
  CapacityReservationCancellationQuoteIds = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL,
  Filters = NULL
)
}
\arguments{
\item{CapacityReservationCancellationQuoteIds}{The IDs of the cancellation quotes to describe.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token to use to retrieve the next page of results.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{One or more filters. Filter names and values are case-sensitive.}
}
\description{
Describes one or more Capacity Reservation cancellation quotes. The results describe only the quotes that you have previously generated by using the \code{\link[=ec2_create_capacity_reservation_cancellation_quote]{create_capacity_reservation_cancellation_quote}} action.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_capacity_reservation_cancellation_quotes/} for full documentation.
}
\keyword{internal}
