% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_capacity_reservation_date_change_quotes}
\alias{ec2_describe_capacity_reservation_date_change_quotes}
\title{Describes one or more Capacity Reservation date change quotes that you
generated by using the CreateCapacityReservationDateChangeQuote
operation}
\usage{
ec2_describe_capacity_reservation_date_change_quotes(
  CapacityReservationModificationQuoteIds = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL,
  Filters = NULL
)
}
\arguments{
\item{CapacityReservationModificationQuoteIds}{The IDs of the date change quotes to describe.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token to use to retrieve the next page of results.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{One or more filters. Filter names and values are case-sensitive.}
}
\description{
Describes one or more Capacity Reservation date change quotes that you generated by using the \code{\link[=ec2_create_capacity_reservation_date_change_quote]{create_capacity_reservation_date_change_quote}} operation.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_capacity_reservation_date_change_quotes/} for full documentation.
}
\keyword{internal}
