% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_export_tasks}
\alias{ec2_describe_export_tasks}
\title{Describes the specified export instance tasks or all of your export
instance tasks}
\usage{
ec2_describe_export_tasks(Filters = NULL, ExportTaskIds = NULL)
}
\arguments{
\item{Filters}{the filters for the export tasks.}

\item{ExportTaskIds}{The export task IDs.}
}
\description{
Describes the specified export instance tasks or all of your export instance tasks.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_export_tasks/} for full documentation.
}
\keyword{internal}
