% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_iam_instance_profile_associations}
\alias{ec2_describe_iam_instance_profile_associations}
\title{Describes your IAM instance profile associations}
\usage{
ec2_describe_iam_instance_profile_associations(AssociationIds, Filters,
  MaxResults, NextToken)
}
\arguments{
\item{AssociationIds}{The IAM instance profile associations.}

\item{Filters}{The filters.
\itemize{
\item \code{instance-id} - The ID of the instance.
\item \code{state} - The state of the association (\code{associating} | \code{associated}
| \code{disassociating}).
}}

\item{MaxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value.}

\item{NextToken}{The token to request the next page of results.}
}
\description{
Describes your IAM instance profile associations.
}
\section{Request syntax}{
\preformatted{svc$describe_iam_instance_profile_associations(
  AssociationIds = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\examples{
\dontrun{
# This example describes the specified IAM instance profile association.
svc$describe_iam_instance_profile_associations(
  AssociationIds = list(
    "iip-assoc-0db249b1f25fa24b8"
  )
)
}

}
\keyword{internal}
