% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_image_references}
\alias{ec2_describe_image_references}
\title{Describes your Amazon Web Services resources that are referencing the
specified images}
\usage{
ec2_describe_image_references(
  ImageIds,
  IncludeAllResourceTypes = NULL,
  ResourceTypes = NULL,
  NextToken = NULL,
  DryRun = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ImageIds}{[required] The IDs of the images to check for resource references.}

\item{IncludeAllResourceTypes}{Specifies whether to check all supported Amazon Web Services resource
types for image references. When specified, default values are applied
for \code{ResourceTypeOptions}. For the default values, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-ami-references.html#how-ami-references-works}{How AMI reference checks work}
in the \emph{Amazon EC2 User Guide}. If you also specify \code{ResourceTypes} with
\code{ResourceTypeOptions}, your specified values override the default
values.

Supported resource types: \code{ec2:Instance} | \code{ec2:LaunchTemplate} |
\code{ssm:Parameter} | \code{imagebuilder:ImageRecipe} |
\code{imagebuilder:ContainerRecipe}

Either \code{IncludeAllResourceTypes} or \code{ResourceTypes} must be specified.}

\item{ResourceTypes}{The Amazon Web Services resource types to check for image references.

Either \code{IncludeAllResourceTypes} or \code{ResourceTypes} must be specified.}

\item{NextToken}{The token returned from a previous paginated request. Pagination
continues from the end of the items returned by the previous request.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}
}
\description{
Describes your Amazon Web Services resources that are referencing the specified images.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_image_references/} for full documentation.
}
\keyword{internal}
