% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_image_usage_report_entries}
\alias{ec2_describe_image_usage_report_entries}
\title{Describes the entries in image usage reports, showing how your images
are used across other Amazon Web Services accounts}
\usage{
ec2_describe_image_usage_report_entries(
  ImageIds = NULL,
  ReportIds = NULL,
  NextToken = NULL,
  Filters = NULL,
  DryRun = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ImageIds}{The IDs of the images for filtering the report entries. If specified,
only report entries containing these images are returned.}

\item{ReportIds}{The IDs of the usage reports.}

\item{NextToken}{The token returned from a previous paginated request. Pagination
continues from the end of the items returned by the previous request.}

\item{Filters}{The filters.
\itemize{
\item \code{account-id} - A 12-digit Amazon Web Services account ID.
\item \code{creation-time} - The time when the report was created, in the ISO
8601 format in the UTC time zone (YYYY-MM-DDThh:mm:ss.sssZ), for
example, \verb{2025-11-29T11:04:43.305Z}. You can use a wildcard (\code{*}),
for example, \verb{2025-11-29T*}, which matches an entire day.
\item \code{resource-type} - The resource type (\code{ec2:Instance} |
\code{ec2:LaunchTemplate}).
}}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}
}
\description{
Describes the entries in image usage reports, showing how your images are used across other Amazon Web Services accounts.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_image_usage_report_entries/} for full documentation.
}
\keyword{internal}
