% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_image_usage_reports}
\alias{ec2_describe_image_usage_reports}
\title{Describes the configuration and status of image usage reports, filtered
by report IDs or image IDs}
\usage{
ec2_describe_image_usage_reports(
  ImageIds = NULL,
  ReportIds = NULL,
  NextToken = NULL,
  Filters = NULL,
  DryRun = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ImageIds}{The IDs of the images for filtering the reports. If specified, only
reports containing these images are returned.}

\item{ReportIds}{The IDs of the image usage reports.}

\item{NextToken}{The token returned from a previous paginated request. Pagination
continues from the end of the items returned by the previous request.}

\item{Filters}{The filters.
\itemize{
\item \code{creation-time} - The time when the report was created, in the ISO
8601 format in the UTC time zone (YYYY-MM-DDThh:mm:ss.sssZ), for
example, \verb{2025-11-29T11:04:43.305Z}. You can use a wildcard (\code{*}),
for example, \verb{2025-11-29T*}, which matches an entire day.
\item \code{state} - The state of the report (\code{available} | \code{pending} |
\code{error}).
\item \verb{tag:<key>} - The key/value combination of a tag assigned to the
resource. Use the tag key in the filter name and the tag value as
the filter value. For example, to find all resources that have a tag
with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for
the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this
filter to find all resources assigned a tag with a specific key,
regardless of the tag value.
}}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}
}
\description{
Describes the configuration and status of image usage reports, filtered by report IDs or image IDs.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_image_usage_reports/} for full documentation.
}
\keyword{internal}
