% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_instance_event_notification_attributes}
\alias{ec2_describe_instance_event_notification_attributes}
\title{Describes the tag keys that are registered to appear in scheduled event
notifications for resources in the current Region}
\usage{
ec2_describe_instance_event_notification_attributes(DryRun)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  InstanceTagAttribute = list(
    InstanceTagKeys = list(
      "string"
    ),
    IncludeAllTagsOfInstance = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the tag keys that are registered to appear in scheduled event
notifications for resources in the current Region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_instance_event_notification_attributes(
  DryRun = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
