% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_instance_sql_ha_history_states}
\alias{ec2_describe_instance_sql_ha_history_states}
\title{Describes the historical SQL Server High Availability states for Amazon
EC2 instances that are enabled for Amazon EC2 High Availability for SQL
Server monitoring}
\usage{
ec2_describe_instance_sql_ha_history_states(
  InstanceIds = NULL,
  StartTime = NULL,
  EndTime = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Filters = NULL,
  DryRun = NULL
)
}
\arguments{
\item{InstanceIds}{The IDs of the SQL Server High Availability instances to describe. If
omitted, the API returns historical states for all SQL Server High
Availability instances.}

\item{StartTime}{The start data and time of the period for which to get the historical
SQL Server High Availability states. If omitted, the API returns all
available historical states.

Timezone: UTC

Format: \code{YYYY-MM-DDThh:mm:ss.sssZ}}

\item{EndTime}{The end data and time of the period for which to get historical SQL
Server High Availability states. If omitted, the API returns historical
states up to the current date and time.

Timezone: UTC

Format: \code{YYYY-MM-DDThh:mm:ss.sssZ}}

\item{NextToken}{The token to use to retrieve the next page of results.}

\item{MaxResults}{The maximum number of results to return for the request in a single
page. The remaining results can be seen by sending another request with
the returned \code{nextToken} value.}

\item{Filters}{One or more filters to apply to the results. Supported filters include:
\itemize{
\item \verb{tag:<key>} - The tag key and value pair assigned to the instance.
For example, to find all instances tagged with \code{Owner:TeamA},
specify \code{tag:Owner} for the filter name and \code{TeamA} for the filter
value.
\item \code{tag-key} - The tag key assigned to the instance.
\item \code{haStatus} - The SQL Server High Availability status of the SQL
Server High Availability instance (\code{processing} | \code{active} |
\code{standby} | \code{invalid}).
\item \code{sqlServerLicenseUsage} - The license type for the SQL Server
license (\code{full} | \code{waived}).
}}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the historical SQL Server High Availability states for Amazon EC2 instances that are enabled for Amazon EC2 High Availability for SQL Server monitoring.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_instance_sql_ha_history_states/} for full documentation.
}
\keyword{internal}
