% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_ipam_pool_allocations}
\alias{ec2_describe_ipam_pool_allocations}
\title{Describes IPAM pool allocations}
\usage{
ec2_describe_ipam_pool_allocations(
  DryRun = NULL,
  IpamPoolAllocationIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action without actually making the request and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPoolAllocationIds}{The IDs of the IPAM pool allocations you want to describe.}

\item{Filters}{One or more filters for the request. For more information about filtering, see \href{https://docs.aws.amazon.com/cli/latest/userguide/cli-usage-filter.html}{Filtering CLI output}.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Describes IPAM pool allocations. You can describe all allocations owned by you across all pools, or you can describe specific allocations by ID.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_ipam_pool_allocations/} for full documentation.
}
\keyword{internal}
