% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_ipam_pools}
\alias{ec2_describe_ipam_pools}
\title{Get information about your IPAM pools}
\usage{
ec2_describe_ipam_pools(
  DryRun = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  IpamPoolIds = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{One or more filters for the request. For more information about
filtering, see \href{https://docs.aws.amazon.com/cli/latest/userguide/cli-usage-filter.html}{Filtering CLI output}.}

\item{MaxResults}{The maximum number of results to return in the request.}

\item{NextToken}{The token for the next page of results.}

\item{IpamPoolIds}{The IDs of the IPAM pools you would like information on.}
}
\description{
Get information about your IPAM pools.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_ipam_pools/} for full documentation.
}
\keyword{internal}
