% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_ipam_resource_discoveries}
\alias{ec2_describe_ipam_resource_discoveries}
\title{Describes IPAM resource discoveries}
\usage{
ec2_describe_ipam_resource_discoveries(
  DryRun = NULL,
  IpamResourceDiscoveryIds = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Filters = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamResourceDiscoveryIds}{The IPAM resource discovery IDs.}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of resource discoveries to return in one page of
results.}

\item{Filters}{The resource discovery filters.}
}
\description{
Describes IPAM resource discoveries. A resource discovery is an IPAM component that enables IPAM to manage and monitor resources that belong to the owning account.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_ipam_resource_discoveries/} for full documentation.
}
\keyword{internal}
