% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_ipam_resource_discovery_associations}
\alias{ec2_describe_ipam_resource_discovery_associations}
\title{Describes resource discovery association with an Amazon VPC IPAM}
\usage{
ec2_describe_ipam_resource_discovery_associations(
  DryRun = NULL,
  IpamResourceDiscoveryAssociationIds = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Filters = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamResourceDiscoveryAssociationIds}{The resource discovery association IDs.}

\item{NextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{MaxResults}{The maximum number of resource discovery associations to return in one
page of results.}

\item{Filters}{The resource discovery association filters.}
}
\description{
Describes resource discovery association with an Amazon VPC IPAM. An associated resource discovery is a resource discovery that has been associated with an IPAM..

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_ipam_resource_discovery_associations/} for full documentation.
}
\keyword{internal}
