% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_local_gateway_virtual_interfaces}
\alias{ec2_describe_local_gateway_virtual_interfaces}
\title{Describes the specified local gateway virtual interfaces}
\usage{
ec2_describe_local_gateway_virtual_interfaces(
  LocalGatewayVirtualInterfaceIds, Filters, MaxResults, NextToken, DryRun)
}
\arguments{
\item{LocalGatewayVirtualInterfaceIds}{The IDs of the virtual interfaces.}

\item{Filters}{One or more filters.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:\preformatted{list(
  LocalGatewayVirtualInterfaces = list(
    list(
      LocalGatewayVirtualInterfaceId = "string",
      LocalGatewayId = "string",
      Vlan = 123,
      LocalAddress = "string",
      PeerAddress = "string",
      LocalBgpAsn = 123,
      PeerBgpAsn = 123,
      OwnerId = "string",
      Tags = list(
        list(
          Key = "string",
          Value = "string"
        )
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Describes the specified local gateway virtual interfaces.
}
\section{Request syntax}{
\preformatted{svc$describe_local_gateway_virtual_interfaces(
  LocalGatewayVirtualInterfaceIds = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string",
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
