% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_local_gateways}
\alias{ec2_describe_local_gateways}
\title{Describes one or more local gateways}
\usage{
ec2_describe_local_gateways(LocalGatewayIds, Filters, MaxResults,
  NextToken, DryRun)
}
\arguments{
\item{LocalGatewayIds}{One or more filters.
\itemize{
\item \code{local-gateway-id} - The ID of a local gateway.
\item \code{local-gateway-route-table-id} - The ID of the local gateway route
table.
\item \code{local-gateway-route-table-virtual-interface-group-association-id} -
The ID of the association.
\item \code{local-gateway-route-table-virtual-interface-group-id} - The ID of
the virtual interface group.
\item \code{outpost-arn} - The Amazon Resource Name (ARN) of the Outpost.
\item \code{state} - The state of the association.
}}

\item{Filters}{One or more filters.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  LocalGateways = list(
    list(
      LocalGatewayId = "string",
      OutpostArn = "string",
      OwnerId = "string",
      State = "string",
      Tags = list(
        list(
          Key = "string",
          Value = "string"
        )
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes one or more local gateways. By default, all local gateways are
described. Alternatively, you can filter the results.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_local_gateways(
  LocalGatewayIds = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string",
  DryRun = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
