% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_mac_hosts}
\alias{ec2_describe_mac_hosts}
\title{Describes the specified EC2 Mac Dedicated Host or all of your EC2 Mac
Dedicated Hosts}
\usage{
ec2_describe_mac_hosts(
  Filters = NULL,
  HostIds = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{The filters.
\itemize{
\item \code{availability-zone} - The Availability Zone of the EC2 Mac Dedicated
Host.
\item \code{instance-type} - The instance type size that the EC2 Mac Dedicated
Host is configured to support.
}}

\item{HostIds}{The IDs of the EC2 Mac Dedicated Hosts.}

\item{MaxResults}{The maximum number of results to return for the request in a single
page. The remaining results can be seen by sending another request with
the returned \code{nextToken} value. This value can be between 5 and 500. If
\code{maxResults} is given a larger value than 500, you receive an error.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\description{
Describes the specified EC2 Mac Dedicated Host or all of your EC2 Mac Dedicated Hosts.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_mac_hosts/} for full documentation.
}
\keyword{internal}
