% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_mac_modification_tasks}
\alias{ec2_describe_mac_modification_tasks}
\title{Describes a System Integrity Protection (SIP) modification task or
volume ownership delegation task for an Amazon EC2 Mac instance}
\usage{
ec2_describe_mac_modification_tasks(
  DryRun = NULL,
  Filters = NULL,
  MacModificationTaskIds = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{Specifies one or more filters for the request:
\itemize{
\item \code{instance-id} - The ID of the instance for which the task was
created.
\item \code{task-state} - The state of the task (\code{successful} | \code{failed} |
\verb{in-progress} | \code{pending}).
\item \code{mac-system-integrity-protection-configuration.sip-status} - The
overall SIP state requested in the task (\code{enabled} | \code{disabled}).
\item \code{start-time} - The date and time the task was created.
\item \code{task-type} - The type of task (\code{sip-modification} |
\code{volume-ownership-delegation}).
}}

\item{MacModificationTaskIds}{The ID of task.}

\item{MaxResults}{The maximum number of results to return for the request in a single
page. The remaining results can be seen by sending another request with
the returned \code{nextToken} value. This value can be between 5 and 500. If
\code{maxResults} is given a larger value than 500, you receive an error.}

\item{NextToken}{The token to use to retrieve the next page of results.}
}
\description{
Describes a System Integrity Protection (SIP) modification task or volume ownership delegation task for an Amazon EC2 Mac instance. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/mac-sip-settings.html#mac-sip-configure}{Configure SIP for Amazon EC2 instances} in the \emph{Amazon EC2 User Guide}.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_mac_modification_tasks/} for full documentation.
}
\keyword{internal}
