% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_network_insights_access_scope_analyses}
\alias{ec2_describe_network_insights_access_scope_analyses}
\title{Describes the specified Network Access Scope analyses}
\usage{
ec2_describe_network_insights_access_scope_analyses(
  NetworkInsightsAccessScopeAnalysisIds = NULL,
  NetworkInsightsAccessScopeId = NULL,
  AnalysisStartTimeBegin = NULL,
  AnalysisStartTimeEnd = NULL,
  Filters = NULL,
  MaxResults = NULL,
  DryRun = NULL,
  NextToken = NULL
)
}
\arguments{
\item{NetworkInsightsAccessScopeAnalysisIds}{The IDs of the Network Access Scope analyses.}

\item{NetworkInsightsAccessScopeId}{The ID of the Network Access Scope.}

\item{AnalysisStartTimeBegin}{Filters the results based on the start time. The analysis must have
started on or after this time.}

\item{AnalysisStartTimeEnd}{Filters the results based on the start time. The analysis must have
started on or before this time.}

\item{Filters}{There are no supported filters.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Describes the specified Network Access Scope analyses.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_network_insights_access_scope_analyses/} for full documentation.
}
\keyword{internal}
