% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_network_insights_access_scopes}
\alias{ec2_describe_network_insights_access_scopes}
\title{Describes the specified Network Access Scopes}
\usage{
ec2_describe_network_insights_access_scopes(
  NetworkInsightsAccessScopeIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  DryRun = NULL,
  NextToken = NULL
)
}
\arguments{
\item{NetworkInsightsAccessScopeIds}{The IDs of the Network Access Scopes.}

\item{Filters}{There are no supported filters.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Describes the specified Network Access Scopes.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_network_insights_access_scopes/} for full documentation.
}
\keyword{internal}
