% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_route_server_peers}
\alias{ec2_describe_route_server_peers}
\title{Describes one or more route server peers}
\usage{
ec2_describe_route_server_peers(
  RouteServerPeerIds = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Filters = NULL,
  DryRun = NULL
)
}
\arguments{
\item{RouteServerPeerIds}{The IDs of the route server peers to describe.}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return with a single call.}

\item{Filters}{One or more filters to apply to the describe request.}

\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes one or more route server peers.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_route_server_peers/} for full documentation.
}
\keyword{internal}
