% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_secondary_interfaces}
\alias{ec2_describe_secondary_interfaces}
\title{Describes one or more of your secondary interfaces}
\usage{
ec2_describe_secondary_interfaces(
  DryRun = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  SecondaryInterfaceIds = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{The filters.
\itemize{
\item \code{attachment.attachment-id} - The ID of the secondary interface
attachment.
\item \code{attachment.instance-id} - The ID of the instance to which the
secondary interface is attached.
\item \code{attachment.instance-owner-id} - The ID of the Amazon Web Services
account that owns the instance to which the secondary interface is
attached.
\item \code{attachment.status} - The attachment status (\code{attaching} |
\code{attached} | \code{detaching} | \code{detached}).
\item \code{private-ipv4-addresses.private-ip-address} - The private IPv4
address associated with the secondary interface.
\item \code{owner-id} - The ID of the Amazon Web Services account that owns the
secondary interface.
\item \code{secondary-interface-arn} - The ARN of the secondary interface.
\item \code{secondary-interface-id} - The ID of the secondary interface.
\item \code{secondary-interface-type} - The type of secondary interface
(\code{secondary}).
\item \code{secondary-network-id} - The ID of the secondary network.
\item \code{secondary-network-type} - The type of the secondary network
(\code{rdma}).
\item \code{secondary-subnet-id} - The ID of the secondary subnet.
\item \code{status} - The status of the secondary interface (\code{available} |
\verb{in-use}).
\item \code{tag}:\<key\> - The key/value combination of a tag assigned to the
resource. Use the tag key in the filter name and the tag value as
the filter value. For example, to find all resources that have a tag
with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for
the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this
filter to find all resources assigned a tag with a specific key,
regardless of the tag value.
}}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{SecondaryInterfaceIds}{The IDs of the secondary interfaces.}
}
\description{
Describes one or more of your secondary interfaces.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_secondary_interfaces/} for full documentation.
}
\keyword{internal}
