% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_secondary_subnets}
\alias{ec2_describe_secondary_subnets}
\title{Describes one or more of your secondary subnets}
\usage{
ec2_describe_secondary_subnets(
  DryRun = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  SecondarySubnetIds = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{The filters.
\itemize{
\item \code{ipv4-cidr-block-association.association-id} - The association ID
for an IPv4 CIDR block associated with the secondary subnet.
\item \code{ipv4-cidr-block-association.cidr-block} - An IPv4 CIDR block
associated with the secondary subnet.
\item \code{ipv4-cidr-block-association.state} - The state of an IPv4 CIDR
block associated with the secondary subnet.
\item \code{owner-id} - The ID of the Amazon Web Services account that owns the
secondary subnet.
\item \code{secondary-network-id} - The ID of the secondary network.
\item \code{secondary-network-type} - The type of the secondary network
(\code{rdma}).
\item \code{secondary-subnet-id} - The ID of the secondary subnet.
\item \code{secondary-subnet-arn} - The ARN of the secondary subnet.
\item \code{state} - The state of the secondary subnet (\verb{create-in-progress} |
\code{create-complete} | \code{create-failed} | \verb{delete-in-progress} |
\code{delete-complete} | \code{delete-failed}).
\item \code{tag}:\<key\> - The key/value combination of a tag assigned to the
resource. Use the tag key in the filter name and the tag value as
the filter value. For example, to find all resources that have a tag
with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for
the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this
filter to find all resources assigned a tag with a specific key,
regardless of the tag value.
}}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{SecondarySubnetIds}{The IDs of the secondary subnets.}
}
\description{
Describes one or more of your secondary subnets.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_secondary_subnets/} for full documentation.
}
\keyword{internal}
