% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_spot_fleet_requests}
\alias{ec2_describe_spot_fleet_requests}
\title{Describes your Spot Fleet requests}
\usage{
ec2_describe_spot_fleet_requests(
  DryRun = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  SpotFleetRequestIds = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token to include in another request to get the next page of items.
This value is \code{null} when there are no more items to return.}

\item{SpotFleetRequestIds}{The IDs of the Spot Fleet requests.}
}
\description{
Describes your Spot Fleet requests.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_spot_fleet_requests/} for full documentation.
}
\keyword{internal}
