% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_store_image_tasks}
\alias{ec2_describe_store_image_tasks}
\title{Describes the progress of the AMI store tasks}
\usage{
ec2_describe_store_image_tasks(ImageIds, DryRun, Filters, NextToken,
  MaxResults)
}
\arguments{
\item{ImageIds}{The AMI IDs for which to show progress. Up to 20 AMI IDs can be included
in a request.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{The filters.
\itemize{
\item \code{task-state} - Returns tasks in a certain state (\code{InProgress} |
\code{Completed} | \code{Failed})
\item \code{bucket} - Returns task information for tasks that targeted a
specific bucket. For the filter value, specify the bucket name.
}}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value. This value can be between 1 and 200. You cannot specify this
parameter and the \code{ImageIDs} parameter in the same call.}
}
\value{
A list with the following syntax:\preformatted{list(
  StoreImageTaskResults = list(
    list(
      AmiId = "string",
      TaskStartTime = as.POSIXct(
        "2015-01-01"
      ),
      Bucket = "string",
      S3objectKey = "string",
      ProgressPercentage = 123,
      StoreTaskState = "string",
      StoreTaskFailureReason = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Describes the progress of the AMI store tasks. You can describe the
store tasks for specified AMIs. If you don't specify the AMIs, you get a
paginated list of store tasks from the last 31 days.

For each AMI task, the response indicates if the task is \code{InProgress},
\code{Completed}, or \code{Failed}. For tasks \code{InProgress}, the response shows the
estimated progress as a percentage.

Tasks are listed in reverse chronological order. Currently, only tasks
from the past 31 days can be viewed.

To use this API, you must have the required permissions. For more
information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ami-store-restore.html#ami-s3-permissions}{Permissions for storing and restoring AMIs using S3}
in the \emph{Amazon Elastic Compute Cloud User Guide}.

For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ami-store-restore.html}{Store and restore an AMI using S3}
in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\section{Request syntax}{
\preformatted{svc$describe_store_image_tasks(
  ImageIds = list(
    "string"
  ),
  DryRun = TRUE|FALSE,
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
