% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_transit_gateway_multicast_domains}
\alias{ec2_describe_transit_gateway_multicast_domains}
\title{Describes one or more transit gateway multicast domains}
\usage{
ec2_describe_transit_gateway_multicast_domains(
  TransitGatewayMulticastDomainIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayMulticastDomainIds}{The ID of the transit gateway multicast domain.}

\item{Filters}{One or more filters. The possible values are:
\itemize{
\item \code{state} - The state of the transit gateway multicast domain. Valid
values are \code{pending} | \code{available} | \code{deleting} | \code{deleted}.
\item \code{transit-gateway-id} - The ID of the transit gateway.
\item \code{transit-gateway-multicast-domain-id} - The ID of the transit
gateway multicast domain.
}}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes one or more transit gateway multicast domains.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_transit_gateway_multicast_domains/} for full documentation.
}
\keyword{internal}
