% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_trunk_interface_associations}
\alias{ec2_describe_trunk_interface_associations}
\title{Describes one or more network interface trunk associations}
\usage{
ec2_describe_trunk_interface_associations(
  AssociationIds = NULL,
  DryRun = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AssociationIds}{The IDs of the associations.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{One or more filters.
\itemize{
\item \code{gre-key} - The ID of a trunk interface association.
\item \code{interface-protocol} - The interface protocol. Valid values are
\code{VLAN} and \code{GRE}.
}}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}
}
\description{
Describes one or more network interface trunk associations.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_trunk_interface_associations/} for full documentation.
}
\keyword{internal}
