% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_verified_access_instances}
\alias{ec2_describe_verified_access_instances}
\title{Describes the specified Amazon Web Services Verified Access instances}
\usage{
ec2_describe_verified_access_instances(
  VerifiedAccessInstanceIds = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Filters = NULL,
  DryRun = NULL
)
}
\arguments{
\item{VerifiedAccessInstanceIds}{The IDs of the Verified Access instances.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{Filters}{One or more filters. Filter names and values are case-sensitive.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the specified Amazon Web Services Verified Access instances.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_verified_access_instances/} for full documentation.
}
\keyword{internal}
