% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_volumes_modifications}
\alias{ec2_describe_volumes_modifications}
\title{Describes the most recent volume modification request for the specified
EBS volumes}
\usage{
ec2_describe_volumes_modifications(
  DryRun = NULL,
  VolumeIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  IncludeManagedResources = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{VolumeIds}{The IDs of the volumes.}

\item{Filters}{The filters.
\itemize{
\item \code{modification-state} - The current modification state (modifying | optimizing | completed | failed).
\item \code{original-iops} - The original IOPS rate of the volume.
\item \code{original-size} - The original size of the volume, in GiB.
\item \code{original-volume-type} - The original volume type of the volume (standard | io1 | io2 | gp2 | sc1 | st1).
\item \code{originalMultiAttachEnabled} - Indicates whether Multi-Attach support was enabled (true | false).
\item \code{start-time} - The modification start time.
\item \code{target-iops} - The target IOPS rate of the volume.
\item \code{target-size} - The target size of the volume, in GiB.
\item \code{target-volume-type} - The target volume type of the volume (standard | io1 | io2 | gp2 | sc1 | st1).
\item \code{targetMultiAttachEnabled} - Indicates whether Multi-Attach support is to be enabled (true | false).
\item \code{volume-id} - The ID of the volume.
}}

\item{NextToken}{The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.}

\item{MaxResults}{The maximum number of results (up to a limit of 500) to be returned in a paginated request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/Query-Requests.html#api-pagination}{Pagination}.}

\item{IncludeManagedResources}{Indicates whether to include managed resources in the output. If this parameter is set to \code{true}, the output includes resources that are managed by Amazon Web Services services, even if managed resource visibility is set to hidden.}
}
\description{
Describes the most recent volume modification request for the specified EBS volumes.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_volumes_modifications/} for full documentation.
}
\keyword{internal}
