% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_vpc_block_public_access_options}
\alias{ec2_describe_vpc_block_public_access_options}
\title{Describe VPC Block Public Access (BPA) options}
\usage{
ec2_describe_vpc_block_public_access_options(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describe VPC Block Public Access (BPA) options. VPC Block Public Access (BPA) enables you to block resources in VPCs and subnets that you own in a Region from reaching or being reached from the internet through internet gateways and egress-only internet gateways. To learn more about VPC BPA, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/security-vpc-bpa.html}{Block public access to VPCs and subnets} in the \emph{Amazon VPC User Guide}.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_vpc_block_public_access_options/} for full documentation.
}
\keyword{internal}
