% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_vpc_endpoint_service_configurations}
\alias{ec2_describe_vpc_endpoint_service_configurations}
\title{Describes the VPC endpoint service configurations in your account (your
services)}
\usage{
ec2_describe_vpc_endpoint_service_configurations(
  DryRun = NULL,
  ServiceIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ServiceIds}{The IDs of the endpoint services.}

\item{Filters}{The filters.
\itemize{
\item \code{service-name} - The name of the service.
\item \code{service-id} - The ID of the service.
\item \code{service-state} - The state of the service (\code{Pending} | \code{Available}
| \code{Deleting} | \code{Deleted} | \code{Failed}).
\item \code{supported-ip-address-types} - The IP address type (\code{ipv4} |
\code{ipv6}).
\item \code{tag}:\<key\> - The key/value combination of a tag assigned to the
resource. Use the tag key in the filter name and the tag value as
the filter value. For example, to find all resources that have a tag
with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for
the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this
filter to find all resources assigned a tag with a specific key,
regardless of the tag value.
}}

\item{MaxResults}{The maximum number of results to return for the request in a single
page. The remaining results of the initial request can be seen by
sending another request with the returned \code{NextToken} value. This value
can be between 5 and 1,000; if \code{MaxResults} is given a value larger than
1,000, only 1,000 results are returned.}

\item{NextToken}{The token to retrieve the next page of results.}
}
\description{
Describes the VPC endpoint service configurations in your account (your services).

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_vpc_endpoint_service_configurations/} for full documentation.
}
\keyword{internal}
