% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_vpn_concentrators}
\alias{ec2_describe_vpn_concentrators}
\title{Describes one or more of your VPN concentrators}
\usage{
ec2_describe_vpn_concentrators(
  VpnConcentratorIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{VpnConcentratorIds}{One or more VPN concentrator IDs.}

\item{Filters}{One or more filters to limit the results.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes one or more of your VPN concentrators.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_vpn_concentrators/} for full documentation.
}
\keyword{internal}
