% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_detach_image_watermark}
\alias{ec2_detach_image_watermark}
\title{Removes a watermark from the specified AMI}
\usage{
ec2_detach_image_watermark(ImageId, WatermarkKey, DryRun = NULL)
}
\arguments{
\item{ImageId}{[required] The ID of the AMI.}

\item{WatermarkKey}{[required] The watermark key to remove, in \code{accountId:watermarkName} format (for example, \code{123456789012:approvedAmi}).}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Removes a watermark from the specified AMI. This is an idempotent operation. It succeeds even if the watermark does not exist on the image.

See \url{https://www.paws-r-sdk.com/docs/ec2_detach_image_watermark/} for full documentation.
}
\keyword{internal}
