% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disable_allowed_images_settings}
\alias{ec2_disable_allowed_images_settings}
\title{Disables Allowed AMIs for your account in the specified Amazon Web
Services Region}
\usage{
ec2_disable_allowed_images_settings(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Disables Allowed AMIs for your account in the specified Amazon Web Services Region. When set to \code{disabled}, the image criteria in your Allowed AMIs settings do not apply, and no restrictions are placed on AMI discoverability or usage. Users in your account can launch instances using any public AMI or AMI shared with your account.

See \url{https://www.paws-r-sdk.com/docs/ec2_disable_allowed_images_settings/} for full documentation.
}
\keyword{internal}
