% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disable_capacity_manager}
\alias{ec2_disable_capacity_manager}
\title{Disables EC2 Capacity Manager for your account}
\usage{
ec2_disable_capacity_manager(DryRun = NULL, ClientToken = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Disables EC2 Capacity Manager for your account. This stops data ingestion and removes access to capacity analytics and optimization recommendations. Previously collected data is retained but no new data will be processed.

See \url{https://www.paws-r-sdk.com/docs/ec2_disable_capacity_manager/} for full documentation.
}
\keyword{internal}
