% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disable_fast_snapshot_restores}
\alias{ec2_disable_fast_snapshot_restores}
\title{Disables fast snapshot restores for the specified snapshots in the
specified Availability Zones}
\usage{
ec2_disable_fast_snapshot_restores(AvailabilityZones, SourceSnapshotIds,
  DryRun)
}
\arguments{
\item{AvailabilityZones}{[required] One or more Availability Zones. For example, \verb{us-east-2a}.}

\item{SourceSnapshotIds}{[required] The IDs of one or more snapshots. For example, \verb{snap-1234567890abcdef0}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Successful = list(
    list(
      SnapshotId = "string",
      AvailabilityZone = "string",
      State = "enabling"|"optimizing"|"enabled"|"disabling"|"disabled",
      StateTransitionReason = "string",
      OwnerId = "string",
      OwnerAlias = "string",
      EnablingTime = as.POSIXct(
        "2015-01-01"
      ),
      OptimizingTime = as.POSIXct(
        "2015-01-01"
      ),
      EnabledTime = as.POSIXct(
        "2015-01-01"
      ),
      DisablingTime = as.POSIXct(
        "2015-01-01"
      ),
      DisabledTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  Unsuccessful = list(
    list(
      SnapshotId = "string",
      FastSnapshotRestoreStateErrors = list(
        list(
          AvailabilityZone = "string",
          Error = list(
            Code = "string",
            Message = "string"
          )
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Disables fast snapshot restores for the specified snapshots in the
specified Availability Zones.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disable_fast_snapshot_restores(
  AvailabilityZones = list(
    "string"
  ),
  SourceSnapshotIds = list(
    "string"
  ),
  DryRun = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
