% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disable_image}
\alias{ec2_disable_image}
\title{Sets the AMI state to disabled and removes all launch permissions from
the AMI}
\usage{
ec2_disable_image(ImageId, DryRun = NULL)
}
\arguments{
\item{ImageId}{[required] The ID of the AMI.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Sets the AMI state to \code{disabled} and removes all launch permissions from the AMI. A disabled AMI can't be used for instance launches.

See \url{https://www.paws-r-sdk.com/docs/ec2_disable_image/} for full documentation.
}
\keyword{internal}
