% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disable_image_deregistration_protection}
\alias{ec2_disable_image_deregistration_protection}
\title{Disables deregistration protection for an AMI}
\usage{
ec2_disable_image_deregistration_protection(ImageId, DryRun = NULL)
}
\arguments{
\item{ImageId}{[required] The ID of the AMI.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Disables deregistration protection for an AMI. When deregistration protection is disabled, the AMI can be deregistered.

See \url{https://www.paws-r-sdk.com/docs/ec2_disable_image_deregistration_protection/} for full documentation.
}
\keyword{internal}
