% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disable_ipam_organization_admin_account}
\alias{ec2_disable_ipam_organization_admin_account}
\title{Disable the IPAM account}
\usage{
ec2_disable_ipam_organization_admin_account(
  DryRun = NULL,
  DelegatedAdminAccountId
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{DelegatedAdminAccountId}{[required] The Organizations member account ID that you want to disable as IPAM
account.}
}
\description{
Disable the IPAM account. For more information, see \href{https://docs.aws.amazon.com/vpc/latest/ipam/enable-integ-ipam.html}{Enable integration with Organizations} in the \emph{Amazon VPC IPAM User Guide}.

See \url{https://www.paws-r-sdk.com/docs/ec2_disable_ipam_organization_admin_account/} for full documentation.
}
\keyword{internal}
