% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disable_ipam_policy}
\alias{ec2_disable_ipam_policy}
\title{Disables an IPAM policy}
\usage{
ec2_disable_ipam_policy(
  DryRun = NULL,
  IpamPolicyId,
  OrganizationTargetId = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPolicyId}{[required] The ID of the IPAM policy to disable.}

\item{OrganizationTargetId}{The ID of the Amazon Web Services Organizations target for which to
disable the IPAM policy. This parameter is required only when IPAM is
integrated with Amazon Web Services Organizations. When IPAM is not
integrated with Amazon Web Services Organizations, omit this parameter
and the policy will be disabled for the current account.

A target can be an individual Amazon Web Services account or an entity
within an Amazon Web Services Organization to which an IPAM policy can
be applied.}
}
\description{
Disables an IPAM policy.

See \url{https://www.paws-r-sdk.com/docs/ec2_disable_ipam_policy/} for full documentation.
}
\keyword{internal}
