% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disable_route_server_propagation}
\alias{ec2_disable_route_server_propagation}
\title{Disables route propagation from a route server to a specified route
table}
\usage{
ec2_disable_route_server_propagation(
  RouteServerId,
  RouteTableId,
  DryRun = NULL
)
}
\arguments{
\item{RouteServerId}{[required] The ID of the route server for which to disable propagation.}

\item{RouteTableId}{[required] The ID of the route table for which to disable route server propagation.}

\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Disables route propagation from a route server to a specified route table.

See \url{https://www.paws-r-sdk.com/docs/ec2_disable_route_server_propagation/} for full documentation.
}
\keyword{internal}
