% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disable_snapshot_block_public_access}
\alias{ec2_disable_snapshot_block_public_access}
\title{Disables the block public access for snapshots setting at the account
level for the specified Amazon Web Services Region}
\usage{
ec2_disable_snapshot_block_public_access(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Disables the \emph{block public access for snapshots} setting at the account level for the specified Amazon Web Services Region. After you disable block public access for snapshots in a Region, users can publicly share snapshots in that Region.

See \url{https://www.paws-r-sdk.com/docs/ec2_disable_snapshot_block_public_access/} for full documentation.
}
\keyword{internal}
