% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_enable_application_status_check_suppression}
\alias{ec2_enable_application_status_check_suppression}
\title{Suppresses application status checks for the specified instances}
\usage{
ec2_enable_application_status_check_suppression(
  InstanceIds = NULL,
  DurationSeconds = NULL,
  ClientToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{InstanceIds}{The IDs of the instances for which to suppress application status checks.}

\item{DurationSeconds}{The duration, in seconds, for which to suppress application status checks. If omitted, the application status check is suppressed indefinitely until you call \code{\link[=ec2_disable_application_status_check_suppression]{disable_application_status_check_suppression}}.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure that the operation completes no more than one time. If you retry a request with the same token, the service ignores the request but does not return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Suppresses application status checks for the specified instances. While suppressed, health checks continue to run but do not affect the instance-level application status. The following rules apply:

See \url{https://www.paws-r-sdk.com/docs/ec2_enable_application_status_check_suppression/} for full documentation.
}
\keyword{internal}
