% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_enable_aws_network_performance_metric_subscription}
\alias{ec2_enable_aws_network_performance_metric_subscription}
\title{Enables Infrastructure Performance subscriptions}
\usage{
ec2_enable_aws_network_performance_metric_subscription(
  Source = NULL,
  Destination = NULL,
  Metric = NULL,
  Statistic = NULL,
  DryRun = NULL
)
}
\arguments{
\item{Source}{The source Region or Availability Zone that the metric subscription is
enabled for. For example, \code{us-east-1}.}

\item{Destination}{The target Region or Availability Zone that the metric subscription is
enabled for. For example, \code{eu-west-1}.}

\item{Metric}{The metric used for the enabled subscription.}

\item{Statistic}{The statistic used for the enabled subscription.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Enables Infrastructure Performance subscriptions.

See \url{https://paws-r.github.io/docs/ec2/enable_aws_network_performance_metric_subscription.html} for full documentation.
}
\keyword{internal}
