% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_enable_capacity_manager}
\alias{ec2_enable_capacity_manager}
\title{Enables EC2 Capacity Manager for your account}
\usage{
ec2_enable_capacity_manager(
  OrganizationsAccess = NULL,
  DryRun = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{OrganizationsAccess}{Specifies whether to enable cross-account access for Amazon Web Services
Organizations. When enabled, Capacity Manager can aggregate data from
all accounts in your organization. Default is false.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Enables EC2 Capacity Manager for your account. This starts data ingestion for your EC2 capacity usage across On-Demand, Spot, and Capacity Reservations. Initial data processing may take several hours to complete.

See \url{https://www.paws-r-sdk.com/docs/ec2_enable_capacity_manager/} for full documentation.
}
\keyword{internal}
