% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_enable_image}
\alias{ec2_enable_image}
\title{Re-enables a disabled AMI}
\usage{
ec2_enable_image(ImageId, DryRun = NULL)
}
\arguments{
\item{ImageId}{[required] The ID of the AMI.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Re-enables a disabled AMI. The re-enabled AMI is marked as \code{available} and can be used for instance launches, appears in describe operations, and can be shared. Amazon Web Services accounts, organizations, and Organizational Units that lost access to the AMI when it was disabled do not regain access automatically. Once the AMI is available, it can be shared with them again.

See \url{https://www.paws-r-sdk.com/docs/ec2_enable_image/} for full documentation.
}
\keyword{internal}
