% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_enable_ipam_internet_registry_association}
\alias{ec2_enable_ipam_internet_registry_association}
\title{Enables Resource Public Key Infrastructure (RPKI) on an existing IPAM
internet registry association by providing BGP Public Key Infrastructure
(BPKI) certificate details}
\usage{
ec2_enable_ipam_internet_registry_association(
  DryRun = NULL,
  IpamInternetRegistryAssociationId,
  RpkiVersion,
  ServiceUri,
  ChildHandle,
  ParentHandle,
  ParentBpkiTa,
  ClientToken = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamInternetRegistryAssociationId}{[required] The ID of the IPAM internet registry association to enable.}

\item{RpkiVersion}{[required] The RPKI version to use from the Parent Response XML.}

\item{ServiceUri}{[required] The RPKI service URI for the publication point from the Parent Response XML.}

\item{ChildHandle}{[required] The child handle for the BPKI certificate hierarchy from the Parent Response XML.}

\item{ParentHandle}{[required] The parent handle for the BPKI certificate hierarchy from the Parent Response XML.}

\item{ParentBpkiTa}{[required] The parent BPKI Trust Anchor certificate in PEM format from the Parent Response XML.}

\item{ClientToken}{A unique, case-sensitive identifier to ensure that the operation completes no more than one time. If this token matches a previous request, the operation ignores the request, but does not return an error.}
}
\description{
Enables Resource Public Key Infrastructure (RPKI) on an existing IPAM internet registry association by providing BGP Public Key Infrastructure (BPKI) certificate details. After enabling, you can create Route Origin Authorizations (ROAs) for prefixes registered with the internet registry.

See \url{https://www.paws-r-sdk.com/docs/ec2_enable_ipam_internet_registry_association/} for full documentation.
}
\keyword{internal}
