% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_enable_reachability_analyzer_organization_sharing}
\alias{ec2_enable_reachability_analyzer_organization_sharing}
\title{Establishes a trust relationship between Reachability Analyzer and
Organizations}
\usage{
ec2_enable_reachability_analyzer_organization_sharing(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Establishes a trust relationship between Reachability Analyzer and Organizations. This operation must be performed by the management account for the organization.

See \url{https://www.paws-r-sdk.com/docs/ec2_enable_reachability_analyzer_organization_sharing/} for full documentation.
}
\keyword{internal}
