% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_enable_vpc_classic_link_dns_support}
\alias{ec2_enable_vpc_classic_link_dns_support}
\title{We are retiring EC2-Classic}
\usage{
ec2_enable_vpc_classic_link_dns_support(VpcId = NULL)
}
\arguments{
\item{VpcId}{The ID of the VPC.}
}
\description{
We are retiring EC2-Classic. We recommend that you migrate from EC2-Classic to a VPC. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/vpc-migrate.html}{Migrate from EC2-Classic to a VPC} in the \emph{Amazon Elastic Compute Cloud User Guide}.

See \url{https://paws-r.github.io/docs/ec2/enable_vpc_classic_link_dns_support.html} for full documentation.
}
\keyword{internal}
