% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_export_verified_access_instance_client_configuration}
\alias{ec2_export_verified_access_instance_client_configuration}
\title{Exports the client configuration for a Verified Access instance}
\usage{
ec2_export_verified_access_instance_client_configuration(
  VerifiedAccessInstanceId,
  DryRun = NULL
)
}
\arguments{
\item{VerifiedAccessInstanceId}{[required] The ID of the Verified Access instance.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Exports the client configuration for a Verified Access instance.

See \url{https://www.paws-r-sdk.com/docs/ec2_export_verified_access_instance_client_configuration/} for full documentation.
}
\keyword{internal}
