% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_active_vpn_tunnel_status}
\alias{ec2_get_active_vpn_tunnel_status}
\title{Returns the currently negotiated security parameters for an active VPN
tunnel, including IKE version, DH groups, encryption algorithms, and
integrity algorithms}
\usage{
ec2_get_active_vpn_tunnel_status(
  VpnConnectionId,
  VpnTunnelOutsideIpAddress,
  DryRun = NULL
)
}
\arguments{
\item{VpnConnectionId}{[required] The ID of the VPN connection for which to retrieve the active tunnel
status.}

\item{VpnTunnelOutsideIpAddress}{[required] The external IP address of the VPN tunnel for which to retrieve the
active status.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request.}
}
\description{
Returns the currently negotiated security parameters for an active VPN tunnel, including IKE version, DH groups, encryption algorithms, and integrity algorithms.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_active_vpn_tunnel_status/} for full documentation.
}
\keyword{internal}
