% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_capacity_manager_attributes}
\alias{ec2_get_capacity_manager_attributes}
\title{Retrieves the current configuration and status of EC2 Capacity Manager
for your account, including enablement status, Organizations access
settings, and data ingestion status}
\usage{
ec2_get_capacity_manager_attributes(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Retrieves the current configuration and status of EC2 Capacity Manager for your account, including enablement status, Organizations access settings, and data ingestion status.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_capacity_manager_attributes/} for full documentation.
}
\keyword{internal}
