% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_capacity_manager_metric_dimensions}
\alias{ec2_get_capacity_manager_metric_dimensions}
\title{Retrieves the available dimension values for capacity metrics within a
specified time range}
\usage{
ec2_get_capacity_manager_metric_dimensions(
  GroupBy,
  FilterBy = NULL,
  StartTime,
  EndTime,
  MetricNames,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{GroupBy}{[required] The dimensions to group by when retrieving available dimension values.
This determines which dimension combinations are returned. Required
parameter.}

\item{FilterBy}{Conditions to filter which dimension values are returned. Each filter
specifies a dimension, comparison operator, and values to match against.}

\item{StartTime}{[required] The start time for the dimension query, in ISO 8601 format. Only
dimensions with data in this time range will be returned.}

\item{EndTime}{[required] The end time for the dimension query, in ISO 8601 format. Only
dimensions with data in this time range will be returned.}

\item{MetricNames}{[required] The metric names to use as an additional filter when retrieving
dimensions. Only dimensions that have data for these metrics will be
returned. Required parameter with maximum size of 1 for v1.}

\item{MaxResults}{The maximum number of dimension combinations to return. Valid range is 1
to 1000. Use with NextToken for pagination.}

\item{NextToken}{The token for the next page of results. Use this value in a subsequent
call to retrieve additional dimension values.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Retrieves the available dimension values for capacity metrics within a specified time range. This is useful for discovering what accounts, regions, instance families, and other dimensions have data available for filtering and grouping.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_capacity_manager_metric_dimensions/} for full documentation.
}
\keyword{internal}
