% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_client_vpn_endpoint_authorization_policy}
\alias{ec2_get_client_vpn_endpoint_authorization_policy}
\title{Describes the authorization policy for a Client VPN endpoint}
\usage{
ec2_get_client_vpn_endpoint_authorization_policy(
  ClientVpnEndpointId,
  DryRun = NULL
)
}
\arguments{
\item{ClientVpnEndpointId}{[required] The ID of the Client VPN endpoint.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the authorization policy for a Client VPN endpoint.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_client_vpn_endpoint_authorization_policy/} for full documentation.
}
\keyword{internal}
