% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_console_screenshot}
\alias{ec2_get_console_screenshot}
\title{Retrieve a JPG-format screenshot of a running instance to help with
troubleshooting}
\usage{
ec2_get_console_screenshot(DryRun, InstanceId, WakeUp)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{InstanceId}{[required] The ID of the instance.}

\item{WakeUp}{When set to \code{true}, acts as keystroke input and wakes up an instance
that's in standby or "sleep" mode.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ImageData = "string",
  InstanceId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieve a JPG-format screenshot of a running instance to help with
troubleshooting.

The returned content is Base64-encoded.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_console_screenshot(
  DryRun = TRUE|FALSE,
  InstanceId = "string",
  WakeUp = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
