% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_declarative_policies_report_summary}
\alias{ec2_get_declarative_policies_report_summary}
\title{Retrieves a summary of the account status report}
\usage{
ec2_get_declarative_policies_report_summary(DryRun = NULL, ReportId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ReportId}{[required] The ID of the report.}
}
\description{
Retrieves a summary of the account status report.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_declarative_policies_report_summary/} for full documentation.
}
\keyword{internal}
