% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_flow_logs_integration_template}
\alias{ec2_get_flow_logs_integration_template}
\title{Generates a CloudFormation template that streamlines and automates the
integration of VPC flow logs with Amazon Athena}
\usage{
ec2_get_flow_logs_integration_template(DryRun, FlowLogId,
  ConfigDeliveryS3DestinationArn, IntegrateServices)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{FlowLogId}{[required] The ID of the flow log.}

\item{ConfigDeliveryS3DestinationArn}{[required] To store the CloudFormation template in Amazon S3, specify the location
in Amazon S3.}

\item{IntegrateServices}{[required] Information about the service integration.}
}
\value{
A list with the following syntax:\preformatted{list(
  Result = "string"
)
}
}
\description{
Generates a CloudFormation template that streamlines and automates the
integration of VPC flow logs with Amazon Athena. This make it easier for
you to query and gain insights from VPC flow logs data. Based on the
information that you provide, we configure resources in the template to
do the following:
\itemize{
\item Create a table in Athena that maps fields to a custom log format
\item Create a Lambda function that updates the table with new partitions
on a daily, weekly, or monthly basis
\item Create a table partitioned between two timestamps in the past
\item Create a set of named queries in Athena that you can use to get
started quickly
}
}
\section{Request syntax}{
\preformatted{svc$get_flow_logs_integration_template(
  DryRun = TRUE|FALSE,
  FlowLogId = "string",
  ConfigDeliveryS3DestinationArn = "string",
  IntegrateServices = list(
    AthenaIntegrations = list(
      list(
        IntegrationResultS3DestinationArn = "string",
        PartitionLoadFrequency = "none"|"daily"|"weekly"|"monthly",
        PartitionStartDate = as.POSIXct(
          "2015-01-01"
        ),
        PartitionEndDate = as.POSIXct(
          "2015-01-01"
        )
      )
    )
  )
)
}
}

\keyword{internal}
