% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_instance_metadata_defaults}
\alias{ec2_get_instance_metadata_defaults}
\title{Gets the default instance metadata service (IMDS) settings that are set
at the account level in the specified Amazon Web Services  Region}
\usage{
ec2_get_instance_metadata_defaults(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Gets the default instance metadata service (IMDS) settings that are set at the account level in the specified Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_instance_metadata_defaults/} for full documentation.
}
\keyword{internal}
