% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_instance_uefi_data}
\alias{ec2_get_instance_uefi_data}
\title{A binary representation of the UEFI variable store}
\usage{
ec2_get_instance_uefi_data(InstanceId, DryRun = NULL)
}
\arguments{
\item{InstanceId}{[required] The ID of the instance from which to retrieve the UEFI data.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
A binary representation of the UEFI variable store. Only non-volatile variables are stored. This is a base64 encoded and zlib compressed binary value that must be properly encoded.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_instance_uefi_data/} for full documentation.
}
\keyword{internal}
