% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_ipam_discovered_routes}
\alias{ec2_get_ipam_discovered_routes}
\title{Retrieves Border Gateway Protocol (BGP) routes discovered by IPAM
resource discovery for a specified Region}
\usage{
ec2_get_ipam_discovered_routes(
  DryRun = NULL,
  IpamResourceDiscoveryId,
  ResourceRegion,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamResourceDiscoveryId}{[required] The ID of the IPAM resource discovery.}

\item{ResourceRegion}{[required] The Amazon Web Services Region to retrieve discovered routes for.}

\item{Filters}{One or more filters to apply to the results.}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call. If not specified, all available results are returned. To retrieve the remaining results, make another call with the returned \code{nextToken} value.}
}
\description{
Retrieves Border Gateway Protocol (BGP) routes discovered by IPAM resource discovery for a specified Region. Use this operation to view the Bring Your Own IP (BYOIP) address ranges that are currently advertised through BGP. We recommend using pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_ipam_discovered_routes/} for full documentation.
}
\keyword{internal}
