% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_ipam_policy_allocation_rules}
\alias{ec2_get_ipam_policy_allocation_rules}
\title{Gets the allocation rules for an IPAM policy}
\usage{
ec2_get_ipam_policy_allocation_rules(
  DryRun = NULL,
  IpamPolicyId,
  Filters = NULL,
  Locale = NULL,
  ResourceType = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPolicyId}{[required] The ID of the IPAM policy for which to get allocation rules.}

\item{Filters}{One or more filters for the allocation rules.}

\item{Locale}{The locale for which to get the allocation rules.}

\item{ResourceType}{The resource type for which to get the allocation rules.

The Amazon Web Services service or resource type that can use IP
addresses through IPAM policies. Supported services and resource types
include:
\itemize{
\item Elastic IP addresses
}}

\item{MaxResults}{The maximum number of results to return in a single call.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Gets the allocation rules for an IPAM policy.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_ipam_policy_allocation_rules/} for full documentation.
}
\keyword{internal}
