% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_ipam_prefix_list_resolver_rules}
\alias{ec2_get_ipam_prefix_list_resolver_rules}
\title{Retrieves the CIDR selection rules for an IPAM prefix list resolver}
\usage{
ec2_get_ipam_prefix_list_resolver_rules(
  DryRun = NULL,
  IpamPrefixListResolverId,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPrefixListResolverId}{[required] The ID of the IPAM prefix list resolver whose rules you want to
retrieve.}

\item{Filters}{One or more filters to limit the results.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Retrieves the CIDR selection rules for an IPAM prefix list resolver. Use this operation to view the business logic that determines which CIDRs are selected for synchronization with prefix lists.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_ipam_prefix_list_resolver_rules/} for full documentation.
}
\keyword{internal}
