% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_ipam_prefix_list_resolver_version_entries}
\alias{ec2_get_ipam_prefix_list_resolver_version_entries}
\title{Retrieves the CIDR entries for a specific version of an IPAM prefix list
resolver}
\usage{
ec2_get_ipam_prefix_list_resolver_version_entries(
  DryRun = NULL,
  IpamPrefixListResolverId,
  IpamPrefixListResolverVersion,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPrefixListResolverId}{[required] The ID of the IPAM prefix list resolver whose version entries you want
to retrieve.}

\item{IpamPrefixListResolverVersion}{[required] The version number of the resolver for which to retrieve CIDR entries.
If not specified, the latest version is used.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Retrieves the CIDR entries for a specific version of an IPAM prefix list resolver. This shows the actual CIDRs that were selected and synchronized at a particular point in time.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_ipam_prefix_list_resolver_version_entries/} for full documentation.
}
\keyword{internal}
