% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_ipam_route_origin_authorizations}
\alias{ec2_get_ipam_route_origin_authorizations}
\title{Retrieves the current Route Origin Authorizations (ROAs) published to
the RPKI for an IPAM internet registry association}
\usage{
ec2_get_ipam_route_origin_authorizations(
  DryRun = NULL,
  IpamInternetRegistryAssociationId,
  Cidr = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamInternetRegistryAssociationId}{[required] The ID of the IPAM internet registry association.}

\item{Cidr}{Filter results to a specific CIDR prefix.}

\item{MaxResults}{The maximum number of results to return in a single call. If not specified, all available results are returned. To retrieve the remaining results, make another call with the returned \code{nextToken} value.}

\item{NextToken}{The token for the next page of results.}
}
\description{
Retrieves the current Route Origin Authorizations (ROAs) published to the RPKI for an IPAM internet registry association. We recommend using pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_ipam_route_origin_authorizations/} for full documentation.
}
\keyword{internal}
