% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_ipam_route_protection_findings}
\alias{ec2_get_ipam_route_protection_findings}
\title{Retrieves route protection findings for an IPAM}
\usage{
ec2_get_ipam_route_protection_findings(
  DryRun = NULL,
  IpamId,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamId}{[required] The ID of the IPAM to retrieve route protection findings for.}

\item{Filters}{One or more filters to apply to the results.}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call. If not specified, all available results are returned. To retrieve the remaining results, make another call with the returned \code{nextToken} value.}
}
\description{
Retrieves route protection findings for an IPAM. Route protection findings show the Resource Public Key Infrastructure (RPKI) validation status of your Bring Your Own IP (BYOIP) routes. Findings identify routes that have valid, invalid, or unknown validation states. We recommend using pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_ipam_route_protection_findings/} for full documentation.
}
\keyword{internal}
