% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_managed_resource_visibility}
\alias{ec2_get_managed_resource_visibility}
\title{Retrieves the managed resource visibility configuration for the account}
\usage{
ec2_get_managed_resource_visibility(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the operation,
without actually making the request, and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Retrieves the managed resource visibility configuration for the account. The response indicates whether managed resources are hidden or visible by default.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_managed_resource_visibility/} for full documentation.
}
\keyword{internal}
