% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_network_insights_access_scope_analysis_findings}
\alias{ec2_get_network_insights_access_scope_analysis_findings}
\title{Gets the findings for the specified Network Access Scope analysis}
\usage{
ec2_get_network_insights_access_scope_analysis_findings(
  NetworkInsightsAccessScopeAnalysisId,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{NetworkInsightsAccessScopeAnalysisId}{[required] The ID of the Network Access Scope analysis.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Gets the findings for the specified Network Access Scope analysis.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_network_insights_access_scope_analysis_findings/} for full documentation.
}
\keyword{internal}
