% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_network_insights_access_scope_content}
\alias{ec2_get_network_insights_access_scope_content}
\title{Gets the content for the specified Network Access Scope}
\usage{
ec2_get_network_insights_access_scope_content(
  NetworkInsightsAccessScopeId,
  DryRun = NULL
)
}
\arguments{
\item{NetworkInsightsAccessScopeId}{[required] The ID of the Network Access Scope.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Gets the content for the specified Network Access Scope.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_network_insights_access_scope_content/} for full documentation.
}
\keyword{internal}
