% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_password_data}
\alias{ec2_get_password_data}
\title{Retrieves the encrypted administrator password for a running Windows
instance}
\usage{
ec2_get_password_data(InstanceId, DryRun = NULL)
}
\arguments{
\item{InstanceId}{[required] The ID of the Windows instance.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Retrieves the encrypted administrator password for a running Windows instance.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_password_data/} for full documentation.
}
\keyword{internal}
