% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_route_server_associations}
\alias{ec2_get_route_server_associations}
\title{Gets information about the associations for the specified route server}
\usage{
ec2_get_route_server_associations(RouteServerId, DryRun = NULL)
}
\arguments{
\item{RouteServerId}{[required] The ID of the route server for which to get association information.}

\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Gets information about the associations for the specified route server.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_route_server_associations/} for full documentation.
}
\keyword{internal}
