% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_snapshot_block_public_access_state}
\alias{ec2_get_snapshot_block_public_access_state}
\title{Gets the current state of block public access for snapshots setting for
the account and Region}
\usage{
ec2_get_snapshot_block_public_access_state(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Gets the current state of \emph{block public access for snapshots} setting for the account and Region.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_snapshot_block_public_access_state/} for full documentation.
}
\keyword{internal}
