% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_transit_gateway_metering_policy_entries}
\alias{ec2_get_transit_gateway_metering_policy_entries}
\title{Retrieves the entries for a transit gateway metering policy}
\usage{
ec2_get_transit_gateway_metering_policy_entries(
  TransitGatewayMeteringPolicyId,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayMeteringPolicyId}{[required] The ID of the transit gateway metering policy to retrieve entries for.}

\item{Filters}{One or more filters to apply when retrieving metering policy entries.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Retrieves the entries for a transit gateway metering policy.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_transit_gateway_metering_policy_entries/} for full documentation.
}
\keyword{internal}
