% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_verified_access_endpoint_policy}
\alias{ec2_get_verified_access_endpoint_policy}
\title{Get the Verified Access policy associated with the endpoint}
\usage{
ec2_get_verified_access_endpoint_policy(
  VerifiedAccessEndpointId,
  DryRun = NULL
)
}
\arguments{
\item{VerifiedAccessEndpointId}{[required] The ID of the Verified Access endpoint.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Get the Verified Access policy associated with the endpoint.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_verified_access_endpoint_policy/} for full documentation.
}
\keyword{internal}
