% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_verified_access_endpoint_targets}
\alias{ec2_get_verified_access_endpoint_targets}
\title{Gets the targets for the specified network CIDR endpoint for Verified
Access}
\usage{
ec2_get_verified_access_endpoint_targets(
  VerifiedAccessEndpointId,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{VerifiedAccessEndpointId}{[required] The ID of the network CIDR endpoint.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Gets the targets for the specified network CIDR endpoint for Verified Access.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_verified_access_endpoint_targets/} for full documentation.
}
\keyword{internal}
