% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_verified_access_group_policy}
\alias{ec2_get_verified_access_group_policy}
\title{Shows the contents of the Verified Access policy associated with the
group}
\usage{
ec2_get_verified_access_group_policy(VerifiedAccessGroupId, DryRun = NULL)
}
\arguments{
\item{VerifiedAccessGroupId}{[required] The ID of the Verified Access group.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Shows the contents of the Verified Access policy associated with the group.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_verified_access_group_policy/} for full documentation.
}
\keyword{internal}
