% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_vpc_resources_blocking_encryption_enforcement}
\alias{ec2_get_vpc_resources_blocking_encryption_enforcement}
\title{Gets information about resources in a VPC that are blocking encryption
enforcement}
\usage{
ec2_get_vpc_resources_blocking_encryption_enforcement(
  VpcId,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{VpcId}{[required] The ID of the VPC to check for resources blocking encryption
enforcement.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token returned from a previous paginated request. Pagination
continues from the end of the items returned by the previous request.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Gets information about resources in a VPC that are blocking encryption enforcement.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_vpc_resources_blocking_encryption_enforcement/} for full documentation.
}
\keyword{internal}
