% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_vpn_tunnel_replacement_status}
\alias{ec2_get_vpn_tunnel_replacement_status}
\title{Get details of available tunnel endpoint maintenance}
\usage{
ec2_get_vpn_tunnel_replacement_status(
  VpnConnectionId,
  VpnTunnelOutsideIpAddress,
  DryRun = NULL
)
}
\arguments{
\item{VpnConnectionId}{[required] The ID of the Site-to-Site VPN connection.}

\item{VpnTunnelOutsideIpAddress}{[required] The external IP address of the VPN tunnel.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Get details of available tunnel endpoint maintenance.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_vpn_tunnel_replacement_status/} for full documentation.
}
\keyword{internal}
