% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_import_volume}
\alias{ec2_import_volume}
\title{Creates an import volume task using metadata from the specified disk
image}
\usage{
ec2_import_volume(AvailabilityZone, Description, DryRun, Image, Volume)
}
\arguments{
\item{AvailabilityZone}{[required] The Availability Zone for the resulting EBS volume.}

\item{Description}{A description of the volume.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Image}{[required] The disk image.}

\item{Volume}{[required] The volume size.}
}
\value{
A list with the following syntax:\preformatted{list(
  ConversionTask = list(
    ConversionTaskId = "string",
    ExpirationTime = "string",
    ImportInstance = list(
      Description = "string",
      InstanceId = "string",
      Platform = "Windows",
      Volumes = list(
        list(
          AvailabilityZone = "string",
          BytesConverted = 123,
          Description = "string",
          Image = list(
            Checksum = "string",
            Format = "VMDK"|"RAW"|"VHD",
            ImportManifestUrl = "string",
            Size = 123
          ),
          Status = "string",
          StatusMessage = "string",
          Volume = list(
            Id = "string",
            Size = 123
          )
        )
      )
    ),
    ImportVolume = list(
      AvailabilityZone = "string",
      BytesConverted = 123,
      Description = "string",
      Image = list(
        Checksum = "string",
        Format = "VMDK"|"RAW"|"VHD",
        ImportManifestUrl = "string",
        Size = 123
      ),
      Volume = list(
        Id = "string",
        Size = 123
      )
    ),
    State = "active"|"cancelling"|"cancelled"|"completed",
    StatusMessage = "string",
    Tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}
}
\description{
Creates an import volume task using metadata from the specified disk
image.

This API action supports only single-volume VMs. To import multi-volume
VMs, use \code{\link[=ec2_import_image]{import_image}} instead. To import a disk to
a snapshot, use \code{\link[=ec2_import_snapshot]{import_snapshot}} instead.

This API action is not supported by the AWS Command Line Interface (AWS
CLI). For information about using the Amazon EC2 CLI, which is
deprecated, see \href{https://awsdocs.s3.amazonaws.com/EC2/ec2-clt.pdf#importing-your-volumes-into-amazon-ebs}{Importing Disks to Amazon EBS}
in the \emph{Amazon EC2 CLI Reference} PDF file.

For information about the import manifest referenced by this API action,
see \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/manifest.html}{VM Import Manifest}.
}
\section{Request syntax}{
\preformatted{svc$import_volume(
  AvailabilityZone = "string",
  Description = "string",
  DryRun = TRUE|FALSE,
  Image = list(
    Bytes = 123,
    Format = "VMDK"|"RAW"|"VHD",
    ImportManifestUrl = "string"
  ),
  Volume = list(
    Size = 123
  )
)
}
}

\keyword{internal}
