% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_account_vpc_encryption_control}
\alias{ec2_modify_account_vpc_encryption_control}
\title{Modifies the account-level VPC Encryption Control configuration}
\usage{
ec2_modify_account_vpc_encryption_control(
  DryRun = NULL,
  Mode = NULL,
  InternetGateway = NULL,
  EgressOnlyInternetGateway = NULL,
  NatGateway = NULL,
  VirtualPrivateGateway = NULL,
  VpcPeering = NULL,
  Lambda = NULL,
  VpcLattice = NULL,
  ElasticFileSystem = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Mode}{The encryption mode for the account encryption control configuration.}

\item{InternetGateway}{Specifies whether to exclude internet gateway resource from account-level encryption enforcement.}

\item{EgressOnlyInternetGateway}{Specifies whether to exclude egress-only internet gateway resource from account-level encryption enforcement.}

\item{NatGateway}{Specifies whether to exclude NAT gateway resource from account-level encryption enforcement.}

\item{VirtualPrivateGateway}{Specifies whether to exclude virtual private gateway resource from account-level encryption enforcement.}

\item{VpcPeering}{Specifies whether to exclude VPC peering connection resource from account-level encryption enforcement.}

\item{Lambda}{Specifies whether to exclude Lambda service from account-level encryption enforcement.}

\item{VpcLattice}{Specifies whether to exclude VPC Lattice service from account-level encryption enforcement.}

\item{ElasticFileSystem}{Specifies whether to exclude Elastic File System service from account-level encryption enforcement.}
}
\description{
Modifies the account-level VPC Encryption Control configuration. This sets the encryption control mode and resource exclusions that apply to the VPCs in your account. VPC Encryption Control enables you to enforce encryption for all data in transit within and between VPCs to meet compliance requirements.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_account_vpc_encryption_control/} for full documentation.
}
\keyword{internal}
