% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_client_vpn_endpoint}
\alias{ec2_modify_client_vpn_endpoint}
\title{Modifies the specified Client VPN endpoint}
\usage{
ec2_modify_client_vpn_endpoint(
  ClientVpnEndpointId,
  ServerCertificateArn = NULL,
  ConnectionLogOptions = NULL,
  DnsServers = NULL,
  VpnPort = NULL,
  Description = NULL,
  SplitTunnel = NULL,
  DryRun = NULL,
  SecurityGroupIds = NULL,
  VpcId = NULL,
  SelfServicePortal = NULL,
  ClientConnectOptions = NULL,
  SessionTimeoutHours = NULL,
  ClientLoginBannerOptions = NULL,
  ClientRouteEnforcementOptions = NULL,
  DisconnectOnSessionTimeout = NULL,
  TransitGatewayConfiguration = NULL,
  DevicePostureOptions = NULL
)
}
\arguments{
\item{ClientVpnEndpointId}{[required] The ID of the Client VPN endpoint to modify.}

\item{ServerCertificateArn}{The ARN of the server certificate to be used. The server certificate must be provisioned in Certificate Manager (ACM).}

\item{ConnectionLogOptions}{Information about the client connection logging options.

If you enable client connection logging, data about client connections is sent to a Cloudwatch Logs log stream. The following information is logged:
\itemize{
\item Client connection requests
\item Client connection results (successful and unsuccessful)
\item Reasons for unsuccessful client connection requests
\item Client connection termination time
}}

\item{DnsServers}{Information about the DNS servers to be used by Client VPN connections. A Client VPN endpoint can have up to two DNS servers.}

\item{VpnPort}{The port number to assign to the Client VPN endpoint for TCP and UDP traffic.

Valid Values: \code{443} | \code{1194}

Default Value: \code{443}}

\item{Description}{A brief description of the Client VPN endpoint.}

\item{SplitTunnel}{Indicates whether the VPN is split-tunnel.

For information about split-tunnel VPN endpoints, see \href{https://docs.aws.amazon.com/vpn/latest/clientvpn-admin/split-tunnel-vpn.html}{Split-tunnel Client VPN endpoint} in the \emph{Client VPN Administrator Guide}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{SecurityGroupIds}{The IDs of one or more security groups to apply to the target network.}

\item{VpcId}{The ID of the VPC to associate with the Client VPN endpoint.}

\item{SelfServicePortal}{Specify whether to enable the self-service portal for the Client VPN endpoint.}

\item{ClientConnectOptions}{The options for managing connection authorization for new client connections.}

\item{SessionTimeoutHours}{The maximum VPN session duration time in hours.

Valid values: \code{8 | 10 | 12 | 24}

Default value: \code{24}}

\item{ClientLoginBannerOptions}{Options for enabling a customizable text banner that will be displayed on Amazon Web Services provided clients when a VPN session is established.}

\item{ClientRouteEnforcementOptions}{Client route enforcement is a feature of the Client VPN service that helps enforce administrator defined routes on devices connected through the VPN. T his feature helps improve your security posture by ensuring that network traffic originating from a connected client is not inadvertently sent outside the VPN tunnel.

Client route enforcement works by monitoring the route table of a connected device for routing policy changes to the VPN connection. If the feature detects any VPN routing policy modifications, it will automatically force an update to the route table, reverting it back to the expected route configurations.}

\item{DisconnectOnSessionTimeout}{Indicates whether the client VPN session is disconnected after the maximum timeout specified in \code{sessionTimeoutHours} is reached. If \code{true}, users are prompted to reconnect client VPN. If \code{false}, client VPN attempts to reconnect automatically. The default value is \code{true}.}

\item{TransitGatewayConfiguration}{The Transit Gateway configuration for the Client VPN endpoint. This option is currently not supported.}

\item{DevicePostureOptions}{The device posture options for the Client VPN endpoint. Specifying this parameter replaces the entire device posture configuration for the endpoint. To remove all device trust providers, specify an empty list.}
}
\description{
Modifies the specified Client VPN endpoint. Modifying the DNS server resets existing client connections.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_client_vpn_endpoint/} for full documentation.
}
\keyword{internal}
