% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_client_vpn_endpoint_authorization_policy}
\alias{ec2_modify_client_vpn_endpoint_authorization_policy}
\title{Creates or updates the authorization policy for a Client VPN endpoint}
\usage{
ec2_modify_client_vpn_endpoint_authorization_policy(
  ClientVpnEndpointId,
  PolicyDocument = NULL,
  Description = NULL,
  ShadowMode = NULL,
  ClientToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{ClientVpnEndpointId}{[required] The ID of the Client VPN endpoint.}

\item{PolicyDocument}{The authorization policy document, written in the Cedar policy language. This parameter is required when you create the authorization policy for a Client VPN endpoint that does not already have one.}

\item{Description}{A brief description of the authorization policy.}

\item{ShadowMode}{Specifies whether the authorization policy is evaluated in shadow mode. Possible values include:
\itemize{
\item \code{enabled} - The authorization policy is evaluated and the results are logged, but access is not enforced.
\item \code{disabled} - The authorization policy is enforced.
}

The default value is \code{disabled}.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates or updates the authorization policy for a Client VPN endpoint. A Client VPN endpoint can have one authorization policy. If a policy already exists for the endpoint, the values that you specify replace the corresponding values in the existing policy, and values that you do not specify remain unchanged.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_client_vpn_endpoint_authorization_policy/} for full documentation.
}
\keyword{internal}
