% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_ipam_policy_allocation_rules}
\alias{ec2_modify_ipam_policy_allocation_rules}
\title{Modifies the allocation rules in an IPAM policy}
\usage{
ec2_modify_ipam_policy_allocation_rules(
  DryRun = NULL,
  IpamPolicyId,
  Locale,
  ResourceType,
  AllocationRules = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPolicyId}{[required] The ID of the IPAM policy whose allocation rules you want to modify.}

\item{Locale}{[required] The locale for which to modify the allocation rules.}

\item{ResourceType}{[required] The resource type for which to modify the allocation rules.

The Amazon Web Services service or resource type that can use IP
addresses through IPAM policies. Supported services and resource types
include:
\itemize{
\item Elastic IP addresses
}}

\item{AllocationRules}{The new allocation rules to apply to the IPAM policy.

Allocation rules are optional configurations within an IPAM policy that
map Amazon Web Services resource types to specific IPAM pools. If no
rules are defined, the resource types default to using Amazon-provided
IP addresses.}
}
\description{
Modifies the allocation rules in an IPAM policy.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_ipam_policy_allocation_rules/} for full documentation.
}
\keyword{internal}
