% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_ipam_pool_allocation}
\alias{ec2_modify_ipam_pool_allocation}
\title{Modifies the description of an IPAM pool allocation}
\usage{
ec2_modify_ipam_pool_allocation(
  DryRun = NULL,
  IpamPoolAllocationId,
  Description = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action without actually making the request and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPoolAllocationId}{[required] The ID of the IPAM pool allocation you want to modify.}

\item{Description}{The new description for the IPAM pool allocation. If you submit a \code{null} value, the description is removed from the allocation.}
}
\description{
Modifies the description of an IPAM pool allocation. For more information, see \href{https://docs.aws.amazon.com/vpc/latest/ipam/modify-alloc-ipam.html}{Modify an IPAM pool allocation} in the \emph{Amazon VPC IPAM User Guide}.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_ipam_pool_allocation/} for full documentation.
}
\keyword{internal}
