% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_ipam_prefix_list_resolver}
\alias{ec2_modify_ipam_prefix_list_resolver}
\title{Modifies an IPAM prefix list resolver}
\usage{
ec2_modify_ipam_prefix_list_resolver(
  DryRun = NULL,
  IpamPrefixListResolverId,
  Description = NULL,
  Rules = NULL
)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamPrefixListResolverId}{[required] The ID of the IPAM prefix list resolver to modify.}

\item{Description}{A new description for the IPAM prefix list resolver.}

\item{Rules}{The updated CIDR selection rules for the resolver. These rules replace
the existing rules entirely.}
}
\description{
Modifies an IPAM prefix list resolver. You can update the description and CIDR selection rules. Changes to rules will trigger re-evaluation and potential updates to associated prefix lists.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_ipam_prefix_list_resolver/} for full documentation.
}
\keyword{internal}
