% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_ipam_routing_policy_registration}
\alias{ec2_modify_ipam_routing_policy_registration}
\title{Modifies an existing routing policy registration}
\usage{
ec2_modify_ipam_routing_policy_registration(
  DryRun = NULL,
  IpamInternetRegistryAssociationId,
  Cidr,
  Asns,
  PermitMoreSpecificAnnouncements = NULL,
  MaxLength = NULL,
  Description = NULL,
  Force = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamInternetRegistryAssociationId}{[required] The ID of the IPAM internet registry association.}

\item{Cidr}{[required] The IP address prefix in CIDR notation identifying the routing policy registration to modify.}

\item{Asns}{[required] The updated list of Autonomous System Numbers (ASNs) authorized to originate the prefix.}

\item{PermitMoreSpecificAnnouncements}{Specifies whether to permit more specific route announcements than the CIDR prefix. Default: \code{false}.}

\item{MaxLength}{The new maximum prefix length that the ASNs are authorized to announce. Must be greater than or equal to the prefix length of the CIDR.}

\item{Description}{A new description for the routing policy registration.}

\item{Force}{Forces the modification even if it conflicts with an announced route. Default: \code{false}.}

\item{ClientToken}{A unique, case-sensitive identifier to ensure that the operation completes no more than one time. If this token matches a previous request, the operation ignores the request, but does not return an error.}
}
\description{
Modifies an existing routing policy registration. You can update the authorized ASNs, maximum prefix length, and other properties of a Route Origin Authorization (ROA).

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_ipam_routing_policy_registration/} for full documentation.
}
\keyword{internal}
