% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_managed_resource_visibility}
\alias{ec2_modify_managed_resource_visibility}
\title{Modifies the managed resource visibility configuration for the account}
\usage{
ec2_modify_managed_resource_visibility(DryRun = NULL, DefaultVisibility)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the operation,
without actually making the request, and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{DefaultVisibility}{[required] The default visibility setting for managed resources. Valid values:
\code{hidden} | \code{visible}.}
}
\description{
Modifies the managed resource visibility configuration for the account. Use this operation to control whether managed resources are hidden or visible by default. Visibility settings are account-wide and affect all IAM principals uniformly. Hidden resources remain fully operational and billable.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_managed_resource_visibility/} for full documentation.
}
\keyword{internal}
