% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_route_server}
\alias{ec2_modify_route_server}
\title{Modifies the configuration of an existing route server}
\usage{
ec2_modify_route_server(
  RouteServerId,
  PersistRoutes = NULL,
  PersistRoutesDuration = NULL,
  SnsNotificationsEnabled = NULL,
  DryRun = NULL
)
}
\arguments{
\item{RouteServerId}{[required] The ID of the route server to modify.}

\item{PersistRoutes}{Specifies whether to persist routes after all BGP sessions are
terminated.
\itemize{
\item enable: Routes will be persisted in FIB and RIB after all BGP
sessions are terminated.
\item disable: Routes will not be persisted in FIB and RIB after all BGP
sessions are terminated.
\item reset: If a route server has persisted routes due to all BGP
sessions having ended, reset will withdraw all routes and reset
route server to an empty FIB and RIB.
}}

\item{PersistRoutesDuration}{The number of minutes a route server will wait after BGP is
re-established to unpersist the routes in the FIB and RIB. Value must be
in the range of 1-5. Required if PersistRoutes is \code{enabled}.

If you set the duration to 1 minute, then when your network appliance
re-establishes BGP with route server, it has 1 minute to relearn it's
adjacent network and advertise those routes to route server before route
server resumes normal functionality. In most cases, 1 minute is probably
sufficient. If, however, you have concerns that your BGP network may not
be capable of fully re-establishing and re-learning everything in 1
minute, you can increase the duration up to 5 minutes.}

\item{SnsNotificationsEnabled}{Specifies whether to enable SNS notifications for route server events.
Enabling SNS notifications persists BGP status changes to an SNS topic
provisioned by Amazon Web Services.}

\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Modifies the configuration of an existing route server.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_route_server/} for full documentation.
}
\keyword{internal}
