% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_security_group_rules}
\alias{ec2_modify_security_group_rules}
\title{Modifies the rules of a security group}
\usage{
ec2_modify_security_group_rules(GroupId, SecurityGroupRules, DryRun = NULL)
}
\arguments{
\item{GroupId}{[required] The ID of the security group.}

\item{SecurityGroupRules}{[required] Information about the security group properties to update.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Modifies the rules of a security group.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_security_group_rules/} for full documentation.
}
\keyword{internal}
