% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_spot_fleet_request}
\alias{ec2_modify_spot_fleet_request}
\title{Modifies the specified Spot Fleet request}
\usage{
ec2_modify_spot_fleet_request(
  ExcessCapacityTerminationPolicy = NULL,
  LaunchTemplateConfigs = NULL,
  SpotFleetRequestId,
  TargetCapacity = NULL,
  OnDemandTargetCapacity = NULL,
  Context = NULL
)
}
\arguments{
\item{ExcessCapacityTerminationPolicy}{Indicates whether running instances should be terminated if the target
capacity of the Spot Fleet request is decreased below the current size
of the Spot Fleet.

Supported only for fleets of type \code{maintain}.}

\item{LaunchTemplateConfigs}{The launch template and overrides. You can only use this parameter if
you specified a launch template (\code{LaunchTemplateConfigs}) in your Spot
Fleet request. If you specified \code{LaunchSpecifications} in your Spot
Fleet request, then omit this parameter.}

\item{SpotFleetRequestId}{[required] The ID of the Spot Fleet request.}

\item{TargetCapacity}{The size of the fleet.}

\item{OnDemandTargetCapacity}{The number of On-Demand Instances in the fleet.}

\item{Context}{Reserved.}
}
\description{
Modifies the specified Spot Fleet request.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_spot_fleet_request/} for full documentation.
}
\keyword{internal}
