% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_transit_gateway}
\alias{ec2_modify_transit_gateway}
\title{Modifies the specified transit gateway}
\usage{
ec2_modify_transit_gateway(
  TransitGatewayId,
  Description = NULL,
  Options = NULL,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayId}{[required] The ID of the transit gateway.}

\item{Description}{The description for the transit gateway.}

\item{Options}{The options to modify.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Modifies the specified transit gateway. When you modify a transit gateway, the modified options are applied to new transit gateway attachments only. Your existing transit gateway attachments are not modified.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_transit_gateway/} for full documentation.
}
\keyword{internal}
