% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_transit_gateway_policy_table_entry}
\alias{ec2_modify_transit_gateway_policy_table_entry}
\title{Modifies the specified transit gateway policy table entry}
\usage{
ec2_modify_transit_gateway_policy_table_entry(
  TransitGatewayPolicyTableId,
  PolicyRuleNumber,
  PolicyRule = NULL,
  TargetRouteTableId = NULL,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayPolicyTableId}{[required] The ID of the transit gateway policy table.}

\item{PolicyRuleNumber}{[required] The rule number of the policy table entry to modify.}

\item{PolicyRule}{The updated matching criteria for the policy table entry. Unspecified fields retain their current values.}

\item{TargetRouteTableId}{The ID of the transit gateway route table to use for traffic matching this rule.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Modifies the specified transit gateway policy table entry.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_transit_gateway_policy_table_entry/} for full documentation.
}
\keyword{internal}
