% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_verified_access_group_policy}
\alias{ec2_modify_verified_access_group_policy}
\title{Modifies the specified Amazon Web Services Verified Access group policy}
\usage{
ec2_modify_verified_access_group_policy(
  VerifiedAccessGroupId,
  PolicyEnabled,
  PolicyDocument = NULL,
  ClientToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{VerifiedAccessGroupId}{[required] The ID of the Verified Access group.}

\item{PolicyEnabled}{[required] The status of the Verified Access policy.}

\item{PolicyDocument}{The Verified Access policy document.}

\item{ClientToken}{A unique, case-sensitive token that you provide to ensure idempotency of
your modification request. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Modifies the specified Amazon Web Services Verified Access group policy.

See \url{https://paws-r.github.io/docs/ec2/modify_verified_access_group_policy.html} for full documentation.
}
\keyword{internal}
