% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_verified_access_trust_provider}
\alias{ec2_modify_verified_access_trust_provider}
\title{Modifies the configuration of the specified Amazon Web Services Verified
Access trust provider}
\usage{
ec2_modify_verified_access_trust_provider(
  VerifiedAccessTrustProviderId,
  OidcOptions = NULL,
  Description = NULL,
  DryRun = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{VerifiedAccessTrustProviderId}{[required] The ID of the Verified Access trust provider.}

\item{OidcOptions}{The options for an OpenID Connect-compatible user-identity trust
provider.}

\item{Description}{A description for the Verified Access trust provider.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ClientToken}{A unique, case-sensitive token that you provide to ensure idempotency of
your modification request. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.}
}
\description{
Modifies the configuration of the specified Amazon Web Services Verified Access trust provider.

See \url{https://paws-r.github.io/docs/ec2/modify_verified_access_trust_provider.html} for full documentation.
}
\keyword{internal}
