% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_vpc_endpoint_payer_responsibility}
\alias{ec2_modify_vpc_endpoint_payer_responsibility}
\title{Modifies the billing account for VPC endpoint usage/charges}
\usage{
ec2_modify_vpc_endpoint_payer_responsibility(
  DryRun = NULL,
  ServiceId = NULL,
  VpcEndpointId,
  PayerResponsibility,
  Scope
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{ServiceId}{The ID of the VPC endpoint service.}

\item{VpcEndpointId}{[required] The ID of the VPC endpoint.}

\item{PayerResponsibility}{[required] The Amazon Web Services account to which the usage of VPC endpoint is charged.}

\item{Scope}{[required] The scope of usage/charges for which the billing account is being modified.}
}
\description{
Modifies the billing account for VPC endpoint usage/charges.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_vpc_endpoint_payer_responsibility/} for full documentation.
}
\keyword{internal}
