% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_vpc_endpoint_service_payer_responsibility}
\alias{ec2_modify_vpc_endpoint_service_payer_responsibility}
\title{Modifies the payer responsibility for your VPC endpoint service}
\usage{
ec2_modify_vpc_endpoint_service_payer_responsibility(
  DryRun = NULL,
  ServiceId,
  PayerResponsibility
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ServiceId}{[required] The ID of the service.}

\item{PayerResponsibility}{[required] The entity that is responsible for the endpoint costs. The default is
the endpoint owner. If you set the payer responsibility to the service
owner, you cannot set it back to the endpoint owner.}
}
\description{
Modifies the payer responsibility for your VPC endpoint service.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_vpc_endpoint_service_payer_responsibility/} for full documentation.
}
\keyword{internal}
