% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_vpc_tenancy}
\alias{ec2_modify_vpc_tenancy}
\title{Modifies the instance tenancy attribute of the specified VPC}
\usage{
ec2_modify_vpc_tenancy(VpcId, InstanceTenancy, DryRun = NULL)
}
\arguments{
\item{VpcId}{[required] The ID of the VPC.}

\item{InstanceTenancy}{[required] The instance tenancy attribute for the VPC.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Modifies the instance tenancy attribute of the specified VPC. You can change the instance tenancy attribute of a VPC to \code{default} only. You cannot change the instance tenancy attribute to \code{dedicated}.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_vpc_tenancy/} for full documentation.
}
\keyword{internal}
