% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_purchase_capacity_block_extension}
\alias{ec2_purchase_capacity_block_extension}
\title{Purchase the Capacity Block extension for use with your account}
\usage{
ec2_purchase_capacity_block_extension(
  CapacityBlockExtensionOfferingId,
  CapacityReservationId,
  DryRun = NULL
)
}
\arguments{
\item{CapacityBlockExtensionOfferingId}{[required] The ID of the Capacity Block extension offering to purchase.}

\item{CapacityReservationId}{[required] The ID of the Capacity reservation to be extended.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Purchase the Capacity Block extension for use with your account. You must specify the ID of the Capacity Block extension offering you are purchasing.

See \url{https://www.paws-r-sdk.com/docs/ec2_purchase_capacity_block_extension/} for full documentation.
}
\keyword{internal}
