% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_reject_capacity_reservation_billing_ownership}
\alias{ec2_reject_capacity_reservation_billing_ownership}
\title{Rejects a request to assign billing of the available capacity of a
shared Capacity Reservation to your account}
\usage{
ec2_reject_capacity_reservation_billing_ownership(
  DryRun = NULL,
  CapacityReservationId
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{CapacityReservationId}{[required] The ID of the Capacity Reservation for which to reject the request.}
}
\description{
Rejects a request to assign billing of the available capacity of a shared Capacity Reservation to your account. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/assign-billing.html}{Billing assignment for shared Amazon EC2 Capacity Reservations}.

See \url{https://www.paws-r-sdk.com/docs/ec2_reject_capacity_reservation_billing_ownership/} for full documentation.
}
\keyword{internal}
