% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_reject_transit_gateway_client_vpn_attachment}
\alias{ec2_reject_transit_gateway_client_vpn_attachment}
\title{Rejects a Transit Gateway attachment request for a Client VPN endpoint}
\usage{
ec2_reject_transit_gateway_client_vpn_attachment(
  TransitGatewayAttachmentId,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayAttachmentId}{[required] The ID of the Transit Gateway attachment.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Rejects a Transit Gateway attachment request for a Client VPN endpoint. The Transit Gateway owner can reject the attachment request to prevent the Client VPN endpoint from routing traffic through the Transit Gateway.

See \url{https://www.paws-r-sdk.com/docs/ec2_reject_transit_gateway_client_vpn_attachment/} for full documentation.
}
\keyword{internal}
