% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_reject_transit_gateway_multicast_domain_associations}
\alias{ec2_reject_transit_gateway_multicast_domain_associations}
\title{Rejects a request to associate cross-account subnets with a transit
gateway multicast domain}
\usage{
ec2_reject_transit_gateway_multicast_domain_associations(
  TransitGatewayMulticastDomainId, TransitGatewayAttachmentId, SubnetIds,
  DryRun)
}
\arguments{
\item{TransitGatewayMulticastDomainId}{The ID of the transit gateway multicast domain.}

\item{TransitGatewayAttachmentId}{The ID of the transit gateway attachment.}

\item{SubnetIds}{The IDs of the subnets to associate with the transit gateway multicast
domain.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:\preformatted{list(
  Associations = list(
    TransitGatewayMulticastDomainId = "string",
    TransitGatewayAttachmentId = "string",
    ResourceId = "string",
    ResourceType = "vpc"|"vpn"|"direct-connect-gateway"|"connect"|"peering"|"tgw-peering",
    ResourceOwnerId = "string",
    Subnets = list(
      list(
        SubnetId = "string",
        State = "pendingAcceptance"|"associating"|"associated"|"disassociating"|"disassociated"|"rejected"|"failed"
      )
    )
  )
)
}
}
\description{
Rejects a request to associate cross-account subnets with a transit
gateway multicast domain.
}
\section{Request syntax}{
\preformatted{svc$reject_transit_gateway_multicast_domain_associations(
  TransitGatewayMulticastDomainId = "string",
  TransitGatewayAttachmentId = "string",
  SubnetIds = list(
    "string"
  ),
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
