% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_reject_transit_gateway_peering_attachment}
\alias{ec2_reject_transit_gateway_peering_attachment}
\title{Rejects a transit gateway peering attachment request}
\usage{
ec2_reject_transit_gateway_peering_attachment(
  TransitGatewayAttachmentId,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayAttachmentId}{[required] The ID of the transit gateway peering attachment.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Rejects a transit gateway peering attachment request.

See \url{https://www.paws-r-sdk.com/docs/ec2_reject_transit_gateway_peering_attachment/} for full documentation.
}
\keyword{internal}
