% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_replace_image_criteria_in_allowed_images_settings}
\alias{ec2_replace_image_criteria_in_allowed_images_settings}
\title{Sets or replaces the criteria for Allowed AMIs}
\usage{
ec2_replace_image_criteria_in_allowed_images_settings(
  ImageCriteria = NULL,
  DryRun = NULL
)
}
\arguments{
\item{ImageCriteria}{The list of criteria that are evaluated to determine whether AMIs are
discoverable and usable in the account in the specified Amazon Web
Services Region.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Sets or replaces the criteria for Allowed AMIs.

See \url{https://www.paws-r-sdk.com/docs/ec2_replace_image_criteria_in_allowed_images_settings/} for full documentation.
}
\keyword{internal}
