% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_replace_image_instance_type_specification}
\alias{ec2_replace_image_instance_type_specification}
\title{Replaces or removes the instance type specification for an AMI}
\usage{
ec2_replace_image_instance_type_specification(
  ImageId,
  InstanceTypeSpecification = NULL,
  DryRun = NULL
)
}
\arguments{
\item{ImageId}{[required] The ID of the AMI.}

\item{InstanceTypeSpecification}{The instance type specification to set on the AMI. Omit this parameter to remove the existing instance type specification.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Replaces or removes the instance type specification for an AMI. The instance type specification defines which instance types are compatible with the AMI.

See \url{https://www.paws-r-sdk.com/docs/ec2_replace_image_instance_type_specification/} for full documentation.
}
\keyword{internal}
