% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_replace_network_acl_association}
\alias{ec2_replace_network_acl_association}
\title{Changes which network ACL a subnet is associated with}
\usage{
ec2_replace_network_acl_association(AssociationId, DryRun, NetworkAclId)
}
\arguments{
\item{AssociationId}{[required] The ID of the current association between the original network ACL and
the subnet.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{NetworkAclId}{[required] The ID of the new network ACL to associate with the subnet.}
}
\description{
Changes which network ACL a subnet is associated with. By default when
you create a subnet, it's automatically associated with the default
network ACL. For more information, see \href{https://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_ACLs.html}{Network ACLs}
in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\details{
This is an idempotent operation.
}
\section{Request syntax}{
\preformatted{svc$replace_network_acl_association(
  AssociationId = "string",
  DryRun = TRUE|FALSE,
  NetworkAclId = "string"
)
}
}

\examples{
# This example associates the specified network ACL with the subnet for
# the specified network ACL association.
\donttest{svc$replace_network_acl_association(
  AssociationId = "aclassoc-e5b95c8c",
  NetworkAclId = "acl-5fb85d36"
)}

}
\keyword{internal}
