% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_replace_vpn_tunnel}
\alias{ec2_replace_vpn_tunnel}
\title{Trigger replacement of specified VPN tunnel}
\usage{
ec2_replace_vpn_tunnel(
  VpnConnectionId,
  VpnTunnelOutsideIpAddress,
  ApplyPendingMaintenance = NULL,
  DryRun = NULL
)
}
\arguments{
\item{VpnConnectionId}{[required] The ID of the Site-to-Site VPN connection.}

\item{VpnTunnelOutsideIpAddress}{[required] The external IP address of the VPN tunnel.}

\item{ApplyPendingMaintenance}{Trigger pending tunnel endpoint maintenance.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Trigger replacement of specified VPN tunnel.

See \url{https://www.paws-r-sdk.com/docs/ec2_replace_vpn_tunnel/} for full documentation.
}
\keyword{internal}
