% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_reset_address_attribute}
\alias{ec2_reset_address_attribute}
\title{Resets the attribute of the specified IP address}
\usage{
ec2_reset_address_attribute(AllocationId, Attribute, DryRun)
}
\arguments{
\item{AllocationId}{[required] [EC2-VPC] The allocation ID.}

\item{Attribute}{[required] The attribute of the IP address.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:\preformatted{list(
  Address = list(
    PublicIp = "string",
    AllocationId = "string",
    PtrRecord = "string",
    PtrRecordUpdate = list(
      Value = "string",
      Status = "string",
      Reason = "string"
    )
  )
)
}
}
\description{
Resets the attribute of the specified IP address. For requirements, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/elastic-ip-addresses-eip.html#Using_Elastic_Addressing_Reverse_DNS}{Using reverse DNS for email applications}.
}
\section{Request syntax}{
\preformatted{svc$reset_address_attribute(
  AllocationId = "string",
  Attribute = "domain-name",
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
