% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_reset_fpga_image_attribute}
\alias{ec2_reset_fpga_image_attribute}
\title{Resets the specified attribute of the specified Amazon FPGA Image (AFI)
to its default value}
\usage{
ec2_reset_fpga_image_attribute(DryRun, FpgaImageId, Attribute)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{FpgaImageId}{[required] The ID of the AFI.}

\item{Attribute}{The attribute.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Return = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
Resets the specified attribute of the specified Amazon FPGA Image (AFI)
to its default value. You can only reset the load permission attribute.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$reset_fpga_image_attribute(
  DryRun = TRUE|FALSE,
  FpgaImageId = "string",
  Attribute = "loadPermission"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
