% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_reset_snapshot_attribute}
\alias{ec2_reset_snapshot_attribute}
\title{Resets permission settings for the specified snapshot}
\usage{
ec2_reset_snapshot_attribute(Attribute, SnapshotId, DryRun)
}
\arguments{
\item{Attribute}{[required] The attribute to reset. Currently, only the attribute for permission to
create volumes can be reset.}

\item{SnapshotId}{[required] The ID of the snapshot.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
An empty list.
}
\description{
Resets permission settings for the specified snapshot.

For more information about modifying snapshot permissions, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ebs-modifying-snapshot-permissions.html}{Sharing snapshots}
in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$reset_snapshot_attribute(
  Attribute = "productCodes"|"createVolumePermission",
  SnapshotId = "string",
  DryRun = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# This example resets the create volume permissions for snapshot
# `snap-1234567890abcdef0`. If the command succeeds, no output is
# returned.
svc$reset_snapshot_attribute(
  Attribute = "createVolumePermission",
  SnapshotId = "snap-1234567890abcdef0"
)
}

}
\keyword{internal}
