% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_restore_managed_prefix_list_version}
\alias{ec2_restore_managed_prefix_list_version}
\title{Restores the entries from a previous version of a managed prefix list to
a new version of the prefix list}
\usage{
ec2_restore_managed_prefix_list_version(DryRun, PrefixListId,
  PreviousVersion, CurrentVersion)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{PrefixListId}{[required] The ID of the prefix list.}

\item{PreviousVersion}{[required] The version to restore.}

\item{CurrentVersion}{[required] The current version number for the prefix list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PrefixList = list(
    PrefixListId = "string",
    AddressFamily = "string",
    State = "create-in-progress"|"create-complete"|"create-failed"|"modify-in-progress"|"modify-complete"|"modify-failed"|"restore-in-progress"|"restore-complete"|"restore-failed"|"delete-in-progress"|"delete-complete"|"delete-failed",
    StateMessage = "string",
    PrefixListArn = "string",
    PrefixListName = "string",
    MaxEntries = 123,
    Version = 123,
    Tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    OwnerId = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Restores the entries from a previous version of a managed prefix list to
a new version of the prefix list.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$restore_managed_prefix_list_version(
  DryRun = TRUE|FALSE,
  PrefixListId = "string",
  PreviousVersion = 123,
  CurrentVersion = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
