% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_start_instances}
\alias{ec2_start_instances}
\title{Starts an Amazon EBS-backed instance that you've previously stopped}
\usage{
ec2_start_instances(InstanceIds, AdditionalInfo = NULL, DryRun = NULL)
}
\arguments{
\item{InstanceIds}{[required] The IDs of the instances.}

\item{AdditionalInfo}{Reserved.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Starts an Amazon EBS-backed instance that you've previously stopped.

See \url{https://www.paws-r-sdk.com/docs/ec2_start_instances/} for full documentation.
}
\keyword{internal}
