% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_start_network_insights_access_scope_analysis}
\alias{ec2_start_network_insights_access_scope_analysis}
\title{Starts analyzing the specified Network Access Scope}
\usage{
ec2_start_network_insights_access_scope_analysis(
  NetworkInsightsAccessScopeId,
  DryRun = NULL,
  TagSpecifications = NULL,
  ClientToken
)
}
\arguments{
\item{NetworkInsightsAccessScopeId}{[required] The ID of the Network Access Scope.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{TagSpecifications}{The tags to apply.}

\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{How to ensure idempotency}.}
}
\description{
Starts analyzing the specified Network Access Scope.

See \url{https://www.paws-r-sdk.com/docs/ec2_start_network_insights_access_scope_analysis/} for full documentation.
}
\keyword{internal}
