% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_unlock_snapshot}
\alias{ec2_unlock_snapshot}
\title{Unlocks a snapshot that is locked in governance mode or that is locked
in compliance mode but still in the cooling-off period}
\usage{
ec2_unlock_snapshot(SnapshotId, DryRun = NULL)
}
\arguments{
\item{SnapshotId}{[required] The ID of the snapshot to unlock.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Unlocks a snapshot that is locked in governance mode or that is locked in compliance mode but still in the cooling-off period. You can't unlock a snapshot that is locked in compliance mode after the cooling-off period has expired.

See \url{https://www.paws-r-sdk.com/docs/ec2_unlock_snapshot/} for full documentation.
}
\keyword{internal}
