% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_unmonitor_instances}
\alias{ec2_unmonitor_instances}
\title{Disables detailed monitoring for a running instance}
\usage{
ec2_unmonitor_instances(InstanceIds, DryRun)
}
\arguments{
\item{InstanceIds}{[required] The IDs of the instances.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  InstanceMonitorings = list(
    list(
      InstanceId = "string",
      Monitoring = list(
        State = "disabled"|"disabling"|"enabled"|"pending"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Disables detailed monitoring for a running instance. For more
information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-cloudwatch.html}{Monitoring your instances and volumes}
in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$unmonitor_instances(
  InstanceIds = list(
    "string"
  ),
  DryRun = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
