% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_update_capacity_manager_monitored_tag_keys}
\alias{ec2_update_capacity_manager_monitored_tag_keys}
\title{Activates or deactivates tag keys for monitoring by EC2 Capacity Manager}
\usage{
ec2_update_capacity_manager_monitored_tag_keys(
  ActivateTagKeys = NULL,
  DeactivateTagKeys = NULL,
  DryRun = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{ActivateTagKeys}{The tag keys to activate for monitoring. Once activated, these tag keys
will be included as dimensions in capacity metric data.}

\item{DeactivateTagKeys}{The tag keys to deactivate. Deactivated tag keys will no longer be
included as dimensions in capacity metric data.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Activates or deactivates tag keys for monitoring by EC2 Capacity Manager. Activated tag keys are included as dimensions in capacity metric data, enabling you to group and filter metrics by tag values.

See \url{https://www.paws-r-sdk.com/docs/ec2_update_capacity_manager_monitored_tag_keys/} for full documentation.
}
\keyword{internal}
