% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_update_capacity_manager_organizations_access}
\alias{ec2_update_capacity_manager_organizations_access}
\title{Updates the Organizations access setting for EC2 Capacity Manager}
\usage{
ec2_update_capacity_manager_organizations_access(
  OrganizationsAccess,
  DryRun = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{OrganizationsAccess}{[required] Specifies whether to enable or disable cross-account access for Amazon
Web Services Organizations. When enabled, Capacity Manager aggregates
data from all accounts in your organization.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Updates the Organizations access setting for EC2 Capacity Manager. This controls whether Capacity Manager can aggregate data from all accounts in your Amazon Web Services Organization or only from the current account.

See \url{https://www.paws-r-sdk.com/docs/ec2_update_capacity_manager_organizations_access/} for full documentation.
}
\keyword{internal}
