% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_validate_security_group_quotas_for_interface}
\alias{ec2_validate_security_group_quotas_for_interface}
\title{Validates whether the specified security groups can be associated with a
single network interface}
\usage{
ec2_validate_security_group_quotas_for_interface(
  SecurityGroupIds,
  DryRun = NULL
)
}
\arguments{
\item{SecurityGroupIds}{[required] The IDs of the security groups to validate for association with a single network interface. You must specify at least one ID, and each ID must be unique. The number of IDs cannot exceed the maximum number of security groups allowed per network interface.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Validates whether the specified security groups can be associated with a single network interface. The operation checks Amazon Virtual Private Cloud (Amazon VPC) quotas. It checks inbound or outbound rules per security group and security groups per network interface. Only authorized Amazon Web Services services can call this operation.

See \url{https://www.paws-r-sdk.com/docs/ec2_validate_security_group_quotas_for_interface/} for full documentation.
}
\keyword{internal}
