% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_batch_get_repository_scanning_configuration}
\alias{ecr_batch_get_repository_scanning_configuration}
\title{Gets the scanning configuration for one or more repositories}
\usage{
ecr_batch_get_repository_scanning_configuration(repositoryNames)
}
\arguments{
\item{repositoryNames}{[required] One or more repository names to get the scanning configuration for.}
}
\description{
Gets the scanning configuration for one or more repositories.

See \url{https://www.paws-r-sdk.com/docs/ecr_batch_get_repository_scanning_configuration/} for full documentation.
}
\keyword{internal}
