% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_create_pull_through_cache_rule}
\alias{ecr_create_pull_through_cache_rule}
\title{Creates a pull through cache rule}
\usage{
ecr_create_pull_through_cache_rule(
  ecrRepositoryPrefix,
  upstreamRegistryUrl,
  registryId = NULL
)
}
\arguments{
\item{ecrRepositoryPrefix}{[required] The repository name prefix to use when caching images from the source
registry.}

\item{upstreamRegistryUrl}{[required] The registry URL of the upstream public registry to use as the source
for the pull through cache rule.}

\item{registryId}{The Amazon Web Services account ID associated with the registry to
create the pull through cache rule for. If you do not specify a
registry, the default registry is assumed.}
}
\description{
Creates a pull through cache rule. A pull through cache rule provides a way to cache images from an external public registry in your Amazon ECR private registry.

See \url{https://www.paws-r-sdk.com/docs/ecr_create_pull_through_cache_rule/} for full documentation.
}
\keyword{internal}
