% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_delete_lifecycle_policy}
\alias{ecr_delete_lifecycle_policy}
\title{Deletes the lifecycle policy associated with the specified repository}
\usage{
ecr_delete_lifecycle_policy(registryId, repositoryName)
}
\arguments{
\item{registryId}{The AWS account ID associated with the registry that contains the
repository. If you do not specify a registry, the default registry is
assumed.}

\item{repositoryName}{[required] The name of the repository.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  registryId = "string",
  repositoryName = "string",
  lifecyclePolicyText = "string",
  lastEvaluatedAt = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the lifecycle policy associated with the specified repository.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_lifecycle_policy(
  registryId = "string",
  repositoryName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
