% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_delete_pull_through_cache_rule}
\alias{ecr_delete_pull_through_cache_rule}
\title{Deletes a pull through cache rule}
\usage{
ecr_delete_pull_through_cache_rule(ecrRepositoryPrefix, registryId = NULL)
}
\arguments{
\item{ecrRepositoryPrefix}{[required] The Amazon ECR repository prefix associated with the pull through cache
rule to delete.}

\item{registryId}{The Amazon Web Services account ID associated with the registry that
contains the pull through cache rule. If you do not specify a registry,
the default registry is assumed.}
}
\description{
Deletes a pull through cache rule.

See \url{https://www.paws-r-sdk.com/docs/ecr_delete_pull_through_cache_rule/} for full documentation.
}
\keyword{internal}
