% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_delete_registry_policy}
\alias{ecr_delete_registry_policy}
\title{Deletes the registry permissions policy}
\usage{
ecr_delete_registry_policy()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  registryId = "string",
  policyText = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the registry permissions policy.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_registry_policy()
}\if{html}{\out{</div>}}
}

\keyword{internal}
