% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_delete_repository_creation_template}
\alias{ecr_delete_repository_creation_template}
\title{Deletes a repository creation template}
\usage{
ecr_delete_repository_creation_template(prefix)
}
\arguments{
\item{prefix}{[required] The repository namespace prefix associated with the repository creation
template.}
}
\description{
Deletes a repository creation template.

See \url{https://www.paws-r-sdk.com/docs/ecr_delete_repository_creation_template/} for full documentation.
}
\keyword{internal}
