% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_deregister_pull_time_update_exclusion}
\alias{ecr_deregister_pull_time_update_exclusion}
\title{Removes a principal from the pull time update exclusion list for a
registry}
\usage{
ecr_deregister_pull_time_update_exclusion(principalArn)
}
\arguments{
\item{principalArn}{[required] The ARN of the IAM principal to remove from the pull time update
exclusion list.}
}
\description{
Removes a principal from the pull time update exclusion list for a registry. Once removed, Amazon ECR will resume updating the pull time if the specified principal pulls an image.

See \url{https://www.paws-r-sdk.com/docs/ecr_deregister_pull_time_update_exclusion/} for full documentation.
}
\keyword{internal}
