% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_describe_image_replication_status}
\alias{ecr_describe_image_replication_status}
\title{Returns the replication status for a specified image}
\usage{
ecr_describe_image_replication_status(
  repositoryName,
  imageId,
  registryId = NULL
)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that the image is in.}

\item{imageId}{[required]}

\item{registryId}{The Amazon Web Services account ID associated with the registry. If you
do not specify a registry, the default registry is assumed.}
}
\description{
Returns the replication status for a specified image.

See \url{https://www.paws-r-sdk.com/docs/ecr_describe_image_replication_status/} for full documentation.
}
\keyword{internal}
