% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_describe_image_signing_status}
\alias{ecr_describe_image_signing_status}
\title{Returns the signing status for a specified image}
\usage{
ecr_describe_image_signing_status(repositoryName, imageId, registryId = NULL)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the image.}

\item{imageId}{[required] An object containing identifying information for an image.}

\item{registryId}{The Amazon Web Services account ID associated with the registry that
contains the repository. If you do not specify a registry, the default
registry is assumed.}
}
\description{
Returns the signing status for a specified image. If the image matched signing rules that reference different signing profiles, a status is returned for each profile.

See \url{https://www.paws-r-sdk.com/docs/ecr_describe_image_signing_status/} for full documentation.
}
\keyword{internal}
