% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_describe_registry}
\alias{ecr_describe_registry}
\title{Describes the settings for a registry}
\usage{
ecr_describe_registry()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  registryId = "string",
  replicationConfiguration = list(
    rules = list(
      list(
        destinations = list(
          list(
            region = "string",
            registryId = "string"
          )
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the settings for a registry. The replication configuration for
a repository can be created or updated with the
\code{\link[=ecr_put_replication_configuration]{put_replication_configuration}} API
action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_registry()
}\if{html}{\out{</div>}}
}

\keyword{internal}
