% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_get_account_setting}
\alias{ecr_get_account_setting}
\title{Retrieves the account setting value for the specified setting name}
\usage{
ecr_get_account_setting(name)
}
\arguments{
\item{name}{[required] The name of the account setting, such as \code{BASIC_SCAN_TYPE_VERSION} or
\code{REGISTRY_POLICY_SCOPE}.}
}
\description{
Retrieves the account setting value for the specified setting name.

See \url{https://www.paws-r-sdk.com/docs/ecr_get_account_setting/} for full documentation.
}
\keyword{internal}
