% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_get_registry_policy}
\alias{ecr_get_registry_policy}
\title{Retrieves the permissions policy for a registry}
\usage{
ecr_get_registry_policy()
}
\value{
A list with the following syntax:\preformatted{list(
  registryId = "string",
  policyText = "string"
)
}
}
\description{
Retrieves the permissions policy for a registry.
}
\section{Request syntax}{
\preformatted{svc$get_registry_policy()
}
}

\keyword{internal}
