% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_get_registry_scanning_configuration}
\alias{ecr_get_registry_scanning_configuration}
\title{Retrieves the scanning configuration for a registry}
\usage{
ecr_get_registry_scanning_configuration()
}
\description{
Retrieves the scanning configuration for a registry.

See \url{https://www.paws-r-sdk.com/docs/ecr_get_registry_scanning_configuration/} for full documentation.
}
\keyword{internal}
