% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_get_repository_policy}
\alias{ecr_get_repository_policy}
\title{Retrieves the repository policy for the specified repository}
\usage{
ecr_get_repository_policy(registryId, repositoryName)
}
\arguments{
\item{registryId}{The AWS account ID associated with the registry that contains the
repository. If you do not specify a registry, the default registry is
assumed.}

\item{repositoryName}{[required] The name of the repository with the policy to retrieve.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  registryId = "string",
  repositoryName = "string",
  policyText = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the repository policy for the specified repository.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_repository_policy(
  registryId = "string",
  repositoryName = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# This example obtains the repository policy for the repository named
# ubuntu.
svc$get_repository_policy(
  repositoryName = "ubuntu"
)
}

}
\keyword{internal}
