% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_get_signing_configuration}
\alias{ecr_get_signing_configuration}
\title{Retrieves the registry's signing configuration, which defines rules for
automatically signing images using Amazon Web Services Signer}
\usage{
ecr_get_signing_configuration()
}
\description{
Retrieves the registry's signing configuration, which defines rules for automatically signing images using Amazon Web Services Signer.

See \url{https://www.paws-r-sdk.com/docs/ecr_get_signing_configuration/} for full documentation.
}
\keyword{internal}
