% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_list_image_referrers}
\alias{ecr_list_image_referrers}
\title{Lists the artifacts associated with a specified subject image}
\usage{
ecr_list_image_referrers(
  registryId = NULL,
  repositoryName,
  subjectId,
  filter = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID associated with the registry that
contains the repository in which to list image referrers. If you do not
specify a registry, the default registry is assumed.}

\item{repositoryName}{[required] The name of the repository that contains the subject image.}

\item{subjectId}{[required] An object containing the image digest of the subject image for which to
retrieve associated artifacts.}

\item{filter}{The filter key and value with which to filter your
\code{\link[=ecr_list_image_referrers]{list_image_referrers}} results. If no filter
is specified, only artifacts with \code{ACTIVE} status are returned.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{\link[=ecr_list_image_referrers]{list_image_referrers}} request where
\code{maxResults} was used and the results exceeded the value of that
parameter. Pagination continues from the end of the previous results
that returned the \code{nextToken} value. This value is \code{null} when there are
no more results to return.

This token should be treated as an opaque identifier that is only used
to retrieve the next items in a list and not for other programmatic
purposes.}

\item{maxResults}{The maximum number of image referrer results returned by
\code{\link[=ecr_list_image_referrers]{list_image_referrers}} in paginated output.
When this parameter is used,
\code{\link[=ecr_list_image_referrers]{list_image_referrers}} only returns
\code{maxResults} results in a single page along with a \code{nextToken} response
element. The remaining results of the initial request can be seen by
sending another \code{\link[=ecr_list_image_referrers]{list_image_referrers}}
request with the returned \code{nextToken} value. This value can be between 1
and 50. If this parameter is not used, then
\code{\link[=ecr_list_image_referrers]{list_image_referrers}} returns up to 50
results and a \code{nextToken} value, if applicable.}
}
\description{
Lists the artifacts associated with a specified subject image.

See \url{https://www.paws-r-sdk.com/docs/ecr_list_image_referrers/} for full documentation.
}
\keyword{internal}
