% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_list_pull_time_update_exclusions}
\alias{ecr_list_pull_time_update_exclusions}
\title{Lists the IAM principals that are excluded from having their image pull
times recorded}
\usage{
ecr_list_pull_time_update_exclusions(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of pull time update exclusion results returned by
\code{\link[=ecr_list_pull_time_update_exclusions]{list_pull_time_update_exclusions}}
in paginated output. When this parameter is used,
\code{\link[=ecr_list_pull_time_update_exclusions]{list_pull_time_update_exclusions}}
only returns \code{maxResults} results in a single page along with a
\code{nextToken} response element. The remaining results of the initial
request can be seen by sending another
\code{\link[=ecr_list_pull_time_update_exclusions]{list_pull_time_update_exclusions}}
request with the returned \code{nextToken} value. This value can be between 1
and 1000. If this parameter is not used, then
\code{\link[=ecr_list_pull_time_update_exclusions]{list_pull_time_update_exclusions}}
returns up to 100 results and a \code{nextToken} value, if applicable.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{\link[=ecr_list_pull_time_update_exclusions]{list_pull_time_update_exclusions}}
request where \code{maxResults} was used and the results exceeded the value
of that parameter. Pagination continues from the end of the previous
results that returned the \code{nextToken} value. This value is \code{null} when
there are no more results to return.

This token should be treated as an opaque identifier that is only used
to retrieve the next items in a list and not for other programmatic
purposes.}
}
\description{
Lists the IAM principals that are excluded from having their image pull times recorded.

See \url{https://www.paws-r-sdk.com/docs/ecr_list_pull_time_update_exclusions/} for full documentation.
}
\keyword{internal}
