% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_put_account_setting}
\alias{ecr_put_account_setting}
\title{Allows you to change the basic scan type version or registry policy
scope}
\usage{
ecr_put_account_setting(name, value)
}
\arguments{
\item{name}{[required] The name of the account setting, such as \code{BASIC_SCAN_TYPE_VERSION},
\code{REGISTRY_POLICY_SCOPE}, or \code{BLOB_MOUNTING}.}

\item{value}{[required] Setting value that is specified. Valid value for basic scan type:
\code{AWS_NATIVE}. Valid values for registry policy scope: \code{V2}. Valid values
for blob mounting: \code{ENABLED} or \code{DISABLED}.}
}
\description{
Allows you to change the basic scan type version or registry policy scope.

See \url{https://www.paws-r-sdk.com/docs/ecr_put_account_setting/} for full documentation.
}
\keyword{internal}
