% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_put_registry_policy}
\alias{ecr_put_registry_policy}
\title{Creates or updates the permissions policy for your registry}
\usage{
ecr_put_registry_policy(policyText)
}
\arguments{
\item{policyText}{[required] The JSON policy text to apply to your registry. The policy text follows
the same format as IAM policy text. For more information, see \href{https://docs.aws.amazon.com/AmazonECR/latest/userguide/registry-permissions.html}{Registry permissions}
in the \emph{Amazon Elastic Container Registry User Guide}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  registryId = "string",
  policyText = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates or updates the permissions policy for your registry.

A registry policy is used to specify permissions for another AWS account
and is used when configuring cross-account replication. For more
information, see \href{https://docs.aws.amazon.com/AmazonECR/latest/userguide/registry-permissions.html}{Registry permissions}
in the \emph{Amazon Elastic Container Registry User Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_registry_policy(
  policyText = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
