% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_put_signing_configuration}
\alias{ecr_put_signing_configuration}
\title{Creates or updates the registry's signing configuration, which defines
rules for automatically signing images with Amazon Web Services Signer}
\usage{
ecr_put_signing_configuration(signingConfiguration)
}
\arguments{
\item{signingConfiguration}{[required] The signing configuration to assign to the registry.}
}
\description{
Creates or updates the registry's signing configuration, which defines rules for automatically signing images with Amazon Web Services Signer.

See \url{https://www.paws-r-sdk.com/docs/ecr_put_signing_configuration/} for full documentation.
}
\keyword{internal}
