% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_register_pull_time_update_exclusion}
\alias{ecr_register_pull_time_update_exclusion}
\title{Adds an IAM principal to the pull time update exclusion list for a
registry}
\usage{
ecr_register_pull_time_update_exclusion(principalArn)
}
\arguments{
\item{principalArn}{[required] The ARN of the IAM principal to exclude from having image pull times
recorded.}
}
\description{
Adds an IAM principal to the pull time update exclusion list for a registry. Amazon ECR will not record the pull time if an excluded principal pulls an image.

See \url{https://www.paws-r-sdk.com/docs/ecr_register_pull_time_update_exclusion/} for full documentation.
}
\keyword{internal}
