% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_update_image_storage_class}
\alias{ecr_update_image_storage_class}
\title{Transitions an image between storage classes}
\usage{
ecr_update_image_storage_class(
  registryId = NULL,
  repositoryName,
  imageId,
  targetStorageClass
)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID associated with the registry that
contains the image to transition. If you do not specify a registry, the
default registry is assumed.}

\item{repositoryName}{[required] The name of the repository that contains the image to transition.}

\item{imageId}{[required]}

\item{targetStorageClass}{[required] The target storage class for the image.}
}
\description{
Transitions an image between storage classes. You can transition images from Amazon ECR standard storage class to Amazon ECR archival storage class for long-term storage, or restore archived images back to Amazon ECR standard.

See \url{https://www.paws-r-sdk.com/docs/ecr_update_image_storage_class/} for full documentation.
}
\keyword{internal}
