% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_validate_pull_through_cache_rule}
\alias{ecr_validate_pull_through_cache_rule}
\title{Validates an existing pull through cache rule for an upstream registry
that requires authentication}
\usage{
ecr_validate_pull_through_cache_rule(ecrRepositoryPrefix, registryId = NULL)
}
\arguments{
\item{ecrRepositoryPrefix}{[required] The repository name prefix associated with the pull through cache rule.}

\item{registryId}{The registry ID associated with the pull through cache rule. If you do
not specify a registry, the default registry is assumed.}
}
\description{
Validates an existing pull through cache rule for an upstream registry that requires authentication. This will retrieve the contents of the Amazon Web Services Secrets Manager secret, verify the syntax, and then validate that authentication to the upstream registry is successful.

See \url{https://www.paws-r-sdk.com/docs/ecr_validate_pull_through_cache_rule/} for full documentation.
}
\keyword{internal}
