% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_complete_layer_upload}
\alias{ecrpublic_complete_layer_upload}
\title{Informs Amazon ECR that the image layer upload is complete for a
specified public registry, repository name, and upload ID}
\usage{
ecrpublic_complete_layer_upload(
  registryId = NULL,
  repositoryName,
  uploadId,
  layerDigests
)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID, or registry alias, associated with
the registry where layers are uploaded. If you do not specify a
registry, the default public registry is assumed.}

\item{repositoryName}{[required] The name of the repository in a public registry to associate with the
image layer.}

\item{uploadId}{[required] The upload ID from a previous
\code{\link[=ecrpublic_initiate_layer_upload]{initiate_layer_upload}} operation to
associate with the image layer.}

\item{layerDigests}{[required] The \code{sha256} digest of the image layer.}
}
\description{
Informs Amazon ECR that the image layer upload is complete for a specified public registry, repository name, and upload ID. You can optionally provide a \code{sha256} digest of the image layer for data validation purposes.

See \url{https://www.paws-r-sdk.com/docs/ecrpublic_complete_layer_upload/} for full documentation.
}
\keyword{internal}
