% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_delete_repository}
\alias{ecrpublic_delete_repository}
\title{Deletes a repository in a public registry}
\usage{
ecrpublic_delete_repository(registryId = NULL, repositoryName, force = NULL)
}
\arguments{
\item{registryId}{The AWS account ID associated with the public registry that contains the
repository to delete. If you do not specify a registry, the default
public registry is assumed.}

\item{repositoryName}{[required] The name of the repository to delete.}

\item{force}{If a repository contains images, forces the deletion.}
}
\description{
Deletes a repository in a public registry. If the repository contains images, you must either delete all images in the repository or use the \code{force} option which deletes all images on your behalf before deleting the repository.

See \url{https://paws-r.github.io/docs/ecrpublic/delete_repository.html} for full documentation.
}
\keyword{internal}
