% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_describe_images}
\alias{ecrpublic_describe_images}
\title{Returns metadata that's related to the images in a repository in a
public registry}
\usage{
ecrpublic_describe_images(
  registryId = NULL,
  repositoryName,
  imageIds = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID that's associated with the public
registry that contains the repository where images are described. If you
do not specify a registry, the default public registry is assumed.}

\item{repositoryName}{[required] The repository that contains the images to describe.}

\item{imageIds}{The list of image IDs for the requested repository.}

\item{nextToken}{The \code{nextToken} value that's returned from a previous paginated
\code{\link[=ecrpublic_describe_images]{describe_images}} request where
\code{maxResults} was used and the results exceeded the value of that
parameter. Pagination continues from the end of the previous results
that returned the \code{nextToken} value. If there are no more results to
return, this value is \code{null}. If you specify images with \code{imageIds}, you
can't use this option.}

\item{maxResults}{The maximum number of repository results that's returned by
\code{\link[=ecrpublic_describe_images]{describe_images}} in paginated output. When
this parameter is used, \code{\link[=ecrpublic_describe_images]{describe_images}}
only returns \code{maxResults} results in a single page along with a
\code{nextToken} response element. You can see the remaining results of the
initial request by sending another
\code{\link[=ecrpublic_describe_images]{describe_images}} request with the returned
\code{nextToken} value. This value can be between 1 and 1000. If this
parameter isn't used, then
\code{\link[=ecrpublic_describe_images]{describe_images}} returns up to 100 results
and a \code{nextToken} value, if applicable. If you specify images with
\code{imageIds}, you can't use this option.}
}
\description{
Returns metadata that's related to the images in a repository in a public registry.

See \url{https://www.paws-r-sdk.com/docs/ecrpublic_describe_images/} for full documentation.
}
\keyword{internal}
