% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_get_registry_catalog_data}
\alias{ecrpublic_get_registry_catalog_data}
\title{Retrieves catalog metadata for a public registry}
\usage{
ecrpublic_get_registry_catalog_data()
}
\description{
Retrieves catalog metadata for a public registry.

See \url{https://www.paws-r-sdk.com/docs/ecrpublic_get_registry_catalog_data/} for full documentation.
}
\keyword{internal}
