% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_get_repository_catalog_data}
\alias{ecrpublic_get_repository_catalog_data}
\title{Retrieve catalog metadata for a repository in a public registry}
\usage{
ecrpublic_get_repository_catalog_data(registryId = NULL, repositoryName)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID that's associated with the registry
that contains the repositories to be described. If you do not specify a
registry, the default public registry is assumed.}

\item{repositoryName}{[required] The name of the repository to retrieve the catalog metadata for.}
}
\description{
Retrieve catalog metadata for a repository in a public registry. This metadata is displayed publicly in the Amazon ECR Public Gallery.

See \url{https://www.paws-r-sdk.com/docs/ecrpublic_get_repository_catalog_data/} for full documentation.
}
\keyword{internal}
