% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_get_repository_policy}
\alias{ecrpublic_get_repository_policy}
\title{Retrieves the repository policy for the specified repository}
\usage{
ecrpublic_get_repository_policy(registryId = NULL, repositoryName)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID that's associated with the public
registry that contains the repository. If you do not specify a registry,
the default public registry is assumed.}

\item{repositoryName}{[required] The name of the repository with the policy to retrieve.}
}
\description{
Retrieves the repository policy for the specified repository.

See \url{https://www.paws-r-sdk.com/docs/ecrpublic_get_repository_policy/} for full documentation.
}
\keyword{internal}
