% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_initiate_layer_upload}
\alias{ecrpublic_initiate_layer_upload}
\title{Notifies Amazon ECR that you intend to upload an image layer}
\usage{
ecrpublic_initiate_layer_upload(registryId = NULL, repositoryName)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID, or registry alias, that's associated
with the registry to which you intend to upload layers. If you do not
specify a registry, the default public registry is assumed.}

\item{repositoryName}{[required] The name of the repository that you want to upload layers to.}
}
\description{
Notifies Amazon ECR that you intend to upload an image layer.

See \url{https://www.paws-r-sdk.com/docs/ecrpublic_initiate_layer_upload/} for full documentation.
}
\keyword{internal}
