% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_put_registry_catalog_data}
\alias{ecrpublic_put_registry_catalog_data}
\title{Create or update the catalog data for a public registry}
\usage{
ecrpublic_put_registry_catalog_data(displayName = NULL)
}
\arguments{
\item{displayName}{The display name for a public registry. The display name is shown as the
repository author in the Amazon ECR Public Gallery.

The registry display name is only publicly visible in the Amazon ECR
Public Gallery for verified accounts.}
}
\description{
Create or update the catalog data for a public registry.

See \url{https://www.paws-r-sdk.com/docs/ecrpublic_put_registry_catalog_data/} for full documentation.
}
\keyword{internal}
