% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_put_repository_catalog_data}
\alias{ecrpublic_put_repository_catalog_data}
\title{Creates or updates the catalog data for a repository in a public
registry}
\usage{
ecrpublic_put_repository_catalog_data(
  registryId = NULL,
  repositoryName,
  catalogData
)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID that's associated with the public
registry the repository is in. If you do not specify a registry, the
default public registry is assumed.}

\item{repositoryName}{[required] The name of the repository to create or update the catalog data for.}

\item{catalogData}{[required] An object containing the catalog data for a repository. This data is
publicly visible in the Amazon ECR Public Gallery.}
}
\description{
Creates or updates the catalog data for a repository in a public registry.

See \url{https://www.paws-r-sdk.com/docs/ecrpublic_put_repository_catalog_data/} for full documentation.
}
\keyword{internal}
