% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_continue_service_deployment}
\alias{ecs_continue_service_deployment}
\title{Continues or rolls back an Amazon ECS service deployment that is paused
at a lifecycle hook}
\usage{
ecs_continue_service_deployment(serviceDeploymentArn, hookId, action = NULL)
}
\arguments{
\item{serviceDeploymentArn}{[required] The ARN of the service deployment to continue or roll back.}

\item{hookId}{[required] The ID of the paused lifecycle hook to act on. You can find the \code{hookId} by calling \code{\link[=ecs_describe_service_deployments]{describe_service_deployments}} and inspecting the \code{lifecycleHookDetails} field of the service deployment.}

\item{action}{The action to take on the paused lifecycle hook. Valid values are:
\itemize{
\item \code{CONTINUE} - Proceeds the deployment to the next lifecycle stage.
\item \code{ROLLBACK} - Rolls back the deployment to the previous service revision.
}

If no value is specified, the default action is \code{CONTINUE}.}
}
\description{
Continues or rolls back an Amazon ECS service deployment that is paused at a lifecycle hook.

See \url{https://www.paws-r-sdk.com/docs/ecs_continue_service_deployment/} for full documentation.
}
\keyword{internal}
