% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_delete_express_gateway_service}
\alias{ecs_delete_express_gateway_service}
\title{Deletes an Express service and removes all associated Amazon Web
Services resources}
\usage{
ecs_delete_express_gateway_service(serviceArn)
}
\arguments{
\item{serviceArn}{[required] The Amazon Resource Name (ARN) of the Express service to delete. The ARN
uniquely identifies the service within your Amazon Web Services account
and region.}
}
\description{
Deletes an Express service and removes all associated Amazon Web Services resources. This operation stops service tasks, removes the Application Load Balancer, target groups, security groups, auto-scaling policies, and other managed infrastructure components.

See \url{https://www.paws-r-sdk.com/docs/ecs_delete_express_gateway_service/} for full documentation.
}
\keyword{internal}
