% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_delete_task_definitions}
\alias{ecs_delete_task_definitions}
\title{Deletes one or more task definitions}
\usage{
ecs_delete_task_definitions(taskDefinitions)
}
\arguments{
\item{taskDefinitions}{[required] The \code{family} and \code{revision} (\code{family:revision}) or full Amazon Resource
Name (ARN) of the task definition to delete. You must specify a
\code{revision}.

You can specify up to 10 task definitions as a comma separated list.}
}
\description{
Deletes one or more task definitions.

See \url{https://www.paws-r-sdk.com/docs/ecs_delete_task_definitions/} for full documentation.
}
\keyword{internal}
