% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_describe_daemon_deployments}
\alias{ecs_describe_daemon_deployments}
\title{Describes one or more of your daemon deployments}
\usage{
ecs_describe_daemon_deployments(daemonDeploymentArns)
}
\arguments{
\item{daemonDeploymentArns}{[required] The ARN of the daemon deployments to describe. You can specify up to 20
ARNs.}
}
\description{
Describes one or more of your daemon deployments.

See \url{https://www.paws-r-sdk.com/docs/ecs_describe_daemon_deployments/} for full documentation.
}
\keyword{internal}
