% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_describe_daemon_revisions}
\alias{ecs_describe_daemon_revisions}
\title{Describes one or more of your daemon revisions}
\usage{
ecs_describe_daemon_revisions(daemonRevisionArns)
}
\arguments{
\item{daemonRevisionArns}{[required] The ARN of the daemon revisions to describe. You can specify up to 20
ARNs.}
}
\description{
Describes one or more of your daemon revisions.

See \url{https://www.paws-r-sdk.com/docs/ecs_describe_daemon_revisions/} for full documentation.
}
\keyword{internal}
