% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_describe_express_gateway_service}
\alias{ecs_describe_express_gateway_service}
\title{Retrieves detailed information about an Express service, including
current status, configuration, managed infrastructure, and service
revisions}
\usage{
ecs_describe_express_gateway_service(serviceArn, include = NULL)
}
\arguments{
\item{serviceArn}{[required] The Amazon Resource Name (ARN) of the Express service to describe. The
ARN uniquely identifies the service within your Amazon Web Services
account and region.}

\item{include}{Specifies additional information to include in the response. Valid
values are \code{TAGS} to include resource tags associated with the Express
service.}
}
\description{
Retrieves detailed information about an Express service, including current status, configuration, managed infrastructure, and service revisions.

See \url{https://www.paws-r-sdk.com/docs/ecs_describe_express_gateway_service/} for full documentation.
}
\keyword{internal}
