% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_describe_service_deployments}
\alias{ecs_describe_service_deployments}
\title{Describes one or more of your service deployments}
\usage{
ecs_describe_service_deployments(serviceDeploymentArns)
}
\arguments{
\item{serviceDeploymentArns}{[required] The ARN of the service deployment.

You can specify a maximum of 20 ARNs.}
}
\description{
Describes one or more of your service deployments.

See \url{https://www.paws-r-sdk.com/docs/ecs_describe_service_deployments/} for full documentation.
}
\keyword{internal}
