% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_describe_service_revisions}
\alias{ecs_describe_service_revisions}
\title{Describes one or more service revisions}
\usage{
ecs_describe_service_revisions(serviceRevisionArns)
}
\arguments{
\item{serviceRevisionArns}{[required] The ARN of the service revision.

You can specify a maximum of 20 ARNs.

You can call \code{\link[=ecs_list_service_deployments]{list_service_deployments}}
to get the ARNs.}
}
\description{
Describes one or more service revisions.

See \url{https://www.paws-r-sdk.com/docs/ecs_describe_service_revisions/} for full documentation.
}
\keyword{internal}
