% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_get_task_protection}
\alias{ecs_get_task_protection}
\title{Retrieves the protection status of tasks in an Amazon ECS service}
\usage{
ecs_get_task_protection(cluster, tasks = NULL)
}
\arguments{
\item{cluster}{[required] The short name or full Amazon Resource Name (ARN) of the cluster that
hosts the service that the task sets exist in.}

\item{tasks}{A list of up to 100 task IDs or full ARN entries.}
}
\description{
Retrieves the protection status of tasks in an Amazon ECS service.

See \url{https://www.paws-r-sdk.com/docs/ecs_get_task_protection/} for full documentation.
}
\keyword{internal}
