% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_list_daemon_deployments}
\alias{ecs_list_daemon_deployments}
\title{Returns a list of daemon deployments for a specified daemon}
\usage{
ecs_list_daemon_deployments(
  daemonArn,
  status = NULL,
  createdAt = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{daemonArn}{[required] The Amazon Resource Name (ARN) of the daemon to list deployments for.}

\item{status}{An optional filter to narrow the
\code{\link[=ecs_list_daemon_deployments]{list_daemon_deployments}} results by
deployment status. If you don't specify a status, all deployments are
returned.}

\item{createdAt}{An optional filter to narrow the
\code{\link[=ecs_list_daemon_deployments]{list_daemon_deployments}} results by
creation time. If you don't specify a time range, all deployments are
returned.}

\item{maxResults}{The maximum number of daemon deployment results that
\code{\link[=ecs_list_daemon_deployments]{list_daemon_deployments}} returned in
paginated output. When this parameter is used,
\code{\link[=ecs_list_daemon_deployments]{list_daemon_deployments}} only returns
\code{maxResults} results in a single page along with a \code{nextToken} response
element. The remaining results of the initial request can be seen by
sending another \code{\link[=ecs_list_daemon_deployments]{list_daemon_deployments}}
request with the returned \code{nextToken} value. This value can be between 1
and 100. If this parameter isn't used, then
\code{\link[=ecs_list_daemon_deployments]{list_daemon_deployments}} returns up to
20 results and a \code{nextToken} value if applicable.}

\item{nextToken}{The \code{nextToken} value returned from a
\code{\link[=ecs_list_daemon_deployments]{list_daemon_deployments}} request
indicating that more results are available to fulfill the request and
further calls will be needed. If \code{maxResults} was provided, it's
possible for the number of results to be fewer than \code{maxResults}.

This token should be treated as an opaque identifier that is only used
to retrieve the next items in a list and not for other programmatic
purposes.}
}
\description{
Returns a list of daemon deployments for a specified daemon. You can filter the results by status or creation time.

See \url{https://www.paws-r-sdk.com/docs/ecs_list_daemon_deployments/} for full documentation.
}
\keyword{internal}
