% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_stop_service_deployment}
\alias{ecs_stop_service_deployment}
\title{Stops an ongoing service deployment}
\usage{
ecs_stop_service_deployment(serviceDeploymentArn, stopType = NULL)
}
\arguments{
\item{serviceDeploymentArn}{[required] The ARN of the service deployment that you want to stop.}

\item{stopType}{How you want Amazon ECS to stop the service.

The valid values are \code{ROLLBACK}.}
}
\description{
Stops an ongoing service deployment.

See \url{https://www.paws-r-sdk.com/docs/ecs_stop_service_deployment/} for full documentation.
}
\keyword{internal}
