% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_submit_attachment_state_changes}
\alias{ecs_submit_attachment_state_changes}
\title{This action is only used by the Amazon ECS agent, and it is not intended
for use outside of the agent}
\usage{
ecs_submit_attachment_state_changes(cluster, attachments)
}
\arguments{
\item{cluster}{The short name or full ARN of the cluster that hosts the container
instance the attachment belongs to.}

\item{attachments}{[required] Any attachments associated with the state change request.}
}
\value{
A list with the following syntax:\preformatted{list(
  acknowledgment = "string"
)
}
}
\description{
This action is only used by the Amazon ECS agent, and it is not intended
for use outside of the agent.

Sent to acknowledge that an attachment changed states.
}
\section{Request syntax}{
\preformatted{svc$submit_attachment_state_changes(
  cluster = "string",
  attachments = list(
    list(
      attachmentArn = "string",
      status = "string"
    )
  )
)
}
}

\keyword{internal}
