% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_submit_task_state_change}
\alias{ecs_submit_task_state_change}
\title{This action is only used by the Amazon ECS agent, and it is not intended
for use outside of the agent}
\usage{
ecs_submit_task_state_change(
  cluster = NULL,
  task = NULL,
  status = NULL,
  reason = NULL,
  containers = NULL,
  attachments = NULL,
  managedAgents = NULL,
  pullStartedAt = NULL,
  pullStoppedAt = NULL,
  executionStoppedAt = NULL
)
}
\arguments{
\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster that
hosts the task.}

\item{task}{The task ID or full ARN of the task in the state change request.}

\item{status}{The status of the state change request.}

\item{reason}{The reason for the state change request.}

\item{containers}{Any containers that's associated with the state change request.}

\item{attachments}{Any attachments associated with the state change request.}

\item{managedAgents}{The details for the managed agent that's associated with the task.}

\item{pullStartedAt}{The Unix timestamp for the time when the container image pull started.}

\item{pullStoppedAt}{The Unix timestamp for the time when the container image pull completed.}

\item{executionStoppedAt}{The Unix timestamp for the time when the task execution stopped.}
}
\description{
This action is only used by the Amazon ECS agent, and it is not intended for use outside of the agent.

See \url{https://www.paws-r-sdk.com/docs/ecs_submit_task_state_change/} for full documentation.
}
\keyword{internal}
