% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_update_cluster}
\alias{ecs_update_cluster}
\title{Updates the cluster}
\usage{
ecs_update_cluster(cluster, settings = NULL, configuration = NULL)
}
\arguments{
\item{cluster}{[required] The name of the cluster to modify the settings for.}

\item{settings}{The cluster settings for your cluster.}

\item{configuration}{The execute command configuration for the cluster.}
}
\description{
Updates the cluster.

See \url{https://paws-r.github.io/docs/ecs/update_cluster.html} for full documentation.
}
\keyword{internal}
