% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_update_cluster_settings}
\alias{ecs_update_cluster_settings}
\title{Modifies the settings to use for a cluster}
\usage{
ecs_update_cluster_settings(cluster, settings)
}
\arguments{
\item{cluster}{[required] The name of the cluster to modify the settings for.}

\item{settings}{[required] The setting to use by default for a cluster. This parameter is used to
enable CloudWatch Container Insights for a cluster. If this value is
specified, it will override the \code{containerInsights} value set with
PutAccountSetting or PutAccountSettingDefault.}
}
\description{
Modifies the settings to use for a cluster.
}
\section{Request syntax}{
\preformatted{svc$update_cluster_settings(
  cluster = "string",
  settings = list(
    list(
      name = "containerInsights",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
