% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_activate_certificate_authority}
\alias{eks_activate_certificate_authority}
\title{Activates a successor certificate authority (CA) as the signing
certificate authority for your cluster, completing a CA rotation}
\usage{
eks_activate_certificate_authority(
  clusterName,
  certificateAuthorityId,
  clientRequestToken = NULL
)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{certificateAuthorityId}{[required] The ID of the certificate authority to activate as the cluster's signing certificate authority. This certificate authority must already exist on the cluster and have a \code{distributionStatus} of \code{COMPLETE}.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Activates a successor certificate authority (CA) as the signing certificate authority for your cluster, completing a CA rotation.

See \url{https://www.paws-r-sdk.com/docs/eks_activate_certificate_authority/} for full documentation.
}
\keyword{internal}
