% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_associate_identity_provider_config}
\alias{eks_associate_identity_provider_config}
\title{Associate an identity provider configuration to a cluster}
\usage{
eks_associate_identity_provider_config(clusterName, oidc, tags,
  clientRequestToken)
}
\arguments{
\item{clusterName}{[required] The name of the cluster to associate the configuration to.}

\item{oidc}{[required] An object that represents an OpenID Connect (OIDC) identity provider
configuration.}

\item{tags}{The metadata to apply to the configuration to assist with categorization
and organization. Each tag consists of a key and an optional value, both
of which you define.}

\item{clientRequestToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\value{
A list with the following syntax:\preformatted{list(
  update = list(
    id = "string",
    status = "InProgress"|"Failed"|"Cancelled"|"Successful",
    type = "VersionUpdate"|"EndpointAccessUpdate"|"LoggingUpdate"|"ConfigUpdate"|"AssociateIdentityProviderConfig"|"DisassociateIdentityProviderConfig"|"AssociateEncryptionConfig"|"AddonUpdate",
    params = list(
      list(
        type = "Version"|"PlatformVersion"|"EndpointPrivateAccess"|"EndpointPublicAccess"|"ClusterLogging"|"DesiredSize"|"LabelsToAdd"|"LabelsToRemove"|"TaintsToAdd"|"TaintsToRemove"|"MaxSize"|"MinSize"|"ReleaseVersion"|"PublicAccessCidrs"|"LaunchTemplateName"|"LaunchTemplateVersion"|"IdentityProviderConfig"|"EncryptionConfig"|"AddonVersion"|"ServiceAccountRoleArn"|"ResolveConflicts",
        value = "string"
      )
    ),
    createdAt = as.POSIXct(
      "2015-01-01"
    ),
    errors = list(
      list(
        errorCode = "SubnetNotFound"|"SecurityGroupNotFound"|"EniLimitReached"|"IpNotAvailable"|"AccessDenied"|"OperationNotPermitted"|"VpcIdNotFound"|"Unknown"|"NodeCreationFailure"|"PodEvictionFailure"|"InsufficientFreeAddresses"|"ClusterUnreachable"|"InsufficientNumberOfReplicas"|"ConfigurationConflict"|"AdmissionRequestDenied",
        errorMessage = "string",
        resourceIds = list(
          "string"
        )
      )
    )
  ),
  tags = list(
    "string"
  )
)
}
}
\description{
Associate an identity provider configuration to a cluster.

If you want to authenticate identities using an identity provider, you
can create an identity provider configuration and associate it to your
cluster. After configuring authentication to your cluster you can create
Kubernetes \code{roles} and \code{clusterroles} to assign permissions to the
roles, and then bind the roles to the identities using Kubernetes
\code{rolebindings} and \code{clusterrolebindings}. For more information see
\href{https://kubernetes.io/docs/reference/access-authn-authz/rbac/}{Using RBAC Authorization}
in the Kubernetes documentation.
}
\section{Request syntax}{
\preformatted{svc$associate_identity_provider_config(
  clusterName = "string",
  oidc = list(
    identityProviderConfigName = "string",
    issuerUrl = "string",
    clientId = "string",
    usernameClaim = "string",
    usernamePrefix = "string",
    groupsClaim = "string",
    groupsPrefix = "string",
    requiredClaims = list(
      "string"
    )
  ),
  tags = list(
    "string"
  ),
  clientRequestToken = "string"
)
}
}

\keyword{internal}
