% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_cancel_update}
\alias{eks_cancel_update}
\title{Cancels an in-progress update to an Amazon EKS cluster on a best-effort
basis}
\usage{
eks_cancel_update(name, updateId, clientRequestToken = NULL)
}
\arguments{
\item{name}{[required] The name of the Amazon EKS cluster associated with the update.}

\item{updateId}{[required] The ID of the update to cancel.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Cancels an in-progress update to an Amazon EKS cluster on a best-effort basis. Cancellation is only performed if the update can be cancelled. Currently, this is supported for \code{VersionRollback} update types on EKS Auto Mode clusters when nodes are rolling back.

See \url{https://www.paws-r-sdk.com/docs/eks_cancel_update/} for full documentation.
}
\keyword{internal}
