% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_create_certificate_authority}
\alias{eks_create_certificate_authority}
\title{Appends a successor certificate authority (CA) to your cluster,
beginning the CA rotation process}
\usage{
eks_create_certificate_authority(clusterName, clientRequestToken = NULL)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Appends a successor certificate authority (CA) to your cluster, beginning the CA rotation process.

See \url{https://www.paws-r-sdk.com/docs/eks_create_certificate_authority/} for full documentation.
}
\keyword{internal}
