% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_delete_capability}
\alias{eks_delete_capability}
\title{Deletes a managed capability from your Amazon EKS cluster}
\usage{
eks_delete_capability(clusterName, capabilityName)
}
\arguments{
\item{clusterName}{[required] The name of the Amazon EKS cluster that contains the capability you want
to delete.}

\item{capabilityName}{[required] The name of the capability to delete.}
}
\description{
Deletes a managed capability from your Amazon EKS cluster. When you delete a capability, Amazon EKS removes the capability infrastructure but retains all resources that were managed by the capability.

See \url{https://www.paws-r-sdk.com/docs/eks_delete_capability/} for full documentation.
}
\keyword{internal}
