% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_delete_certificate_authority}
\alias{eks_delete_certificate_authority}
\title{Deletes a certificate authority (CA) from your cluster}
\usage{
eks_delete_certificate_authority(
  clusterName,
  certificateAuthorityId,
  clientRequestToken = NULL
)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{certificateAuthorityId}{[required] The ID of the certificate authority to delete. You can't delete the certificate authority that's currently signing certificates for the cluster.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Deletes a certificate authority (CA) from your cluster.

See \url{https://www.paws-r-sdk.com/docs/eks_delete_certificate_authority/} for full documentation.
}
\keyword{internal}
