% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_delete_pod_identity_association}
\alias{eks_delete_pod_identity_association}
\title{Deletes a EKS Pod Identity association}
\usage{
eks_delete_pod_identity_association(clusterName, associationId)
}
\arguments{
\item{clusterName}{[required] The cluster name that}

\item{associationId}{[required] The ID of the association to be deleted.}
}
\description{
Deletes a EKS Pod Identity association.

See \url{https://www.paws-r-sdk.com/docs/eks_delete_pod_identity_association/} for full documentation.
}
\keyword{internal}
