% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_deregister_cluster}
\alias{eks_deregister_cluster}
\title{Deregisters a connected cluster to remove it from the Amazon EKS control
plane}
\usage{
eks_deregister_cluster(name)
}
\arguments{
\item{name}{[required] The name of the connected cluster to deregister.}
}
\description{
Deregisters a connected cluster to remove it from the Amazon EKS control plane.

See \url{https://www.paws-r-sdk.com/docs/eks_deregister_cluster/} for full documentation.
}
\keyword{internal}
