% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_access_entry}
\alias{eks_describe_access_entry}
\title{Describes an access entry}
\usage{
eks_describe_access_entry(clusterName, principalArn)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{principalArn}{[required] The ARN of the IAM principal for the \code{AccessEntry}.}
}
\description{
Describes an access entry.

See \url{https://www.paws-r-sdk.com/docs/eks_describe_access_entry/} for full documentation.
}
\keyword{internal}
