% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_addon_configuration}
\alias{eks_describe_addon_configuration}
\title{Returns configuration options}
\usage{
eks_describe_addon_configuration(addonName, addonVersion)
}
\arguments{
\item{addonName}{[required] The name of the add-on. The name must match one of the names that
\href{https://docs.aws.amazon.com/eks/latest/APIReference/API_DescribeAddonVersions.html}{\code{describe_addon_versions}}
returns.}

\item{addonVersion}{[required] The version of the add-on. The version must match one of the versions
returned by
\href{https://docs.aws.amazon.com/eks/latest/APIReference/API_DescribeAddonVersions.html}{\code{describe_addon_versions}}
.}
}
\description{
Returns configuration options.

See \url{https://paws-r.github.io/docs/eks/describe_addon_configuration.html} for full documentation.
}
\keyword{internal}
