% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_capability}
\alias{eks_describe_capability}
\title{Returns detailed information about a specific managed capability in your
Amazon EKS cluster, including its current status, configuration, health
information, and any issues that may be affecting its operation}
\usage{
eks_describe_capability(clusterName, capabilityName)
}
\arguments{
\item{clusterName}{[required] The name of the Amazon EKS cluster that contains the capability you want
to describe.}

\item{capabilityName}{[required] The name of the capability to describe.}
}
\description{
Returns detailed information about a specific managed capability in your Amazon EKS cluster, including its current status, configuration, health information, and any issues that may be affecting its operation.

See \url{https://www.paws-r-sdk.com/docs/eks_describe_capability/} for full documentation.
}
\keyword{internal}
