% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_certificate_authority}
\alias{eks_describe_certificate_authority}
\title{Returns detailed information about a certificate authority (CA) in your
cluster, including its validity period, signing and distribution status,
provenance, scheduled auto-activation events, and public certificate
data}
\usage{
eks_describe_certificate_authority(clusterName, certificateAuthorityId)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{certificateAuthorityId}{[required] The ID of the certificate authority to describe.}
}
\description{
Returns detailed information about a certificate authority (CA) in your cluster, including its validity period, signing and distribution status, provenance, scheduled auto-activation events, and public certificate data.

See \url{https://www.paws-r-sdk.com/docs/eks_describe_certificate_authority/} for full documentation.
}
\keyword{internal}
