% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_cluster}
\alias{eks_describe_cluster}
\title{Returns descriptive information about an Amazon EKS cluster}
\usage{
eks_describe_cluster(name)
}
\arguments{
\item{name}{[required] The name of the cluster to describe.}
}
\description{
Returns descriptive information about an Amazon EKS cluster.

See \url{https://www.paws-r-sdk.com/docs/eks_describe_cluster/} for full documentation.
}
\keyword{internal}
