% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_cluster_versions}
\alias{eks_describe_cluster_versions}
\title{Lists available Kubernetes versions for Amazon EKS clusters}
\usage{
eks_describe_cluster_versions(
  clusterType = NULL,
  maxResults = NULL,
  nextToken = NULL,
  defaultOnly = NULL,
  includeAll = NULL,
  clusterVersions = NULL,
  status = NULL
)
}
\arguments{
\item{clusterType}{The type of cluster to filter versions by.}

\item{maxResults}{Maximum number of results to return.}

\item{nextToken}{Pagination token for the next set of results.}

\item{defaultOnly}{Filter to show only default versions.}

\item{includeAll}{Include all available versions in the response.}

\item{clusterVersions}{List of specific cluster versions to describe.}

\item{status}{Filter versions by their current status.}
}
\description{
Lists available Kubernetes versions for Amazon EKS clusters.

See \url{https://www.paws-r-sdk.com/docs/eks_describe_cluster_versions/} for full documentation.
}
\keyword{internal}
