% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_eks_anywhere_subscription}
\alias{eks_describe_eks_anywhere_subscription}
\title{Returns descriptive information about a subscription}
\usage{
eks_describe_eks_anywhere_subscription(id)
}
\arguments{
\item{id}{[required] The ID of the subscription.}
}
\description{
Returns descriptive information about a subscription.

See \url{https://www.paws-r-sdk.com/docs/eks_describe_eks_anywhere_subscription/} for full documentation.
}
\keyword{internal}
