% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_identity_provider_config}
\alias{eks_describe_identity_provider_config}
\title{Returns descriptive information about an identity provider configuration}
\usage{
eks_describe_identity_provider_config(clusterName,
  identityProviderConfig)
}
\arguments{
\item{clusterName}{[required] The cluster name that the identity provider configuration is associated
to.}

\item{identityProviderConfig}{[required] An object that represents an identity provider configuration.}
}
\value{
A list with the following syntax:\preformatted{list(
  identityProviderConfig = list(
    oidc = list(
      identityProviderConfigName = "string",
      identityProviderConfigArn = "string",
      clusterName = "string",
      issuerUrl = "string",
      clientId = "string",
      usernameClaim = "string",
      usernamePrefix = "string",
      groupsClaim = "string",
      groupsPrefix = "string",
      requiredClaims = list(
        "string"
      ),
      tags = list(
        "string"
      ),
      status = "CREATING"|"DELETING"|"ACTIVE"
    )
  )
)
}
}
\description{
Returns descriptive information about an identity provider
configuration.
}
\section{Request syntax}{
\preformatted{svc$describe_identity_provider_config(
  clusterName = "string",
  identityProviderConfig = list(
    type = "string",
    name = "string"
  )
)
}
}

\keyword{internal}
