% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_insight}
\alias{eks_describe_insight}
\title{Returns details about an insight that you specify using its ID}
\usage{
eks_describe_insight(clusterName, id)
}
\arguments{
\item{clusterName}{[required] The name of the cluster to describe the insight for.}

\item{id}{[required] The identity of the insight to describe.}
}
\description{
Returns details about an insight that you specify using its ID.

See \url{https://www.paws-r-sdk.com/docs/eks_describe_insight/} for full documentation.
}
\keyword{internal}
