% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_insights_refresh}
\alias{eks_describe_insights_refresh}
\title{Returns the status of the latest on-demand cluster insights refresh
operation}
\usage{
eks_describe_insights_refresh(clusterName)
}
\arguments{
\item{clusterName}{[required] The name of the cluster associated with the insights refresh operation.}
}
\description{
Returns the status of the latest on-demand cluster insights refresh operation.

See \url{https://www.paws-r-sdk.com/docs/eks_describe_insights_refresh/} for full documentation.
}
\keyword{internal}
