% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_pod_identity_association}
\alias{eks_describe_pod_identity_association}
\title{Returns descriptive information about an EKS Pod Identity association}
\usage{
eks_describe_pod_identity_association(clusterName, associationId)
}
\arguments{
\item{clusterName}{[required] The name of the cluster that the association is in.}

\item{associationId}{[required] The ID of the association that you want the description of.}
}
\description{
Returns descriptive information about an EKS Pod Identity association.

See \url{https://www.paws-r-sdk.com/docs/eks_describe_pod_identity_association/} for full documentation.
}
\keyword{internal}
