% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_disassociate_access_policy}
\alias{eks_disassociate_access_policy}
\title{Disassociates an access policy from an access entry}
\usage{
eks_disassociate_access_policy(clusterName, principalArn, policyArn)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{principalArn}{[required] The ARN of the IAM principal for the \code{AccessEntry}.}

\item{policyArn}{[required] The ARN of the policy to disassociate from the access entry. For a list
of associated policies ARNs, use
\code{\link[=eks_list_associated_access_policies]{list_associated_access_policies}}.}
}
\description{
Disassociates an access policy from an access entry.

See \url{https://www.paws-r-sdk.com/docs/eks_disassociate_access_policy/} for full documentation.
}
\keyword{internal}
