% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_list_capabilities}
\alias{eks_list_capabilities}
\title{Lists all managed capabilities in your Amazon EKS cluster}
\usage{
eks_list_capabilities(clusterName, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{clusterName}{[required] The name of the Amazon EKS cluster for which you want to list
capabilities.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated request, where
\code{maxResults} was used and the results exceeded the value of that
parameter. Pagination continues from the end of the previous results
that returned the \code{nextToken} value. This value is null when there are
no more results to return.}

\item{maxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value. If you don't specify a value, the default is 100 results.}
}
\description{
Lists all managed capabilities in your Amazon EKS cluster. You can use this operation to get an overview of all capabilities and their current status.

See \url{https://www.paws-r-sdk.com/docs/eks_list_capabilities/} for full documentation.
}
\keyword{internal}
