% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_list_certificate_authorities}
\alias{eks_list_certificate_authorities}
\title{Lists the certificate authorities (CAs) for your cluster}
\usage{
eks_list_certificate_authorities(
  clusterName,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{maxResults}{The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned \code{nextToken} value. If you don't specify a value, the default is 100 results.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated request, where \code{maxResults} was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the \code{nextToken} value. This value is null when there are no more results to return.

This token should be treated as an opaque identifier that is used only to retrieve the next items in a list and not for other programmatic purposes.}
}
\description{
Lists the certificate authorities (CAs) for your cluster. A cluster has at most two certificate authorities: the outgoing CA that's currently signing and, during a rotation, one successor CA.

See \url{https://www.paws-r-sdk.com/docs/eks_list_certificate_authorities/} for full documentation.
}
\keyword{internal}
