% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_register_cluster}
\alias{eks_register_cluster}
\title{Connects a Kubernetes cluster to the Amazon EKS control plane}
\usage{
eks_register_cluster(
  name,
  connectorConfig,
  clientRequestToken = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] A unique name for this cluster in your Amazon Web Services Region.}

\item{connectorConfig}{[required] The configuration settings required to connect the Kubernetes cluster to
the Amazon EKS control plane.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{tags}{Metadata that assists with categorization and organization. Each tag
consists of a key and an optional value. You define both. Tags don't
propagate to any other cluster or Amazon Web Services resources.}
}
\description{
Connects a Kubernetes cluster to the Amazon EKS control plane.

See \url{https://www.paws-r-sdk.com/docs/eks_register_cluster/} for full documentation.
}
\keyword{internal}
