% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_start_insights_refresh}
\alias{eks_start_insights_refresh}
\title{Initiates an on-demand refresh operation for cluster insights, getting
the latest analysis outside of the standard refresh schedule}
\usage{
eks_start_insights_refresh(clusterName)
}
\arguments{
\item{clusterName}{[required] The name of the cluster for the refresh insights operation.}
}
\description{
Initiates an on-demand refresh operation for cluster insights, getting the latest analysis outside of the standard refresh schedule.

See \url{https://www.paws-r-sdk.com/docs/eks_start_insights_refresh/} for full documentation.
}
\keyword{internal}
