% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_update_capability}
\alias{eks_update_capability}
\title{Updates the configuration of a managed capability in your Amazon EKS
cluster}
\usage{
eks_update_capability(
  clusterName,
  capabilityName,
  roleArn = NULL,
  configuration = NULL,
  clientRequestToken = NULL,
  deletePropagationPolicy = NULL
)
}
\arguments{
\item{clusterName}{[required] The name of the Amazon EKS cluster that contains the capability you want
to update configuration for.}

\item{capabilityName}{[required] The name of the capability to update configuration for.}

\item{roleArn}{The Amazon Resource Name (ARN) of the IAM role that the capability uses
to interact with Amazon Web Services services. If you specify a new role
ARN, the capability will start using the new role for all subsequent
operations.}

\item{configuration}{The updated configuration settings for the capability. You only need to
specify the configuration parameters you want to change. For Argo CD
capabilities, you can update RBAC role mappings and network access
settings.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. This token is valid for 24 hours after
creation.}

\item{deletePropagationPolicy}{The updated delete propagation policy for the capability. Currently, the
only supported value is \code{RETAIN}.}
}
\description{
Updates the configuration of a managed capability in your Amazon EKS cluster. You can update the IAM role, configuration settings, and delete propagation policy for a capability.

See \url{https://www.paws-r-sdk.com/docs/eks_update_capability/} for full documentation.
}
\keyword{internal}
