% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_update_eks_anywhere_subscription}
\alias{eks_update_eks_anywhere_subscription}
\title{Update an EKS Anywhere Subscription}
\usage{
eks_update_eks_anywhere_subscription(id, autoRenew, clientRequestToken = NULL)
}
\arguments{
\item{id}{[required] The ID of the subscription.}

\item{autoRenew}{[required] A boolean indicating whether or not to automatically renew the
subscription.}

\item{clientRequestToken}{Unique, case-sensitive identifier to ensure the idempotency of the
request.}
}
\description{
Update an EKS Anywhere Subscription. Only auto renewal and tags can be updated after subscription creation.

See \url{https://www.paws-r-sdk.com/docs/eks_update_eks_anywhere_subscription/} for full documentation.
}
\keyword{internal}
