% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_associate_environment_operations_role}
\alias{elasticbeanstalk_associate_environment_operations_role}
\title{The operations role feature of Elastic Beanstalk is in beta release and
is subject to change}
\usage{
elasticbeanstalk_associate_environment_operations_role(
  EnvironmentName,
  OperationsRole
)
}
\arguments{
\item{EnvironmentName}{[required] The name of the environment to which to set the operations role.}

\item{OperationsRole}{[required] The Amazon Resource Name (ARN) of an existing IAM role to be used as the environment's operations role.}
}
\description{
The operations role feature of Elastic Beanstalk is in beta release and is subject to change.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_associate_environment_operations_role/} for full documentation.
}
\keyword{internal}
