% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_check_dns_availability}
\alias{elasticbeanstalk_check_dns_availability}
\title{Checks if the specified CNAME is available}
\usage{
elasticbeanstalk_check_dns_availability(CNAMEPrefix)
}
\arguments{
\item{CNAMEPrefix}{[required] The prefix used when this CNAME is reserved.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Available = TRUE|FALSE,
  FullyQualifiedCNAME = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Checks if the specified CNAME is available.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$check_dns_availability(
  CNAMEPrefix = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following operation checks the availability of the subdomain
# my-cname:
svc$check_dns_availability(
  CNAMEPrefix = "my-cname"
)
}

}
\keyword{internal}
