% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_delete_platform_version}
\alias{elasticbeanstalk_delete_platform_version}
\title{Deletes the specified version of a custom platform}
\usage{
elasticbeanstalk_delete_platform_version(PlatformArn)
}
\arguments{
\item{PlatformArn}{The ARN of the version of the custom platform.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PlatformSummary = list(
    PlatformArn = "string",
    PlatformOwner = "string",
    PlatformStatus = "Creating"|"Failed"|"Ready"|"Deleting"|"Deleted",
    PlatformCategory = "string",
    OperatingSystemName = "string",
    OperatingSystemVersion = "string",
    SupportedTierList = list(
      "string"
    ),
    SupportedAddonList = list(
      "string"
    ),
    PlatformLifecycleState = "string",
    PlatformVersion = "string",
    PlatformBranchName = "string",
    PlatformBranchLifecycleState = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified version of a custom platform.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_platform_version(
  PlatformArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
