% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_describe_applications}
\alias{elasticbeanstalk_describe_applications}
\title{Returns the descriptions of existing applications}
\usage{
elasticbeanstalk_describe_applications(ApplicationNames = NULL)
}
\arguments{
\item{ApplicationNames}{If specified, Elastic Beanstalk restricts the returned descriptions to only include those with the specified names.}
}
\description{
Returns the descriptions of existing applications.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_describe_applications/} for full documentation.
}
\keyword{internal}
