% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_describe_environment_health}
\alias{elasticbeanstalk_describe_environment_health}
\title{Returns information about the overall health of the specified
environment}
\usage{
elasticbeanstalk_describe_environment_health(
  EnvironmentName = NULL,
  EnvironmentId = NULL,
  AttributeNames = NULL
)
}
\arguments{
\item{EnvironmentName}{Specify the environment by name.

You must specify either this or an EnvironmentName, or both.}

\item{EnvironmentId}{Specify the environment by ID.

You must specify either this or an EnvironmentName, or both.}

\item{AttributeNames}{Specify the response elements to return. To retrieve all attributes, set to \code{All}. If no attribute names are specified, returns the name of the environment.}
}
\description{
Returns information about the overall health of the specified environment. The \strong{DescribeEnvironmentHealth} operation is only available with Elastic Beanstalk Enhanced Health.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_describe_environment_health/} for full documentation.
}
\keyword{internal}
