% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_describe_environment_managed_actions}
\alias{elasticbeanstalk_describe_environment_managed_actions}
\title{Lists an environment's upcoming and in-progress managed actions}
\usage{
elasticbeanstalk_describe_environment_managed_actions(EnvironmentName,
  EnvironmentId, Status)
}
\arguments{
\item{EnvironmentName}{The name of the target environment.}

\item{EnvironmentId}{The environment ID of the target environment.}

\item{Status}{To show only actions with a particular status, specify a status.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ManagedActions = list(
    list(
      ActionId = "string",
      ActionDescription = "string",
      ActionType = "InstanceRefresh"|"PlatformUpdate"|"Unknown",
      Status = "Scheduled"|"Pending"|"Running"|"Unknown",
      WindowStartTime = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists an environment's upcoming and in-progress managed actions.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_environment_managed_actions(
  EnvironmentName = "string",
  EnvironmentId = "string",
  Status = "Scheduled"|"Pending"|"Running"|"Unknown"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
