% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_disassociate_environment_operations_role}
\alias{elasticbeanstalk_disassociate_environment_operations_role}
\title{The operations role feature of Elastic Beanstalk is in beta release and
is subject to change}
\usage{
elasticbeanstalk_disassociate_environment_operations_role(EnvironmentName)
}
\arguments{
\item{EnvironmentName}{[required] The name of the environment from which to disassociate the operations role.}
}
\description{
The operations role feature of Elastic Beanstalk is in beta release and is subject to change.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_disassociate_environment_operations_role/} for full documentation.
}
\keyword{internal}
