% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_update_application_resource_lifecycle}
\alias{elasticbeanstalk_update_application_resource_lifecycle}
\title{Modifies lifecycle settings for an application}
\usage{
elasticbeanstalk_update_application_resource_lifecycle(
  ApplicationName,
  ResourceLifecycleConfig
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application.}

\item{ResourceLifecycleConfig}{[required] The lifecycle configuration.}
}
\description{
Modifies lifecycle settings for an application.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_update_application_resource_lifecycle/} for full documentation.
}
\keyword{internal}
