% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_update_configuration_template}
\alias{elasticbeanstalk_update_configuration_template}
\title{Updates the specified configuration template to have the specified
properties or configuration option values}
\usage{
elasticbeanstalk_update_configuration_template(ApplicationName,
  TemplateName, Description, OptionSettings, OptionsToRemove)
}
\arguments{
\item{ApplicationName}{[required] The name of the application associated with the configuration template
to update.

If no application is found with this name,
\code{\link[=elasticbeanstalk_update_configuration_template]{update_configuration_template}}
returns an \code{InvalidParameterValue} error.}

\item{TemplateName}{[required] The name of the configuration template to update.

If no configuration template is found with this name,
\code{\link[=elasticbeanstalk_update_configuration_template]{update_configuration_template}}
returns an \code{InvalidParameterValue} error.}

\item{Description}{A new description for the configuration.}

\item{OptionSettings}{A list of configuration option settings to update with the new specified
option value.}

\item{OptionsToRemove}{A list of configuration options to remove from the configuration set.

Constraint: You can remove only \code{UserDefined} configuration options.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SolutionStackName = "string",
  PlatformArn = "string",
  ApplicationName = "string",
  TemplateName = "string",
  Description = "string",
  EnvironmentName = "string",
  DeploymentStatus = "deployed"|"pending"|"failed",
  DateCreated = as.POSIXct(
    "2015-01-01"
  ),
  DateUpdated = as.POSIXct(
    "2015-01-01"
  ),
  OptionSettings = list(
    list(
      ResourceName = "string",
      Namespace = "string",
      OptionName = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates the specified configuration template to have the specified
properties or configuration option values.

If a property (for example, \code{ApplicationName}) is not provided, its
value remains unchanged. To clear such properties, specify an empty
string.

Related Topics
\itemize{
\item \code{\link[=elasticbeanstalk_describe_configuration_options]{describe_configuration_options}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_configuration_template(
  ApplicationName = "string",
  TemplateName = "string",
  Description = "string",
  OptionSettings = list(
    list(
      ResourceName = "string",
      Namespace = "string",
      OptionName = "string",
      Value = "string"
    )
  ),
  OptionsToRemove = list(
    list(
      ResourceName = "string",
      Namespace = "string",
      OptionName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following operation removes the configured CloudWatch custom health
# metrics configuration ConfigDocument from a saved configuration template
# named my-template:
svc$update_configuration_template(
  ApplicationName = "my-app",
  OptionsToRemove = list(
    list(
      Namespace = "aws:elasticbeanstalk:healthreporting:system",
      OptionName = "ConfigDocument"
    )
  ),
  TemplateName = "my-template"
)
}

}
\keyword{internal}
