% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_create_job_template}
\alias{emrcontainers_create_job_template}
\title{Creates a job template}
\usage{
emrcontainers_create_job_template(
  name,
  clientToken,
  jobTemplateData,
  tags = NULL,
  kmsKeyArn = NULL
)
}
\arguments{
\item{name}{[required] The specified name of the job template.}

\item{clientToken}{[required] The client token of the job template.}

\item{jobTemplateData}{[required] The job template data which holds values of StartJobRun API request.}

\item{tags}{The tags that are associated with the job template.}

\item{kmsKeyArn}{The KMS key ARN used to encrypt the job template.}
}
\description{
Creates a job template. Job template stores values of StartJobRun API request in a template and can be used to start a job run. Job template allows two use cases: avoid repeating recurring StartJobRun API request values, enforcing certain values in StartJobRun API request.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_create_job_template/} for full documentation.
}
\keyword{internal}
