% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_create_security_configuration}
\alias{emrcontainers_create_security_configuration}
\title{Creates a security configuration}
\usage{
emrcontainers_create_security_configuration(
  clientToken,
  name,
  containerProvider = NULL,
  securityConfigurationData,
  tags = NULL
)
}
\arguments{
\item{clientToken}{[required] The client idempotency token to use when creating the security
configuration.}

\item{name}{[required] The name of the security configuration.}

\item{containerProvider}{The container provider associated with the security configuration.}

\item{securityConfigurationData}{[required] Security configuration input for the request.}

\item{tags}{The tags to add to the security configuration.}
}
\description{
Creates a security configuration. Security configurations in Amazon EMR on EKS are templates for different security setups. You can use security configurations to configure the Lake Formation integration setup. You can also create a security configuration to re-use a security setup each time you create a virtual cluster.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_create_security_configuration/} for full documentation.
}
\keyword{internal}
