% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_create_virtual_cluster}
\alias{emrcontainers_create_virtual_cluster}
\title{Creates a virtual cluster}
\usage{
emrcontainers_create_virtual_cluster(
  name,
  containerProvider,
  clientToken,
  tags = NULL,
  securityConfigurationId = NULL,
  sessionEnabled = NULL,
  schedulerConfiguration = NULL
)
}
\arguments{
\item{name}{[required] The specified name of the virtual cluster.}

\item{containerProvider}{[required] The container provider of the virtual cluster.}

\item{clientToken}{[required] The client token of the virtual cluster.}

\item{tags}{The tags assigned to the virtual cluster.}

\item{securityConfigurationId}{The ID of the security configuration.}

\item{sessionEnabled}{Indicates whether the virtual cluster has session support enabled.}

\item{schedulerConfiguration}{The scheduler configuration (concurrency and queue limits) to apply to the virtual cluster at creation time. When omitted, no limits are applied.}
}
\description{
Creates a virtual cluster. Virtual cluster is a managed entity on Amazon EMR on EKS. You can create, update, describe, list and delete virtual clusters. They do not consume any additional resource in your system. A single virtual cluster maps to a single Kubernetes namespace. Given this relationship, you can model virtual clusters the same way you model Kubernetes namespaces to meet your requirements.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_create_virtual_cluster/} for full documentation.
}
\keyword{internal}
