% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_delete_managed_endpoint}
\alias{emrcontainers_delete_managed_endpoint}
\title{Deletes a managed endpoint}
\usage{
emrcontainers_delete_managed_endpoint(id, virtualClusterId)
}
\arguments{
\item{id}{[required] The ID of the managed endpoint.}

\item{virtualClusterId}{[required] The ID of the endpoint's virtual cluster.}
}
\description{
Deletes a managed endpoint. A managed endpoint is a gateway that connects Amazon EMR Studio to Amazon EMR on EKS so that Amazon EMR Studio can communicate with your virtual cluster.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_delete_managed_endpoint/} for full documentation.
}
\keyword{internal}
