% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_describe_job_run}
\alias{emrcontainers_describe_job_run}
\title{Displays detailed information about a job run}
\usage{
emrcontainers_describe_job_run(id, virtualClusterId)
}
\arguments{
\item{id}{[required] The ID of the job run request.}

\item{virtualClusterId}{[required] The ID of the virtual cluster for which the job run is submitted.}
}
\description{
Displays detailed information about a job run. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_describe_job_run/} for full documentation.
}
\keyword{internal}
