% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_describe_job_template}
\alias{emrcontainers_describe_job_template}
\title{Displays detailed information about a specified job template}
\usage{
emrcontainers_describe_job_template(id)
}
\arguments{
\item{id}{[required] The ID of the job template that will be described.}
}
\description{
Displays detailed information about a specified job template. Job template stores values of StartJobRun API request in a template and can be used to start a job run. Job template allows two use cases: avoid repeating recurring StartJobRun API request values, enforcing certain values in StartJobRun API request.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_describe_job_template/} for full documentation.
}
\keyword{internal}
