% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_describe_security_configuration}
\alias{emrcontainers_describe_security_configuration}
\title{Displays detailed information about a specified security configuration}
\usage{
emrcontainers_describe_security_configuration(id)
}
\arguments{
\item{id}{[required] The ID of the security configuration.}
}
\description{
Displays detailed information about a specified security configuration. Security configurations in Amazon EMR on EKS are templates for different security setups. You can use security configurations to configure the Lake Formation integration setup. You can also create a security configuration to re-use a security setup each time you create a virtual cluster.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_describe_security_configuration/} for full documentation.
}
\keyword{internal}
