% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_list_job_templates}
\alias{emrcontainers_list_job_templates}
\title{Lists job templates based on a set of parameters}
\usage{
emrcontainers_list_job_templates(
  createdAfter = NULL,
  createdBefore = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{createdAfter}{The date and time after which the job templates were created.}

\item{createdBefore}{The date and time before which the job templates were created.}

\item{maxResults}{The maximum number of job templates that can be listed.}

\item{nextToken}{The token for the next set of job templates to return.}
}
\description{
Lists job templates based on a set of parameters. Job template stores values of StartJobRun API request in a template and can be used to start a job run. Job template allows two use cases: avoid repeating recurring StartJobRun API request values, enforcing certain values in StartJobRun API request.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_list_job_templates/} for full documentation.
}
\keyword{internal}
