% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_list_security_configurations}
\alias{emrcontainers_list_security_configurations}
\title{Lists security configurations based on a set of parameters}
\usage{
emrcontainers_list_security_configurations(
  createdAfter = NULL,
  createdBefore = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{createdAfter}{The date and time after which the security configuration was created.}

\item{createdBefore}{The date and time before which the security configuration was created.}

\item{maxResults}{The maximum number of security configurations the operation can list.}

\item{nextToken}{The token for the next set of security configurations to return.}
}
\description{
Lists security configurations based on a set of parameters. Security configurations in Amazon EMR on EKS are templates for different security setups. You can use security configurations to configure the Lake Formation integration setup. You can also create a security configuration to re-use a security setup each time you create a virtual cluster.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_list_security_configurations/} for full documentation.
}
\keyword{internal}
