% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_list_tags_for_resource}
\alias{emrcontainers_list_tags_for_resource}
\title{Lists the tags assigned to the resources}
\usage{
emrcontainers_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of tagged resources.}
}
\description{
Lists the tags assigned to the resources.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
