% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_tag_resource}
\alias{emrcontainers_tag_resource}
\title{Assigns tags to resources}
\usage{
emrcontainers_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of resources.}

\item{tags}{[required] The tags assigned to resources.}
}
\description{
Assigns tags to resources. A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value, both of which you define. Tags enable you to categorize your Amazon Web Services resources by attributes such as purpose, owner, or environment. When you have many resources of the same type, you can quickly identify a specific resource based on the tags you've assigned to it. For example, you can define a set of tags for your Amazon EMR on EKS clusters to help you track each cluster's owner and stack level. We recommend that you devise a consistent set of tag keys for each resource type. You can then search and filter the resources based on the tags that you add.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_tag_resource/} for full documentation.
}
\keyword{internal}
