% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_update_virtual_cluster}
\alias{emrcontainers_update_virtual_cluster}
\title{Updates a virtual cluster}
\usage{
emrcontainers_update_virtual_cluster(
  id,
  schedulerConfiguration = NULL,
  clientToken
)
}
\arguments{
\item{id}{[required] The ID of the virtual cluster to update.}

\item{schedulerConfiguration}{The scheduler configuration to apply to the virtual cluster. The new configuration fully replaces the existing one. If you omit a field, the corresponding limit is removed.}

\item{clientToken}{[required] A unique, case-sensitive identifier that you provide to ensure that the operation completes no more than one time. If this token matches a previous request, the service ignores the request, but does not return an error.}
}
\description{
Updates a virtual cluster. Virtual cluster is a managed entity on Amazon EMR on EKS. You can create, update, describe, list and delete virtual clusters. They do not consume any additional resource in your system. A single virtual cluster maps to a single Kubernetes namespace. Given this relationship, you can model virtual clusters the same way you model Kubernetes namespaces to meet your requirements.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_update_virtual_cluster/} for full documentation.
}
\keyword{internal}
