% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_delete_application}
\alias{emrserverless_delete_application}
\title{Deletes an application}
\usage{
emrserverless_delete_application(applicationId)
}
\arguments{
\item{applicationId}{[required] The ID of the application that will be deleted.}
}
\description{
Deletes an application. An application has to be in a stopped or created state in order to be deleted.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_delete_application/} for full documentation.
}
\keyword{internal}
