% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_get_application}
\alias{emrserverless_get_application}
\title{Displays detailed information about a specified application}
\usage{
emrserverless_get_application(applicationId)
}
\arguments{
\item{applicationId}{[required] The ID of the application that will be described.}
}
\description{
Displays detailed information about a specified application.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_get_application/} for full documentation.
}
\keyword{internal}
