% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_get_job_run}
\alias{emrserverless_get_job_run}
\title{Displays detailed information about a job run}
\usage{
emrserverless_get_job_run(applicationId, jobRunId)
}
\arguments{
\item{applicationId}{[required] The ID of the application on which the job run is submitted.}

\item{jobRunId}{[required] The ID of the job run.}
}
\description{
Displays detailed information about a job run.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_get_job_run/} for full documentation.
}
\keyword{internal}
