% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_get_resource_dashboard}
\alias{emrserverless_get_resource_dashboard}
\title{Returns a URL that you can use to access the application UIs for a
specified resource, such as a session}
\usage{
emrserverless_get_resource_dashboard(applicationId, resourceId, resourceType)
}
\arguments{
\item{applicationId}{[required] The ID of the application that the resource belongs to.}

\item{resourceId}{[required] The ID of the resource.}

\item{resourceType}{[required] The type of resource to access the dashboard for. Currently, only
\code{Session} is supported.}
}
\description{
Returns a URL that you can use to access the application UIs for a specified resource, such as a session.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_get_resource_dashboard/} for full documentation.
}
\keyword{internal}
