% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_get_session}
\alias{emrserverless_get_session}
\title{Displays detailed information about a session}
\usage{
emrserverless_get_session(applicationId, sessionId)
}
\arguments{
\item{applicationId}{[required] The ID of the application that the session belongs to.}

\item{sessionId}{[required] The ID of the session.}
}
\description{
Displays detailed information about a session.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_get_session/} for full documentation.
}
\keyword{internal}
