% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_get_session_endpoint}
\alias{emrserverless_get_session_endpoint}
\title{Returns the session endpoint URL and a time-limited authentication token
for the specified session}
\usage{
emrserverless_get_session_endpoint(applicationId, sessionId)
}
\arguments{
\item{applicationId}{[required] The ID of the application that the session belongs to.}

\item{sessionId}{[required] The ID of the session.}
}
\description{
Returns the session endpoint URL and a time-limited authentication token for the specified session. Use the endpoint and token to connect a client to the session. Call this operation again when the authentication token expires to obtain a new token.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_get_session_endpoint/} for full documentation.
}
\keyword{internal}
