% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_list_applications}
\alias{emrserverless_list_applications}
\title{Lists applications based on a set of parameters}
\usage{
emrserverless_list_applications(
  nextToken = NULL,
  maxResults = NULL,
  states = NULL
)
}
\arguments{
\item{nextToken}{The token for the next set of application results.}

\item{maxResults}{The maximum number of applications that can be listed.}

\item{states}{An optional filter for application states. Note that if this filter
contains multiple states, the resulting list will be grouped by the
state.}
}
\description{
Lists applications based on a set of parameters.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_list_applications/} for full documentation.
}
\keyword{internal}
