% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_list_job_runs}
\alias{emrserverless_list_job_runs}
\title{Lists job runs based on a set of parameters}
\usage{
emrserverless_list_job_runs(
  applicationId,
  nextToken = NULL,
  maxResults = NULL,
  createdAtAfter = NULL,
  createdAtBefore = NULL,
  states = NULL
)
}
\arguments{
\item{applicationId}{[required] The ID of the application for which to list the job run.}

\item{nextToken}{The token for the next set of job run results.}

\item{maxResults}{The maximum number of job runs that can be listed.}

\item{createdAtAfter}{The lower bound of the option to filter by creation date and time.}

\item{createdAtBefore}{The upper bound of the option to filter by creation date and time.}

\item{states}{An optional filter for job run states. Note that if this filter contains
multiple states, the resulting list will be grouped by the state.}
}
\description{
Lists job runs based on a set of parameters.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_list_job_runs/} for full documentation.
}
\keyword{internal}
