% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_list_sessions}
\alias{emrserverless_list_sessions}
\title{Lists sessions for the specified application}
\usage{
emrserverless_list_sessions(
  applicationId,
  nextToken = NULL,
  maxResults = NULL,
  states = NULL,
  createdAtAfter = NULL,
  createdAtBefore = NULL
)
}
\arguments{
\item{applicationId}{[required] The ID of the application to list sessions for.}

\item{nextToken}{The token for the next set of session results.}

\item{maxResults}{The maximum number of sessions to return in each page of results.}

\item{states}{An optional filter for session states. Note that if this filter contains
multiple states, the resulting list will be grouped by the state.}

\item{createdAtAfter}{The lower bound of the option to filter by creation date and time.}

\item{createdAtBefore}{The upper bound of the option to filter by creation date and time.}
}
\description{
Lists sessions for the specified application. You can filter sessions by state and creation time.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_list_sessions/} for full documentation.
}
\keyword{internal}
