% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_list_tags_for_resource}
\alias{emrserverless_list_tags_for_resource}
\title{Lists the tags assigned to the resources}
\usage{
emrserverless_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource to list the
tags for. Currently, the supported resources are Amazon EMR Serverless
applications and job runs.}
}
\description{
Lists the tags assigned to the resources.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
