% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_start_application}
\alias{emrserverless_start_application}
\title{Starts a specified application and initializes initial capacity if
configured}
\usage{
emrserverless_start_application(applicationId)
}
\arguments{
\item{applicationId}{[required] The ID of the application to start.}
}
\description{
Starts a specified application and initializes initial capacity if configured.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_start_application/} for full documentation.
}
\keyword{internal}
