% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_start_session}
\alias{emrserverless_start_session}
\title{Creates and starts a new session on the specified application}
\usage{
emrserverless_start_session(
  applicationId,
  clientToken,
  executionRoleArn,
  configurationOverrides = NULL,
  tags = NULL,
  idleTimeoutMinutes = NULL,
  name = NULL
)
}
\arguments{
\item{applicationId}{[required] The ID of the application on which to start the session.}

\item{clientToken}{[required] A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you retry a request that completed
successfully using the same client token, the server returns the
successful response without performing the operation again.}

\item{executionRoleArn}{[required] The execution role ARN for the session. Amazon EMR Serverless uses this
role to access Amazon Web Services resources on your behalf during
session execution.}

\item{configurationOverrides}{The configuration overrides for the session. Only runtime configuration
overrides are supported.}

\item{tags}{The tags to assign to the session.}

\item{idleTimeoutMinutes}{The idle timeout in minutes for the session. After the session remains
idle for this duration, Amazon EMR Serverless automatically terminates
it.}

\item{name}{The optional name for the session.}
}
\description{
Creates and starts a new session on the specified application. The application must be in the \code{STARTED} state or have \code{AutoStart} enabled, and have interactive sessions enabled. This operation is supported for EMR release 7.13.0 and later.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_start_session/} for full documentation.
}
\keyword{internal}
