% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_stop_application}
\alias{emrserverless_stop_application}
\title{Stops a specified application and releases initial capacity if
configured}
\usage{
emrserverless_stop_application(applicationId)
}
\arguments{
\item{applicationId}{[required] The ID of the application to stop.}
}
\description{
Stops a specified application and releases initial capacity if configured. All scheduled and running jobs must be completed or cancelled before stopping an application.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_stop_application/} for full documentation.
}
\keyword{internal}
