% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_tag_resource}
\alias{emrserverless_tag_resource}
\title{Assigns tags to resources}
\usage{
emrserverless_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource to list the
tags for. Currently, the supported resources are Amazon EMR Serverless
applications and job runs.}

\item{tags}{[required] The tags to add to the resource. A tag is an array of key-value pairs.}
}
\description{
Assigns tags to resources. A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value, both of which you define. Tags enable you to categorize your Amazon Web Services resources by attributes such as purpose, owner, or environment. When you have many resources of the same type, you can quickly identify a specific resource based on the tags you've assigned to it.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_tag_resource/} for full documentation.
}
\keyword{internal}
