% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_terminate_session}
\alias{emrserverless_terminate_session}
\title{Terminates the specified session}
\usage{
emrserverless_terminate_session(applicationId, sessionId)
}
\arguments{
\item{applicationId}{[required] The ID of the application that the session belongs to.}

\item{sessionId}{[required] The ID of the session to terminate.}
}
\description{
Terminates the specified session. After you terminate a session, it enters the \code{TERMINATING} state and then the \code{TERMINATED} state. You can still access the Spark History Server for a terminated session through the \code{\link[=emrserverless_get_resource_dashboard]{get_resource_dashboard}} operation.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_terminate_session/} for full documentation.
}
\keyword{internal}
