% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_untag_resource}
\alias{emrserverless_untag_resource}
\title{Removes tags from resources}
\usage{
emrserverless_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource to list the
tags for. Currently, the supported resources are Amazon EMR Serverless
applications and job runs.}

\item{tagKeys}{[required] The keys of the tags to be removed.}
}
\description{
Removes tags from resources.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_untag_resource/} for full documentation.
}
\keyword{internal}
