% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_delete_component}
\alias{imagebuilder_delete_component}
\title{Deletes a component build version}
\usage{
imagebuilder_delete_component(componentBuildVersionArn)
}
\arguments{
\item{componentBuildVersionArn}{[required] The Amazon Resource Name (ARN) of the component build version to delete.}
}
\description{
Deletes a component build version. The request fails with \code{ResourceDependencyException} if an image recipe or container recipe references this component version. It also fails if the component build version is shared with other accounts.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_delete_component/} for full documentation.
}
\keyword{internal}
