% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_delete_image_pipeline}
\alias{imagebuilder_delete_image_pipeline}
\title{Deletes an image pipeline}
\usage{
imagebuilder_delete_image_pipeline(imagePipelineArn)
}
\arguments{
\item{imagePipelineArn}{[required] The Amazon Resource Name (ARN) of the image pipeline to delete.}
}
\description{
Deletes an image pipeline. Images that the pipeline created aren't deleted - remove those separately with \code{\link[=imagebuilder_delete_image]{delete_image}}. You can delete a pipeline while a build that it started is still running. The build continues independently.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_delete_image_pipeline/} for full documentation.
}
\keyword{internal}
