% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_delete_infrastructure_configuration}
\alias{imagebuilder_delete_infrastructure_configuration}
\title{Deletes an infrastructure configuration}
\usage{
imagebuilder_delete_infrastructure_configuration(
  infrastructureConfigurationArn
)
}
\arguments{
\item{infrastructureConfigurationArn}{[required] The Amazon Resource Name (ARN) of the infrastructure configuration to delete.}
}
\description{
Deletes an infrastructure configuration. You can't delete a configuration that an image pipeline still references. The request fails with \code{ResourceDependencyException}. Update or delete the referencing pipelines first.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_delete_infrastructure_configuration/} for full documentation.
}
\keyword{internal}
