% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_container_recipe_policy}
\alias{imagebuilder_get_container_recipe_policy}
\title{Retrieves the policy for a container recipe}
\usage{
imagebuilder_get_container_recipe_policy(containerRecipeArn)
}
\arguments{
\item{containerRecipeArn}{[required] The Amazon Resource Name (ARN) of the container recipe for the policy
being requested.}
}
\description{
Retrieves the policy for a container recipe.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_container_recipe_policy/} for full documentation.
}
\keyword{internal}
