% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_distribution_configuration}
\alias{imagebuilder_get_distribution_configuration}
\title{Retrieves a distribution configuration}
\usage{
imagebuilder_get_distribution_configuration(distributionConfigurationArn)
}
\arguments{
\item{distributionConfigurationArn}{[required] The Amazon Resource Name (ARN) of the distribution configuration that you want to retrieve.}
}
\description{
Retrieves a distribution configuration.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_distribution_configuration/} for full documentation.
}
\keyword{internal}
