% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_lifecycle_execution}
\alias{imagebuilder_get_lifecycle_execution}
\title{Retrieves runtime information for a lifecycle execution – a single run
of lifecycle actions that a lifecycle policy or a
StartResourceStateUpdate request started}
\usage{
imagebuilder_get_lifecycle_execution(lifecycleExecutionId)
}
\arguments{
\item{lifecycleExecutionId}{[required] The unique identifier for a runtime instance of the lifecycle policy.}
}
\description{
Retrieves runtime information for a lifecycle execution – a single run of lifecycle actions that a lifecycle policy or a \code{\link[=imagebuilder_start_resource_state_update]{start_resource_state_update}} request started.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_lifecycle_execution/} for full documentation.
}
\keyword{internal}
