% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_marketplace_resource}
\alias{imagebuilder_get_marketplace_resource}
\title{Verifies the subscription and performs resource dependency checks on the
requested Amazon Web Services Marketplace resource}
\usage{
imagebuilder_get_marketplace_resource(
  resourceType,
  resourceArn,
  resourceLocation = NULL
)
}
\arguments{
\item{resourceType}{[required] Specifies which type of Amazon Web Services Marketplace resource Image Builder retrieves.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies an Amazon Web Services Marketplace resource.}

\item{resourceLocation}{The Amazon S3 location of the component artifact to retrieve, in \verb{s3://bucket/key} form.}
}
\description{
Verifies the subscription and performs resource dependency checks on the requested Amazon Web Services Marketplace resource. The caller must be entitled to the resource. For Amazon Web Services Marketplace components, the response contains fields to download the components and their artifacts.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_marketplace_resource/} for full documentation.
}
\keyword{internal}
