% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_workflow_execution}
\alias{imagebuilder_get_workflow_execution}
\title{Retrieves runtime information for a specific runtime instance of the
workflow}
\usage{
imagebuilder_get_workflow_execution(workflowExecutionId)
}
\arguments{
\item{workflowExecutionId}{[required] Use the unique identifier for a runtime instance of the workflow to get runtime details.}
}
\description{
Retrieves runtime information for a specific runtime instance of the workflow.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_workflow_execution/} for full documentation.
}
\keyword{internal}
