% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_workflow_step_execution}
\alias{imagebuilder_get_workflow_step_execution}
\title{Retrieves runtime information for a specific runtime instance of the
workflow step}
\usage{
imagebuilder_get_workflow_step_execution(stepExecutionId)
}
\arguments{
\item{stepExecutionId}{[required] The unique identifier for the runtime instance of the workflow step that you want to get runtime details for. To get the identifiers for the steps that ran in a workflow, call \code{\link[=imagebuilder_list_workflow_step_executions]{list_workflow_step_executions}}.}
}
\description{
Retrieves runtime information for a specific runtime instance of the workflow step.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_workflow_step_execution/} for full documentation.
}
\keyword{internal}
