% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_infrastructure_configurations}
\alias{imagebuilder_list_infrastructure_configurations}
\title{Returns a list of infrastructure configurations}
\usage{
imagebuilder_list_infrastructure_configurations(
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filters}{You can filter on \code{name} to streamline results.}

\item{maxResults}{The maximum number of items to return in a single request.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}
}
\description{
Returns a list of infrastructure configurations.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_infrastructure_configurations/} for full documentation.
}
\keyword{internal}
