% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_lifecycle_execution_resources}
\alias{imagebuilder_list_lifecycle_execution_resources}
\title{Lists resources that the runtime instance of the image lifecycle
identified for lifecycle actions}
\usage{
imagebuilder_list_lifecycle_execution_resources(
  lifecycleExecutionId,
  parentResourceId = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{lifecycleExecutionId}{[required] The unique identifier for a runtime instance of the lifecycle policy.}

\item{parentResourceId}{The Amazon Resource Name (ARN) of an image build version to get the output resources for, such as AMIs or container images in Amazon ECR. You can get this value from the \code{resourceId} in the top-level response. If you leave this property empty, the response lists the Image Builder resources that the lifecycle execution identified for lifecycle actions. If the image build version that you specify in \code{parentResourceId} wasn't part of this lifecycle execution, the response contains an empty list.}

\item{maxResults}{The maximum number of items to return in a single request.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}
}
\description{
Lists resources that the runtime instance of the image lifecycle identified for lifecycle actions.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_lifecycle_execution_resources/} for full documentation.
}
\keyword{internal}
