% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_tags_for_resource}
\alias{imagebuilder_list_tags_for_resource}
\title{Returns the list of tags for the specified resource}
\usage{
imagebuilder_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource whose tags you want to
retrieve.}
}
\description{
Returns the list of tags for the specified resource.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
