% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_waiting_workflow_steps}
\alias{imagebuilder_list_waiting_workflow_steps}
\title{Lists the workflow steps in your Amazon Web Services account that have
paused at a WaitForAction step, and are waiting for you to respond}
\usage{
imagebuilder_list_waiting_workflow_steps(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of items to return in a single request.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}
}
\description{
Lists the workflow steps in your Amazon Web Services account that have paused at a \code{WaitForAction} step, and are waiting for you to respond. To send a response, call \code{\link[=imagebuilder_send_workflow_step_action]{send_workflow_step_action}}.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_waiting_workflow_steps/} for full documentation.
}
\keyword{internal}
