% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_workflow_build_versions}
\alias{imagebuilder_list_workflow_build_versions}
\title{Returns a list of build versions for a specific workflow resource}
\usage{
imagebuilder_list_workflow_build_versions(
  workflowVersionArn = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{workflowVersionArn}{The Amazon Resource Name (ARN) of the workflow resource for which to get a list of build versions. The version segments can contain wildcards (\code{x}) to match multiple versions of the workflow. If you don't specify an ARN, the response lists build versions for all of the workflows in your account.}

\item{maxResults}{The maximum number of items to return in a single request.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}
}
\description{
Returns a list of build versions for a specific workflow resource.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_workflow_build_versions/} for full documentation.
}
\keyword{internal}
