% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_put_image_policy}
\alias{imagebuilder_put_image_policy}
\title{Applies a policy to an image}
\usage{
imagebuilder_put_image_policy(imageArn, policy)
}
\arguments{
\item{imageArn}{[required] The Amazon Resource Name (ARN) of the image that this policy should be applied to.}

\item{policy}{[required] The resource policy to apply to the image, as a JSON policy document. Image Builder validates the policy with Amazon Web Services RAM before applying it, and rejects invalid policies with \code{InvalidParameterValueException}.}
}
\description{
Applies a policy to an image. The preferred way to share resources is with the RAM API \href{https://docs.aws.amazon.com/ram/latest/APIReference/API_CreateResourceShare.html}{CreateResourceShare}. If you use the PutImagePolicy operation instead, you must also call the RAM API \href{https://docs.aws.amazon.com/ram/latest/APIReference/API_PromoteResourceShareCreatedFromPolicy.html}{PromoteResourceShareCreatedFromPolicy}. Otherwise, the resource isn't visible to the principals that it's shared with.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_put_image_policy/} for full documentation.
}
\keyword{internal}
