% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_tag_resource}
\alias{imagebuilder_tag_resource}
\title{Adds a tag to a resource}
\usage{
imagebuilder_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to tag.}

\item{tags}{[required] The tags to apply to the resource.}
}
\description{
Adds a tag to a resource.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_tag_resource/} for full documentation.
}
\keyword{internal}
