% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_untag_resource}
\alias{imagebuilder_untag_resource}
\title{Removes a tag from a resource}
\usage{
imagebuilder_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to untag.}

\item{tagKeys}{[required] The tag keys to remove from the resource.}
}
\description{
Removes a tag from a resource.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_untag_resource/} for full documentation.
}
\keyword{internal}
