% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_update_distribution_configuration}
\alias{imagebuilder_update_distribution_configuration}
\title{Updates a distribution configuration}
\usage{
imagebuilder_update_distribution_configuration(
  distributionConfigurationArn,
  description = NULL,
  distributions,
  clientToken
)
}
\arguments{
\item{distributionConfigurationArn}{[required] The Amazon Resource Name (ARN) of the distribution configuration that you want to update.}

\item{description}{The description of the distribution configuration.}

\item{distributions}{[required] The distribution settings for the configuration. Each entry defines how output images are distributed in one target Amazon Web Services Region. A Region can appear at most once in the list. This list replaces the configuration's existing distributions entirely.}

\item{clientToken}{[required] A unique, case-sensitive identifier you provide to ensure that the operation runs no more than one time. If you retry a request with the same client token, Image Builder returns the original response without running the operation again. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency} in the \emph{Amazon EC2 API Reference}.}
}
\description{
Updates a distribution configuration. Distribution configurations define and configure the outputs for your images, including the target Regions, accounts, and settings for each Region.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_update_distribution_configuration/} for full documentation.
}
\keyword{internal}
