% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_delete_capacity_provider}
\alias{lambda_delete_capacity_provider}
\title{Deletes a capacity provider}
\usage{
lambda_delete_capacity_provider(CapacityProviderName)
}
\arguments{
\item{CapacityProviderName}{[required] The name of the capacity provider to delete.}
}
\description{
Deletes a capacity provider. You cannot delete a capacity provider that is currently being used by Lambda functions.

See \url{https://www.paws-r-sdk.com/docs/lambda_delete_capacity_provider/} for full documentation.
}
\keyword{internal}
