% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_delete_code_signing_config}
\alias{lambda_delete_code_signing_config}
\title{Deletes the code signing configuration}
\usage{
lambda_delete_code_signing_config(CodeSigningConfigArn)
}
\arguments{
\item{CodeSigningConfigArn}{[required] The The Amazon Resource Name (ARN) of the code signing configuration.}
}
\value{
An empty list.
}
\description{
Deletes the code signing configuration. You can delete the code signing
configuration only if no function is using it.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_code_signing_config(
  CodeSigningConfigArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
