% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_delete_resource_policy}
\alias{lambda_delete_resource_policy}
\title{Deletes a resource-based policy from a Lambda resource}
\usage{
lambda_delete_resource_policy(ResourceArn, RevisionId = NULL)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Lambda resource you want to delete the policy from. You can use a qualified or an unqualified ARN. The value must be a complete ARN, and the operation does not accept wildcard characters.}

\item{RevisionId}{The revision ID that the existing policy must match for the deletion to proceed. If the revision ID doesn't match, the operation fails with a \code{PreconditionFailedException} error. To retrieve the current revision ID, use the \code{\link[=lambda_get_resource_policy]{get_resource_policy}} operation.}
}
\description{
Deletes a \href{https://docs.aws.amazon.com/lambda/latest/dg/access-control-resource-based.html}{resource-based policy} from a Lambda resource.

See \url{https://www.paws-r-sdk.com/docs/lambda_delete_resource_policy/} for full documentation.
}
\keyword{internal}
