% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_account_settings}
\alias{lambda_get_account_settings}
\title{Retrieves details about your account's limits and usage in an Amazon Web
Services Region}
\usage{
lambda_get_account_settings()
}
\description{
Retrieves details about your account's \href{https://docs.aws.amazon.com/lambda/latest/dg/gettingstarted-limits.html}{limits} and usage in an Amazon Web Services Region.

See \url{https://paws-r.github.io/docs/lambda/get_account_settings.html} for full documentation.
}
\keyword{internal}
