% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_alias}
\alias{lambda_get_alias}
\title{Returns details about a Lambda function alias}
\usage{
lambda_get_alias(FunctionName, Name)
}
\arguments{
\item{FunctionName}{[required] The name of the Lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} -
\verb{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \verb{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only
the function name, it is limited to 64 characters in length.}

\item{Name}{[required] The name of the alias.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AliasArn = "string",
  Name = "string",
  FunctionVersion = "string",
  Description = "string",
  RoutingConfig = list(
    AdditionalVersionWeights = list(
      123.0
    )
  ),
  RevisionId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns details about a Lambda function
\href{https://docs.aws.amazon.com/lambda/latest/dg/configuration-versions.html}{alias}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_alias(
  FunctionName = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example returns details about an alias named BLUE for a
# function named my-function
svc$get_alias(
  FunctionName = "my-function",
  Name = "BLUE"
)
}

}
\keyword{internal}
