% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_capacity_provider}
\alias{lambda_get_capacity_provider}
\title{Retrieves information about a specific capacity provider, including its
configuration, state, and associated resources}
\usage{
lambda_get_capacity_provider(CapacityProviderName)
}
\arguments{
\item{CapacityProviderName}{[required] The name of the capacity provider to retrieve.}
}
\description{
Retrieves information about a specific capacity provider, including its configuration, state, and associated resources.

See \url{https://www.paws-r-sdk.com/docs/lambda_get_capacity_provider/} for full documentation.
}
\keyword{internal}
