% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_function_recursion_config}
\alias{lambda_get_function_recursion_config}
\title{Returns your function's recursive loop detection configuration}
\usage{
lambda_get_function_recursion_config(FunctionName)
}
\arguments{
\item{FunctionName}{[required] The name of the function.}
}
\description{
Returns your function's \href{https://docs.aws.amazon.com/lambda/latest/dg/invocation-recursion.html}{recursive loop detection} configuration.

See \url{https://www.paws-r-sdk.com/docs/lambda_get_function_recursion_config/} for full documentation.
}
\keyword{internal}
