% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_function_scaling_config}
\alias{lambda_get_function_scaling_config}
\title{Retrieves the scaling configuration for a Lambda Managed Instances
function}
\usage{
lambda_get_function_scaling_config(FunctionName, Qualifier)
}
\arguments{
\item{FunctionName}{[required] The name or ARN of the Lambda function.}

\item{Qualifier}{[required] Specify a version or alias to get the scaling configuration for a
published version of the function.}
}
\description{
Retrieves the scaling configuration for a Lambda Managed Instances function.

See \url{https://www.paws-r-sdk.com/docs/lambda_get_function_scaling_config/} for full documentation.
}
\keyword{internal}
