% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_function_url_config}
\alias{lambda_get_function_url_config}
\title{Returns details about a Lambda function URL}
\usage{
lambda_get_function_url_config(FunctionName, Qualifier = NULL)
}
\arguments{
\item{FunctionName}{[required] The name or ARN of the Lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} – \verb{my-function}.
\item \strong{Function ARN} –
\verb{arn:aws:lambda:us-west-2:123456789012:function:my-function}.
\item \strong{Partial ARN} – \verb{123456789012:function:my-function}.
}

The length constraint applies only to the full ARN. If you specify only
the function name, it is limited to 64 characters in length.}

\item{Qualifier}{The alias name.}
}
\description{
Returns details about a Lambda function URL.

See \url{https://www.paws-r-sdk.com/docs/lambda_get_function_url_config/} for full documentation.
}
\keyword{internal}
