% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_resource_policy}
\alias{lambda_get_resource_policy}
\title{Retrieves the resource-based policy attached to a Lambda resource}
\usage{
lambda_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Lambda resource you want to retrieve the policy for. You can use a qualified or an unqualified ARN. The value must be a complete ARN, and the operation does not accept wildcard characters.}
}
\description{
Retrieves the \href{https://docs.aws.amazon.com/lambda/latest/dg/access-control-resource-based.html}{resource-based policy} attached to a Lambda resource.

See \url{https://www.paws-r-sdk.com/docs/lambda_get_resource_policy/} for full documentation.
}
\keyword{internal}
