% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_list_aliases}
\alias{lambda_list_aliases}
\title{Returns a list of aliases for a Lambda function}
\usage{
lambda_list_aliases(FunctionName, FunctionVersion, Marker, MaxItems)
}
\arguments{
\item{FunctionName}{[required] The name of the Lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \code{MyFunction}.
\item \strong{Function ARN} -
\verb{arn:aws:lambda:us-west-2:123456789012:function:MyFunction}.
\item \strong{Partial ARN} - \verb{123456789012:function:MyFunction}.
}

The length constraint applies only to the full ARN. If you specify only
the function name, it is limited to 64 characters in length.}

\item{FunctionVersion}{Specify a function version to only list aliases that invoke that
version.}

\item{Marker}{Specify the pagination token that's returned by a previous request to
retrieve the next page of results.}

\item{MaxItems}{Limit the number of aliases returned.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NextMarker = "string",
  Aliases = list(
    list(
      AliasArn = "string",
      Name = "string",
      FunctionVersion = "string",
      Description = "string",
      RoutingConfig = list(
        AdditionalVersionWeights = list(
          123.0
        )
      ),
      RevisionId = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of
\href{https://docs.aws.amazon.com/lambda/latest/dg/configuration-versions.html}{aliases}
for a Lambda function.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_aliases(
  FunctionName = "string",
  FunctionVersion = "string",
  Marker = "string",
  MaxItems = 123
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example returns a list of aliases for a function named
# my-function.
svc$list_aliases(
  FunctionName = "my-function"
)
}

}
\keyword{internal}
