% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_list_capacity_providers}
\alias{lambda_list_capacity_providers}
\title{Returns a list of capacity providers in your account}
\usage{
lambda_list_capacity_providers(State = NULL, Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{State}{Filter capacity providers by their current state.}

\item{Marker}{Specify the pagination token that's returned by a previous request to
retrieve the next page of results.}

\item{MaxItems}{The maximum number of capacity providers to return.}
}
\description{
Returns a list of capacity providers in your account.

See \url{https://www.paws-r-sdk.com/docs/lambda_list_capacity_providers/} for full documentation.
}
\keyword{internal}
