% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_list_code_signing_configs}
\alias{lambda_list_code_signing_configs}
\title{Returns a list of code signing configurations}
\usage{
lambda_list_code_signing_configs(Marker, MaxItems)
}
\arguments{
\item{Marker}{Specify the pagination token that's returned by a previous request to
retrieve the next page of results.}

\item{MaxItems}{Maximum number of items to return.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextMarker = "string",
  CodeSigningConfigs = list(
    list(
      CodeSigningConfigId = "string",
      CodeSigningConfigArn = "string",
      Description = "string",
      AllowedPublishers = list(
        SigningProfileVersionArns = list(
          "string"
        )
      ),
      CodeSigningPolicies = list(
        UntrustedArtifactOnDeployment = "Warn"|"Enforce"
      ),
      LastModified = "string"
    )
  )
)
}
}
\description{
Returns a list of \href{https://docs.aws.amazon.com/lambda/latest/dg/}{code signing configurations}. A
request returns up to 10,000 configurations per call. You can use the
\code{MaxItems} parameter to return fewer configurations per call.
}
\section{Request syntax}{
\preformatted{svc$list_code_signing_configs(
  Marker = "string",
  MaxItems = 123
)
}
}

\keyword{internal}
