% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_list_durable_executions_by_function}
\alias{lambda_list_durable_executions_by_function}
\title{Returns a list of durable executions for a specified Lambda function}
\usage{
lambda_list_durable_executions_by_function(
  FunctionName,
  Qualifier = NULL,
  DurableExecutionName = NULL,
  Statuses = NULL,
  StartedAfter = NULL,
  StartedBefore = NULL,
  ReverseOrder = NULL,
  Marker = NULL,
  MaxItems = NULL
)
}
\arguments{
\item{FunctionName}{[required] The name or ARN of the Lambda function. You can specify a function name, a partial ARN, or a full ARN.}

\item{Qualifier}{The function version or alias. If not specified, lists executions for the $LATEST version.}

\item{DurableExecutionName}{Filter executions by name. Only executions with names that matches this string are returned.}

\item{Statuses}{Filter executions by status. Valid values: RUNNING, SUCCEEDED, FAILED, TIMED_OUT, STOPPED.}

\item{StartedAfter}{Filter executions that started after this timestamp (ISO 8601 format).}

\item{StartedBefore}{Filter executions that started before this timestamp (ISO 8601 format).}

\item{ReverseOrder}{Set to true to return results in chronological order (oldest first). Default is false.}

\item{Marker}{Pagination token from a previous request to continue retrieving results.}

\item{MaxItems}{Maximum number of executions to return (1-1000). Default is 100.}
}
\description{
Returns a list of \href{https://docs.aws.amazon.com/lambda/latest/dg/durable-functions.html}{durable executions} for a specified Lambda function. You can filter the results by execution name, status, and start time range. This API supports pagination for large result sets.

See \url{https://www.paws-r-sdk.com/docs/lambda_list_durable_executions_by_function/} for full documentation.
}
\keyword{internal}
