% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_list_function_versions_by_capacity_provider}
\alias{lambda_list_function_versions_by_capacity_provider}
\title{Returns a list of function versions that are configured to use a
specific capacity provider}
\usage{
lambda_list_function_versions_by_capacity_provider(
  CapacityProviderName,
  Marker = NULL,
  MaxItems = NULL
)
}
\arguments{
\item{CapacityProviderName}{[required] The name of the capacity provider to list function versions for.}

\item{Marker}{Specify the pagination token that's returned by a previous request to
retrieve the next page of results.}

\item{MaxItems}{The maximum number of function versions to return in the response.}
}
\description{
Returns a list of function versions that are configured to use a specific capacity provider.

See \url{https://www.paws-r-sdk.com/docs/lambda_list_function_versions_by_capacity_provider/} for full documentation.
}
\keyword{internal}
