% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_put_provisioned_concurrency_config}
\alias{lambda_put_provisioned_concurrency_config}
\title{Adds a provisioned concurrency configuration to a function's alias or
version}
\usage{
lambda_put_provisioned_concurrency_config(FunctionName, Qualifier,
  ProvisionedConcurrentExecutions)
}
\arguments{
\item{FunctionName}{[required] The name of the Lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \verb{my-function}.
\item \strong{Function ARN} -
\verb{arn:aws:lambda:us-west-2:123456789012:function:my-function}.
\item \strong{Partial ARN} - \verb{123456789012:function:my-function}.
}

The length constraint applies only to the full ARN. If you specify only
the function name, it is limited to 64 characters in length.}

\item{Qualifier}{[required] The version number or alias name.}

\item{ProvisionedConcurrentExecutions}{[required] The amount of provisioned concurrency to allocate for the version or
alias.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RequestedProvisionedConcurrentExecutions = 123,
  AvailableProvisionedConcurrentExecutions = 123,
  AllocatedProvisionedConcurrentExecutions = 123,
  Status = "IN_PROGRESS"|"READY"|"FAILED",
  StatusReason = "string",
  LastModified = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Adds a provisioned concurrency configuration to a function's alias or
version.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_provisioned_concurrency_config(
  FunctionName = "string",
  Qualifier = "string",
  ProvisionedConcurrentExecutions = 123
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example allocates 100 provisioned concurrency for the BLUE
# alias of the specified function.
svc$put_provisioned_concurrency_config(
  FunctionName = "my-function",
  ProvisionedConcurrentExecutions = 100L,
  Qualifier = "BLUE"
)
}

}
\keyword{internal}
