% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_put_resource_policy}
\alias{lambda_put_resource_policy}
\title{Adds a resource-based policy to a Lambda resource}
\usage{
lambda_put_resource_policy(ResourceArn, Policy, RevisionId = NULL)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Lambda resource you want to add the policy to. You can use a qualified or an unqualified ARN. The value must be a complete ARN, and the operation does not accept wildcard characters.}

\item{Policy}{[required] The policy document you want to add to your Lambda resource. This is formatted as a JSON string.

For more information, see \href{https://docs.aws.amazon.com/lambda/latest/dg/access-control-resource-based.html}{Working with resource-based policies in Lambda} in the \emph{Lambda Developer Guide}.}

\item{RevisionId}{The revision ID that the existing policy must match for the replacement to proceed. If the revision ID doesn't match, the operation fails with a \code{PreconditionFailedException} error. To retrieve the current revision ID, use the \code{\link[=lambda_get_resource_policy]{get_resource_policy}} operation.}
}
\description{
Adds a \href{https://docs.aws.amazon.com/lambda/latest/dg/access-control-resource-based.html}{resource-based policy} to a Lambda resource. Resource-based policies grant access to other \href{https://docs.aws.amazon.com/lambda/latest/dg/permissions-function-cross-account.html}{Amazon Web Services accounts}, \href{https://docs.aws.amazon.com/lambda/latest/dg/permissions-function-organization.html}{organizations}, or \href{https://docs.aws.amazon.com/lambda/latest/dg/permissions-function-services.html}{services}. Resource-based policies apply to a single Lambda resource (for example, a function, function version, or function alias).

See \url{https://www.paws-r-sdk.com/docs/lambda_put_resource_policy/} for full documentation.
}
\keyword{internal}
