% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_send_durable_execution_callback_heartbeat}
\alias{lambda_send_durable_execution_callback_heartbeat}
\title{Sends a heartbeat signal for a long-running callback operation to
prevent timeout}
\usage{
lambda_send_durable_execution_callback_heartbeat(CallbackId)
}
\arguments{
\item{CallbackId}{[required] The unique identifier for the callback operation.}
}
\description{
Sends a heartbeat signal for a long-running callback operation to prevent timeout. Use this API to extend the callback timeout period while the external operation is still in progress.

See \url{https://www.paws-r-sdk.com/docs/lambda_send_durable_execution_callback_heartbeat/} for full documentation.
}
\keyword{internal}
