% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_send_durable_execution_callback_success}
\alias{lambda_send_durable_execution_callback_success}
\title{Sends a successful completion response for a callback operation in a
durable execution}
\usage{
lambda_send_durable_execution_callback_success(CallbackId, Result = NULL)
}
\arguments{
\item{CallbackId}{[required] The unique identifier for the callback operation.}

\item{Result}{The result data from the successful callback operation. Maximum size is
256 KB.}
}
\description{
Sends a successful completion response for a callback operation in a durable execution. Use this API when an external system has successfully completed a callback operation.

See \url{https://www.paws-r-sdk.com/docs/lambda_send_durable_execution_callback_success/} for full documentation.
}
\keyword{internal}
