% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_update_code_signing_config}
\alias{lambda_update_code_signing_config}
\title{Update the code signing configuration}
\usage{
lambda_update_code_signing_config(CodeSigningConfigArn, Description,
  AllowedPublishers, CodeSigningPolicies)
}
\arguments{
\item{CodeSigningConfigArn}{[required] The The Amazon Resource Name (ARN) of the code signing configuration.}

\item{Description}{Descriptive name for this code signing configuration.}

\item{AllowedPublishers}{Signing profiles for this code signing configuration.}

\item{CodeSigningPolicies}{The code signing policy.}
}
\description{
Update the code signing configuration. Changes to the code signing
configuration take effect the next time a user tries to deploy a code
package to the function.
}
\section{Request syntax}{
\preformatted{svc$update_code_signing_config(
  CodeSigningConfigArn = "string",
  Description = "string",
  AllowedPublishers = list(
    SigningProfileVersionArns = list(
      "string"
    )
  ),
  CodeSigningPolicies = list(
    UntrustedArtifactOnDeployment = "Warn"|"Enforce"
  )
)
}
}

\keyword{internal}
