% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_allocate_static_ip}
\alias{lightsail_allocate_static_ip}
\title{Allocates a static IP address}
\usage{
lightsail_allocate_static_ip(staticIpName)
}
\arguments{
\item{staticIpName}{[required] The name of the static IP address.}
}
\description{
Allocates a static IP address.

See \url{https://www.paws-r-sdk.com/docs/lightsail_allocate_static_ip/} for full documentation.
}
\keyword{internal}
