% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_attach_static_ip}
\alias{lightsail_attach_static_ip}
\title{Attaches a static IP address to a specific Amazon Lightsail instance}
\usage{
lightsail_attach_static_ip(staticIpName, instanceName)
}
\arguments{
\item{staticIpName}{[required] The name of the static IP.}

\item{instanceName}{[required] The instance name to which you want to attach the static IP address.}
}
\description{
Attaches a static IP address to a specific Amazon Lightsail instance.

See \url{https://www.paws-r-sdk.com/docs/lightsail_attach_static_ip/} for full documentation.
}
\keyword{internal}
