% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_create_gui_session_access_details}
\alias{lightsail_create_gui_session_access_details}
\title{Creates two URLs that are used to access a virtual computer’s graphical
user interface (GUI) session}
\usage{
lightsail_create_gui_session_access_details(resourceName)
}
\arguments{
\item{resourceName}{[required] The resource name.}
}
\description{
Creates two URLs that are used to access a virtual computer’s graphical user interface (GUI) session. The primary URL initiates a web-based NICE DCV session to the virtual computer's application. The secondary URL initiates a web-based NICE DCV session to the virtual computer's operating session.

See \url{https://www.paws-r-sdk.com/docs/lightsail_create_gui_session_access_details/} for full documentation.
}
\keyword{internal}
