% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_bucket_access_key}
\alias{lightsail_delete_bucket_access_key}
\title{Deletes an access key for the specified Amazon Lightsail bucket}
\usage{
lightsail_delete_bucket_access_key(bucketName, accessKeyId)
}
\arguments{
\item{bucketName}{[required] The name of the bucket that the access key belongs to.}

\item{accessKeyId}{[required] The ID of the access key to delete.

Use the \code{\link[=lightsail_get_bucket_access_keys]{get_bucket_access_keys}}
action to get a list of access key IDs that you can specify.}
}
\description{
Deletes an access key for the specified Amazon Lightsail bucket.

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_bucket_access_key/} for full documentation.
}
\keyword{internal}
