% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_certificate}
\alias{lightsail_delete_certificate}
\title{Deletes an SSL/TLS certificate for your Amazon Lightsail content
delivery network (CDN) distribution}
\usage{
lightsail_delete_certificate(certificateName)
}
\arguments{
\item{certificateName}{[required] The name of the certificate to delete.

Use the \code{GetCertificates} action to get a list of certificate names that
you can specify.}
}
\description{
Deletes an SSL/TLS certificate for your Amazon Lightsail content
delivery network (CDN) distribution.

Certificates that are currently attached to a distribution cannot be
deleted. Use the \code{DetachCertificateFromDistribution} action to detach a
certificate from a distribution.
}
\section{Request syntax}{
\preformatted{svc$delete_certificate(
  certificateName = "string"
)
}
}

\keyword{internal}
