% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_container_service}
\alias{lightsail_delete_container_service}
\title{Deletes your Amazon Lightsail container service}
\usage{
lightsail_delete_container_service(serviceName)
}
\arguments{
\item{serviceName}{[required] The name of the container service to delete.}
}
\description{
Deletes your Amazon Lightsail container service.

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_container_service/} for full documentation.
}
\keyword{internal}
