% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_distribution}
\alias{lightsail_delete_distribution}
\title{Deletes your Amazon Lightsail content delivery network (CDN)
distribution}
\usage{
lightsail_delete_distribution(distributionName = NULL)
}
\arguments{
\item{distributionName}{The name of the distribution to delete.

Use the \code{\link[=lightsail_get_distributions]{get_distributions}} action to get
a list of distribution names that you can specify.}
}
\description{
Deletes your Amazon Lightsail content delivery network (CDN) distribution.

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_distribution/} for full documentation.
}
\keyword{internal}
