% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_domain}
\alias{lightsail_delete_domain}
\title{Deletes the specified domain recordset and all of its domain records}
\usage{
lightsail_delete_domain(domainName)
}
\arguments{
\item{domainName}{[required] The specific domain name to delete.}
}
\description{
Deletes the specified domain recordset and all of its domain records.
}
\details{
The \verb{delete domain} operation supports tag-based access control via
resource tags applied to the resource identified by \verb{domain name}. For
more information, see the Lightsail Dev Guide.
}
\section{Request syntax}{
\preformatted{svc$delete_domain(
  domainName = "string"
)
}
}

\keyword{internal}
