% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_instance_snapshot}
\alias{lightsail_delete_instance_snapshot}
\title{Deletes a specific snapshot of a virtual private server (or instance)}
\usage{
lightsail_delete_instance_snapshot(instanceSnapshotName)
}
\arguments{
\item{instanceSnapshotName}{[required] The name of the snapshot to delete.}
}
\description{
Deletes a specific snapshot of a virtual private server (or \emph{instance}).

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_instance_snapshot/} for full documentation.
}
\keyword{internal}
