% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_key_pair}
\alias{lightsail_delete_key_pair}
\title{Deletes a specific SSH key pair}
\usage{
lightsail_delete_key_pair(keyPairName)
}
\arguments{
\item{keyPairName}{[required] The name of the key pair to delete.}
}
\description{
Deletes a specific SSH key pair.
}
\details{
The \verb{delete key pair} operation supports tag-based access control via
resource tags applied to the resource identified by \verb{key pair name}. For
more information, see the Lightsail Dev Guide.
}
\section{Request syntax}{
\preformatted{svc$delete_key_pair(
  keyPairName = "string"
)
}
}

\keyword{internal}
