% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_relational_database_snapshot}
\alias{lightsail_delete_relational_database_snapshot}
\title{Deletes a database snapshot in Amazon Lightsail}
\usage{
lightsail_delete_relational_database_snapshot(
  relationalDatabaseSnapshotName)
}
\arguments{
\item{relationalDatabaseSnapshotName}{[required] The name of the database snapshot that you are deleting.}
}
\description{
Deletes a database snapshot in Amazon Lightsail.
}
\details{
The \code{delete relational database snapshot} operation supports tag-based
access control via resource tags applied to the resource identified by
relationalDatabaseName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\section{Request syntax}{
\preformatted{svc$delete_relational_database_snapshot(
  relationalDatabaseSnapshotName = "string"
)
}
}

\keyword{internal}
