% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_disable_add_on}
\alias{lightsail_disable_add_on}
\title{Disables an add-on for an Amazon Lightsail resource}
\usage{
lightsail_disable_add_on(addOnType, resourceName)
}
\arguments{
\item{addOnType}{[required] The add-on type to disable.}

\item{resourceName}{[required] The name of the source resource for which to disable the add-on.}
}
\description{
Disables an add-on for an Amazon Lightsail resource. For more
information, see the Lightsail Dev Guide.
}
\section{Request syntax}{
\preformatted{svc$disable_add_on(
  addOnType = "AutoSnapshot",
  resourceName = "string"
)
}
}

\keyword{internal}
