% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_download_default_key_pair}
\alias{lightsail_download_default_key_pair}
\title{Downloads the regional Amazon Lightsail default key pair}
\usage{
lightsail_download_default_key_pair()
}
\description{
Downloads the regional Amazon Lightsail default key pair.

See \url{https://paws-r.github.io/docs/lightsail/download_default_key_pair.html} for full documentation.
}
\keyword{internal}
