% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_enable_add_on}
\alias{lightsail_enable_add_on}
\title{Enables or modifies an add-on for an Amazon Lightsail resource}
\usage{
lightsail_enable_add_on(resourceName, addOnRequest)
}
\arguments{
\item{resourceName}{[required] The name of the source resource for which to enable or modify the
add-on.}

\item{addOnRequest}{[required] An array of strings representing the add-on to enable or modify.}
}
\description{
Enables or modifies an add-on for an Amazon Lightsail resource. For more information, see the Amazon Lightsail Developer Guide.

See \url{https://paws-r.github.io/docs/lightsail/enable_add_on.html} for full documentation.
}
\keyword{internal}
