% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_active_names}
\alias{lightsail_get_active_names}
\title{Returns the names of all active (not deleted) resources}
\usage{
lightsail_get_active_names(pageToken)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_active_names]{get_active_names}} request. If your
results are paginated, the response will return a next page token that
you can specify as the page token in a subsequent request.}
}
\description{
Returns the names of all active (not deleted) resources.
}
\section{Request syntax}{
\preformatted{svc$get_active_names(
  pageToken = "string"
)
}
}

\keyword{internal}
