% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_bucket_access_keys}
\alias{lightsail_get_bucket_access_keys}
\title{Returns the existing access key IDs for the specified Amazon Lightsail
bucket}
\usage{
lightsail_get_bucket_access_keys(bucketName)
}
\arguments{
\item{bucketName}{[required] The name of the bucket for which to return access keys.}
}
\description{
Returns the existing access key IDs for the specified Amazon Lightsail bucket.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_bucket_access_keys/} for full documentation.
}
\keyword{internal}
