% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_bucket_bundles}
\alias{lightsail_get_bucket_bundles}
\title{Returns the bundles that you can apply to a Amazon Lightsail bucket}
\usage{
lightsail_get_bucket_bundles(includeInactive = NULL)
}
\arguments{
\item{includeInactive}{A Boolean value that indicates whether to include inactive (unavailable)
bundles in the response of your request.}
}
\description{
Returns the bundles that you can apply to a Amazon Lightsail bucket.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_bucket_bundles/} for full documentation.
}
\keyword{internal}
