% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_container_api_metadata}
\alias{lightsail_get_container_api_metadata}
\title{Returns information about Amazon Lightsail containers, such as the
current version of the Lightsail Control (lightsailctl) plugin}
\usage{
lightsail_get_container_api_metadata()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  metadata = list(
    list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about Amazon Lightsail containers, such as the
current version of the Lightsail Control (lightsailctl) plugin.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_container_api_metadata()
}\if{html}{\out{</div>}}
}

\keyword{internal}
