% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_container_images}
\alias{lightsail_get_container_images}
\title{Returns the container images that are registered to your Amazon
Lightsail container service}
\usage{
lightsail_get_container_images(serviceName)
}
\arguments{
\item{serviceName}{[required] The name of the container service for which to return registered
container images.}
}
\value{
A list with the following syntax:\preformatted{list(
  containerImages = list(
    list(
      image = "string",
      digest = "string",
      createdAt = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Returns the container images that are registered to your Amazon
Lightsail container service.

If you created a deployment on your Lightsail container service that
uses container images from a public registry like Docker Hub, those
images are not returned as part of this action. Those images are not
registered to your Lightsail container service.
}
\section{Request syntax}{
\preformatted{svc$get_container_images(
  serviceName = "string"
)
}
}

\keyword{internal}
