% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_container_service_powers}
\alias{lightsail_get_container_service_powers}
\title{Returns the list of powers that can be specified for your Amazon
Lightsail container services}
\usage{
lightsail_get_container_service_powers()
}
\description{
Returns the list of powers that can be specified for your Amazon Lightsail container services.

See \url{https://paws-r.github.io/docs/lightsail/get_container_service_powers.html} for full documentation.
}
\keyword{internal}
