% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_disk}
\alias{lightsail_get_disk}
\title{Returns information about a specific block storage disk}
\usage{
lightsail_get_disk(diskName)
}
\arguments{
\item{diskName}{[required] The name of the disk (\code{my-disk}).}
}
\description{
Returns information about a specific block storage disk.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_disk/} for full documentation.
}
\keyword{internal}
