% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_disks}
\alias{lightsail_get_disks}
\title{Returns information about all block storage disks in your AWS account
and region}
\usage{
lightsail_get_disks(pageToken)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_disks]{get_disks}} request. If your results are
paginated, the response will return a next page token that you can
specify as the page token in a subsequent request.}
}
\description{
Returns information about all block storage disks in your AWS account
and region.
}
\section{Request syntax}{
\preformatted{svc$get_disks(
  pageToken = "string"
)
}
}

\keyword{internal}
