% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_distribution_bundles}
\alias{lightsail_get_distribution_bundles}
\title{Returns the list bundles that can be applied to you Amazon Lightsail
content delivery network (CDN) distributions}
\usage{
lightsail_get_distribution_bundles()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  bundles = list(
    list(
      bundleId = "string",
      name = "string",
      price = 123.0,
      transferPerMonthInGb = 123,
      isActive = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the list bundles that can be applied to you Amazon Lightsail
content delivery network (CDN) distributions.

A distribution bundle specifies the monthly network transfer quota and
monthly cost of your dsitribution.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_distribution_bundles()
}\if{html}{\out{</div>}}
}

\keyword{internal}
