% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_load_balancer}
\alias{lightsail_get_load_balancer}
\title{Returns information about the specified Lightsail load balancer}
\usage{
lightsail_get_load_balancer(loadBalancerName)
}
\arguments{
\item{loadBalancerName}{[required] The name of the load balancer.}
}
\description{
Returns information about the specified Lightsail load balancer.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_load_balancer/} for full documentation.
}
\keyword{internal}
