% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_load_balancers}
\alias{lightsail_get_load_balancers}
\title{Returns information about all load balancers in an account}
\usage{
lightsail_get_load_balancers(pageToken)
}
\arguments{
\item{pageToken}{A token used for paginating the results from your GetLoadBalancers
request.}
}
\description{
Returns information about all load balancers in an account.
}
\details{
If you are describing a long list of load balancers, you can paginate
the output to make the list more manageable. You can use the pageToken
and nextPageToken values to retrieve the next items in the list.
}
\section{Request syntax}{
\preformatted{svc$get_load_balancers(
  pageToken = "string"
)
}
}

\keyword{internal}
