% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_operation}
\alias{lightsail_get_operation}
\title{Returns information about a specific operation}
\usage{
lightsail_get_operation(operationId)
}
\arguments{
\item{operationId}{[required] A GUID used to identify the operation.}
}
\description{
Returns information about a specific operation. Operations include events such as when you create an instance, allocate a static IP, attach a static IP, and so on.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_operation/} for full documentation.
}
\keyword{internal}
