% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_profile}
\alias{lightsail_get_profile}
\title{Returns information about the profile of the Amazon Lightsail account
that makes the request}
\usage{
lightsail_get_profile()
}
\description{
Returns information about the profile of the Amazon Lightsail account that makes the request. The response includes the profile type and, for accounts enrolled in the Lightsail partner program, the partner membership details.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_profile/} for full documentation.
}
\keyword{internal}
