% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_database_log_events}
\alias{lightsail_get_relational_database_log_events}
\title{Returns a list of log events for a database in Amazon Lightsail}
\usage{
lightsail_get_relational_database_log_events(
  relationalDatabaseName,
  logStreamName,
  startTime = NULL,
  endTime = NULL,
  startFromHead = NULL,
  pageToken = NULL
)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database for which to get log events.}

\item{logStreamName}{[required] The name of the log stream.

Use the \verb{get relational database log streams} operation to get a list of
available log streams.}

\item{startTime}{The start of the time interval from which to get log events.

Constraints:
\itemize{
\item Specified in Coordinated Universal Time (UTC).
\item Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at
8 PM UTC, then you input \code{1538424000} as the start time.
}}

\item{endTime}{The end of the time interval from which to get log events.

Constraints:
\itemize{
\item Specified in Coordinated Universal Time (UTC).
\item Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 8
PM UTC, then you input \code{1538424000} as the end time.
}}

\item{startFromHead}{Parameter to specify if the log should start from head or tail. If
\code{true} is specified, the log event starts from the head of the log. If
\code{false} is specified, the log event starts from the tail of the log.

For PostgreSQL, the default value of \code{false} is the only option
available.}

\item{pageToken}{The token to advance to the next or previous page of results from your
request.

To get a page token, perform an initial
\code{\link[=lightsail_get_relational_database_log_events]{get_relational_database_log_events}}
request. If your results are paginated, the response will return a next
forward token and/or next backward token that you can specify as the
page token in a subsequent request.}
}
\description{
Returns a list of log events for a database in Amazon Lightsail.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_relational_database_log_events/} for full documentation.
}
\keyword{internal}
