% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_setup_history}
\alias{lightsail_get_setup_history}
\title{Returns detailed information for five of the most recent
SetupInstanceHttps requests that were ran on the target instance}
\usage{
lightsail_get_setup_history(resourceName, pageToken = NULL)
}
\arguments{
\item{resourceName}{[required] The name of the resource for which you are requesting information.}

\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_setup_history]{get_setup_history}} request. If your
results are paginated, the response will return a next page token that
you can specify as the page token in a subsequent request.}
}
\description{
Returns detailed information for five of the most recent \code{\link[=lightsail_setup_instance_https]{setup_instance_https}} requests that were ran on the target instance.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_setup_history/} for full documentation.
}
\keyword{internal}
