% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_is_vpc_peered}
\alias{lightsail_is_vpc_peered}
\title{Returns a Boolean value indicating whether your Lightsail VPC is peered}
\usage{
lightsail_is_vpc_peered()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  isPeered = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
Returns a Boolean value indicating whether your Lightsail VPC is peered.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$is_vpc_peered()
}\if{html}{\out{</div>}}
}

\keyword{internal}
