% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_reboot_instance}
\alias{lightsail_reboot_instance}
\title{Restarts a specific instance}
\usage{
lightsail_reboot_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the instance to reboot.}
}
\description{
Restarts a specific instance.

The \verb{reboot instance} operation supports tag-based access control via
resource tags applied to the resource identified by \verb{instance name}. For
more information, see the Lightsail Dev Guide.
}
\section{Request syntax}{
\preformatted{svc$reboot_instance(
  instanceName = "string"
)
}
}

\keyword{internal}
