% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_reboot_relational_database}
\alias{lightsail_reboot_relational_database}
\title{Restarts a specific database in Amazon Lightsail}
\usage{
lightsail_reboot_relational_database(relationalDatabaseName)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database to reboot.}
}
\description{
Restarts a specific database in Amazon Lightsail.

See \url{https://www.paws-r-sdk.com/docs/lightsail_reboot_relational_database/} for full documentation.
}
\keyword{internal}
