% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_reset_distribution_cache}
\alias{lightsail_reset_distribution_cache}
\title{Deletes currently cached content from your Amazon Lightsail content
delivery network (CDN) distribution}
\usage{
lightsail_reset_distribution_cache(distributionName)
}
\arguments{
\item{distributionName}{The name of the distribution for which to reset cache.

Use the \code{GetDistributions} action to get a list of distribution names
that you can specify.}
}
\description{
Deletes currently cached content from your Amazon Lightsail content
delivery network (CDN) distribution.

After resetting the cache, the next time a content request is made, your
distribution pulls, serves, and caches it from the origin.
}
\section{Request syntax}{
\preformatted{svc$reset_distribution_cache(
  distributionName = "string"
)
}
}

\keyword{internal}
