% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_setup_instance_https}
\alias{lightsail_setup_instance_https}
\title{Creates an SSL/TLS certificate that secures traffic for your website}
\usage{
lightsail_setup_instance_https(
  instanceName,
  emailAddress,
  domainNames,
  certificateProvider
)
}
\arguments{
\item{instanceName}{[required] The name of the Lightsail instance.}

\item{emailAddress}{[required] The contact method for SSL/TLS certificate renewal alerts. You can enter
one email address.}

\item{domainNames}{[required] The name of the domain and subdomains that were specified for the
SSL/TLS certificate.}

\item{certificateProvider}{[required] The certificate authority that issues the SSL/TLS certificate.}
}
\description{
Creates an SSL/TLS certificate that secures traffic for your website. After the certificate is created, it is installed on the specified Lightsail instance.

See \url{https://www.paws-r-sdk.com/docs/lightsail_setup_instance_https/} for full documentation.
}
\keyword{internal}
