% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_start_gui_session}
\alias{lightsail_start_gui_session}
\title{Initiates a graphical user interface (GUI) session that’s used to access
a virtual computer’s operating system and application}
\usage{
lightsail_start_gui_session(resourceName)
}
\arguments{
\item{resourceName}{[required] The resource name.}
}
\description{
Initiates a graphical user interface (GUI) session that’s used to access a virtual computer’s operating system and application. The session will be active for 1 hour. Use this action to resume the session after it expires.

See \url{https://www.paws-r-sdk.com/docs/lightsail_start_gui_session/} for full documentation.
}
\keyword{internal}
