% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_stop_gui_session}
\alias{lightsail_stop_gui_session}
\title{Terminates a web-based NICE DCV session that’s used to access a virtual
computer’s operating system or application}
\usage{
lightsail_stop_gui_session(resourceName)
}
\arguments{
\item{resourceName}{[required] The resource name.}
}
\description{
Terminates a web-based NICE DCV session that’s used to access a virtual computer’s operating system or application. The session will close and any unsaved data will be lost.

See \url{https://www.paws-r-sdk.com/docs/lightsail_stop_gui_session/} for full documentation.
}
\keyword{internal}
