% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_stop_relational_database}
\alias{lightsail_stop_relational_database}
\title{Stops a specific database that is currently running in Amazon Lightsail}
\usage{
lightsail_stop_relational_database(
  relationalDatabaseName,
  relationalDatabaseSnapshotName = NULL
)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database to stop.}

\item{relationalDatabaseSnapshotName}{The name of your new database snapshot to be created before stopping
your database.}
}
\description{
Stops a specific database that is currently running in Amazon Lightsail.

See \url{https://paws-r.github.io/docs/lightsail/stop_relational_database.html} for full documentation.
}
\keyword{internal}
