% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_unpeer_vpc}
\alias{lightsail_unpeer_vpc}
\title{Attempts to unpeer the Lightsail VPC from the user's default VPC}
\usage{
lightsail_unpeer_vpc()
}
\description{
Attempts to unpeer the Lightsail VPC from the user\'s default VPC.
}
\section{Request syntax}{
\preformatted{svc$unpeer_vpc()
}
}

\keyword{internal}
