% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_update_bucket_bundle}
\alias{lightsail_update_bucket_bundle}
\title{Updates the bundle, or storage plan, of an existing Amazon Lightsail
bucket}
\usage{
lightsail_update_bucket_bundle(bucketName, bundleId)
}
\arguments{
\item{bucketName}{[required] The name of the bucket for which to update the bundle.}

\item{bundleId}{[required] The ID of the new bundle to apply to the bucket.

Use the \code{\link[=lightsail_get_bucket_bundles]{get_bucket_bundles}} action to
get a list of bundle IDs that you can specify.}
}
\description{
Updates the bundle, or storage plan, of an existing Amazon Lightsail bucket.

See \url{https://www.paws-r-sdk.com/docs/lightsail_update_bucket_bundle/} for full documentation.
}
\keyword{internal}
