% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_update_relational_database}
\alias{lightsail_update_relational_database}
\title{Allows the update of one or more attributes of a database in Amazon
Lightsail}
\usage{
lightsail_update_relational_database(relationalDatabaseName,
  masterUserPassword, rotateMasterUserPassword, preferredBackupWindow,
  preferredMaintenanceWindow, enableBackupRetention,
  disableBackupRetention, publiclyAccessible, applyImmediately,
  caCertificateIdentifier)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database to update.}

\item{masterUserPassword}{The password for the master user of your database. The password can
include any printable ASCII character except "/", """, or "@".

Constraints: Must contain 8 to 41 characters.}

\item{rotateMasterUserPassword}{When \code{true}, the master user password is changed to a new strong
password generated by Lightsail.

Use the \verb{get relational database master user password} operation to get
the new password.}

\item{preferredBackupWindow}{The daily time range during which automated backups are created for your
database if automated backups are enabled.

Constraints:
\itemize{
\item Must be in the \code{hh24:mi-hh24:mi} format.

Example: \code{16:00-16:30}
\item Specified in Coordinated Universal Time (UTC).
\item Must not conflict with the preferred maintenance window.
\item Must be at least 30 minutes.
}}

\item{preferredMaintenanceWindow}{The weekly time range during which system maintenance can occur on your
database.

The default is a 30-minute window selected at random from an 8-hour
block of time for each AWS Region, occurring on a random day of the
week.

Constraints:
\itemize{
\item Must be in the \code{ddd:hh24:mi-ddd:hh24:mi} format.
\item Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.
\item Must be at least 30 minutes.
\item Specified in Coordinated Universal Time (UTC).
\item Example: \code{Tue:17:00-Tue:17:30}
}}

\item{enableBackupRetention}{When \code{true}, enables automated backup retention for your database.

Updates are applied during the next maintenance window because this can
result in an outage.}

\item{disableBackupRetention}{When \code{true}, disables automated backup retention for your database.

Disabling backup retention deletes all automated database backups.
Before disabling this, you may want to create a snapshot of your
database using the \verb{create relational database snapshot} operation.

Updates are applied during the next maintenance window because this can
result in an outage.}

\item{publiclyAccessible}{Specifies the accessibility options for your database. A value of \code{true}
specifies a database that is available to resources outside of your
Lightsail account. A value of \code{false} specifies a database that is
available only to your Lightsail resources in the same region as your
database.}

\item{applyImmediately}{When \code{true}, applies changes immediately. When \code{false}, applies changes
during the preferred maintenance window. Some changes may cause an
outage.

Default: \code{false}}

\item{caCertificateIdentifier}{Indicates the certificate that needs to be associated with the database.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  operations = list(
    list(
      id = "string",
      resourceName = "string",
      resourceType = "ContainerService"|"Instance"|"StaticIp"|"KeyPair"|"InstanceSnapshot"|"Domain"|"PeeredVpc"|"LoadBalancer"|"LoadBalancerTlsCertificate"|"Disk"|"DiskSnapshot"|"RelationalDatabase"|"RelationalDatabaseSnapshot"|"ExportSnapshotRecord"|"CloudFormationStackRecord"|"Alarm"|"ContactMethod"|"Distribution"|"Certificate",
      createdAt = as.POSIXct(
        "2015-01-01"
      ),
      location = list(
        availabilityZone = "string",
        regionName = "us-east-1"|"us-east-2"|"us-west-1"|"us-west-2"|"eu-west-1"|"eu-west-2"|"eu-west-3"|"eu-central-1"|"ca-central-1"|"ap-south-1"|"ap-southeast-1"|"ap-southeast-2"|"ap-northeast-1"|"ap-northeast-2"
      ),
      isTerminal = TRUE|FALSE,
      operationDetails = "string",
      operationType = "DeleteKnownHostKeys"|"DeleteInstance"|"CreateInstance"|"StopInstance"|"StartInstance"|"RebootInstance"|"OpenInstancePublicPorts"|"PutInstancePublicPorts"|"CloseInstancePublicPorts"|"AllocateStaticIp"|"ReleaseStaticIp"|"AttachStaticIp"|"DetachStaticIp"|"UpdateDomainEntry"|"DeleteDomainEntry"|"CreateDomain"|"DeleteDomain"|"CreateInstanceSnapshot"|"DeleteInstanceSnapshot"|"CreateInstancesFromSnapshot"|"CreateLoadBalancer"|"DeleteLoadBalancer"|"AttachInstancesToLoadBalancer"|"DetachInstancesFromLoadBalancer"|"UpdateLoadBalancerAttribute"|"CreateLoadBalancerTlsCertificate"|"DeleteLoadBalancerTlsCertificate"|"AttachLoadBalancerTlsCertificate"|"CreateDisk"|"DeleteDisk"|"AttachDisk"|"DetachDisk"|"CreateDiskSnapshot"|"DeleteDiskSnapshot"|"CreateDiskFromSnapshot"|"CreateRelationalDatabase"|"UpdateRelationalDatabase"|"DeleteRelationalDatabase"|"CreateRelationalDatabaseFromSnapshot"|"CreateRelationalDatabaseSnapshot"|"DeleteRelationalDatabaseSnapshot"|"UpdateRelationalDatabaseParameters"|"StartRelationalDatabase"|"RebootRelationalDatabase"|"StopRelationalDatabase"|"EnableAddOn"|"DisableAddOn"|"PutAlarm"|"GetAlarms"|"DeleteAlarm"|"TestAlarm"|"CreateContactMethod"|"GetContactMethods"|"SendContactMethodVerification"|"DeleteContactMethod"|"CreateDistribution"|"UpdateDistribution"|"DeleteDistribution"|"ResetDistributionCache"|"AttachCertificateToDistribution"|"DetachCertificateFromDistribution"|"UpdateDistributionBundle"|"CreateCertificate"|"DeleteCertificate"|"CreateContainerService"|"UpdateContainerService"|"DeleteContainerService"|"CreateContainerServiceDeployment"|"CreateContainerServiceRegistryLogin"|"RegisterContainerImage"|"DeleteContainerImage",
      status = "NotStarted"|"Started"|"Failed"|"Completed"|"Succeeded",
      statusChangedAt = as.POSIXct(
        "2015-01-01"
      ),
      errorCode = "string",
      errorDetails = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Allows the update of one or more attributes of a database in Amazon
Lightsail.

Updates are applied immediately, or in cases where the updates could
result in an outage, are applied during the database's predefined
maintenance window.

The \verb{update relational database} operation supports tag-based access
control via resource tags applied to the resource identified by
relationalDatabaseName. For more information, see the Lightsail Dev
Guide.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_relational_database(
  relationalDatabaseName = "string",
  masterUserPassword = "string",
  rotateMasterUserPassword = TRUE|FALSE,
  preferredBackupWindow = "string",
  preferredMaintenanceWindow = "string",
  enableBackupRetention = TRUE|FALSE,
  disableBackupRetention = TRUE|FALSE,
  publiclyAccessible = TRUE|FALSE,
  applyImmediately = TRUE|FALSE,
  caCertificateIdentifier = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
