% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_accept_environment_account_connection}
\alias{proton_accept_environment_account_connection}
\title{In a management account, an environment account connection request is
accepted}
\usage{
proton_accept_environment_account_connection(id)
}
\arguments{
\item{id}{[required] The ID of the environment account connection.}
}
\description{
In a management account, an environment account connection request is accepted. When the environment account connection request is accepted, Proton can use the associated IAM role to provision environment infrastructure resources in the associated environment account.

See \url{https://www.paws-r-sdk.com/docs/proton_accept_environment_account_connection/} for full documentation.
}
\keyword{internal}
