% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_cancel_component_deployment}
\alias{proton_cancel_component_deployment}
\title{Attempts to cancel a component deployment (for a component that is in
the IN_PROGRESS deployment status)}
\usage{
proton_cancel_component_deployment(componentName)
}
\arguments{
\item{componentName}{[required] The name of the component with the deployment to cancel.}
}
\description{
Attempts to cancel a component deployment (for a component that is in the \code{IN_PROGRESS} deployment status).

See \url{https://www.paws-r-sdk.com/docs/proton_cancel_component_deployment/} for full documentation.
}
\keyword{internal}
