% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_cancel_service_pipeline_deployment}
\alias{proton_cancel_service_pipeline_deployment}
\title{Attempts to cancel a service pipeline deployment on an
UpdateServicePipeline action, if the deployment is IN_PROGRESS}
\usage{
proton_cancel_service_pipeline_deployment(serviceName)
}
\arguments{
\item{serviceName}{[required] The name of the service with the service pipeline deployment to cancel.}
}
\description{
Attempts to cancel a service pipeline deployment on an \code{\link[=proton_update_service_pipeline]{update_service_pipeline}} action, if the deployment is \code{IN_PROGRESS}. For more information, see \href{https://docs.aws.amazon.com/proton/latest/userguide/ag-svc-pipeline-update.html}{Update a service pipeline} in the \emph{Proton User guide}.

See \url{https://www.paws-r-sdk.com/docs/proton_cancel_service_pipeline_deployment/} for full documentation.
}
\keyword{internal}
