% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_create_service_sync_config}
\alias{proton_create_service_sync_config}
\title{Create the Proton Ops configuration file}
\usage{
proton_create_service_sync_config(
  branch,
  filePath,
  repositoryName,
  repositoryProvider,
  serviceName
)
}
\arguments{
\item{branch}{[required] The repository branch for your Proton Ops file.}

\item{filePath}{[required] The path to the Proton Ops file.}

\item{repositoryName}{[required] The repository name.}

\item{repositoryProvider}{[required] The provider type for your repository.}

\item{serviceName}{[required] The name of the service the Proton Ops file is for.}
}
\description{
Create the Proton Ops configuration file.

See \url{https://www.paws-r-sdk.com/docs/proton_create_service_sync_config/} for full documentation.
}
\keyword{internal}
