% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_delete_component}
\alias{proton_delete_component}
\title{Delete an Proton component resource}
\usage{
proton_delete_component(name)
}
\arguments{
\item{name}{[required] The name of the component to delete.}
}
\description{
Delete an Proton component resource.

See \url{https://www.paws-r-sdk.com/docs/proton_delete_component/} for full documentation.
}
\keyword{internal}
