% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_delete_environment_template}
\alias{proton_delete_environment_template}
\title{If no other major or minor versions of an environment template exist,
delete the environment template}
\usage{
proton_delete_environment_template(name)
}
\arguments{
\item{name}{[required] The name of the environment template to delete.}
}
\description{
If no other major or minor versions of an environment template exist, delete the environment template.

See \url{https://www.paws-r-sdk.com/docs/proton_delete_environment_template/} for full documentation.
}
\keyword{internal}
