% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_delete_environment_template_version}
\alias{proton_delete_environment_template_version}
\title{If no other minor versions of an environment template exist, delete a
major version of the environment template if it's not the Recommended
version}
\usage{
proton_delete_environment_template_version(
  majorVersion,
  minorVersion,
  templateName
)
}
\arguments{
\item{majorVersion}{[required] The environment template major version to delete.}

\item{minorVersion}{[required] The environment template minor version to delete.}

\item{templateName}{[required] The name of the environment template.}
}
\description{
If no other minor versions of an environment template exist, delete a major version of the environment template if it's not the \code{Recommended} version. Delete the \code{Recommended} version of the environment template if no other major versions or minor versions of the environment template exist. A major version of an environment template is a version that's not backward compatible.

See \url{https://www.paws-r-sdk.com/docs/proton_delete_environment_template_version/} for full documentation.
}
\keyword{internal}
