% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_delete_repository}
\alias{proton_delete_repository}
\title{De-register and unlink your repository}
\usage{
proton_delete_repository(name, provider)
}
\arguments{
\item{name}{[required] The repository name.}

\item{provider}{[required] The repository provider.}
}
\description{
De-register and unlink your repository.

See \url{https://www.paws-r-sdk.com/docs/proton_delete_repository/} for full documentation.
}
\keyword{internal}
