% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_delete_service_sync_config}
\alias{proton_delete_service_sync_config}
\title{Delete the Proton Ops file}
\usage{
proton_delete_service_sync_config(serviceName)
}
\arguments{
\item{serviceName}{[required] The name of the service that you want to delete the service sync
configuration for.}
}
\description{
Delete the Proton Ops file.

See \url{https://www.paws-r-sdk.com/docs/proton_delete_service_sync_config/} for full documentation.
}
\keyword{internal}
