% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_delete_service_template}
\alias{proton_delete_service_template}
\title{If no other major or minor versions of the service template exist,
delete the service template}
\usage{
proton_delete_service_template(name)
}
\arguments{
\item{name}{[required] The name of the service template to delete.}
}
\description{
If no other major or minor versions of the service template exist, delete the service template.

See \url{https://www.paws-r-sdk.com/docs/proton_delete_service_template/} for full documentation.
}
\keyword{internal}
