% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_component}
\alias{proton_get_component}
\title{Get detailed data for a component}
\usage{
proton_get_component(name)
}
\arguments{
\item{name}{[required] The name of the component that you want to get the detailed data for.}
}
\description{
Get detailed data for a component.

See \url{https://www.paws-r-sdk.com/docs/proton_get_component/} for full documentation.
}
\keyword{internal}
