% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_deployment}
\alias{proton_get_deployment}
\title{Get detailed data for a deployment}
\usage{
proton_get_deployment(
  componentName = NULL,
  environmentName = NULL,
  id,
  serviceInstanceName = NULL,
  serviceName = NULL
)
}
\arguments{
\item{componentName}{The name of a component that you want to get the detailed data for.}

\item{environmentName}{The name of a environment that you want to get the detailed data for.}

\item{id}{[required] The ID of the deployment that you want to get the detailed data for.}

\item{serviceInstanceName}{The name of the service instance associated with the given deployment
ID. \code{serviceName} must be specified to identify the service instance.}

\item{serviceName}{The name of the service associated with the given deployment ID.}
}
\description{
Get detailed data for a deployment.

See \url{https://www.paws-r-sdk.com/docs/proton_get_deployment/} for full documentation.
}
\keyword{internal}
