% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_environment_account_connection}
\alias{proton_get_environment_account_connection}
\title{In an environment account, get the detailed data for an environment
account connection}
\usage{
proton_get_environment_account_connection(id)
}
\arguments{
\item{id}{[required] The ID of the environment account connection that you want to get the
detailed data for.}
}
\description{
In an environment account, get the detailed data for an environment account connection.

See \url{https://www.paws-r-sdk.com/docs/proton_get_environment_account_connection/} for full documentation.
}
\keyword{internal}
