% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_environment_template}
\alias{proton_get_environment_template}
\title{Get detailed data for an environment template}
\usage{
proton_get_environment_template(name)
}
\arguments{
\item{name}{[required] The name of the environment template that you want to get the detailed
data for.}
}
\description{
Get detailed data for an environment template.

See \url{https://www.paws-r-sdk.com/docs/proton_get_environment_template/} for full documentation.
}
\keyword{internal}
