% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_environment_template_version}
\alias{proton_get_environment_template_version}
\title{Get detailed data for a major or minor version of an environment
template}
\usage{
proton_get_environment_template_version(
  majorVersion,
  minorVersion,
  templateName
)
}
\arguments{
\item{majorVersion}{[required] To get environment template major version detail data, include
\verb{major Version}.}

\item{minorVersion}{[required] To get environment template minor version detail data, include
\code{minorVersion}.}

\item{templateName}{[required] The name of the environment template a version of which you want to get
detailed data for.}
}
\description{
Get detailed data for a major or minor version of an environment template.

See \url{https://www.paws-r-sdk.com/docs/proton_get_environment_template_version/} for full documentation.
}
\keyword{internal}
