% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_repository}
\alias{proton_get_repository}
\title{Get detail data for a repository}
\usage{
proton_get_repository(name, provider)
}
\arguments{
\item{name}{[required] The repository name, for example \code{myrepos/myrepo}.}

\item{provider}{[required] The repository provider.}
}
\description{
Get detail data for a repository.

See \url{https://paws-r.github.io/docs/proton/get_repository.html} for full documentation.
}
\keyword{internal}
