% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_repository_sync_status}
\alias{proton_get_repository_sync_status}
\title{Get the sync status of a repository used for Proton template sync}
\usage{
proton_get_repository_sync_status(
  branch,
  repositoryName,
  repositoryProvider,
  syncType
)
}
\arguments{
\item{branch}{[required] The repository branch.}

\item{repositoryName}{[required] The repository name.}

\item{repositoryProvider}{[required] The repository provider.}

\item{syncType}{[required] The repository sync type.}
}
\description{
Get the sync status of a repository used for Proton template sync. For more information about template sync, see .

See \url{https://www.paws-r-sdk.com/docs/proton_get_repository_sync_status/} for full documentation.
}
\keyword{internal}
