% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_resources_summary}
\alias{proton_get_resources_summary}
\title{Get counts of Proton resources}
\usage{
proton_get_resources_summary()
}
\description{
Get counts of Proton resources.

See \url{https://paws-r.github.io/docs/proton/get_resources_summary.html} for full documentation.
}
\keyword{internal}
