% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_service_instance}
\alias{proton_get_service_instance}
\title{Get detailed data for a service instance}
\usage{
proton_get_service_instance(name, serviceName)
}
\arguments{
\item{name}{[required] The name of a service instance that you want to get the detailed data
for.}

\item{serviceName}{[required] The name of the service that you want the service instance input for.}
}
\description{
Get detailed data for a service instance. A service instance is an instantiation of service template and it runs in a specific environment.

See \url{https://www.paws-r-sdk.com/docs/proton_get_service_instance/} for full documentation.
}
\keyword{internal}
