% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_service_instance_sync_status}
\alias{proton_get_service_instance_sync_status}
\title{Get the status of the synced service instance}
\usage{
proton_get_service_instance_sync_status(serviceInstanceName, serviceName)
}
\arguments{
\item{serviceInstanceName}{[required] The name of the service instance that you want the sync status input
for.}

\item{serviceName}{[required] The name of the service that the service instance belongs to.}
}
\description{
Get the status of the synced service instance.

See \url{https://www.paws-r-sdk.com/docs/proton_get_service_instance_sync_status/} for full documentation.
}
\keyword{internal}
