% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_service_sync_blocker_summary}
\alias{proton_get_service_sync_blocker_summary}
\title{Get detailed data for the service sync blocker summary}
\usage{
proton_get_service_sync_blocker_summary(
  serviceInstanceName = NULL,
  serviceName
)
}
\arguments{
\item{serviceInstanceName}{The name of the service instance that you want to get the service sync
blocker summary for. If given bothe the instance name and the service
name, only the instance is blocked.}

\item{serviceName}{[required] The name of the service that you want to get the service sync blocker
summary for. If given only the service name, all instances are blocked.}
}
\description{
Get detailed data for the service sync blocker summary.

See \url{https://www.paws-r-sdk.com/docs/proton_get_service_sync_blocker_summary/} for full documentation.
}
\keyword{internal}
