% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_service_sync_config}
\alias{proton_get_service_sync_config}
\title{Get detailed information for the service sync configuration}
\usage{
proton_get_service_sync_config(serviceName)
}
\arguments{
\item{serviceName}{[required] The name of the service that you want to get the service sync
configuration for.}
}
\description{
Get detailed information for the service sync configuration.

See \url{https://www.paws-r-sdk.com/docs/proton_get_service_sync_config/} for full documentation.
}
\keyword{internal}
