% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_service_template}
\alias{proton_get_service_template}
\title{Get detailed data for a service template}
\usage{
proton_get_service_template(name)
}
\arguments{
\item{name}{[required] The name of the service template that you want to get detailed data for.}
}
\description{
Get detailed data for a service template.

See \url{https://www.paws-r-sdk.com/docs/proton_get_service_template/} for full documentation.
}
\keyword{internal}
