% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_template_sync_config}
\alias{proton_get_template_sync_config}
\title{Get detail data for a template sync configuration}
\usage{
proton_get_template_sync_config(templateName, templateType)
}
\arguments{
\item{templateName}{[required] The template name.}

\item{templateType}{[required] The template type.}
}
\description{
Get detail data for a template sync configuration.

See \url{https://www.paws-r-sdk.com/docs/proton_get_template_sync_config/} for full documentation.
}
\keyword{internal}
