% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_template_sync_status}
\alias{proton_get_template_sync_status}
\title{Get the status of a template sync}
\usage{
proton_get_template_sync_status(templateName, templateType, templateVersion)
}
\arguments{
\item{templateName}{[required] The template name.}

\item{templateType}{[required] The template type.}

\item{templateVersion}{[required] The template major version.}
}
\description{
Get the status of a template sync.

See \url{https://www.paws-r-sdk.com/docs/proton_get_template_sync_status/} for full documentation.
}
\keyword{internal}
