% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_component_outputs}
\alias{proton_list_component_outputs}
\title{Get a list of component Infrastructure as Code (IaC) outputs}
\usage{
proton_list_component_outputs(
  componentName,
  deploymentId = NULL,
  nextToken = NULL
)
}
\arguments{
\item{componentName}{[required] The name of the component whose outputs you want.}

\item{deploymentId}{The ID of the deployment whose outputs you want.}

\item{nextToken}{A token that indicates the location of the next output in the array of
outputs, after the list of outputs that was previously requested.}
}
\description{
Get a list of component Infrastructure as Code (IaC) outputs.

See \url{https://www.paws-r-sdk.com/docs/proton_list_component_outputs/} for full documentation.
}
\keyword{internal}
