% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_component_provisioned_resources}
\alias{proton_list_component_provisioned_resources}
\title{List provisioned resources for a component with details}
\usage{
proton_list_component_provisioned_resources(componentName, nextToken = NULL)
}
\arguments{
\item{componentName}{[required] The name of the component whose provisioned resources you want.}

\item{nextToken}{A token that indicates the location of the next provisioned resource in
the array of provisioned resources, after the list of provisioned
resources that was previously requested.}
}
\description{
List provisioned resources for a component with details.

See \url{https://www.paws-r-sdk.com/docs/proton_list_component_provisioned_resources/} for full documentation.
}
\keyword{internal}
