% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_deployments}
\alias{proton_list_deployments}
\title{List deployments}
\usage{
proton_list_deployments(
  componentName = NULL,
  environmentName = NULL,
  maxResults = NULL,
  nextToken = NULL,
  serviceInstanceName = NULL,
  serviceName = NULL
)
}
\arguments{
\item{componentName}{The name of a component for result list filtering. Proton returns
deployments associated with that component.}

\item{environmentName}{The name of an environment for result list filtering. Proton returns
deployments associated with the environment.}

\item{maxResults}{The maximum number of deployments to list.}

\item{nextToken}{A token that indicates the location of the next deployment in the array
of deployment, after the list of deployment that was previously
requested.}

\item{serviceInstanceName}{The name of a service instance for result list filtering. Proton returns
the deployments associated with the service instance.}

\item{serviceName}{The name of a service for result list filtering. Proton returns
deployments associated with service instances of the service.}
}
\description{
List deployments. You can filter the result list by environment, service, or a single service instance.

See \url{https://www.paws-r-sdk.com/docs/proton_list_deployments/} for full documentation.
}
\keyword{internal}
