% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_environment_account_connections}
\alias{proton_list_environment_account_connections}
\title{View a list of environment account connections}
\usage{
proton_list_environment_account_connections(
  environmentName = NULL,
  maxResults = NULL,
  nextToken = NULL,
  requestedBy,
  statuses = NULL
)
}
\arguments{
\item{environmentName}{The environment name that's associated with each listed environment
account connection.}

\item{maxResults}{The maximum number of environment account connections to list.}

\item{nextToken}{A token that indicates the location of the next environment account
connection in the array of environment account connections, after the
list of environment account connections that was previously requested.}

\item{requestedBy}{[required] The type of account making the
\code{\link[=proton_list_environment_account_connections]{list_environment_account_connections}}
request.}

\item{statuses}{The status details for each listed environment account connection.}
}
\description{
View a list of environment account connections.

See \url{https://www.paws-r-sdk.com/docs/proton_list_environment_account_connections/} for full documentation.
}
\keyword{internal}
