% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_environment_provisioned_resources}
\alias{proton_list_environment_provisioned_resources}
\title{List the provisioned resources for your environment}
\usage{
proton_list_environment_provisioned_resources(
  environmentName,
  nextToken = NULL
)
}
\arguments{
\item{environmentName}{[required] The environment name.}

\item{nextToken}{A token that indicates the location of the next environment provisioned
resource in the array of environment provisioned resources, after the
list of environment provisioned resources that was previously requested.}
}
\description{
List the provisioned resources for your environment.

See \url{https://www.paws-r-sdk.com/docs/proton_list_environment_provisioned_resources/} for full documentation.
}
\keyword{internal}
