% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_environments}
\alias{proton_list_environments}
\title{List environments with detail data summaries}
\usage{
proton_list_environments(
  environmentTemplates = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{environmentTemplates}{An array of the versions of the environment template.}

\item{maxResults}{The maximum number of environments to list.}

\item{nextToken}{A token that indicates the location of the next environment in the array
of environments, after the list of environments that was previously
requested.}
}
\description{
List environments with detail data summaries.

See \url{https://www.paws-r-sdk.com/docs/proton_list_environments/} for full documentation.
}
\keyword{internal}
