% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_repositories}
\alias{proton_list_repositories}
\title{List repositories with detail data}
\usage{
proton_list_repositories(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of repositories to list.}

\item{nextToken}{A token that indicates the location of the next repository in the array
of repositories, after the list of repositories previously requested.}
}
\description{
List repositories with detail data.

See \url{https://paws-r.github.io/docs/proton/list_repositories.html} for full documentation.
}
\keyword{internal}
