% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_repository_sync_definitions}
\alias{proton_list_repository_sync_definitions}
\title{List repository sync definitions with detail data}
\usage{
proton_list_repository_sync_definitions(
  nextToken = NULL,
  repositoryName,
  repositoryProvider,
  syncType
)
}
\arguments{
\item{nextToken}{A token that indicates the location of the next repository sync
definition in the array of repository sync definitions, after the list
of repository sync definitions previously requested.}

\item{repositoryName}{[required] The repository name.}

\item{repositoryProvider}{[required] The repository provider.}

\item{syncType}{[required] The sync type. The only supported value is \code{TEMPLATE_SYNC}.}
}
\description{
List repository sync definitions with detail data.

See \url{https://www.paws-r-sdk.com/docs/proton_list_repository_sync_definitions/} for full documentation.
}
\keyword{internal}
