% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_service_instance_outputs}
\alias{proton_list_service_instance_outputs}
\title{Get a list service of instance Infrastructure as Code (IaC) outputs}
\usage{
proton_list_service_instance_outputs(
  deploymentId = NULL,
  nextToken = NULL,
  serviceInstanceName,
  serviceName
)
}
\arguments{
\item{deploymentId}{The ID of the deployment whose outputs you want.}

\item{nextToken}{A token that indicates the location of the next output in the array of
outputs, after the list of outputs that was previously requested.}

\item{serviceInstanceName}{[required] The name of the service instance whose outputs you want.}

\item{serviceName}{[required] The name of the service that \code{serviceInstanceName} is associated to.}
}
\description{
Get a list service of instance Infrastructure as Code (IaC) outputs.

See \url{https://www.paws-r-sdk.com/docs/proton_list_service_instance_outputs/} for full documentation.
}
\keyword{internal}
