% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_service_pipeline_outputs}
\alias{proton_list_service_pipeline_outputs}
\title{Get a list of service pipeline Infrastructure as Code (IaC) outputs}
\usage{
proton_list_service_pipeline_outputs(
  deploymentId = NULL,
  nextToken = NULL,
  serviceName
)
}
\arguments{
\item{deploymentId}{The ID of the deployment you want the outputs for.}

\item{nextToken}{A token that indicates the location of the next output in the array of
outputs, after the list of outputs that was previously requested.}

\item{serviceName}{[required] The name of the service whose pipeline's outputs you want.}
}
\description{
Get a list of service pipeline Infrastructure as Code (IaC) outputs.

See \url{https://www.paws-r-sdk.com/docs/proton_list_service_pipeline_outputs/} for full documentation.
}
\keyword{internal}
