% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_service_pipeline_provisioned_resources}
\alias{proton_list_service_pipeline_provisioned_resources}
\title{List provisioned resources for a service and pipeline with details}
\usage{
proton_list_service_pipeline_provisioned_resources(
  nextToken = NULL,
  serviceName
)
}
\arguments{
\item{nextToken}{A token that indicates the location of the next provisioned resource in
the array of provisioned resources, after the list of provisioned
resources that was previously requested.}

\item{serviceName}{[required] The name of the service whose pipeline's provisioned resources you want.}
}
\description{
List provisioned resources for a service and pipeline with details.

See \url{https://www.paws-r-sdk.com/docs/proton_list_service_pipeline_provisioned_resources/} for full documentation.
}
\keyword{internal}
