% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_service_template_versions}
\alias{proton_list_service_template_versions}
\title{List major or minor versions of a service template with detail data}
\usage{
proton_list_service_template_versions(
  majorVersion = NULL,
  maxResults = NULL,
  nextToken = NULL,
  templateName
)
}
\arguments{
\item{majorVersion}{To view a list of minor of versions under a major version of a service
template, include \verb{major Version}.

To view a list of major versions of a service template, \emph{exclude}
\verb{major Version}.}

\item{maxResults}{The maximum number of major or minor versions of a service template to
list.}

\item{nextToken}{A token that indicates the location of the next major or minor version
in the array of major or minor versions of a service template, after the
list of major or minor versions that was previously requested.}

\item{templateName}{[required] The name of the service template.}
}
\description{
List major or minor versions of a service template with detail data.

See \url{https://www.paws-r-sdk.com/docs/proton_list_service_template_versions/} for full documentation.
}
\keyword{internal}
