% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_service_templates}
\alias{proton_list_service_templates}
\title{List service templates with detail data}
\usage{
proton_list_service_templates(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of service templates to list.}

\item{nextToken}{A token that indicates the location of the next service template in the
array of service templates, after the list of service templates
previously requested.}
}
\description{
List service templates with detail data.

See \url{https://www.paws-r-sdk.com/docs/proton_list_service_templates/} for full documentation.
}
\keyword{internal}
