% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_services}
\alias{proton_list_services}
\title{List services with summaries of detail data}
\usage{
proton_list_services(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of services to list.}

\item{nextToken}{A token that indicates the location of the next service in the array of
services, after the list of services that was previously requested.}
}
\description{
List services with summaries of detail data.

See \url{https://www.paws-r-sdk.com/docs/proton_list_services/} for full documentation.
}
\keyword{internal}
