% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_tags_for_resource}
\alias{proton_list_tags_for_resource}
\title{List tags for a resource}
\usage{
proton_list_tags_for_resource(maxResults = NULL, nextToken = NULL, resourceArn)
}
\arguments{
\item{maxResults}{The maximum number of tags to list.}

\item{nextToken}{A token that indicates the location of the next resource tag in the
array of resource tags, after the list of resource tags that was
previously requested.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource for the listed tags.}
}
\description{
List tags for a resource. For more information, see \href{https://docs.aws.amazon.com/proton/latest/userguide/resources.html}{Proton resources and tagging} in the \emph{Proton User Guide}.

See \url{https://www.paws-r-sdk.com/docs/proton_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
