% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_notify_resource_deployment_status_change}
\alias{proton_notify_resource_deployment_status_change}
\title{Notify Proton of status changes to a provisioned resource when you use
self-managed provisioning}
\usage{
proton_notify_resource_deployment_status_change(
  deploymentId = NULL,
  outputs = NULL,
  resourceArn,
  status = NULL,
  statusMessage = NULL
)
}
\arguments{
\item{deploymentId}{The deployment ID for your provisioned resource.}

\item{outputs}{The provisioned resource state change detail data that's returned by
Proton.}

\item{resourceArn}{[required] The provisioned resource Amazon Resource Name (ARN).}

\item{status}{The status of your provisioned resource.}

\item{statusMessage}{The deployment status message for your provisioned resource.}
}
\description{
Notify Proton of status changes to a provisioned resource when you use self-managed provisioning.

See \url{https://www.paws-r-sdk.com/docs/proton_notify_resource_deployment_status_change/} for full documentation.
}
\keyword{internal}
