% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_tag_resource}
\alias{proton_tag_resource}
\title{Tag a resource}
\usage{
proton_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Proton resource to apply customer
tags to.}

\item{tags}{[required] A list of customer tags to apply to the Proton resource.}
}
\description{
Tag a resource. A tag is a key-value pair of metadata that you associate with an Proton resource.

See \url{https://www.paws-r-sdk.com/docs/proton_tag_resource/} for full documentation.
}
\keyword{internal}
