% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_untag_resource}
\alias{proton_untag_resource}
\title{Remove a customer tag from a resource}
\usage{
proton_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to remove customer tags
from.}

\item{tagKeys}{[required] A list of customer tag keys that indicate the customer tags to be
removed from the resource.}
}
\description{
Remove a customer tag from a resource. A tag is a key-value pair of metadata associated with an Proton resource.

See \url{https://www.paws-r-sdk.com/docs/proton_untag_resource/} for full documentation.
}
\keyword{internal}
