% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_update_environment_template}
\alias{proton_update_environment_template}
\title{Update an environment template}
\usage{
proton_update_environment_template(
  description = NULL,
  displayName = NULL,
  name
)
}
\arguments{
\item{description}{A description of the environment template update.}

\item{displayName}{The name of the environment template to update as displayed in the
developer interface.}

\item{name}{[required] The name of the environment template to update.}
}
\description{
Update an environment template.

See \url{https://www.paws-r-sdk.com/docs/proton_update_environment_template/} for full documentation.
}
\keyword{internal}
