% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_update_environment_template_version}
\alias{proton_update_environment_template_version}
\title{Update a major or minor version of an environment template}
\usage{
proton_update_environment_template_version(
  description = NULL,
  majorVersion,
  minorVersion,
  status = NULL,
  templateName
)
}
\arguments{
\item{description}{A description of environment template version to update.}

\item{majorVersion}{[required] To update a major version of an environment template, include
\verb{major Version}.}

\item{minorVersion}{[required] To update a minor version of an environment template, include
\code{minorVersion}.}

\item{status}{The status of the environment template minor version to update.}

\item{templateName}{[required] The name of the environment template.}
}
\description{
Update a major or minor version of an environment template.

See \url{https://www.paws-r-sdk.com/docs/proton_update_environment_template_version/} for full documentation.
}
\keyword{internal}
