% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_update_service}
\alias{proton_update_service}
\title{Edit a service description or use a spec to add and delete service
instances}
\usage{
proton_update_service(description = NULL, name, spec = NULL)
}
\arguments{
\item{description}{The edited service description.}

\item{name}{[required] The name of the service to edit.}

\item{spec}{Lists the service instances to add and the existing service instances to
remain. Omit the existing service instances to delete from the list.
\emph{Don't} include edits to the existing service instances or pipeline. For
more information, see \href{https://docs.aws.amazon.com/proton/latest/userguide/ag-svc-update.html}{Edit a service}
in the \emph{Proton User Guide}.}
}
\description{
Edit a service description or use a spec to add and delete service instances.

See \url{https://www.paws-r-sdk.com/docs/proton_update_service/} for full documentation.
}
\keyword{internal}
