% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_update_service_sync_config}
\alias{proton_update_service_sync_config}
\title{Update the Proton Ops config file}
\usage{
proton_update_service_sync_config(
  branch,
  filePath,
  repositoryName,
  repositoryProvider,
  serviceName
)
}
\arguments{
\item{branch}{[required] The name of the code repository branch where the Proton Ops file is
found.}

\item{filePath}{[required] The path to the Proton Ops file.}

\item{repositoryName}{[required] The name of the repository where the Proton Ops file is found.}

\item{repositoryProvider}{[required] The name of the repository provider where the Proton Ops file is found.}

\item{serviceName}{[required] The name of the service the Proton Ops file is for.}
}
\description{
Update the Proton Ops config file.

See \url{https://www.paws-r-sdk.com/docs/proton_update_service_sync_config/} for full documentation.
}
\keyword{internal}
