% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_update_service_template}
\alias{proton_update_service_template}
\title{Update a service template}
\usage{
proton_update_service_template(description = NULL, displayName = NULL, name)
}
\arguments{
\item{description}{A description of the service template update.}

\item{displayName}{The name of the service template to update that's displayed in the
developer interface.}

\item{name}{[required] The name of the service template to update.}
}
\description{
Update a service template.

See \url{https://www.paws-r-sdk.com/docs/proton_update_service_template/} for full documentation.
}
\keyword{internal}
