% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_update_template_sync_config}
\alias{proton_update_template_sync_config}
\title{Update template sync configuration parameters, except for the
templateName and templateType}
\usage{
proton_update_template_sync_config(
  branch,
  repositoryName,
  repositoryProvider,
  subdirectory = NULL,
  templateName,
  templateType
)
}
\arguments{
\item{branch}{[required] The repository branch.}

\item{repositoryName}{[required] The name of the repository (for example, \code{myrepos/myrepo}).}

\item{repositoryProvider}{[required] The repository provider.}

\item{subdirectory}{A subdirectory path to your template bundle version. When included,
limits the template bundle search to this repository directory.}

\item{templateName}{[required] The synced template name.}

\item{templateType}{[required] The synced template type.}
}
\description{
Update template sync configuration parameters, except for the \code{templateName} and \code{templateType}.

See \url{https://paws-r.github.io/docs/proton/update_template_sync_config.html} for full documentation.
}
\keyword{internal}
