% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_create_cloud_formation_template}
\alias{serverlessapplicationrepository_create_cloud_formation_template}
\title{Creates an AWS CloudFormation template}
\usage{
serverlessapplicationrepository_create_cloud_formation_template(
  ApplicationId, SemanticVersion)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}

\item{SemanticVersion}{The semantic version of the application:

\url{https://semver.org/}}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ApplicationId = "string",
  CreationTime = "string",
  ExpirationTime = "string",
  SemanticVersion = "string",
  Status = "PREPARING"|"ACTIVE"|"EXPIRED",
  TemplateId = "string",
  TemplateUrl = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates an AWS CloudFormation template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_cloud_formation_template(
  ApplicationId = "string",
  SemanticVersion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
