% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_delete_application}
\alias{serverlessapplicationrepository_delete_application}
\title{Deletes the specified application}
\usage{
serverlessapplicationrepository_delete_application(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}
}
\value{
An empty list.
}
\description{
Deletes the specified application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_application(
  ApplicationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
