% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_list_application_dependencies}
\alias{serverlessapplicationrepository_list_application_dependencies}
\title{Retrieves the list of applications nested in the containing application}
\usage{
serverlessapplicationrepository_list_application_dependencies(
  ApplicationId, MaxItems, NextToken, SemanticVersion)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}

\item{MaxItems}{The total number of items to return.}

\item{NextToken}{A token to specify where to start paginating.}

\item{SemanticVersion}{The semantic version of the application to get.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Dependencies = list(
    list(
      ApplicationId = "string",
      SemanticVersion = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the list of applications nested in the containing application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_application_dependencies(
  ApplicationId = "string",
  MaxItems = 123,
  NextToken = "string",
  SemanticVersion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
