% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_list_applications}
\alias{serverlessapplicationrepository_list_applications}
\title{Lists applications owned by the requester}
\usage{
serverlessapplicationrepository_list_applications(MaxItems, NextToken)
}
\arguments{
\item{MaxItems}{The total number of items to return.}

\item{NextToken}{A token to specify where to start paginating.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Applications = list(
    list(
      ApplicationId = "string",
      Author = "string",
      CreationTime = "string",
      Description = "string",
      HomePageUrl = "string",
      Labels = list(
        "string"
      ),
      Name = "string",
      SpdxLicenseId = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists applications owned by the requester.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_applications(
  MaxItems = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
