% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprunner_operations.R
\name{apprunner_disassociate_custom_domain}
\alias{apprunner_disassociate_custom_domain}
\title{Disassociate a custom domain name from an App Runner service}
\usage{
apprunner_disassociate_custom_domain(ServiceArn, DomainName)
}
\arguments{
\item{ServiceArn}{[required] The Amazon Resource Name (ARN) of the App Runner service that you want
to disassociate a custom domain name from.}

\item{DomainName}{[required] The domain name that you want to disassociate from the App Runner
service.}
}
\description{
Disassociate a custom domain name from an App Runner service.

See \url{https://www.paws-r-sdk.com/docs/apprunner_disassociate_custom_domain/} for full documentation.
}
\keyword{internal}
