% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_cancel_job}
\alias{batch_cancel_job}
\title{Cancels a job in an Batch job queue}
\usage{
batch_cancel_job(jobId, reason)
}
\arguments{
\item{jobId}{[required] The Batch job ID of the job to cancel.}

\item{reason}{[required] A message to attach to the job that explains the reason for canceling
it. This message is returned by future
\code{\link[=batch_describe_jobs]{describe_jobs}} operations on the job. This
message is also recorded in the Batch activity logs.}
}
\description{
Cancels a job in an Batch job queue. Jobs that are in the \code{SUBMITTED} or \code{PENDING} are canceled. A job in\code{RUNNABLE} remains in \code{RUNNABLE} until it reaches the head of the job queue. Then the job status is updated to \code{FAILED}.

See \url{https://www.paws-r-sdk.com/docs/batch_cancel_job/} for full documentation.
}
\keyword{internal}
