% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_describe_compute_environments}
\alias{batch_describe_compute_environments}
\title{Describes one or more of your compute environments}
\usage{
batch_describe_compute_environments(computeEnvironments, maxResults,
  nextToken)
}
\arguments{
\item{computeEnvironments}{A list of up to 100 compute environment names or full Amazon Resource
Name (ARN) entries.}

\item{maxResults}{The maximum number of cluster results returned by
\code{DescribeComputeEnvironments} in paginated output. When this parameter
is used, \code{DescribeComputeEnvironments} only returns \code{maxResults} results
in a single page along with a \code{nextToken} response element. The
remaining results of the initial request can be seen by sending another
\code{DescribeComputeEnvironments} request with the returned \code{nextToken}
value. This value can be between 1 and 100. If this parameter is not
used, then \code{DescribeComputeEnvironments} returns up to 100 results and a
\code{nextToken} value if applicable.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{DescribeComputeEnvironments} request where \code{maxResults} was used and
the results exceeded the value of that parameter. Pagination continues
from the end of the previous results that returned the \code{nextToken}
value. This value is \code{null} when there are no more results to return.

This token should be treated as an opaque identifier that is only used
to retrieve the next items in a list and not for other programmatic
purposes.}
}
\description{
Describes one or more of your compute environments.
}
\details{
If you are using an unmanaged compute environment, you can use the
\code{DescribeComputeEnvironment} operation to determine the \code{ecsClusterArn}
that you should launch your Amazon ECS container instances into.
}
\section{Request syntax}{
\preformatted{svc$describe_compute_environments(
  computeEnvironments = list(
    "string"
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\section{Examples}{
\preformatted{# This example describes the P2OnDemand compute environment.
svc$describe_compute_environments(
  computeEnvironments = list(
    "P2OnDemand"
  )
)
}
}

\keyword{internal}
