% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_get_job_queue_snapshot}
\alias{batch_get_job_queue_snapshot}
\title{Provides a list of the first 100 RUNNABLE jobs associated to a single
job queue}
\usage{
batch_get_job_queue_snapshot(jobQueue)
}
\arguments{
\item{jobQueue}{[required] The job queue’s name or full queue Amazon Resource Name (ARN).}
}
\description{
Provides a list of the first 100 \code{RUNNABLE} jobs associated to a single job queue.

See \url{https://www.paws-r-sdk.com/docs/batch_get_job_queue_snapshot/} for full documentation.
}
\keyword{internal}
