% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_list_tags_for_resource}
\alias{batch_list_tags_for_resource}
\title{Lists the tags for an Batch resource}
\usage{
batch_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource that tags
are listed for. Batch resources that support tags are compute
environments, jobs, job definitions, job queues, and scheduling
policies. ARNs for child jobs of array and multi-node parallel (MNP)
jobs aren't supported.}
}
\description{
Lists the tags for an Batch resource. Batch resources that support tags are compute environments, jobs, job definitions, job queues, and scheduling policies. ARNs for child jobs of array and multi-node parallel (MNP) jobs aren't supported.

See \url{https://www.paws-r-sdk.com/docs/batch_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
