% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_terminate_job}
\alias{batch_terminate_job}
\title{Terminates a job in a job queue}
\usage{
batch_terminate_job(jobId, reason)
}
\arguments{
\item{jobId}{[required] The Batch job ID of the job to terminate.}

\item{reason}{[required] A message to attach to the job that explains the reason for canceling
it. This message is returned by future
\code{\link[=batch_describe_jobs]{describe_jobs}} operations on the job. This
message is also recorded in the Batch activity logs.}
}
\description{
Terminates a job in a job queue. Jobs that are in the \code{STARTING} or \code{RUNNING} state are terminated, which causes them to transition to \code{FAILED}. Jobs that have not progressed to the \code{STARTING} state are cancelled.

See \url{https://www.paws-r-sdk.com/docs/batch_terminate_job/} for full documentation.
}
\keyword{internal}
