% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_operations.R
\name{batch_update_scheduling_policy}
\alias{batch_update_scheduling_policy}
\title{Updates a scheduling policy}
\usage{
batch_update_scheduling_policy(arn, fairsharePolicy = NULL)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the scheduling policy to update.}

\item{fairsharePolicy}{The fair share policy.}
}
\description{
Updates a scheduling policy.

See \url{https://www.paws-r-sdk.com/docs/batch_update_scheduling_policy/} for full documentation.
}
\keyword{internal}
