% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braket_operations.R
\name{braket_create_quantum_task}
\alias{braket_create_quantum_task}
\title{Creates a quantum task}
\usage{
braket_create_quantum_task(
  clientToken,
  deviceArn,
  deviceParameters = NULL,
  shots,
  outputS3Bucket,
  outputS3KeyPrefix,
  action,
  tags = NULL,
  jobToken = NULL,
  associations = NULL,
  experimentalCapabilities = NULL
)
}
\arguments{
\item{clientToken}{[required] The client token associated with the request.}

\item{deviceArn}{[required] The ARN of the device to run the quantum task on.}

\item{deviceParameters}{The parameters for the device to run the quantum task on.}

\item{shots}{[required] The number of shots to use for the quantum task.}

\item{outputS3Bucket}{[required] The S3 bucket to store quantum task result files in.}

\item{outputS3KeyPrefix}{[required] The key prefix for the location in the S3 bucket to store quantum task
results in.}

\item{action}{[required] The action associated with the quantum task.}

\item{tags}{Tags to be added to the quantum task you're creating.}

\item{jobToken}{The token for an Amazon Braket hybrid job that associates it with the
quantum task.}

\item{associations}{The list of Amazon Braket resources associated with the quantum task.}

\item{experimentalCapabilities}{Enable experimental capabilities for the quantum task.}
}
\description{
Creates a quantum task.

See \url{https://www.paws-r-sdk.com/docs/braket_create_quantum_task/} for full documentation.
}
\keyword{internal}
