% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeoptimizer_operations.R
\name{computeoptimizer_get_rds_database_recommendation_projected_metrics}
\alias{computeoptimizer_get_rds_database_recommendation_projected_metrics}
\title{Returns the projected metrics of Aurora and RDS database recommendations}
\usage{
computeoptimizer_get_rds_database_recommendation_projected_metrics(
  resourceArn,
  stat,
  period,
  startTime,
  endTime,
  recommendationPreferences = NULL
)
}
\arguments{
\item{resourceArn}{[required] The ARN that identifies the Amazon Aurora or RDS database.

The following is the format of the ARN:

\code{arn:aws:rds:{region}:{accountId}:db:{resourceName}}}

\item{stat}{[required] The statistic of the projected metrics.}

\item{period}{[required] The granularity, in seconds, of the projected metrics data points.}

\item{startTime}{[required] The timestamp of the first projected metrics data point to return.}

\item{endTime}{[required] The timestamp of the last projected metrics data point to return.}

\item{recommendationPreferences}{}
}
\description{
Returns the projected metrics of Aurora and RDS database recommendations.

See \url{https://www.paws-r-sdk.com/docs/computeoptimizer_get_rds_database_recommendation_projected_metrics/} for full documentation.
}
\keyword{internal}
