% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_associate_dhcp_options}
\alias{ec2_associate_dhcp_options}
\title{Associates a set of DHCP options (that you've previously created) with
the specified VPC, or associates no DHCP options with the VPC}
\usage{
ec2_associate_dhcp_options(DhcpOptionsId, VpcId, DryRun = NULL)
}
\arguments{
\item{DhcpOptionsId}{[required] The ID of the DHCP options set, or \code{default} to associate no DHCP
options with the VPC.}

\item{VpcId}{[required] The ID of the VPC.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Associates a set of DHCP options (that you've previously created) with the specified VPC, or associates no DHCP options with the VPC.

See \url{https://www.paws-r-sdk.com/docs/ec2_associate_dhcp_options/} for full documentation.
}
\keyword{internal}
