% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_associate_iam_instance_profile}
\alias{ec2_associate_iam_instance_profile}
\title{Associates an IAM instance profile with a running or stopped instance}
\usage{
ec2_associate_iam_instance_profile(IamInstanceProfile, InstanceId)
}
\arguments{
\item{IamInstanceProfile}{[required] The IAM instance profile.}

\item{InstanceId}{[required] The ID of the instance.}
}
\description{
Associates an IAM instance profile with a running or stopped instance. You cannot associate more than one IAM instance profile with an instance.

See \url{https://www.paws-r-sdk.com/docs/ec2_associate_iam_instance_profile/} for full documentation.
}
\keyword{internal}
