% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_associate_subnet_cidr_block}
\alias{ec2_associate_subnet_cidr_block}
\title{Associates a CIDR block with your subnet}
\usage{
ec2_associate_subnet_cidr_block(
  Ipv6CidrBlock = NULL,
  SubnetId,
  Ipv6IpamPoolId = NULL,
  Ipv6NetmaskLength = NULL
)
}
\arguments{
\item{Ipv6CidrBlock}{The IPv6 CIDR block for your subnet.}

\item{SubnetId}{[required] The ID of your subnet.}

\item{Ipv6IpamPoolId}{An IPv6 IPAM pool ID.}

\item{Ipv6NetmaskLength}{An IPv6 netmask length.}
}
\description{
Associates a CIDR block with your subnet. You can only associate a single IPv6 CIDR block with your subnet.

See \url{https://www.paws-r-sdk.com/docs/ec2_associate_subnet_cidr_block/} for full documentation.
}
\keyword{internal}
