% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_attach_classic_link_vpc}
\alias{ec2_attach_classic_link_vpc}
\title{This action is deprecated}
\usage{
ec2_attach_classic_link_vpc(DryRun = NULL, Groups, InstanceId, VpcId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Groups}{[required] The IDs of the security groups. You cannot specify security groups from
a different VPC.}

\item{InstanceId}{[required] The ID of the EC2-Classic instance.}

\item{VpcId}{[required] The ID of the ClassicLink-enabled VPC.}
}
\description{
This action is deprecated.

See \url{https://www.paws-r-sdk.com/docs/ec2_attach_classic_link_vpc/} for full documentation.
}
\keyword{internal}
