% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_authorize_security_group_egress}
\alias{ec2_authorize_security_group_egress}
\title{Adds the specified outbound (egress) rules to a security group}
\usage{
ec2_authorize_security_group_egress(
  DryRun = NULL,
  GroupId,
  IpPermissions = NULL,
  TagSpecifications = NULL,
  CidrIp = NULL,
  FromPort = NULL,
  IpProtocol = NULL,
  ToPort = NULL,
  SourceSecurityGroupName = NULL,
  SourceSecurityGroupOwnerId = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{GroupId}{[required] The ID of the security group.}

\item{IpPermissions}{The permissions for the security group rules.}

\item{TagSpecifications}{The tags applied to the security group rule.}

\item{CidrIp}{Not supported. Use IP permissions instead.}

\item{FromPort}{Not supported. Use IP permissions instead.}

\item{IpProtocol}{Not supported. Use IP permissions instead.}

\item{ToPort}{Not supported. Use IP permissions instead.}

\item{SourceSecurityGroupName}{Not supported. Use IP permissions instead.}

\item{SourceSecurityGroupOwnerId}{Not supported. Use IP permissions instead.}
}
\description{
Adds the specified outbound (egress) rules to a security group.

See \url{https://www.paws-r-sdk.com/docs/ec2_authorize_security_group_egress/} for full documentation.
}
\keyword{internal}
