% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_bundle_instance}
\alias{ec2_bundle_instance}
\title{Bundles an Amazon instance store-backed Windows instance}
\usage{
ec2_bundle_instance(InstanceId, Storage, DryRun = NULL)
}
\arguments{
\item{InstanceId}{[required] The ID of the instance to bundle.

Default: None}

\item{Storage}{[required] The bucket in which to store the AMI. You can specify a bucket that you
already own or a new bucket that Amazon EC2 creates on your behalf. If
you specify a bucket that belongs to someone else, Amazon EC2 returns an
error.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Bundles an Amazon instance store-backed Windows instance.

See \url{https://www.paws-r-sdk.com/docs/ec2_bundle_instance/} for full documentation.
}
\keyword{internal}
