% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_cancel_capacity_reservation}
\alias{ec2_cancel_capacity_reservation}
\title{Cancels the specified Capacity Reservation, releases the reserved
capacity, and changes the Capacity Reservation's state to cancelled}
\usage{
ec2_cancel_capacity_reservation(CapacityReservationId, DryRun = NULL)
}
\arguments{
\item{CapacityReservationId}{[required] The ID of the Capacity Reservation to be cancelled.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Cancels the specified Capacity Reservation, releases the reserved capacity, and changes the Capacity Reservation's state to \code{cancelled}.

See \url{https://www.paws-r-sdk.com/docs/ec2_cancel_capacity_reservation/} for full documentation.
}
\keyword{internal}
