% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_cancel_spot_fleet_requests}
\alias{ec2_cancel_spot_fleet_requests}
\title{Cancels the specified Spot Fleet requests}
\usage{
ec2_cancel_spot_fleet_requests(
  DryRun = NULL,
  SpotFleetRequestIds,
  TerminateInstances
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{SpotFleetRequestIds}{[required] The IDs of the Spot Fleet requests.

Constraint: You can specify up to 100 IDs in a single request.}

\item{TerminateInstances}{[required] Indicates whether to terminate the associated instances when the Spot
Fleet request is canceled. The default is to terminate the instances.

To let the instances continue to run after the Spot Fleet request is
canceled, specify \code{no-terminate-instances}.}
}
\description{
Cancels the specified Spot Fleet requests.

See \url{https://www.paws-r-sdk.com/docs/ec2_cancel_spot_fleet_requests/} for full documentation.
}
\keyword{internal}
