% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_flow_logs}
\alias{ec2_create_flow_logs}
\title{Creates one or more flow logs to capture information about IP traffic
for a specific network interface, subnet, or VPC}
\usage{
ec2_create_flow_logs(DryRun, ClientToken, DeliverLogsPermissionArn,
  LogGroupName, ResourceIds, ResourceType, TrafficType,
  LogDestinationType, LogDestination)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Run_Instance_Idempotency.html}{How to Ensure Idempotency}.}

\item{DeliverLogsPermissionArn}{The ARN for the IAM role that's used to post flow logs to a log group.}

\item{LogGroupName}{The name of the log group.}

\item{ResourceIds}{[required] One or more subnet, network interface, or VPC IDs.

Constraints: Maximum of 1000 resources}

\item{ResourceType}{[required] The type of resource on which to create the flow log.}

\item{TrafficType}{[required] The type of traffic to log.}

\item{LogDestinationType}{Specifies the type of destination to which the flow log data is to be
published. Flow log data can be published to CloudWatch Logs or Amazon
S3. To publish flow log data to CloudWatch Logs, specify
\code{cloud-watch-logs}. To publish flow log data to Amazon S3, specify \code{s3}.

Default: \code{cloud-watch-logs}}

\item{LogDestination}{Specifies the destination to which the flow log data is to be published.
Flow log data can be published to an CloudWatch Logs log group or an
Amazon S3 bucket. The value specified for this parameter depends on the
value specified for LogDestinationType.

If LogDestinationType is not specified or \code{cloud-watch-logs}, specify
the Amazon Resource Name (ARN) of the CloudWatch Logs log group.

If LogDestinationType is \code{s3}, specify the ARN of the Amazon S3 bucket.
You can also specify a subfolder in the bucket. To specify a subfolder
in the bucket, use the following ARN format:
\code{bucket_ARN/subfolder_name/}. For example, to specify a subfolder named
\code{my-logs} in a bucket named \code{my-bucket}, use the following ARN:
\code{arn:aws:s3:::my-bucket/my-logs/}. You cannot use \code{AWSLogs} as a
subfolder name. This is a reserved term.}
}
\description{
Creates one or more flow logs to capture information about IP traffic
for a specific network interface, subnet, or VPC.
}
\details{
Flow log data for a monitored network interface is recorded as flow log
records, which are log events consisting of fields that describe the
traffic flow. For more information, see \href{https://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/flow-logs.html#flow-log-records}{Flow Log Records}
in the \emph{Amazon Virtual Private Cloud User Guide}.

When publishing to CloudWatch Logs, flow log records are published to a
log group, and each network interface has a unique log stream in the log
group. When publishing to Amazon S3, flow log records for all of the
monitored network interfaces are published to a single log file object
that is stored in the specified bucket.

For more information, see \href{https://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/flow-logs.html}{VPC Flow Logs}
in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_flow_logs(
  DryRun = TRUE|FALSE,
  ClientToken = "string",
  DeliverLogsPermissionArn = "string",
  LogGroupName = "string",
  ResourceIds = list(
    "string"
  ),
  ResourceType = "VPC"|"Subnet"|"NetworkInterface",
  TrafficType = "ACCEPT"|"REJECT"|"ALL",
  LogDestinationType = "cloud-watch-logs"|"s3",
  LogDestination = "string"
)
}
}

\keyword{internal}
