% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_instance_export_task}
\alias{ec2_create_instance_export_task}
\title{Exports a running or stopped instance to an Amazon S3 bucket}
\usage{
ec2_create_instance_export_task(
  Description = NULL,
  ExportToS3Task,
  InstanceId,
  TargetEnvironment,
  TagSpecifications = NULL
)
}
\arguments{
\item{Description}{A description for the conversion task or the resource being exported.
The maximum length is 255 characters.}

\item{ExportToS3Task}{[required] The format and location for an export instance task.}

\item{InstanceId}{[required] The ID of the instance.}

\item{TargetEnvironment}{[required] The target virtualization environment.}

\item{TagSpecifications}{The tags to apply to the export instance task during creation.}
}
\description{
Exports a running or stopped instance to an Amazon S3 bucket.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_instance_export_task/} for full documentation.
}
\keyword{internal}
