% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_replace_root_volume_task}
\alias{ec2_create_replace_root_volume_task}
\title{Creates a root volume replacement task for an Amazon EC2 instance}
\usage{
ec2_create_replace_root_volume_task(InstanceId, SnapshotId, ClientToken,
  DryRun, TagSpecifications)
}
\arguments{
\item{InstanceId}{[required] The ID of the instance for which to replace the root volume.}

\item{SnapshotId}{The ID of the snapshot from which to restore the replacement root
volume. If you want to restore the volume to the initial launch state,
omit this parameter.}

\item{ClientToken}{Unique, case-sensitive identifier you provide to ensure the idempotency
of the request. If you do not specify a client token, a randomly
generated token is used for the request to ensure idempotency. For more
information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{TagSpecifications}{The tags to apply to the root volume replacement task.}
}
\value{
A list with the following syntax:\preformatted{list(
  ReplaceRootVolumeTask = list(
    ReplaceRootVolumeTaskId = "string",
    InstanceId = "string",
    TaskState = "pending"|"in-progress"|"failing"|"succeeded"|"failed"|"failed-detached",
    StartTime = "string",
    CompleteTime = "string",
    Tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}
}
\description{
Creates a root volume replacement task for an Amazon EC2 instance. The
root volume can either be restored to its initial launch state, or it
can be restored using a specific snapshot.

For more information, see \href{https://docs.aws.amazon.com/}{Replace a root volume} in the \emph{Amazon Elastic Compute
Cloud User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_replace_root_volume_task(
  InstanceId = "string",
  SnapshotId = "string",
  ClientToken = "string",
  DryRun = TRUE|FALSE,
  TagSpecifications = list(
    list(
      ResourceType = "client-vpn-endpoint"|"customer-gateway"|"dedicated-host"|"dhcp-options"|"egress-only-internet-gateway"|"elastic-ip"|"elastic-gpu"|"export-image-task"|"export-instance-task"|"fleet"|"fpga-image"|"host-reservation"|"image"|"import-image-task"|"import-snapshot-task"|"instance"|"internet-gateway"|"key-pair"|"launch-template"|"local-gateway-route-table-vpc-association"|"natgateway"|"network-acl"|"network-interface"|"network-insights-analysis"|"network-insights-path"|"placement-group"|"reserved-instances"|"route-table"|"security-group"|"snapshot"|"spot-fleet-request"|"spot-instances-request"|"subnet"|"traffic-mirror-filter"|"traffic-mirror-session"|"traffic-mirror-target"|"transit-gateway"|"transit-gateway-attachment"|"transit-gateway-connect-peer"|"transit-gateway-multicast-domain"|"transit-gateway-route-table"|"volume"|"vpc"|"vpc-peering-connection"|"vpn-connection"|"vpn-gateway"|"vpc-flow-log",
      Tags = list(
        list(
          Key = "string",
          Value = "string"
        )
      )
    )
  )
)
}
}

\keyword{internal}
