% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_traffic_mirror_filter}
\alias{ec2_create_traffic_mirror_filter}
\title{Creates a Traffic Mirror filter}
\usage{
ec2_create_traffic_mirror_filter(
  Description = NULL,
  TagSpecifications = NULL,
  DryRun = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{Description}{The description of the Traffic Mirror filter.}

\item{TagSpecifications}{The tags to assign to a Traffic Mirror filter.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{How to ensure idempotency}.}
}
\description{
Creates a Traffic Mirror filter.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_traffic_mirror_filter/} for full documentation.
}
\keyword{internal}
