% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_transit_gateway}
\alias{ec2_create_transit_gateway}
\title{Creates a transit gateway}
\usage{
ec2_create_transit_gateway(
  Description = NULL,
  Options = NULL,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{Description}{A description of the transit gateway.}

\item{Options}{The transit gateway options.}

\item{TagSpecifications}{The tags to apply to the transit gateway.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a transit gateway.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_transit_gateway/} for full documentation.
}
\keyword{internal}
