% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_transit_gateway_multicast_domain}
\alias{ec2_create_transit_gateway_multicast_domain}
\title{Creates a multicast domain using the specified transit gateway}
\usage{
ec2_create_transit_gateway_multicast_domain(
  TransitGatewayId,
  Options = NULL,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayId}{[required] The ID of the transit gateway.}

\item{Options}{The options for the transit gateway multicast domain.}

\item{TagSpecifications}{The tags for the transit gateway multicast domain.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a multicast domain using the specified transit gateway.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_transit_gateway_multicast_domain/} for full documentation.
}
\keyword{internal}
