% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_transit_gateway_route_table}
\alias{ec2_create_transit_gateway_route_table}
\title{Creates a route table for the specified transit gateway}
\usage{
ec2_create_transit_gateway_route_table(
  TransitGatewayId,
  TagSpecifications = NULL,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayId}{[required] The ID of the transit gateway.}

\item{TagSpecifications}{The tags to apply to the transit gateway route table.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a route table for the specified transit gateway.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_transit_gateway_route_table/} for full documentation.
}
\keyword{internal}
