% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_create_volume}
\alias{ec2_create_volume}
\title{Creates an EBS volume that can be attached to an instance in the same
Availability Zone}
\usage{
ec2_create_volume(
  AvailabilityZone,
  Encrypted = NULL,
  Iops = NULL,
  KmsKeyId = NULL,
  OutpostArn = NULL,
  Size = NULL,
  SnapshotId = NULL,
  VolumeType = NULL,
  DryRun = NULL,
  TagSpecifications = NULL,
  MultiAttachEnabled = NULL,
  Throughput = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{AvailabilityZone}{[required] The ID of the Availability Zone in which to create the volume. For
example, \verb{us-east-1a}.}

\item{Encrypted}{Indicates whether the volume should be encrypted. The effect of setting
the encryption state to \code{true} depends on the volume origin (new or from
a snapshot), starting encryption state, ownership, and whether
encryption by default is enabled. For more information, see \href{https://docs.aws.amazon.com/ebs/latest/userguide/work-with-ebs-encr.html#encryption-by-default}{Encryption by default}
in the \emph{Amazon EBS User Guide}.

Encrypted Amazon EBS volumes must be attached to instances that support
Amazon EBS encryption. For more information, see \href{https://docs.aws.amazon.com/ebs/latest/userguide/ebs-encryption-requirements.html#ebs-encryption_supported_instances}{Supported instance types}.}

\item{Iops}{The number of I/O operations per second (IOPS). For \code{gp3}, \code{io1}, and
\code{io2} volumes, this represents the number of IOPS that are provisioned
for the volume. For \code{gp2} volumes, this represents the baseline
performance of the volume and the rate at which the volume accumulates
I/O credits for bursting.

The following are the supported values for each volume type:
\itemize{
\item \code{gp3}: 3,000 - 16,000 IOPS
\item \code{io1}: 100 - 64,000 IOPS
\item \code{io2}: 100 - 256,000 IOPS
}

For \code{io2} volumes, you can achieve up to 256,000 IOPS on \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-types.html#ec2-nitro-instances}{instances built on the Nitro System}.
On other instances, you can achieve performance up to 32,000 IOPS.

This parameter is required for \code{io1} and \code{io2} volumes. The default for
\code{gp3} volumes is 3,000 IOPS. This parameter is not supported for \code{gp2},
\code{st1}, \code{sc1}, or \code{standard} volumes.}

\item{KmsKeyId}{The identifier of the Key Management Service (KMS) KMS key to use for
Amazon EBS encryption. If this parameter is not specified, your KMS key
for Amazon EBS is used. If \code{KmsKeyId} is specified, the encrypted state
must be \code{true}.

You can specify the KMS key using any of the following:
\itemize{
\item Key ID. For example, 1234abcd-12ab-34cd-56ef-1234567890ab.
\item Key alias. For example, alias/ExampleAlias.
\item Key ARN. For example,
arn:aws:kms:us-east-1:012345678910:key/1234abcd-12ab-34cd-56ef-1234567890ab.
\item Alias ARN. For example,
arn:aws:kms:us-east-1:012345678910:alias/ExampleAlias.
}

Amazon Web Services authenticates the KMS key asynchronously. Therefore,
if you specify an ID, alias, or ARN that is not valid, the action can
appear to complete, but eventually fails.}

\item{OutpostArn}{The Amazon Resource Name (ARN) of the Outpost.}

\item{Size}{The size of the volume, in GiBs. You must specify either a snapshot ID
or a volume size. If you specify a snapshot, the default is the snapshot
size. You can specify a volume size that is equal to or larger than the
snapshot size.

The following are the supported volumes sizes for each volume type:
\itemize{
\item \code{gp2} and \code{gp3}: 1 - 16,384 GiB
\item \code{io1}: 4 - 16,384 GiB
\item \code{io2}: 4 - 65,536 GiB
\item \code{st1} and \code{sc1}: 125 - 16,384 GiB
\item \code{standard}: 1 - 1024 GiB
}}

\item{SnapshotId}{The snapshot from which to create the volume. You must specify either a
snapshot ID or a volume size.}

\item{VolumeType}{The volume type. This parameter can be one of the following values:
\itemize{
\item General Purpose SSD: \code{gp2} | \code{gp3}
\item Provisioned IOPS SSD: \code{io1} | \code{io2}
\item Throughput Optimized HDD: \code{st1}
\item Cold HDD: \code{sc1}
\item Magnetic: \code{standard}
}

Throughput Optimized HDD (\code{st1}) and Cold HDD (\code{sc1}) volumes can't be
used as boot volumes.

For more information, see \href{https://docs.aws.amazon.com/ebs/latest/userguide/ebs-volume-types.html}{Amazon EBS volume types}
in the \emph{Amazon EBS User Guide}.

Default: \code{gp2}}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{TagSpecifications}{The tags to apply to the volume during creation.}

\item{MultiAttachEnabled}{Indicates whether to enable Amazon EBS Multi-Attach. If you enable
Multi-Attach, you can attach the volume to up to 16 \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-types.html#ec2-nitro-instances}{Instances built on the Nitro System}
in the same Availability Zone. This parameter is supported with \code{io1}
and \code{io2} volumes only. For more information, see \href{https://docs.aws.amazon.com/ebs/latest/userguide/ebs-volumes-multi.html}{Amazon EBS Multi-Attach}
in the \emph{Amazon EBS User Guide}.}

\item{Throughput}{The throughput to provision for a volume, with a maximum of 1,000 MiB/s.

This parameter is valid only for \code{gp3} volumes.

Valid Range: Minimum value of 125. Maximum value of 1000.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensure Idempotency}.}
}
\description{
Creates an EBS volume that can be attached to an instance in the same Availability Zone.

See \url{https://www.paws-r-sdk.com/docs/ec2_create_volume/} for full documentation.
}
\keyword{internal}
