% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_fpga_image}
\alias{ec2_delete_fpga_image}
\title{Deletes the specified Amazon FPGA Image (AFI)}
\usage{
ec2_delete_fpga_image(DryRun = NULL, FpgaImageId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{FpgaImageId}{[required] The ID of the AFI.}
}
\description{
Deletes the specified Amazon FPGA Image (AFI).

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_fpga_image/} for full documentation.
}
\keyword{internal}
