% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_nat_gateway}
\alias{ec2_delete_nat_gateway}
\title{Deletes the specified NAT gateway}
\usage{
ec2_delete_nat_gateway(DryRun = NULL, NatGatewayId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{NatGatewayId}{[required] The ID of the NAT gateway.}
}
\description{
Deletes the specified NAT gateway. Deleting a public NAT gateway disassociates its Elastic IP address, but does not release the address from your account. Deleting a NAT gateway does not delete any NAT gateway routes in your route tables.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_nat_gateway/} for full documentation.
}
\keyword{internal}
