% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_network_acl_entry}
\alias{ec2_delete_network_acl_entry}
\title{Deletes the specified ingress or egress entry (rule) from the specified
network ACL}
\usage{
ec2_delete_network_acl_entry(DryRun = NULL, Egress, NetworkAclId, RuleNumber)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Egress}{[required] Indicates whether the rule is an egress rule.}

\item{NetworkAclId}{[required] The ID of the network ACL.}

\item{RuleNumber}{[required] The rule number of the entry to delete.}
}
\description{
Deletes the specified ingress or egress entry (rule) from the specified network ACL.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_network_acl_entry/} for full documentation.
}
\keyword{internal}
