% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_network_insights_path}
\alias{ec2_delete_network_insights_path}
\title{Deletes the specified path}
\usage{
ec2_delete_network_insights_path(DryRun = NULL, NetworkInsightsPathId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{NetworkInsightsPathId}{[required] The ID of the path.}
}
\description{
Deletes the specified path.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_network_insights_path/} for full documentation.
}
\keyword{internal}
