% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_placement_group}
\alias{ec2_delete_placement_group}
\title{Deletes the specified placement group}
\usage{
ec2_delete_placement_group(DryRun = NULL, GroupName)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{GroupName}{[required] The name of the placement group.}
}
\description{
Deletes the specified placement group. You must terminate all instances in the placement group before you can delete the placement group. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/placement-groups.html}{Placement groups} in the \emph{Amazon EC2 User Guide}.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_placement_group/} for full documentation.
}
\keyword{internal}
