% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_public_ipv_4_pool}
\alias{ec2_delete_public_ipv_4_pool}
\title{Delete a public IPv4 pool}
\usage{
ec2_delete_public_ipv_4_pool(DryRun = NULL, PoolId)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{PoolId}{[required] The ID of the public IPv4 pool you want to delete.}
}
\description{
Delete a public IPv4 pool. A public IPv4 pool is an EC2 IP address pool required for the public IPv4 CIDRs that you own and bring to Amazon Web Services to manage with IPAM. IPv6 addresses you bring to Amazon Web Services, however, use IPAM pools only.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_public_ipv_4_pool/} for full documentation.
}
\keyword{internal}
