% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_queued_reserved_instances}
\alias{ec2_delete_queued_reserved_instances}
\title{Deletes the queued purchases for the specified Reserved Instances}
\usage{
ec2_delete_queued_reserved_instances(DryRun = NULL, ReservedInstancesIds)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ReservedInstancesIds}{[required] The IDs of the Reserved Instances.}
}
\description{
Deletes the queued purchases for the specified Reserved Instances.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_queued_reserved_instances/} for full documentation.
}
\keyword{internal}
