% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_security_group}
\alias{ec2_delete_security_group}
\title{Deletes a security group}
\usage{
ec2_delete_security_group(GroupId = NULL, GroupName = NULL, DryRun = NULL)
}
\arguments{
\item{GroupId}{The ID of the security group.}

\item{GroupName}{[Default VPC] The name of the security group. You can specify either
the security group name or the security group ID. For security groups in
a nondefault VPC, you must specify the security group ID.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes a security group.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_security_group/} for full documentation.
}
\keyword{internal}
