% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_snapshot}
\alias{ec2_delete_snapshot}
\title{Deletes the specified snapshot}
\usage{
ec2_delete_snapshot(SnapshotId, DryRun = NULL)
}
\arguments{
\item{SnapshotId}{[required] The ID of the EBS snapshot.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified snapshot.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_snapshot/} for full documentation.
}
\keyword{internal}
