% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_traffic_mirror_session}
\alias{ec2_delete_traffic_mirror_session}
\title{Deletes the specified Traffic Mirror session}
\usage{
ec2_delete_traffic_mirror_session(TrafficMirrorSessionId, DryRun)
}
\arguments{
\item{TrafficMirrorSessionId}{[required] The ID of the Traffic Mirror session.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:\preformatted{list(
  TrafficMirrorSessionId = "string"
)
}
}
\description{
Deletes the specified Traffic Mirror session.
}
\section{Request syntax}{
\preformatted{svc$delete_traffic_mirror_session(
  TrafficMirrorSessionId = "string",
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
