% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_traffic_mirror_target}
\alias{ec2_delete_traffic_mirror_target}
\title{Deletes the specified Traffic Mirror target}
\usage{
ec2_delete_traffic_mirror_target(TrafficMirrorTargetId, DryRun = NULL)
}
\arguments{
\item{TrafficMirrorTargetId}{[required] The ID of the Traffic Mirror target.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified Traffic Mirror target.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_traffic_mirror_target/} for full documentation.
}
\keyword{internal}
