% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway}
\alias{ec2_delete_transit_gateway}
\title{Deletes the specified transit gateway}
\usage{
ec2_delete_transit_gateway(TransitGatewayId, DryRun = NULL)
}
\arguments{
\item{TransitGatewayId}{[required] The ID of the transit gateway.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified transit gateway.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_transit_gateway/} for full documentation.
}
\keyword{internal}
