% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_connect}
\alias{ec2_delete_transit_gateway_connect}
\title{Deletes the specified Connect attachment}
\usage{
ec2_delete_transit_gateway_connect(TransitGatewayAttachmentId, DryRun = NULL)
}
\arguments{
\item{TransitGatewayAttachmentId}{[required] The ID of the Connect attachment.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified Connect attachment. You must first delete any Connect peers for the attachment.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_transit_gateway_connect/} for full documentation.
}
\keyword{internal}
