% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_connect_peer}
\alias{ec2_delete_transit_gateway_connect_peer}
\title{Deletes the specified Connect peer}
\usage{
ec2_delete_transit_gateway_connect_peer(
  TransitGatewayConnectPeerId,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayConnectPeerId}{[required] The ID of the Connect peer.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified Connect peer.

See \url{https://paws-r.github.io/docs/ec2/delete_transit_gateway_connect_peer.html} for full documentation.
}
\keyword{internal}
