% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_transit_gateway_route_table}
\alias{ec2_delete_transit_gateway_route_table}
\title{Deletes the specified transit gateway route table}
\usage{
ec2_delete_transit_gateway_route_table(TransitGatewayRouteTableId,
  DryRun)
}
\arguments{
\item{TransitGatewayRouteTableId}{[required] The ID of the transit gateway route table.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TransitGatewayRouteTable = list(
    TransitGatewayRouteTableId = "string",
    TransitGatewayId = "string",
    State = "pending"|"available"|"deleting"|"deleted",
    DefaultAssociationRouteTable = TRUE|FALSE,
    DefaultPropagationRouteTable = TRUE|FALSE,
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    Tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified transit gateway route table. You must disassociate
the route table from any transit gateway route tables before you can
delete it.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_transit_gateway_route_table(
  TransitGatewayRouteTableId = "string",
  DryRun = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
