% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_volume}
\alias{ec2_delete_volume}
\title{Deletes the specified EBS volume}
\usage{
ec2_delete_volume(VolumeId, DryRun = NULL)
}
\arguments{
\item{VolumeId}{[required] The ID of the volume.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified EBS volume. The volume must be in the \code{available} state (not attached to an instance).

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_volume/} for full documentation.
}
\keyword{internal}
