% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_vpc_endpoint_connection_notifications}
\alias{ec2_delete_vpc_endpoint_connection_notifications}
\title{Deletes the specified VPC endpoint connection notifications}
\usage{
ec2_delete_vpc_endpoint_connection_notifications(
  DryRun = NULL,
  ConnectionNotificationIds
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ConnectionNotificationIds}{[required] The IDs of the notifications.}
}
\description{
Deletes the specified VPC endpoint connection notifications.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_vpc_endpoint_connection_notifications/} for full documentation.
}
\keyword{internal}
