% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_vpc_endpoint_service_configurations}
\alias{ec2_delete_vpc_endpoint_service_configurations}
\title{Deletes the specified VPC endpoint service configurations}
\usage{
ec2_delete_vpc_endpoint_service_configurations(DryRun = NULL, ServiceIds)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ServiceIds}{[required] The IDs of the services.}
}
\description{
Deletes the specified VPC endpoint service configurations. Before you can delete an endpoint service configuration, you must reject any \code{Available} or \code{PendingAcceptance} interface endpoint connections that are attached to the service.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_vpc_endpoint_service_configurations/} for full documentation.
}
\keyword{internal}
