% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_vpc_endpoints}
\alias{ec2_delete_vpc_endpoints}
\title{Deletes the specified VPC endpoints}
\usage{
ec2_delete_vpc_endpoints(DryRun = NULL, VpcEndpointIds)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{VpcEndpointIds}{[required] The IDs of the VPC endpoints.}
}
\description{
Deletes the specified VPC endpoints.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_vpc_endpoints/} for full documentation.
}
\keyword{internal}
