% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_vpn_connection}
\alias{ec2_delete_vpn_connection}
\title{Deletes the specified VPN connection}
\usage{
ec2_delete_vpn_connection(VpnConnectionId, DryRun = NULL)
}
\arguments{
\item{VpnConnectionId}{[required] The ID of the VPN connection.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified VPN connection.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_vpn_connection/} for full documentation.
}
\keyword{internal}
