% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_delete_vpn_gateway}
\alias{ec2_delete_vpn_gateway}
\title{Deletes the specified virtual private gateway}
\usage{
ec2_delete_vpn_gateway(VpnGatewayId, DryRun = NULL)
}
\arguments{
\item{VpnGatewayId}{[required] The ID of the virtual private gateway.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified virtual private gateway. You must first detach the virtual private gateway from the VPC. Note that you don't need to delete the virtual private gateway if you plan to delete and recreate the VPN connection between your VPC and your network.

See \url{https://www.paws-r-sdk.com/docs/ec2_delete_vpn_gateway/} for full documentation.
}
\keyword{internal}
