% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_aggregate_id_format}
\alias{ec2_describe_aggregate_id_format}
\title{Describes the longer ID format settings for all resource types in a
specific Region}
\usage{
ec2_describe_aggregate_id_format(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the longer ID format settings for all resource types in a specific Region. This request is useful for performing a quick audit to determine whether a specific Region is fully opted in for longer IDs (17-character IDs).

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_aggregate_id_format/} for full documentation.
}
\keyword{internal}
