% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_conversion_tasks}
\alias{ec2_describe_conversion_tasks}
\title{Describes the specified conversion tasks or all your conversion tasks}
\usage{
ec2_describe_conversion_tasks(ConversionTaskIds = NULL, DryRun = NULL)
}
\arguments{
\item{ConversionTaskIds}{The conversion task IDs.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the specified conversion tasks or all your conversion tasks. For more information, see the \href{https://docs.aws.amazon.com/vm-import/latest/userguide/}{VM Import/Export User Guide}.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_conversion_tasks/} for full documentation.
}
\keyword{internal}
