% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_egress_only_internet_gateways}
\alias{ec2_describe_egress_only_internet_gateways}
\title{Describes one or more of your egress-only internet gateways}
\usage{
ec2_describe_egress_only_internet_gateways(
  DryRun = NULL,
  EgressOnlyInternetGatewayIds = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Filters = NULL
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{EgressOnlyInternetGatewayIds}{The IDs of the egress-only internet gateways.}

\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token returned from a previous paginated request. Pagination
continues from the end of the items returned by the previous request.}

\item{Filters}{The filters.
\itemize{
\item \code{tag}:\<key\> - The key/value combination of a tag assigned to the
resource. Use the tag key in the filter name and the tag value as
the filter value. For example, to find all resources that have a tag
with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for
the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this
filter to find all resources assigned a tag with a specific key,
regardless of the tag value.
}}
}
\description{
Describes one or more of your egress-only internet gateways.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_egress_only_internet_gateways/} for full documentation.
}
\keyword{internal}
