% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_iam_instance_profile_associations}
\alias{ec2_describe_iam_instance_profile_associations}
\title{Describes your IAM instance profile associations}
\usage{
ec2_describe_iam_instance_profile_associations(
  AssociationIds = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AssociationIds}{The IAM instance profile associations.}

\item{Filters}{The filters.
\itemize{
\item \code{instance-id} - The ID of the instance.
\item \code{state} - The state of the association (\code{associating} | \code{associated}
| \code{disassociating}).
}}

\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token returned from a previous paginated request. Pagination
continues from the end of the items returned by the previous request.}
}
\description{
Describes your IAM instance profile associations.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_iam_instance_profile_associations/} for full documentation.
}
\keyword{internal}
