% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_instance_event_notification_attributes}
\alias{ec2_describe_instance_event_notification_attributes}
\title{Describes the tag keys that are registered to appear in scheduled event
notifications for resources in the current Region}
\usage{
ec2_describe_instance_event_notification_attributes(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the tag keys that are registered to appear in scheduled event notifications for resources in the current Region.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_instance_event_notification_attributes/} for full documentation.
}
\keyword{internal}
