% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_reserved_instances_listings}
\alias{ec2_describe_reserved_instances_listings}
\title{Describes your account's Reserved Instance listings in the Reserved
Instance Marketplace}
\usage{
ec2_describe_reserved_instances_listings(
  Filters = NULL,
  ReservedInstancesId = NULL,
  ReservedInstancesListingId = NULL
)
}
\arguments{
\item{Filters}{One or more filters.
\itemize{
\item \code{reserved-instances-id} - The ID of the Reserved Instances.
\item \code{reserved-instances-listing-id} - The ID of the Reserved Instances
listing.
\item \code{status} - The status of the Reserved Instance listing (\code{pending} |
\code{active} | \code{cancelled} | \code{closed}).
\item \code{status-message} - The reason for the status.
}}

\item{ReservedInstancesId}{One or more Reserved Instance IDs.}

\item{ReservedInstancesListingId}{One or more Reserved Instance listing IDs.}
}
\description{
Describes your account's Reserved Instance listings in the Reserved Instance Marketplace.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_reserved_instances_listings/} for full documentation.
}
\keyword{internal}
