% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_security_group_references}
\alias{ec2_describe_security_group_references}
\title{Describes the VPCs on the other side of a VPC peering connection that
are referencing the security groups you've specified in this request}
\usage{
ec2_describe_security_group_references(DryRun = NULL, GroupId)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{GroupId}{[required] The IDs of the security groups in your account.}
}
\description{
Describes the VPCs on the other side of a VPC peering connection that are referencing the security groups you've specified in this request.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_security_group_references/} for full documentation.
}
\keyword{internal}
