% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_volume_attribute}
\alias{ec2_describe_volume_attribute}
\title{Describes the specified attribute of the specified volume}
\usage{
ec2_describe_volume_attribute(Attribute, VolumeId, DryRun = NULL)
}
\arguments{
\item{Attribute}{[required] The attribute of the volume. This parameter is required.}

\item{VolumeId}{[required] The ID of the volume.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the specified attribute of the specified volume. You can specify only one attribute at a time.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_volume_attribute/} for full documentation.
}
\keyword{internal}
