% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_vpc_classic_link_dns_support}
\alias{ec2_describe_vpc_classic_link_dns_support}
\title{We are retiring EC2-Classic on August 15, 2022}
\usage{
ec2_describe_vpc_classic_link_dns_support(
  MaxResults = NULL,
  NextToken = NULL,
  VpcIds = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The token for the next page of results.}

\item{VpcIds}{One or more VPC IDs.}
}
\description{
We are retiring EC2-Classic on August 15, 2022. We recommend that you migrate from EC2-Classic to a VPC. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/vpc-migrate.html}{Migrate from EC2-Classic to a VPC} in the \emph{Amazon Elastic Compute Cloud User Guide}.

See \url{https://paws-r.github.io/docs/ec2/describe_vpc_classic_link_dns_support.html} for full documentation.
}
\keyword{internal}
