% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_vpc_classic_link_dns_support}
\alias{ec2_describe_vpc_classic_link_dns_support}
\title{This action is deprecated}
\usage{
ec2_describe_vpc_classic_link_dns_support(
  MaxResults = NULL,
  NextToken = NULL,
  VpcIds = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of items to return for this request. To get the next
page of items, make another request with the token returned in the
output. For more information, see
\href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Query-Requests.html#api-pagination}{Pagination}.}

\item{NextToken}{The token returned from a previous paginated request. Pagination
continues from the end of the items returned by the previous request.}

\item{VpcIds}{The IDs of the VPCs.}
}
\description{
This action is deprecated.

See \url{https://www.paws-r-sdk.com/docs/ec2_describe_vpc_classic_link_dns_support/} for full documentation.
}
\keyword{internal}
