% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disassociate_subnet_cidr_block}
\alias{ec2_disassociate_subnet_cidr_block}
\title{Disassociates a CIDR block from a subnet}
\usage{
ec2_disassociate_subnet_cidr_block(AssociationId)
}
\arguments{
\item{AssociationId}{[required] The association ID for the CIDR block.}
}
\description{
Disassociates a CIDR block from a subnet. Currently, you can disassociate an IPv6 CIDR block only. You must detach or delete all gateways and resources that are associated with the CIDR block before you can disassociate it.

See \url{https://www.paws-r-sdk.com/docs/ec2_disassociate_subnet_cidr_block/} for full documentation.
}
\keyword{internal}
