% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disassociate_transit_gateway_multicast_domain}
\alias{ec2_disassociate_transit_gateway_multicast_domain}
\title{Disassociates the specified subnets from the transit gateway multicast
domain}
\usage{
ec2_disassociate_transit_gateway_multicast_domain(
  TransitGatewayMulticastDomainId,
  TransitGatewayAttachmentId,
  SubnetIds,
  DryRun = NULL
)
}
\arguments{
\item{TransitGatewayMulticastDomainId}{[required] The ID of the transit gateway multicast domain.}

\item{TransitGatewayAttachmentId}{[required] The ID of the attachment.}

\item{SubnetIds}{[required] The IDs of the subnets;}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Disassociates the specified subnets from the transit gateway multicast domain.

See \url{https://www.paws-r-sdk.com/docs/ec2_disassociate_transit_gateway_multicast_domain/} for full documentation.
}
\keyword{internal}
