% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_disassociate_trunk_interface}
\alias{ec2_disassociate_trunk_interface}
\title{Removes an association between a branch network interface with a trunk
network interface}
\usage{
ec2_disassociate_trunk_interface(
  AssociationId,
  ClientToken = NULL,
  DryRun = NULL
)
}
\arguments{
\item{AssociationId}{[required] The ID of the association}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{How to Ensure Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Removes an association between a branch network interface with a trunk network interface.

See \url{https://www.paws-r-sdk.com/docs/ec2_disassociate_trunk_interface/} for full documentation.
}
\keyword{internal}
