% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_enable_aws_network_performance_metric_subscription}
\alias{ec2_enable_aws_network_performance_metric_subscription}
\title{Enables Infrastructure Performance subscriptions}
\usage{
ec2_enable_aws_network_performance_metric_subscription(
  Source = NULL,
  Destination = NULL,
  Metric = NULL,
  Statistic = NULL,
  DryRun = NULL
)
}
\arguments{
\item{Source}{The source Region (like \code{us-east-1}) or Availability Zone ID (like
\code{use1-az1}) that the metric subscription is enabled for. If you use
Availability Zone IDs, the Source and Destination Availability Zones
must be in the same Region.}

\item{Destination}{The target Region (like \code{us-east-2}) or Availability Zone ID (like
\code{use2-az2}) that the metric subscription is enabled for. If you use
Availability Zone IDs, the Source and Destination Availability Zones
must be in the same Region.}

\item{Metric}{The metric used for the enabled subscription.}

\item{Statistic}{The statistic used for the enabled subscription.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Enables Infrastructure Performance subscriptions.

See \url{https://www.paws-r-sdk.com/docs/ec2_enable_aws_network_performance_metric_subscription/} for full documentation.
}
\keyword{internal}
