% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_enable_ebs_encryption_by_default}
\alias{ec2_enable_ebs_encryption_by_default}
\title{Enables EBS encryption by default for your account in the current Region}
\usage{
ec2_enable_ebs_encryption_by_default(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Enables EBS encryption by default for your account in the current Region.

See \url{https://www.paws-r-sdk.com/docs/ec2_enable_ebs_encryption_by_default/} for full documentation.
}
\keyword{internal}
