% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_enable_transit_gateway_route_table_propagation}
\alias{ec2_enable_transit_gateway_route_table_propagation}
\title{Enables the specified attachment to propagate routes to the specified
propagation route table}
\usage{
ec2_enable_transit_gateway_route_table_propagation(
  TransitGatewayRouteTableId,
  TransitGatewayAttachmentId = NULL,
  DryRun = NULL,
  TransitGatewayRouteTableAnnouncementId = NULL
)
}
\arguments{
\item{TransitGatewayRouteTableId}{[required] The ID of the propagation route table.}

\item{TransitGatewayAttachmentId}{The ID of the attachment.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{TransitGatewayRouteTableAnnouncementId}{The ID of the transit gateway route table announcement.}
}
\description{
Enables the specified attachment to propagate routes to the specified propagation route table.

See \url{https://www.paws-r-sdk.com/docs/ec2_enable_transit_gateway_route_table_propagation/} for full documentation.
}
\keyword{internal}
