% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_console_output}
\alias{ec2_get_console_output}
\title{Gets the console output for the specified instance}
\usage{
ec2_get_console_output(InstanceId, DryRun = NULL, Latest = NULL)
}
\arguments{
\item{InstanceId}{[required] The ID of the instance.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Latest}{When enabled, retrieves the latest console output for the instance.

Default: disabled (\code{false})}
}
\description{
Gets the console output for the specified instance. For Linux instances, the instance console output displays the exact console output that would normally be displayed on a physical monitor attached to a computer. For Windows instances, the instance console output includes the last three system event log errors.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_console_output/} for full documentation.
}
\keyword{internal}
