% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_console_screenshot}
\alias{ec2_get_console_screenshot}
\title{Retrieve a JPG-format screenshot of a running instance to help with
troubleshooting}
\usage{
ec2_get_console_screenshot(DryRun = NULL, InstanceId, WakeUp = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{InstanceId}{[required] The ID of the instance.}

\item{WakeUp}{When set to \code{true}, acts as keystroke input and wakes up an instance
that's in standby or "sleep" mode.}
}
\description{
Retrieve a JPG-format screenshot of a running instance to help with troubleshooting.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_console_screenshot/} for full documentation.
}
\keyword{internal}
