% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_default_credit_specification}
\alias{ec2_get_default_credit_specification}
\title{Describes the default credit option for CPU usage of a burstable
performance instance family}
\usage{
ec2_get_default_credit_specification(DryRun = NULL, InstanceFamily)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{InstanceFamily}{[required] The instance family.}
}
\description{
Describes the default credit option for CPU usage of a burstable performance instance family.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_default_credit_specification/} for full documentation.
}
\keyword{internal}
