% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_ebs_default_kms_key_id}
\alias{ec2_get_ebs_default_kms_key_id}
\title{Describes the default KMS key for EBS encryption by default for your
account in this Region}
\usage{
ec2_get_ebs_default_kms_key_id(DryRun = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the default KMS key for EBS encryption by default for your account in this Region.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_ebs_default_kms_key_id/} for full documentation.
}
\keyword{internal}
