% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_host_reservation_purchase_preview}
\alias{ec2_get_host_reservation_purchase_preview}
\title{Preview a reservation purchase with configurations that match those of
your Dedicated Host}
\usage{
ec2_get_host_reservation_purchase_preview(HostIdSet, OfferingId)
}
\arguments{
\item{HostIdSet}{[required] The IDs of the Dedicated Hosts with which the reservation is associated.}

\item{OfferingId}{[required] The offering ID of the reservation.}
}
\description{
Preview a reservation purchase with configurations that match those of your Dedicated Host. You must have active Dedicated Hosts in your account before you purchase a reservation.

See \url{https://www.paws-r-sdk.com/docs/ec2_get_host_reservation_purchase_preview/} for full documentation.
}
\keyword{internal}
