% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_import_instance}
\alias{ec2_import_instance}
\title{Creates an import instance task using metadata from the specified disk
image}
\usage{
ec2_import_instance(
  Description = NULL,
  DiskImages = NULL,
  DryRun = NULL,
  LaunchSpecification = NULL,
  Platform
)
}
\arguments{
\item{Description}{A description for the instance being imported.}

\item{DiskImages}{The disk image.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{LaunchSpecification}{The launch specification.}

\item{Platform}{[required] The instance operating system.}
}
\description{
Creates an import instance task using metadata from the specified disk image.

See \url{https://www.paws-r-sdk.com/docs/ec2_import_instance/} for full documentation.
}
\keyword{internal}
