% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_ipam_scope}
\alias{ec2_modify_ipam_scope}
\title{Modify an IPAM scope}
\usage{
ec2_modify_ipam_scope(DryRun = NULL, IpamScopeId, Description = NULL)
}
\arguments{
\item{DryRun}{A check for whether you have the required permissions for the action
without actually making the request and provides an error response. If
you have the required permissions, the error response is
\code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpamScopeId}{[required] The ID of the scope you want to modify.}

\item{Description}{The description of the scope you want to modify.}
}
\description{
Modify an IPAM scope.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_ipam_scope/} for full documentation.
}
\keyword{internal}
