% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_traffic_mirror_filter_network_services}
\alias{ec2_modify_traffic_mirror_filter_network_services}
\title{Allows or restricts mirroring network services}
\usage{
ec2_modify_traffic_mirror_filter_network_services(
  TrafficMirrorFilterId,
  AddNetworkServices = NULL,
  RemoveNetworkServices = NULL,
  DryRun = NULL
)
}
\arguments{
\item{TrafficMirrorFilterId}{[required] The ID of the Traffic Mirror filter.}

\item{AddNetworkServices}{The network service, for example Amazon DNS, that you want to mirror.}

\item{RemoveNetworkServices}{The network service, for example Amazon DNS, that you no longer want to
mirror.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Allows or restricts mirroring network services.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_traffic_mirror_filter_network_services/} for full documentation.
}
\keyword{internal}
