% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_vpc_peering_connection_options}
\alias{ec2_modify_vpc_peering_connection_options}
\title{We are retiring EC2-Classic}
\usage{
ec2_modify_vpc_peering_connection_options(
  AccepterPeeringConnectionOptions = NULL,
  DryRun = NULL,
  RequesterPeeringConnectionOptions = NULL,
  VpcPeeringConnectionId
)
}
\arguments{
\item{AccepterPeeringConnectionOptions}{The VPC peering connection options for the accepter VPC.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{RequesterPeeringConnectionOptions}{The VPC peering connection options for the requester VPC.}

\item{VpcPeeringConnectionId}{[required] The ID of the VPC peering connection.}
}
\description{
We are retiring EC2-Classic. We recommend that you migrate from EC2-Classic to a VPC. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/vpc-migrate.html}{Migrate from EC2-Classic to a VPC} in the \emph{Amazon Elastic Compute Cloud User Guide}.

See \url{https://paws-r.github.io/docs/ec2/modify_vpc_peering_connection_options.html} for full documentation.
}
\keyword{internal}
