% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_modify_vpc_peering_connection_options}
\alias{ec2_modify_vpc_peering_connection_options}
\title{Modifies the VPC peering connection options on one side of a VPC peering
connection}
\usage{
ec2_modify_vpc_peering_connection_options(
  AccepterPeeringConnectionOptions = NULL,
  DryRun = NULL,
  RequesterPeeringConnectionOptions = NULL,
  VpcPeeringConnectionId
)
}
\arguments{
\item{AccepterPeeringConnectionOptions}{The VPC peering connection options for the accepter VPC.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{RequesterPeeringConnectionOptions}{The VPC peering connection options for the requester VPC.}

\item{VpcPeeringConnectionId}{[required] The ID of the VPC peering connection.}
}
\description{
Modifies the VPC peering connection options on one side of a VPC peering connection.

See \url{https://www.paws-r-sdk.com/docs/ec2_modify_vpc_peering_connection_options/} for full documentation.
}
\keyword{internal}
