% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_purchase_reserved_instances_offering}
\alias{ec2_purchase_reserved_instances_offering}
\title{Purchases a Reserved Instance for use with your account}
\usage{
ec2_purchase_reserved_instances_offering(
  InstanceCount,
  ReservedInstancesOfferingId,
  DryRun = NULL,
  LimitPrice = NULL,
  PurchaseTime = NULL
)
}
\arguments{
\item{InstanceCount}{[required] The number of Reserved Instances to purchase.}

\item{ReservedInstancesOfferingId}{[required] The ID of the Reserved Instance offering to purchase.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{LimitPrice}{Specified for Reserved Instance Marketplace offerings to limit the total
order and ensure that the Reserved Instances are not purchased at
unexpected prices.}

\item{PurchaseTime}{The time at which to purchase the Reserved Instance, in UTC format (for
example, \emph{YYYY}-\emph{MM}-\emph{DD}T\emph{HH}:\emph{MM}:\emph{SS}Z).}
}
\description{
Purchases a Reserved Instance for use with your account. With Reserved Instances, you pay a lower hourly rate compared to On-Demand instance pricing.

See \url{https://www.paws-r-sdk.com/docs/ec2_purchase_reserved_instances_offering/} for full documentation.
}
\keyword{internal}
