% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_register_instance_event_notification_attributes}
\alias{ec2_register_instance_event_notification_attributes}
\title{Registers a set of tag keys to include in scheduled event notifications
for your resources}
\usage{
ec2_register_instance_event_notification_attributes(
  DryRun = NULL,
  InstanceTagAttribute
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{InstanceTagAttribute}{[required] Information about the tag keys to register.}
}
\description{
Registers a set of tag keys to include in scheduled event notifications for your resources.

See \url{https://www.paws-r-sdk.com/docs/ec2_register_instance_event_notification_attributes/} for full documentation.
}
\keyword{internal}
