% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_reject_vpc_endpoint_connections}
\alias{ec2_reject_vpc_endpoint_connections}
\title{Rejects VPC endpoint connection requests to your VPC endpoint service}
\usage{
ec2_reject_vpc_endpoint_connections(DryRun = NULL, ServiceId, VpcEndpointIds)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ServiceId}{[required] The ID of the service.}

\item{VpcEndpointIds}{[required] The IDs of the VPC endpoints.}
}
\description{
Rejects VPC endpoint connection requests to your VPC endpoint service.

See \url{https://www.paws-r-sdk.com/docs/ec2_reject_vpc_endpoint_connections/} for full documentation.
}
\keyword{internal}
