% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_restore_address_to_classic}
\alias{ec2_restore_address_to_classic}
\title{This action is deprecated}
\usage{
ec2_restore_address_to_classic(DryRun = NULL, PublicIp)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{PublicIp}{[required] The Elastic IP address.}
}
\description{
This action is deprecated.

See \url{https://www.paws-r-sdk.com/docs/ec2_restore_address_to_classic/} for full documentation.
}
\keyword{internal}
