% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_start_vpc_endpoint_service_private_dns_verification}
\alias{ec2_start_vpc_endpoint_service_private_dns_verification}
\title{Initiates the verification process to prove that the service provider
owns the private DNS name domain for the endpoint service}
\usage{
ec2_start_vpc_endpoint_service_private_dns_verification(
  DryRun = NULL,
  ServiceId
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ServiceId}{[required] The ID of the endpoint service.}
}
\description{
Initiates the verification process to prove that the service provider owns the private DNS name domain for the endpoint service.

See \url{https://www.paws-r-sdk.com/docs/ec2_start_vpc_endpoint_service_private_dns_verification/} for full documentation.
}
\keyword{internal}
