% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_unassign_ipv_6_addresses}
\alias{ec2_unassign_ipv_6_addresses}
\title{Unassigns the specified IPv6 addresses or Prefix Delegation prefixes
from a network interface}
\usage{
ec2_unassign_ipv_6_addresses(
  Ipv6Prefixes = NULL,
  NetworkInterfaceId,
  Ipv6Addresses = NULL
)
}
\arguments{
\item{Ipv6Prefixes}{The IPv6 prefixes to unassign from the network interface.}

\item{NetworkInterfaceId}{[required] The ID of the network interface.}

\item{Ipv6Addresses}{The IPv6 addresses to unassign from the network interface.}
}
\description{
Unassigns the specified IPv6 addresses or Prefix Delegation prefixes from a network interface.

See \url{https://www.paws-r-sdk.com/docs/ec2_unassign_ipv_6_addresses/} for full documentation.
}
\keyword{internal}
