% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_unassign_private_ip_addresses}
\alias{ec2_unassign_private_ip_addresses}
\title{Unassigns one or more secondary private IP addresses, or IPv4 Prefix
Delegation prefixes from a network interface}
\usage{
ec2_unassign_private_ip_addresses(
  NetworkInterfaceId,
  PrivateIpAddresses = NULL,
  Ipv4Prefixes = NULL
)
}
\arguments{
\item{NetworkInterfaceId}{[required] The ID of the network interface.}

\item{PrivateIpAddresses}{The secondary private IP addresses to unassign from the network
interface. You can specify this option multiple times to unassign more
than one IP address.}

\item{Ipv4Prefixes}{The IPv4 prefixes to unassign from the network interface.}
}
\description{
Unassigns one or more secondary private IP addresses, or IPv4 Prefix Delegation prefixes from a network interface.

See \url{https://www.paws-r-sdk.com/docs/ec2_unassign_private_ip_addresses/} for full documentation.
}
\keyword{internal}
