% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_batch_check_layer_availability}
\alias{ecr_batch_check_layer_availability}
\title{Checks the availability of one or more image layers in a repository}
\usage{
ecr_batch_check_layer_availability(
  registryId = NULL,
  repositoryName,
  layerDigests
)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID associated with the registry that
contains the image layers to check. If you do not specify a registry,
the default registry is assumed.}

\item{repositoryName}{[required] The name of the repository that is associated with the image layers to
check.}

\item{layerDigests}{[required] The digests of the image layers to check.}
}
\description{
Checks the availability of one or more image layers in a repository.

See \url{https://www.paws-r-sdk.com/docs/ecr_batch_check_layer_availability/} for full documentation.
}
\keyword{internal}
