% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_delete_repository_policy}
\alias{ecr_delete_repository_policy}
\title{Deletes the repository policy associated with the specified repository}
\usage{
ecr_delete_repository_policy(registryId = NULL, repositoryName)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID associated with the registry that
contains the repository policy to delete. If you do not specify a
registry, the default registry is assumed.}

\item{repositoryName}{[required] The name of the repository that is associated with the repository policy
to delete.}
}
\description{
Deletes the repository policy associated with the specified repository.

See \url{https://www.paws-r-sdk.com/docs/ecr_delete_repository_policy/} for full documentation.
}
\keyword{internal}
