% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_describe_images}
\alias{ecr_describe_images}
\title{Returns metadata about the images in a repository}
\usage{
ecr_describe_images(
  registryId = NULL,
  repositoryName,
  imageIds = NULL,
  nextToken = NULL,
  maxResults = NULL,
  filter = NULL
)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID associated with the registry that
contains the repository in which to describe images. If you do not
specify a registry, the default registry is assumed.}

\item{repositoryName}{[required] The repository that contains the images to describe.}

\item{imageIds}{The list of image IDs for the requested repository.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{\link[=ecr_describe_images]{describe_images}} request where \code{maxResults} was
used and the results exceeded the value of that parameter. Pagination
continues from the end of the previous results that returned the
\code{nextToken} value. This value is \code{null} when there are no more results
to return. This option cannot be used when you specify images with
\code{imageIds}.}

\item{maxResults}{The maximum number of repository results returned by
\code{\link[=ecr_describe_images]{describe_images}} in paginated output. When this
parameter is used, \code{\link[=ecr_describe_images]{describe_images}} only returns
\code{maxResults} results in a single page along with a \code{nextToken} response
element. The remaining results of the initial request can be seen by
sending another \code{\link[=ecr_describe_images]{describe_images}} request with
the returned \code{nextToken} value. This value can be between 1 and 1000. If
this parameter is not used, then
\code{\link[=ecr_describe_images]{describe_images}} returns up to 100 results and a
\code{nextToken} value, if applicable. This option cannot be used when you
specify images with \code{imageIds}.}

\item{filter}{The filter key and value with which to filter your
\code{\link[=ecr_describe_images]{describe_images}} results.}
}
\description{
Returns metadata about the images in a repository.

See \url{https://www.paws-r-sdk.com/docs/ecr_describe_images/} for full documentation.
}
\keyword{internal}
