% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_get_account_setting}
\alias{ecr_get_account_setting}
\title{Retrieves the basic scan type version name}
\usage{
ecr_get_account_setting(name)
}
\arguments{
\item{name}{[required] Basic scan type version name.}
}
\description{
Retrieves the basic scan type version name.

See \url{https://www.paws-r-sdk.com/docs/ecr_get_account_setting/} for full documentation.
}
\keyword{internal}
