% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_get_authorization_token}
\alias{ecr_get_authorization_token}
\title{Retrieves an authorization token}
\usage{
ecr_get_authorization_token(registryIds = NULL)
}
\arguments{
\item{registryIds}{A list of Amazon Web Services account IDs that are associated with the
registries for which to get AuthorizationData objects. If you do not
specify a registry, the default registry is assumed.}
}
\description{
Retrieves an authorization token. An authorization token represents your IAM authentication credentials and can be used to access any Amazon ECR registry that your IAM principal has access to. The authorization token is valid for 12 hours.

See \url{https://www.paws-r-sdk.com/docs/ecr_get_authorization_token/} for full documentation.
}
\keyword{internal}
