% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_get_download_url_for_layer}
\alias{ecr_get_download_url_for_layer}
\title{Retrieves the pre-signed Amazon S3 download URL corresponding to an
image layer}
\usage{
ecr_get_download_url_for_layer(registryId = NULL, repositoryName, layerDigest)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID associated with the registry that
contains the image layer to download. If you do not specify a registry,
the default registry is assumed.}

\item{repositoryName}{[required] The name of the repository that is associated with the image layer to
download.}

\item{layerDigest}{[required] The digest of the image layer to download.}
}
\description{
Retrieves the pre-signed Amazon S3 download URL corresponding to an image layer. You can only get URLs for image layers that are referenced in an image.

See \url{https://www.paws-r-sdk.com/docs/ecr_get_download_url_for_layer/} for full documentation.
}
\keyword{internal}
