% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_get_registry_policy}
\alias{ecr_get_registry_policy}
\title{Retrieves the permissions policy for a registry}
\usage{
ecr_get_registry_policy()
}
\description{
Retrieves the permissions policy for a registry.

See \url{https://www.paws-r-sdk.com/docs/ecr_get_registry_policy/} for full documentation.
}
\keyword{internal}
