% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_initiate_layer_upload}
\alias{ecr_initiate_layer_upload}
\title{Notifies Amazon ECR that you intend to upload an image layer}
\usage{
ecr_initiate_layer_upload(registryId = NULL, repositoryName)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID associated with the registry to which
you intend to upload layers. If you do not specify a registry, the
default registry is assumed.}

\item{repositoryName}{[required] The name of the repository to which you intend to upload layers.}
}
\description{
Notifies Amazon ECR that you intend to upload an image layer.

See \url{https://www.paws-r-sdk.com/docs/ecr_initiate_layer_upload/} for full documentation.
}
\keyword{internal}
