% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_list_tags_for_resource}
\alias{ecr_list_tags_for_resource}
\title{List the tags for an Amazon ECR resource}
\usage{
ecr_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource for which to
list the tags. Currently, the only supported resource is an Amazon ECR
repository.}
}
\description{
List the tags for an Amazon ECR resource.

See \url{https://www.paws-r-sdk.com/docs/ecr_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
