% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_put_account_setting}
\alias{ecr_put_account_setting}
\title{Allows you to change the basic scan type version by setting the name
parameter to either CLAIR to AWS_NATIVE}
\usage{
ecr_put_account_setting(name, value)
}
\arguments{
\item{name}{[required] Basic scan type version name.}

\item{value}{[required] Setting value that determines what basic scan type is being used:
\code{AWS_NATIVE} or \code{CLAIR}.}
}
\description{
Allows you to change the basic scan type version by setting the \code{name} parameter to either \code{CLAIR} to \code{AWS_NATIVE}.

See \url{https://www.paws-r-sdk.com/docs/ecr_put_account_setting/} for full documentation.
}
\keyword{internal}
