% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_put_lifecycle_policy}
\alias{ecr_put_lifecycle_policy}
\title{Creates or updates the lifecycle policy for the specified repository}
\usage{
ecr_put_lifecycle_policy(
  registryId = NULL,
  repositoryName,
  lifecyclePolicyText
)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID associated with the registry that
contains the repository. If you do not specify a registry, the default
registry is assumed.}

\item{repositoryName}{[required] The name of the repository to receive the policy.}

\item{lifecyclePolicyText}{[required] The JSON repository policy text to apply to the repository.}
}
\description{
Creates or updates the lifecycle policy for the specified repository. For more information, see \href{https://docs.aws.amazon.com/AmazonECR/latest/userguide/LifecyclePolicies.html}{Lifecycle policy template}.

See \url{https://www.paws-r-sdk.com/docs/ecr_put_lifecycle_policy/} for full documentation.
}
\keyword{internal}
