% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_put_registry_policy}
\alias{ecr_put_registry_policy}
\title{Creates or updates the permissions policy for your registry}
\usage{
ecr_put_registry_policy(policyText)
}
\arguments{
\item{policyText}{[required] The JSON policy text to apply to your registry. The policy text follows
the same format as IAM policy text. For more information, see \href{https://docs.aws.amazon.com/AmazonECR/latest/userguide/registry-permissions.html}{Registry permissions}
in the \emph{Amazon Elastic Container Registry User Guide}.}
}
\description{
Creates or updates the permissions policy for your registry.

See \url{https://www.paws-r-sdk.com/docs/ecr_put_registry_policy/} for full documentation.
}
\keyword{internal}
