% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_start_lifecycle_policy_preview}
\alias{ecr_start_lifecycle_policy_preview}
\title{Starts a preview of a lifecycle policy for the specified repository}
\usage{
ecr_start_lifecycle_policy_preview(
  registryId = NULL,
  repositoryName,
  lifecyclePolicyText = NULL
)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID associated with the registry that
contains the repository. If you do not specify a registry, the default
registry is assumed.}

\item{repositoryName}{[required] The name of the repository to be evaluated.}

\item{lifecyclePolicyText}{The policy to be evaluated against. If you do not specify a policy, the
current policy for the repository is used.}
}
\description{
Starts a preview of a lifecycle policy for the specified repository. This allows you to see the results before associating the lifecycle policy with the repository.

See \url{https://www.paws-r-sdk.com/docs/ecr_start_lifecycle_policy_preview/} for full documentation.
}
\keyword{internal}
