% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_operations.R
\name{ecr_update_pull_through_cache_rule}
\alias{ecr_update_pull_through_cache_rule}
\title{Updates an existing pull through cache rule}
\usage{
ecr_update_pull_through_cache_rule(
  registryId = NULL,
  ecrRepositoryPrefix,
  credentialArn
)
}
\arguments{
\item{registryId}{The Amazon Web Services account ID associated with the registry
associated with the pull through cache rule. If you do not specify a
registry, the default registry is assumed.}

\item{ecrRepositoryPrefix}{[required] The repository name prefix to use when caching images from the source
registry.}

\item{credentialArn}{[required] The Amazon Resource Name (ARN) of the Amazon Web Services Secrets
Manager secret that identifies the credentials to authenticate to the
upstream registry.}
}
\description{
Updates an existing pull through cache rule.

See \url{https://www.paws-r-sdk.com/docs/ecr_update_pull_through_cache_rule/} for full documentation.
}
\keyword{internal}
