% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_get_authorization_token}
\alias{ecrpublic_get_authorization_token}
\title{Retrieves an authorization token}
\usage{
ecrpublic_get_authorization_token()
}
\description{
Retrieves an authorization token. An authorization token represents your IAM authentication credentials. You can use it to access any Amazon ECR registry that your IAM principal has access to. The authorization token is valid for 12 hours. This API requires the \code{ecr-public:GetAuthorizationToken} and \code{sts:GetServiceBearerToken} permissions.

See \url{https://www.paws-r-sdk.com/docs/ecrpublic_get_authorization_token/} for full documentation.
}
\keyword{internal}
