% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrpublic_operations.R
\name{ecrpublic_list_tags_for_resource}
\alias{ecrpublic_list_tags_for_resource}
\title{List the tags for an Amazon ECR Public resource}
\usage{
ecrpublic_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource to list the
tags for. Currently, the supported resource is an Amazon ECR Public
repository.}
}
\description{
List the tags for an Amazon ECR Public resource.

See \url{https://www.paws-r-sdk.com/docs/ecrpublic_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
