% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_delete_account_setting}
\alias{ecs_delete_account_setting}
\title{Disables an account setting for a specified IAM user, IAM role, or the
root user for an account}
\usage{
ecs_delete_account_setting(name, principalArn)
}
\arguments{
\item{name}{[required] The resource name for which to disable the account setting. If
\code{serviceLongArnFormat} is specified, the ARN for your Amazon ECS
services is affected. If \code{taskLongArnFormat} is specified, the ARN and
resource ID for your Amazon ECS tasks is affected. If
\code{containerInstanceLongArnFormat} is specified, the ARN and resource ID
for your Amazon ECS container instances is affected. If \code{awsvpcTrunking}
is specified, the ENI limit for your Amazon ECS container instances is
affected.}

\item{principalArn}{The ARN of the principal, which can be an IAM user, IAM role, or the
root user. If you specify the root user, it disables the account setting
for all IAM users, IAM roles, and the root user of the account unless an
IAM user or role explicitly overrides these settings. If this field is
omitted, the setting is changed only for the authenticated user.}
}
\value{
A list with the following syntax:\preformatted{list(
  setting = list(
    name = "serviceLongArnFormat"|"taskLongArnFormat"|"containerInstanceLongArnFormat"|"awsvpcTrunking"|"containerInsights",
    value = "string",
    principalArn = "string"
  )
)
}
}
\description{
Disables an account setting for a specified IAM user, IAM role, or the
root user for an account.
}
\section{Request syntax}{
\preformatted{svc$delete_account_setting(
  name = "serviceLongArnFormat"|"taskLongArnFormat"|"containerInstanceLongArnFormat"|"awsvpcTrunking"|"containerInsights",
  principalArn = "string"
)
}
}

\examples{
\dontrun{
# This example deletes the account setting for your user for the specified
# resource type.
svc$delete_account_setting(
  name = "serviceLongArnFormat"
)

# This example deletes the account setting for a specific IAM user or IAM
# role for the specified resource type. Only the root user can view or
# modify the account settings for another user.
svc$delete_account_setting(
  name = "containerInstanceLongArnFormat",
  principalArn = "arn:aws:iam::<aws_account_id>:user/principalName"
)
}

}
\keyword{internal}
