% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_delete_attributes}
\alias{ecs_delete_attributes}
\title{Deletes one or more custom attributes from an Amazon ECS resource}
\usage{
ecs_delete_attributes(cluster = NULL, attributes)
}
\arguments{
\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster that
contains the resource to delete attributes. If you do not specify a
cluster, the default cluster is assumed.}

\item{attributes}{[required] The attributes to delete from your resource. You can specify up to 10
attributes for each request. For custom attributes, specify the
attribute name and target ID, but don't specify the value. If you
specify the target ID using the short form, you must also specify the
target type.}
}
\description{
Deletes one or more custom attributes from an Amazon ECS resource.

See \url{https://www.paws-r-sdk.com/docs/ecs_delete_attributes/} for full documentation.
}
\keyword{internal}
