% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_delete_capacity_provider}
\alias{ecs_delete_capacity_provider}
\title{Deletes the specified capacity provider}
\usage{
ecs_delete_capacity_provider(capacityProvider)
}
\arguments{
\item{capacityProvider}{[required] The short name or full Amazon Resource Name (ARN) of the capacity
provider to delete.}
}
\description{
Deletes the specified capacity provider.

See \url{https://www.paws-r-sdk.com/docs/ecs_delete_capacity_provider/} for full documentation.
}
\keyword{internal}
