% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_describe_container_instances}
\alias{ecs_describe_container_instances}
\title{Describes one or more container instances}
\usage{
ecs_describe_container_instances(
  cluster = NULL,
  containerInstances,
  include = NULL
)
}
\arguments{
\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster that
hosts the container instances to describe. If you do not specify a
cluster, the default cluster is assumed. This parameter is required if
the container instance or container instances you are describing were
launched in any cluster other than the default cluster.}

\item{containerInstances}{[required] A list of up to 100 container instance IDs or full Amazon Resource Name
(ARN) entries.}

\item{include}{Specifies whether you want to see the resource tags for the container
instance. If \code{TAGS} is specified, the tags are included in the response.
If \code{CONTAINER_INSTANCE_HEALTH} is specified, the container instance
health is included in the response. If this field is omitted, tags and
container instance health status aren't included in the response.}
}
\description{
Describes one or more container instances. Returns metadata about each container instance requested.

See \url{https://www.paws-r-sdk.com/docs/ecs_describe_container_instances/} for full documentation.
}
\keyword{internal}
