% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_describe_task_definition}
\alias{ecs_describe_task_definition}
\title{Describes a task definition}
\usage{
ecs_describe_task_definition(taskDefinition, include = NULL)
}
\arguments{
\item{taskDefinition}{[required] The \code{family} for the latest \code{ACTIVE} revision, \code{family} and \code{revision}
(\code{family:revision}) for a specific revision in the family, or full
Amazon Resource Name (ARN) of the task definition to describe.}

\item{include}{Determines whether to see the resource tags for the task definition. If
\code{TAGS} is specified, the tags are included in the response. If this
field is omitted, tags aren't included in the response.}
}
\description{
Describes a task definition. You can specify a \code{family} and \code{revision} to find information about a specific task definition, or you can simply specify the family to find the latest \code{ACTIVE} revision in that family.

See \url{https://www.paws-r-sdk.com/docs/ecs_describe_task_definition/} for full documentation.
}
\keyword{internal}
