% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_describe_tasks}
\alias{ecs_describe_tasks}
\title{Describes a specified task or tasks}
\usage{
ecs_describe_tasks(cluster = NULL, tasks, include = NULL)
}
\arguments{
\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster that
hosts the task or tasks to describe. If you do not specify a cluster,
the default cluster is assumed.}

\item{tasks}{[required] A list of up to 100 task IDs or full ARN entries.}

\item{include}{Specifies whether you want to see the resource tags for the task. If
\code{TAGS} is specified, the tags are included in the response. If this
field is omitted, tags aren't included in the response.}
}
\description{
Describes a specified task or tasks.

See \url{https://www.paws-r-sdk.com/docs/ecs_describe_tasks/} for full documentation.
}
\keyword{internal}
