% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_list_tags_for_resource}
\alias{ecs_list_tags_for_resource}
\title{List the tags for an Amazon ECS resource}
\usage{
ecs_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource to list the
tags for. Currently, the supported resources are Amazon ECS tasks,
services, task definitions, clusters, and container instances.}
}
\description{
List the tags for an Amazon ECS resource.

See \url{https://www.paws-r-sdk.com/docs/ecs_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
