% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_register_daemon_task_definition}
\alias{ecs_register_daemon_task_definition}
\title{Registers a new daemon task definition from the supplied family and
containerDefinitions}
\usage{
ecs_register_daemon_task_definition(
  family,
  taskRoleArn = NULL,
  executionRoleArn = NULL,
  containerDefinitions,
  cpu = NULL,
  memory = NULL,
  volumes = NULL,
  tags = NULL
)
}
\arguments{
\item{family}{[required] You must specify a \code{family} for a daemon task definition. This family is
used as a name for your daemon task definition. Up to 255 letters
(uppercase and lowercase), numbers, underscores, and hyphens are
allowed.}

\item{taskRoleArn}{The short name or full Amazon Resource Name (ARN) of the IAM role that
containers in this daemon task can assume. All containers in this daemon
task are granted the permissions that are specified in this role.}

\item{executionRoleArn}{The Amazon Resource Name (ARN) of the task execution role that grants
the Amazon ECS container agent permission to make Amazon Web Services
API calls on your behalf. The task execution role is required for daemon
tasks that pull container images from Amazon ECR or send container logs
to CloudWatch.}

\item{containerDefinitions}{[required] A list of container definitions in JSON format that describe the
containers that make up your daemon task.}

\item{cpu}{The number of CPU units used by the daemon task. It can be expressed as
an integer using CPU units (for example, \code{1024}).}

\item{memory}{The amount of memory (in MiB) used by the daemon task. It can be
expressed as an integer using MiB (for example, \code{1024}).}

\item{volumes}{A list of volume definitions in JSON format that containers in your
daemon task can use.}

\item{tags}{The metadata that you apply to the daemon task definition to help you
categorize and organize them. Each tag consists of a key and an optional
value. You define both of them.

The following basic restrictions apply to tags:
\itemize{
\item Maximum number of tags per resource - 50
\item For each resource, each tag key must be unique, and each tag key can
have only one value.
\item Maximum key length - 128 Unicode characters in UTF-8
\item Maximum value length - 256 Unicode characters in UTF-8
\item If your tagging schema is used across multiple services and
resources, remember that other services may have restrictions on
allowed characters. Generally allowed characters are: letters,
numbers, and spaces representable in UTF-8, and the following
characters: + - = . _ : / @.
\item Tag keys and values are case-sensitive.
\item Do not use \verb{aws:}, \verb{AWS:}, or any upper or lowercase combination of
such as a prefix for either keys or values as it is reserved for
Amazon Web Services use. You cannot edit or delete tag keys or
values with this prefix. Tags with this prefix do not count against
your tags per resource limit.
}}
}
\description{
Registers a new daemon task definition from the supplied \code{family} and \code{containerDefinitions}. Optionally, you can add data volumes to your containers with the \code{volumes} parameter. For more information, see \href{https://docs.aws.amazon.com/AmazonECS/latest/developerguide/}{Daemon task definitions} in the \emph{Amazon Elastic Container Service Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/ecs_register_daemon_task_definition/} for full documentation.
}
\keyword{internal}
