% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_stop_task}
\alias{ecs_stop_task}
\title{Stops a running task}
\usage{
ecs_stop_task(cluster = NULL, task, reason = NULL)
}
\arguments{
\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster that
hosts the task to stop. If you do not specify a cluster, the default
cluster is assumed.}

\item{task}{[required] The task ID of the task to stop.}

\item{reason}{An optional message specified when a task is stopped. For example, if
you're using a custom scheduler, you can use this parameter to specify
the reason for stopping the task here, and the message appears in
subsequent \code{\link[=ecs_describe_tasks]{describe_tasks}} API operations on this
task.}
}
\description{
Stops a running task. Any tags associated with the task will be deleted.

See \url{https://www.paws-r-sdk.com/docs/ecs_stop_task/} for full documentation.
}
\keyword{internal}
