% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_submit_attachment_state_changes}
\alias{ecs_submit_attachment_state_changes}
\title{This action is only used by the Amazon ECS agent, and it is not intended
for use outside of the agent}
\usage{
ecs_submit_attachment_state_changes(cluster = NULL, attachments)
}
\arguments{
\item{cluster}{The short name or full ARN of the cluster that hosts the container
instance the attachment belongs to.}

\item{attachments}{[required] Any attachments associated with the state change request.}
}
\description{
This action is only used by the Amazon ECS agent, and it is not intended for use outside of the agent.

See \url{https://www.paws-r-sdk.com/docs/ecs_submit_attachment_state_changes/} for full documentation.
}
\keyword{internal}
