% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_submit_container_state_change}
\alias{ecs_submit_container_state_change}
\title{This action is only used by the Amazon ECS agent, and it is not intended
for use outside of the agent}
\usage{
ecs_submit_container_state_change(
  cluster = NULL,
  task = NULL,
  containerName = NULL,
  runtimeId = NULL,
  status = NULL,
  exitCode = NULL,
  reason = NULL,
  networkBindings = NULL
)
}
\arguments{
\item{cluster}{The short name or full ARN of the cluster that hosts the container.}

\item{task}{The task ID or full Amazon Resource Name (ARN) of the task that hosts
the container.}

\item{containerName}{The name of the container.}

\item{runtimeId}{The ID of the Docker container.}

\item{status}{The status of the state change request.}

\item{exitCode}{The exit code that's returned for the state change request.}

\item{reason}{The reason for the state change request.}

\item{networkBindings}{The network bindings of the container.}
}
\description{
This action is only used by the Amazon ECS agent, and it is not intended for use outside of the agent.

See \url{https://www.paws-r-sdk.com/docs/ecs_submit_container_state_change/} for full documentation.
}
\keyword{internal}
