% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_update_cluster}
\alias{ecs_update_cluster}
\title{Updates the cluster}
\usage{
ecs_update_cluster(cluster, settings, configuration)
}
\arguments{
\item{cluster}{[required] The name of the cluster to modify the settings for.}

\item{settings}{The cluster settings for your cluster.}

\item{configuration}{The execute command configuration for the cluster.}
}
\value{
A list with the following syntax:\preformatted{list(
  cluster = list(
    clusterArn = "string",
    clusterName = "string",
    configuration = list(
      executeCommandConfiguration = list(
        kmsKeyId = "string",
        logging = "NONE"|"DEFAULT"|"OVERRIDE",
        logConfiguration = list(
          cloudWatchLogGroupName = "string",
          cloudWatchEncryptionEnabled = TRUE|FALSE,
          s3BucketName = "string",
          s3EncryptionEnabled = TRUE|FALSE,
          s3KeyPrefix = "string"
        )
      )
    ),
    status = "string",
    registeredContainerInstancesCount = 123,
    runningTasksCount = 123,
    pendingTasksCount = 123,
    activeServicesCount = 123,
    statistics = list(
      list(
        name = "string",
        value = "string"
      )
    ),
    tags = list(
      list(
        key = "string",
        value = "string"
      )
    ),
    settings = list(
      list(
        name = "containerInsights",
        value = "string"
      )
    ),
    capacityProviders = list(
      "string"
    ),
    defaultCapacityProviderStrategy = list(
      list(
        capacityProvider = "string",
        weight = 123,
        base = 123
      )
    ),
    attachments = list(
      list(
        id = "string",
        type = "string",
        status = "string",
        details = list(
          list(
            name = "string",
            value = "string"
          )
        )
      )
    ),
    attachmentsStatus = "string"
  )
)
}
}
\description{
Updates the cluster.
}
\section{Request syntax}{
\preformatted{svc$update_cluster(
  cluster = "string",
  settings = list(
    list(
      name = "containerInsights",
      value = "string"
    )
  ),
  configuration = list(
    executeCommandConfiguration = list(
      kmsKeyId = "string",
      logging = "NONE"|"DEFAULT"|"OVERRIDE",
      logConfiguration = list(
        cloudWatchLogGroupName = "string",
        cloudWatchEncryptionEnabled = TRUE|FALSE,
        s3BucketName = "string",
        s3EncryptionEnabled = TRUE|FALSE,
        s3KeyPrefix = "string"
      )
    )
  )
)
}
}

\keyword{internal}
