% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_update_daemon}
\alias{ecs_update_daemon}
\title{Updates the specified daemon}
\usage{
ecs_update_daemon(
  daemonArn,
  daemonTaskDefinitionArn,
  capacityProviderArns,
  deploymentConfiguration = NULL,
  propagateTags = NULL,
  enableECSManagedTags = NULL,
  enableExecuteCommand = NULL
)
}
\arguments{
\item{daemonArn}{[required] The Amazon Resource Name (ARN) of the daemon to update.}

\item{daemonTaskDefinitionArn}{[required] The Amazon Resource Name (ARN) of the daemon task definition to use for
the updated daemon.}

\item{capacityProviderArns}{[required] The Amazon Resource Names (ARNs) of the capacity providers to associate
with the daemon.}

\item{deploymentConfiguration}{Optional deployment parameters that control how the daemon rolls out
updates, including the drain percentage, alarm-based rollback, and bake
time.}

\item{propagateTags}{Specifies whether to propagate the tags from the daemon to the daemon
tasks. If you don't specify a value, the tags aren't propagated. You can
only propagate tags to daemon tasks during task creation.}

\item{enableECSManagedTags}{Specifies whether to turn on Amazon ECS managed tags for the tasks in
the daemon. For more information, see \href{https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-using-tags.html}{Tagging your Amazon ECS resources}
in the \emph{Amazon Elastic Container Service Developer Guide}.}

\item{enableExecuteCommand}{If \code{true}, the execute command functionality is turned on for all tasks
in the daemon. If \code{false}, the execute command functionality is turned
off.}
}
\description{
Updates the specified daemon. When you update a daemon, a new deployment is triggered that progressively rolls out the changes to the container instances associated with the daemon's capacity providers. For more information, see \href{https://docs.aws.amazon.com/AmazonECS/latest/developerguide/}{Daemon deployments} in the \emph{Amazon Elastic Container Service Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/ecs_update_daemon/} for full documentation.
}
\keyword{internal}
