% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_associate_encryption_config}
\alias{eks_associate_encryption_config}
\title{Associate encryption configuration to an existing cluster}
\usage{
eks_associate_encryption_config(clusterName, encryptionConfig,
  clientRequestToken)
}
\arguments{
\item{clusterName}{[required] The name of the cluster that you are associating with encryption
configuration.}

\item{encryptionConfig}{[required] The configuration you are using for encryption.}

\item{clientRequestToken}{The client request token you are using with the encryption
configuration.}
}
\value{
A list with the following syntax:\preformatted{list(
  update = list(
    id = "string",
    status = "InProgress"|"Failed"|"Cancelled"|"Successful",
    type = "VersionUpdate"|"EndpointAccessUpdate"|"LoggingUpdate"|"ConfigUpdate"|"AssociateIdentityProviderConfig"|"DisassociateIdentityProviderConfig"|"AssociateEncryptionConfig"|"AddonUpdate",
    params = list(
      list(
        type = "Version"|"PlatformVersion"|"EndpointPrivateAccess"|"EndpointPublicAccess"|"ClusterLogging"|"DesiredSize"|"LabelsToAdd"|"LabelsToRemove"|"TaintsToAdd"|"TaintsToRemove"|"MaxSize"|"MinSize"|"ReleaseVersion"|"PublicAccessCidrs"|"LaunchTemplateName"|"LaunchTemplateVersion"|"IdentityProviderConfig"|"EncryptionConfig"|"AddonVersion"|"ServiceAccountRoleArn"|"ResolveConflicts",
        value = "string"
      )
    ),
    createdAt = as.POSIXct(
      "2015-01-01"
    ),
    errors = list(
      list(
        errorCode = "SubnetNotFound"|"SecurityGroupNotFound"|"EniLimitReached"|"IpNotAvailable"|"AccessDenied"|"OperationNotPermitted"|"VpcIdNotFound"|"Unknown"|"NodeCreationFailure"|"PodEvictionFailure"|"InsufficientFreeAddresses"|"ClusterUnreachable"|"InsufficientNumberOfReplicas"|"ConfigurationConflict"|"AdmissionRequestDenied",
        errorMessage = "string",
        resourceIds = list(
          "string"
        )
      )
    )
  )
)
}
}
\description{
Associate encryption configuration to an existing cluster.

You can use this API to enable encryption on existing clusters which do
not have encryption already enabled. This allows you to implement a
defense-in-depth security strategy without migrating applications to new
EKS clusters.
}
\section{Request syntax}{
\preformatted{svc$associate_encryption_config(
  clusterName = "string",
  encryptionConfig = list(
    list(
      resources = list(
        "string"
      ),
      provider = list(
        keyArn = "string"
      )
    )
  ),
  clientRequestToken = "string"
)
}
}

\keyword{internal}
