% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_associate_encryption_config}
\alias{eks_associate_encryption_config}
\title{Associates an encryption configuration to an existing cluster}
\usage{
eks_associate_encryption_config(
  clusterName,
  encryptionConfig,
  clientRequestToken = NULL
)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{encryptionConfig}{[required] The configuration you are using for encryption.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Associates an encryption configuration to an existing cluster.

See \url{https://www.paws-r-sdk.com/docs/eks_associate_encryption_config/} for full documentation.
}
\keyword{internal}
