% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_associate_identity_provider_config}
\alias{eks_associate_identity_provider_config}
\title{Associates an identity provider configuration to a cluster}
\usage{
eks_associate_identity_provider_config(
  clusterName,
  oidc,
  tags = NULL,
  clientRequestToken = NULL
)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{oidc}{[required] An object representing an OpenID Connect (OIDC) identity provider
configuration.}

\item{tags}{Metadata that assists with categorization and organization. Each tag
consists of a key and an optional value. You define both. Tags don't
propagate to any other cluster or Amazon Web Services resources.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Associates an identity provider configuration to a cluster.

See \url{https://www.paws-r-sdk.com/docs/eks_associate_identity_provider_config/} for full documentation.
}
\keyword{internal}
