% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_create_cluster}
\alias{eks_create_cluster}
\title{Creates an Amazon EKS control plane}
\usage{
eks_create_cluster(name, version, roleArn, resourcesVpcConfig,
  kubernetesNetworkConfig, logging, clientRequestToken, tags,
  encryptionConfig)
}
\arguments{
\item{name}{[required] The unique name to give to your cluster.}

\item{version}{The desired Kubernetes version for your cluster. If you don't specify a
value here, the latest version available in Amazon EKS is used.}

\item{roleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that provides permissions
for the Kubernetes control plane to make calls to AWS API operations on
your behalf. For more information, see \href{https://docs.aws.amazon.com/eks/latest/userguide/service_IAM_role.html}{Amazon EKS Service IAM Role}
in the \strong{Amazon EKS User Guide} .}

\item{resourcesVpcConfig}{[required] The VPC configuration used by the cluster control plane. Amazon EKS VPC
resources have specific requirements to work properly with Kubernetes.
For more information, see \href{https://docs.aws.amazon.com/eks/latest/userguide/network_reqs.html}{Cluster VPC Considerations}
and \href{https://docs.aws.amazon.com/eks/latest/userguide/sec-group-reqs.html}{Cluster Security Group Considerations}
in the \emph{Amazon EKS User Guide}. You must specify at least two subnets.
You can specify up to five security groups, but we recommend that you
use a dedicated security group for your cluster control plane.}

\item{kubernetesNetworkConfig}{The Kubernetes network configuration for the cluster.}

\item{logging}{Enable or disable exporting the Kubernetes control plane logs for your
cluster to CloudWatch Logs. By default, cluster control plane logs
aren't exported to CloudWatch Logs. For more information, see \href{https://docs.aws.amazon.com/eks/latest/userguide/control-plane-logs.html}{Amazon EKS Cluster Control Plane Logs}
in the \strong{Amazon EKS User Guide} .

CloudWatch Logs ingestion, archive storage, and data scanning rates
apply to exported control plane logs. For more information, see \href{https://aws.amazon.com/cloudwatch/pricing/}{Amazon CloudWatch Pricing}.}

\item{clientRequestToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{tags}{The metadata to apply to the cluster to assist with categorization and
organization. Each tag consists of a key and an optional value, both of
which you define.}

\item{encryptionConfig}{The encryption configuration for the cluster.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  cluster = list(
    name = "string",
    arn = "string",
    createdAt = as.POSIXct(
      "2015-01-01"
    ),
    version = "string",
    endpoint = "string",
    roleArn = "string",
    resourcesVpcConfig = list(
      subnetIds = list(
        "string"
      ),
      securityGroupIds = list(
        "string"
      ),
      clusterSecurityGroupId = "string",
      vpcId = "string",
      endpointPublicAccess = TRUE|FALSE,
      endpointPrivateAccess = TRUE|FALSE,
      publicAccessCidrs = list(
        "string"
      )
    ),
    kubernetesNetworkConfig = list(
      serviceIpv4Cidr = "string"
    ),
    logging = list(
      clusterLogging = list(
        list(
          types = list(
            "api"|"audit"|"authenticator"|"controllerManager"|"scheduler"
          ),
          enabled = TRUE|FALSE
        )
      )
    ),
    identity = list(
      oidc = list(
        issuer = "string"
      )
    ),
    status = "CREATING"|"ACTIVE"|"DELETING"|"FAILED"|"UPDATING",
    certificateAuthority = list(
      data = "string"
    ),
    clientRequestToken = "string",
    platformVersion = "string",
    tags = list(
      "string"
    ),
    encryptionConfig = list(
      list(
        resources = list(
          "string"
        ),
        provider = list(
          keyArn = "string"
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates an Amazon EKS control plane.

The Amazon EKS control plane consists of control plane instances that
run the Kubernetes software, such as \code{etcd} and the API server. The
control plane runs in an account managed by AWS, and the Kubernetes API
is exposed via the Amazon EKS API server endpoint. Each Amazon EKS
cluster control plane is single-tenant and unique and runs on its own
set of Amazon EC2 instances.

The cluster control plane is provisioned across multiple Availability
Zones and fronted by an Elastic Load Balancing Network Load Balancer.
Amazon EKS also provisions elastic network interfaces in your VPC
subnets to provide connectivity from the control plane instances to the
worker nodes (for example, to support \verb{kubectl exec}, \code{logs}, and
\code{proxy} data flows).

Amazon EKS worker nodes run in your AWS account and connect to your
cluster's control plane via the Kubernetes API server endpoint and a
certificate file that is created for your cluster.

You can use the \code{endpointPublicAccess} and \code{endpointPrivateAccess}
parameters to enable or disable public and private access to your
cluster's Kubernetes API server endpoint. By default, public access is
enabled, and private access is disabled. For more information, see
\href{https://docs.aws.amazon.com/eks/latest/userguide/cluster-endpoint.html}{Amazon EKS Cluster Endpoint Access Control}
in the \strong{Amazon EKS User Guide} .

You can use the \code{logging} parameter to enable or disable exporting the
Kubernetes control plane logs for your cluster to CloudWatch Logs. By
default, cluster control plane logs aren't exported to CloudWatch Logs.
For more information, see \href{https://docs.aws.amazon.com/eks/latest/userguide/control-plane-logs.html}{Amazon EKS Cluster Control Plane Logs}
in the \strong{Amazon EKS User Guide} .

CloudWatch Logs ingestion, archive storage, and data scanning rates
apply to exported control plane logs. For more information, see \href{https://aws.amazon.com/cloudwatch/pricing/}{Amazon CloudWatch Pricing}.

Cluster creation typically takes between 10 and 15 minutes. After you
create an Amazon EKS cluster, you must configure your Kubernetes tooling
to communicate with the API server and launch worker nodes into your
cluster. For more information, see \href{https://docs.aws.amazon.com/eks/latest/userguide/}{Managing Cluster Authentication} and
\href{https://docs.aws.amazon.com/eks/latest/userguide/launch-workers.html}{Launching Amazon EKS Worker Nodes}
in the \emph{Amazon EKS User Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_cluster(
  name = "string",
  version = "string",
  roleArn = "string",
  resourcesVpcConfig = list(
    subnetIds = list(
      "string"
    ),
    securityGroupIds = list(
      "string"
    ),
    endpointPublicAccess = TRUE|FALSE,
    endpointPrivateAccess = TRUE|FALSE,
    publicAccessCidrs = list(
      "string"
    )
  ),
  kubernetesNetworkConfig = list(
    serviceIpv4Cidr = "string"
  ),
  logging = list(
    clusterLogging = list(
      list(
        types = list(
          "api"|"audit"|"authenticator"|"controllerManager"|"scheduler"
        ),
        enabled = TRUE|FALSE
      )
    )
  ),
  clientRequestToken = "string",
  tags = list(
    "string"
  ),
  encryptionConfig = list(
    list(
      resources = list(
        "string"
      ),
      provider = list(
        keyArn = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example creates an Amazon EKS cluster called prod.
svc$create_cluster(
  version = "1.10",
  name = "prod",
  clientRequestToken = "1d2129a1-3d38-460a-9756-e5b91fddb951",
  resourcesVpcConfig = list(
    securityGroupIds = list(
      "sg-6979fe18"
    ),
    subnetIds = list(
      "subnet-6782e71e",
      "subnet-e7e761ac"
    )
  ),
  roleArn = "arn:aws:iam::012345678910:role/eks-service-role-AWSServiceRole..."
)
}

}
\keyword{internal}
