% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_delete_nodegroup}
\alias{eks_delete_nodegroup}
\title{Deletes a managed node group}
\usage{
eks_delete_nodegroup(clusterName, nodegroupName)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{nodegroupName}{[required] The name of the node group to delete.}
}
\description{
Deletes a managed node group.

See \url{https://www.paws-r-sdk.com/docs/eks_delete_nodegroup/} for full documentation.
}
\keyword{internal}
