% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_fargate_profile}
\alias{eks_describe_fargate_profile}
\title{Describes an Fargate profile}
\usage{
eks_describe_fargate_profile(clusterName, fargateProfileName)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{fargateProfileName}{[required] The name of the Fargate profile to describe.}
}
\description{
Describes an Fargate profile.

See \url{https://www.paws-r-sdk.com/docs/eks_describe_fargate_profile/} for full documentation.
}
\keyword{internal}
