% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_nodegroup}
\alias{eks_describe_nodegroup}
\title{Returns descriptive information about an Amazon EKS node group}
\usage{
eks_describe_nodegroup(clusterName, nodegroupName)
}
\arguments{
\item{clusterName}{[required] The name of the Amazon EKS cluster associated with the node group.}

\item{nodegroupName}{[required] The name of the node group to describe.}
}
\value{
A list with the following syntax:\preformatted{list(
  nodegroup = list(
    nodegroupName = "string",
    nodegroupArn = "string",
    clusterName = "string",
    version = "string",
    releaseVersion = "string",
    createdAt = as.POSIXct(
      "2015-01-01"
    ),
    modifiedAt = as.POSIXct(
      "2015-01-01"
    ),
    status = "CREATING"|"ACTIVE"|"UPDATING"|"DELETING"|"CREATE_FAILED"|"DELETE_FAILED"|"DEGRADED",
    capacityType = "ON_DEMAND"|"SPOT",
    scalingConfig = list(
      minSize = 123,
      maxSize = 123,
      desiredSize = 123
    ),
    instanceTypes = list(
      "string"
    ),
    subnets = list(
      "string"
    ),
    remoteAccess = list(
      ec2SshKey = "string",
      sourceSecurityGroups = list(
        "string"
      )
    ),
    amiType = "AL2_x86_64"|"AL2_x86_64_GPU"|"AL2_ARM_64"|"CUSTOM",
    nodeRole = "string",
    labels = list(
      "string"
    ),
    taints = list(
      list(
        key = "string",
        value = "string",
        effect = "NO_SCHEDULE"|"NO_EXECUTE"|"PREFER_NO_SCHEDULE"
      )
    ),
    resources = list(
      autoScalingGroups = list(
        list(
          name = "string"
        )
      ),
      remoteAccessSecurityGroup = "string"
    ),
    diskSize = 123,
    health = list(
      issues = list(
        list(
          code = "AutoScalingGroupNotFound"|"AutoScalingGroupInvalidConfiguration"|"Ec2SecurityGroupNotFound"|"Ec2SecurityGroupDeletionFailure"|"Ec2LaunchTemplateNotFound"|"Ec2LaunchTemplateVersionMismatch"|"Ec2SubnetNotFound"|"Ec2SubnetInvalidConfiguration"|"IamInstanceProfileNotFound"|"IamLimitExceeded"|"IamNodeRoleNotFound"|"NodeCreationFailure"|"AsgInstanceLaunchFailures"|"InstanceLimitExceeded"|"InsufficientFreeAddresses"|"AccessDenied"|"InternalFailure"|"ClusterUnreachable",
          message = "string",
          resourceIds = list(
            "string"
          )
        )
      )
    ),
    launchTemplate = list(
      name = "string",
      version = "string",
      id = "string"
    ),
    tags = list(
      "string"
    )
  )
)
}
}
\description{
Returns descriptive information about an Amazon EKS node group.
}
\section{Request syntax}{
\preformatted{svc$describe_nodegroup(
  clusterName = "string",
  nodegroupName = "string"
)
}
}

\keyword{internal}
