% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_describe_nodegroup}
\alias{eks_describe_nodegroup}
\title{Returns descriptive information about an Amazon EKS node group}
\usage{
eks_describe_nodegroup(clusterName, nodegroupName)
}
\arguments{
\item{clusterName}{[required] The name of the Amazon EKS cluster associated with the node group.}

\item{nodegroupName}{[required] The name of the node group to describe.}
}
\description{
Returns descriptive information about an Amazon EKS node group.

See \url{https://paws-r.github.io/docs/eks/describe_nodegroup.html} for full documentation.
}
\keyword{internal}
