% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_disassociate_identity_provider_config}
\alias{eks_disassociate_identity_provider_config}
\title{Disassociates an identity provider configuration from a cluster}
\usage{
eks_disassociate_identity_provider_config(clusterName,
  identityProviderConfig, clientRequestToken)
}
\arguments{
\item{clusterName}{[required] The name of the cluster to disassociate an identity provider from.}

\item{identityProviderConfig}{[required] An object that represents an identity provider configuration.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\value{
A list with the following syntax:\preformatted{list(
  update = list(
    id = "string",
    status = "InProgress"|"Failed"|"Cancelled"|"Successful",
    type = "VersionUpdate"|"EndpointAccessUpdate"|"LoggingUpdate"|"ConfigUpdate"|"AssociateIdentityProviderConfig"|"DisassociateIdentityProviderConfig"|"AssociateEncryptionConfig"|"AddonUpdate",
    params = list(
      list(
        type = "Version"|"PlatformVersion"|"EndpointPrivateAccess"|"EndpointPublicAccess"|"ClusterLogging"|"DesiredSize"|"LabelsToAdd"|"LabelsToRemove"|"TaintsToAdd"|"TaintsToRemove"|"MaxSize"|"MinSize"|"ReleaseVersion"|"PublicAccessCidrs"|"LaunchTemplateName"|"LaunchTemplateVersion"|"IdentityProviderConfig"|"EncryptionConfig"|"AddonVersion"|"ServiceAccountRoleArn"|"ResolveConflicts",
        value = "string"
      )
    ),
    createdAt = as.POSIXct(
      "2015-01-01"
    ),
    errors = list(
      list(
        errorCode = "SubnetNotFound"|"SecurityGroupNotFound"|"EniLimitReached"|"IpNotAvailable"|"AccessDenied"|"OperationNotPermitted"|"VpcIdNotFound"|"Unknown"|"NodeCreationFailure"|"PodEvictionFailure"|"InsufficientFreeAddresses"|"ClusterUnreachable"|"InsufficientNumberOfReplicas"|"ConfigurationConflict"|"AdmissionRequestDenied",
        errorMessage = "string",
        resourceIds = list(
          "string"
        )
      )
    )
  )
)
}
}
\description{
Disassociates an identity provider configuration from a cluster. If you
disassociate an identity provider from your cluster, users included in
the provider can no longer access the cluster. However, you can still
access the cluster with AWS IAM users.
}
\section{Request syntax}{
\preformatted{svc$disassociate_identity_provider_config(
  clusterName = "string",
  identityProviderConfig = list(
    type = "string",
    name = "string"
  ),
  clientRequestToken = "string"
)
}
}

\keyword{internal}
