% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_list_tags_for_resource}
\alias{eks_list_tags_for_resource}
\title{List the tags for an Amazon EKS resource}
\usage{
eks_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource to list tags
for.}
}
\description{
List the tags for an Amazon EKS resource.

See \url{https://www.paws-r-sdk.com/docs/eks_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
