% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_update_cluster_version}
\alias{eks_update_cluster_version}
\title{Updates an Amazon EKS cluster to the specified Kubernetes version}
\usage{
eks_update_cluster_version(name, version, clientRequestToken = NULL)
}
\arguments{
\item{name}{[required] The name of the Amazon EKS cluster to update.}

\item{version}{[required] The desired Kubernetes version following a successful update.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Updates an Amazon EKS cluster to the specified Kubernetes version. Your cluster continues to function during the update. The response output includes an update ID that you can use to track the status of your cluster update with the \code{\link[=eks_describe_update]{describe_update}} API operation.

See \url{https://www.paws-r-sdk.com/docs/eks_update_cluster_version/} for full documentation.
}
\keyword{internal}
