% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eks_operations.R
\name{eks_update_nodegroup_config}
\alias{eks_update_nodegroup_config}
\title{Updates an Amazon EKS managed node group configuration}
\usage{
eks_update_nodegroup_config(
  clusterName,
  nodegroupName,
  labels = NULL,
  taints = NULL,
  scalingConfig = NULL,
  updateConfig = NULL,
  clientRequestToken = NULL
)
}
\arguments{
\item{clusterName}{[required] The name of your cluster.}

\item{nodegroupName}{[required] The name of the managed node group to update.}

\item{labels}{The Kubernetes \code{labels} to apply to the nodes in the node group after
the update.}

\item{taints}{The Kubernetes taints to be applied to the nodes in the node group after
the update. For more information, see \href{https://docs.aws.amazon.com/eks/latest/userguide/node-taints-managed-node-groups.html}{Node taints on managed node groups}.}

\item{scalingConfig}{The scaling configuration details for the Auto Scaling group after the
update.}

\item{updateConfig}{The node group update configuration.}

\item{clientRequestToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Updates an Amazon EKS managed node group configuration. Your node group continues to function during the update. The response output includes an update ID that you can use to track the status of your node group update with the \code{\link[=eks_describe_update]{describe_update}} API operation. Currently you can update the Kubernetes labels for a node group or the scaling configuration.

See \url{https://www.paws-r-sdk.com/docs/eks_update_nodegroup_config/} for full documentation.
}
\keyword{internal}
