% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_abort_environment_update}
\alias{elasticbeanstalk_abort_environment_update}
\title{Cancels in-progress environment configuration update or application
version deployment}
\usage{
elasticbeanstalk_abort_environment_update(
  EnvironmentId = NULL,
  EnvironmentName = NULL
)
}
\arguments{
\item{EnvironmentId}{This specifies the ID of the environment with the in-progress update
that you want to cancel.}

\item{EnvironmentName}{This specifies the name of the environment with the in-progress update
that you want to cancel.}
}
\description{
Cancels in-progress environment configuration update or application version deployment.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_abort_environment_update/} for full documentation.
}
\keyword{internal}
