% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_apply_environment_managed_action}
\alias{elasticbeanstalk_apply_environment_managed_action}
\title{Applies a scheduled managed action immediately}
\usage{
elasticbeanstalk_apply_environment_managed_action(
  EnvironmentName = NULL,
  EnvironmentId = NULL,
  ActionId
)
}
\arguments{
\item{EnvironmentName}{The name of the target environment.}

\item{EnvironmentId}{The environment ID of the target environment.}

\item{ActionId}{[required] The action ID of the scheduled managed action to execute.}
}
\description{
Applies a scheduled managed action immediately. A managed action can be applied only if its status is \code{Scheduled}. Get the status and action ID of a managed action with \code{\link[=elasticbeanstalk_describe_environment_managed_actions]{describe_environment_managed_actions}}.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_apply_environment_managed_action/} for full documentation.
}
\keyword{internal}
