% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_check_dns_availability}
\alias{elasticbeanstalk_check_dns_availability}
\title{Checks if the specified CNAME is available}
\usage{
elasticbeanstalk_check_dns_availability(CNAMEPrefix)
}
\arguments{
\item{CNAMEPrefix}{[required] The prefix used when this CNAME is reserved.}
}
\description{
Checks if the specified CNAME is available.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_check_dns_availability/} for full documentation.
}
\keyword{internal}
