% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_create_application}
\alias{elasticbeanstalk_create_application}
\title{Creates an application that has one configuration template named default
and no application versions}
\usage{
elasticbeanstalk_create_application(
  ApplicationName,
  Description = NULL,
  ResourceLifecycleConfig = NULL,
  Tags = NULL
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application. Must be unique within your account.}

\item{Description}{Your description of the application.}

\item{ResourceLifecycleConfig}{Specifies an application resource lifecycle configuration to prevent
your application from accumulating too many versions.}

\item{Tags}{Specifies the tags applied to the application.

Elastic Beanstalk applies these tags only to the application.
Environments that you create in the application don't inherit the tags.}
}
\description{
Creates an application that has one configuration template named \code{default} and no application versions.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_create_application/} for full documentation.
}
\keyword{internal}
