% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_delete_configuration_template}
\alias{elasticbeanstalk_delete_configuration_template}
\title{Deletes the specified configuration template}
\usage{
elasticbeanstalk_delete_configuration_template(ApplicationName, TemplateName)
}
\arguments{
\item{ApplicationName}{[required] The name of the application to delete the configuration template from.}

\item{TemplateName}{[required] The name of the configuration template to delete.}
}
\description{
Deletes the specified configuration template.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_delete_configuration_template/} for full documentation.
}
\keyword{internal}
