% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_delete_environment_configuration}
\alias{elasticbeanstalk_delete_environment_configuration}
\title{Deletes the draft configuration associated with the running environment}
\usage{
elasticbeanstalk_delete_environment_configuration(
  ApplicationName,
  EnvironmentName
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application the environment is associated with.}

\item{EnvironmentName}{[required] The name of the environment to delete the draft configuration from.}
}
\description{
Deletes the draft configuration associated with the running environment.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_delete_environment_configuration/} for full documentation.
}
\keyword{internal}
