% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_delete_platform_version}
\alias{elasticbeanstalk_delete_platform_version}
\title{Deletes the specified version of a custom platform}
\usage{
elasticbeanstalk_delete_platform_version(PlatformArn = NULL)
}
\arguments{
\item{PlatformArn}{The ARN of the version of the custom platform.}
}
\description{
Deletes the specified version of a custom platform.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_delete_platform_version/} for full documentation.
}
\keyword{internal}
