% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_describe_account_attributes}
\alias{elasticbeanstalk_describe_account_attributes}
\title{Returns attributes related to AWS Elastic Beanstalk that are associated
with the calling AWS account}
\usage{
elasticbeanstalk_describe_account_attributes()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ResourceQuotas = list(
    ApplicationQuota = list(
      Maximum = 123
    ),
    ApplicationVersionQuota = list(
      Maximum = 123
    ),
    EnvironmentQuota = list(
      Maximum = 123
    ),
    ConfigurationTemplateQuota = list(
      Maximum = 123
    ),
    CustomPlatformQuota = list(
      Maximum = 123
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns attributes related to AWS Elastic Beanstalk that are associated
with the calling AWS account.

The result currently has one set of attributes—resource quotas.
}
\keyword{internal}
