% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_describe_platform_version}
\alias{elasticbeanstalk_describe_platform_version}
\title{Describes a platform version}
\usage{
elasticbeanstalk_describe_platform_version(PlatformArn = NULL)
}
\arguments{
\item{PlatformArn}{The ARN of the platform version.}
}
\description{
Describes a platform version. Provides full details. Compare to \code{\link[=elasticbeanstalk_list_platform_versions]{list_platform_versions}}, which provides summary information about a list of platform versions.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_describe_platform_version/} for full documentation.
}
\keyword{internal}
