% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_disassociate_environment_operations_role}
\alias{elasticbeanstalk_disassociate_environment_operations_role}
\title{Disassociate the operations role from an environment}
\usage{
elasticbeanstalk_disassociate_environment_operations_role(
  EnvironmentName)
}
\arguments{
\item{EnvironmentName}{[required] The name of the environment from which to disassociate the operations
role.}
}
\value{
An empty list.
}
\description{
Disassociate the operations role from an environment. After this call is
made, Elastic Beanstalk uses the caller's permissions for permissions to
downstream services during subsequent calls acting on this environment.
For more information, see \href{https://docs.aws.amazon.com/elasticbeanstalk/latest/dg/}{Operations roles} in the
\emph{AWS Elastic Beanstalk Developer Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disassociate_environment_operations_role(
  EnvironmentName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
