% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_disassociate_environment_operations_role}
\alias{elasticbeanstalk_disassociate_environment_operations_role}
\title{Disassociate the operations role from an environment}
\usage{
elasticbeanstalk_disassociate_environment_operations_role(EnvironmentName)
}
\arguments{
\item{EnvironmentName}{[required] The name of the environment from which to disassociate the operations
role.}
}
\description{
Disassociate the operations role from an environment. After this call is made, Elastic Beanstalk uses the caller's permissions for permissions to downstream services during subsequent calls acting on this environment. For more information, see \href{https://docs.aws.amazon.com/elasticbeanstalk/latest/dg/}{Operations roles} in the \emph{AWS Elastic Beanstalk Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_disassociate_environment_operations_role/} for full documentation.
}
\keyword{internal}
