% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_list_available_solution_stacks}
\alias{elasticbeanstalk_list_available_solution_stacks}
\title{Returns a list of the available solution stack names, with the public
version first and then in reverse chronological order}
\usage{
elasticbeanstalk_list_available_solution_stacks()
}
\description{
Returns a list of the available solution stack names, with the public version first and then in reverse chronological order.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_list_available_solution_stacks/} for full documentation.
}
\keyword{internal}
