% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_list_tags_for_resource}
\alias{elasticbeanstalk_list_tags_for_resource}
\title{Return the tags applied to an AWS Elastic Beanstalk resource}
\usage{
elasticbeanstalk_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resouce for which a tag list is
requested.

Must be the ARN of an Elastic Beanstalk resource.}
}
\description{
Return the tags applied to an AWS Elastic Beanstalk resource. The response contains a list of tag key-value pairs.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
