% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_restart_app_server}
\alias{elasticbeanstalk_restart_app_server}
\title{Causes the environment to restart the application container server
running on each Amazon EC2 instance}
\usage{
elasticbeanstalk_restart_app_server(
  EnvironmentId = NULL,
  EnvironmentName = NULL
)
}
\arguments{
\item{EnvironmentId}{The ID of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentName, or both.
If you do not specify either, AWS Elastic Beanstalk returns
\code{MissingRequiredParameter} error.}

\item{EnvironmentName}{The name of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentId, or both. If
you do not specify either, AWS Elastic Beanstalk returns
\code{MissingRequiredParameter} error.}
}
\description{
Causes the environment to restart the application container server running on each Amazon EC2 instance.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_restart_app_server/} for full documentation.
}
\keyword{internal}
