% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_update_application}
\alias{elasticbeanstalk_update_application}
\title{Updates the specified application to have the specified properties}
\usage{
elasticbeanstalk_update_application(ApplicationName, Description)
}
\arguments{
\item{ApplicationName}{[required] The name of the application to update. If no such application is found,
\code{\link[=elasticbeanstalk_update_application]{update_application}} returns an
\code{InvalidParameterValue} error.}

\item{Description}{A new description for the application.

Default: If not specified, AWS Elastic Beanstalk does not update the
description.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Application = list(
    ApplicationArn = "string",
    ApplicationName = "string",
    Description = "string",
    DateCreated = as.POSIXct(
      "2015-01-01"
    ),
    DateUpdated = as.POSIXct(
      "2015-01-01"
    ),
    Versions = list(
      "string"
    ),
    ConfigurationTemplates = list(
      "string"
    ),
    ResourceLifecycleConfig = list(
      ServiceRole = "string",
      VersionLifecycleConfig = list(
        MaxCountRule = list(
          Enabled = TRUE|FALSE,
          MaxCount = 123,
          DeleteSourceFromS3 = TRUE|FALSE
        ),
        MaxAgeRule = list(
          Enabled = TRUE|FALSE,
          MaxAgeInDays = 123,
          DeleteSourceFromS3 = TRUE|FALSE
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates the specified application to have the specified properties.

If a property (for example, \code{description}) is not provided, the value
remains unchanged. To clear these properties, specify an empty string.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_application(
  ApplicationName = "string",
  Description = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following operation updates the description of an application named
# my-app:
svc$update_application(
  ApplicationName = "my-app",
  Description = "my Elastic Beanstalk application"
)
}

}
\keyword{internal}
