% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_update_application}
\alias{elasticbeanstalk_update_application}
\title{Updates the specified application to have the specified properties}
\usage{
elasticbeanstalk_update_application(ApplicationName, Description = NULL)
}
\arguments{
\item{ApplicationName}{[required] The name of the application to update. If no such application is found,
\code{\link[=elasticbeanstalk_update_application]{update_application}} returns an
\code{InvalidParameterValue} error.}

\item{Description}{A new description for the application.

Default: If not specified, AWS Elastic Beanstalk does not update the
description.}
}
\description{
Updates the specified application to have the specified properties.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_update_application/} for full documentation.
}
\keyword{internal}
