% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_update_application_version}
\alias{elasticbeanstalk_update_application_version}
\title{Updates the specified application version to have the specified
properties}
\usage{
elasticbeanstalk_update_application_version(
  ApplicationName,
  VersionLabel,
  Description = NULL
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application associated with this version.

If no application is found with this name,
\code{\link[=elasticbeanstalk_update_application]{update_application}} returns an
\code{InvalidParameterValue} error.}

\item{VersionLabel}{[required] The name of the version to update.

If no application version is found with this label,
\code{\link[=elasticbeanstalk_update_application]{update_application}} returns an
\code{InvalidParameterValue} error.}

\item{Description}{A new description for this version.}
}
\description{
Updates the specified application version to have the specified properties.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_update_application_version/} for full documentation.
}
\keyword{internal}
