% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbeanstalk_operations.R
\name{elasticbeanstalk_update_tags_for_resource}
\alias{elasticbeanstalk_update_tags_for_resource}
\title{Update the list of tags applied to an AWS Elastic Beanstalk resource}
\usage{
elasticbeanstalk_update_tags_for_resource(
  ResourceArn,
  TagsToAdd = NULL,
  TagsToRemove = NULL
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resouce to be updated.

Must be the ARN of an Elastic Beanstalk resource.}

\item{TagsToAdd}{A list of tags to add or update. If a key of an existing tag is added,
the tag's value is updated.

Specify at least one of these parameters: \code{TagsToAdd}, \code{TagsToRemove}.}

\item{TagsToRemove}{A list of tag keys to remove. If a tag key doesn't exist, it is silently
ignored.

Specify at least one of these parameters: \code{TagsToAdd}, \code{TagsToRemove}.}
}
\description{
Update the list of tags applied to an AWS Elastic Beanstalk resource. Two lists can be passed: \code{TagsToAdd} for tags to add or update, and \code{TagsToRemove}.

See \url{https://www.paws-r-sdk.com/docs/elasticbeanstalk_update_tags_for_resource/} for full documentation.
}
\keyword{internal}
