% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_service.R
\name{emrcontainers}
\alias{emrcontainers}
\title{Amazon EMR Containers}
\usage{
emrcontainers(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Amazon EMR on EKS provides a deployment option for Amazon EMR that allows you to run open-source big data frameworks on Amazon Elastic Kubernetes Service (Amazon EKS). With this deployment option, you can focus on running analytics workloads while Amazon EMR on EKS builds, configures, and manages containers for open-source applications. For more information about Amazon EMR on EKS concepts and tasks, see \href{https://docs.aws.amazon.com/emr/latest/EMR-on-EKS-DevelopmentGuide/emr-eks.html}{What is Amazon EMR on EKS}.

\emph{Amazon EMR containers} is the API name for Amazon EMR on EKS. The \code{emr-containers} prefix is used in the following scenarios:
\itemize{
\item It is the prefix in the CLI commands for Amazon EMR on EKS. For example, \verb{aws emr-containers start-job-run}.
\item It is the prefix before IAM policy actions for Amazon EMR on EKS. For example, \verb{"Action": [ "emr-containers:StartJobRun"]}. For more information, see \href{https://docs.aws.amazon.com/emr/latest/EMR-on-EKS-DevelopmentGuide/security_iam_service-with-iam.html#security_iam_service-with-iam-id-based-policies-actions}{Policy actions for Amazon EMR on EKS}.
\item It is the prefix used in Amazon EMR on EKS service endpoints. For example, \verb{emr-containers.us-east-2.amazonaws.com}. For more information, see \href{https://docs.aws.amazon.com/emr/latest/EMR-on-EKS-DevelopmentGuide/service-quotas.html#service-endpoints}{Amazon EMR on EKSService Endpoints}.
}
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- emrcontainers(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=emrcontainers_cancel_job_run]{cancel_job_run} \tab Cancels a job run\cr
\link[=emrcontainers_create_job_template]{create_job_template} \tab Creates a job template\cr
\link[=emrcontainers_create_managed_endpoint]{create_managed_endpoint} \tab Creates a managed endpoint\cr
\link[=emrcontainers_create_security_configuration]{create_security_configuration} \tab Creates a security configuration\cr
\link[=emrcontainers_create_virtual_cluster]{create_virtual_cluster} \tab Creates a virtual cluster\cr
\link[=emrcontainers_delete_job_template]{delete_job_template} \tab Deletes a job template\cr
\link[=emrcontainers_delete_managed_endpoint]{delete_managed_endpoint} \tab Deletes a managed endpoint\cr
\link[=emrcontainers_delete_security_configuration]{delete_security_configuration} \tab Deletes a security configuration\cr
\link[=emrcontainers_delete_virtual_cluster]{delete_virtual_cluster} \tab Deletes a virtual cluster\cr
\link[=emrcontainers_describe_job_run]{describe_job_run} \tab Displays detailed information about a job run\cr
\link[=emrcontainers_describe_job_template]{describe_job_template} \tab Displays detailed information about a specified job template\cr
\link[=emrcontainers_describe_managed_endpoint]{describe_managed_endpoint} \tab Displays detailed information about a managed endpoint\cr
\link[=emrcontainers_describe_security_configuration]{describe_security_configuration} \tab Displays detailed information about a specified security configuration\cr
\link[=emrcontainers_describe_virtual_cluster]{describe_virtual_cluster} \tab Displays detailed information about a specified virtual cluster\cr
\link[=emrcontainers_get_managed_endpoint_session_credentials]{get_managed_endpoint_session_credentials} \tab Generate a session token to connect to a managed endpoint\cr
\link[=emrcontainers_list_job_runs]{list_job_runs} \tab Lists job runs based on a set of parameters\cr
\link[=emrcontainers_list_job_templates]{list_job_templates} \tab Lists job templates based on a set of parameters\cr
\link[=emrcontainers_list_managed_endpoints]{list_managed_endpoints} \tab Lists managed endpoints based on a set of parameters\cr
\link[=emrcontainers_list_security_configurations]{list_security_configurations} \tab Lists security configurations based on a set of parameters\cr
\link[=emrcontainers_list_tags_for_resource]{list_tags_for_resource} \tab Lists the tags assigned to the resources\cr
\link[=emrcontainers_list_virtual_clusters]{list_virtual_clusters} \tab Lists information about the specified virtual cluster\cr
\link[=emrcontainers_start_job_run]{start_job_run} \tab Starts a job run\cr
\link[=emrcontainers_tag_resource]{tag_resource} \tab Assigns tags to resources\cr
\link[=emrcontainers_untag_resource]{untag_resource} \tab Removes tags from resources\cr
\link[=emrcontainers_update_virtual_cluster]{update_virtual_cluster} \tab Updates a virtual cluster
}
}

\examples{
\dontrun{
svc <- emrcontainers()
svc$cancel_job_run(
  Foo = 123
)
}

}
