% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrcontainers_operations.R
\name{emrcontainers_describe_virtual_cluster}
\alias{emrcontainers_describe_virtual_cluster}
\title{Displays detailed information about a specified virtual cluster}
\usage{
emrcontainers_describe_virtual_cluster(id)
}
\arguments{
\item{id}{[required] The ID of the virtual cluster that will be described.}
}
\description{
Displays detailed information about a specified virtual cluster. Virtual cluster is a managed entity on Amazon EMR on EKS. You can create, describe, list and delete virtual clusters. They do not consume any additional resource in your system. A single virtual cluster maps to a single Kubernetes namespace. Given this relationship, you can model virtual clusters the same way you model Kubernetes namespaces to meet your requirements.

See \url{https://www.paws-r-sdk.com/docs/emrcontainers_describe_virtual_cluster/} for full documentation.
}
\keyword{internal}
