% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrserverless_operations.R
\name{emrserverless_cancel_job_run}
\alias{emrserverless_cancel_job_run}
\title{Cancels a job run}
\usage{
emrserverless_cancel_job_run(
  applicationId,
  jobRunId,
  shutdownGracePeriodInSeconds = NULL
)
}
\arguments{
\item{applicationId}{[required] The ID of the application on which the job run will be canceled.}

\item{jobRunId}{[required] The ID of the job run to cancel.}

\item{shutdownGracePeriodInSeconds}{The duration in seconds to wait before forcefully terminating the job
after cancellation is requested.}
}
\description{
Cancels a job run.

See \url{https://www.paws-r-sdk.com/docs/emrserverless_cancel_job_run/} for full documentation.
}
\keyword{internal}
