% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_create_distribution_configuration}
\alias{imagebuilder_create_distribution_configuration}
\title{Creates a new distribution configuration}
\usage{
imagebuilder_create_distribution_configuration(
  name,
  description = NULL,
  distributions,
  tags = NULL,
  clientToken
)
}
\arguments{
\item{name}{[required] The name of the distribution configuration.}

\item{description}{The description of the distribution configuration.}

\item{distributions}{[required] The distributions of the distribution configuration.}

\item{tags}{The tags of the distribution configuration.}

\item{clientToken}{[required] Unique, case-sensitive identifier you provide to ensure idempotency of
the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}
in the \emph{Amazon EC2 API Reference}.}
}
\description{
Creates a new distribution configuration. Distribution configurations define and configure the outputs of your pipeline.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_create_distribution_configuration/} for full documentation.
}
\keyword{internal}
