% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_delete_infrastructure_configuration}
\alias{imagebuilder_delete_infrastructure_configuration}
\title{Deletes an infrastructure configuration}
\usage{
imagebuilder_delete_infrastructure_configuration(
  infrastructureConfigurationArn
)
}
\arguments{
\item{infrastructureConfigurationArn}{[required] The Amazon Resource Name (ARN) of the infrastructure configuration to
delete.}
}
\description{
Deletes an infrastructure configuration.

See \url{https://paws-r.github.io/docs/imagebuilder/delete_infrastructure_configuration.html} for full documentation.
}
\keyword{internal}
