% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_delete_workflow}
\alias{imagebuilder_delete_workflow}
\title{Deletes a specific workflow resource}
\usage{
imagebuilder_delete_workflow(workflowBuildVersionArn)
}
\arguments{
\item{workflowBuildVersionArn}{[required] The Amazon Resource Name (ARN) of the workflow resource to delete.}
}
\description{
Deletes a specific workflow resource.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_delete_workflow/} for full documentation.
}
\keyword{internal}
