% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_distribute_image}
\alias{imagebuilder_distribute_image}
\title{Distributes an existing AMI to target Regions and accounts without
running the full image build process}
\usage{
imagebuilder_distribute_image(
  sourceImage,
  distributionConfigurationArn,
  executionRole,
  tags = NULL,
  clientToken,
  loggingConfiguration = NULL
)
}
\arguments{
\item{sourceImage}{[required] The source image to distribute. Specify an AMI identifier, SSM parameter path, or Image Builder image Amazon Resource Name (ARN). When you specify an Image Builder image Amazon Resource Name (ARN), the image must be in the \code{AVAILABLE} state.}

\item{distributionConfigurationArn}{[required] The Amazon Resource Name (ARN) of the distribution configuration. The configuration defines target Regions, accounts, and AMI settings. The distribution configuration must be in the same Region as this operation.}

\item{executionRole}{[required] The name or Amazon Resource Name (ARN) of the IAM role that Image Builder assumes to distribute the image.}

\item{tags}{The tags to apply to the distributed image.}

\item{clientToken}{[required] Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency} in the \emph{Amazon EC2 API Reference}.}

\item{loggingConfiguration}{The logging configuration for the distribution.}
}
\description{
Distributes an existing AMI to target Regions and accounts without running the full image build process. This operation only runs the distribution phase on an image that has already been built.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_distribute_image/} for full documentation.
}
\keyword{internal}
