% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_distribute_image}
\alias{imagebuilder_distribute_image}
\title{DistributeImage distributes existing AMIs to additional regions and
accounts without rebuilding the image}
\usage{
imagebuilder_distribute_image(
  sourceImage,
  distributionConfigurationArn,
  executionRole,
  tags = NULL,
  clientToken,
  loggingConfiguration = NULL
)
}
\arguments{
\item{sourceImage}{[required] The source image Amazon Resource Name (ARN) to distribute.}

\item{distributionConfigurationArn}{[required] The Amazon Resource Name (ARN) of the distribution configuration to use.}

\item{executionRole}{[required] The IAM role to use for the distribution.}

\item{tags}{The tags to apply to the distributed image.}

\item{clientToken}{[required] Unique, case-sensitive identifier you provide to ensure idempotency of
the request. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency}
in the \emph{Amazon EC2 API Reference}.}

\item{loggingConfiguration}{The logging configuration for the distribution.}
}
\description{
DistributeImage distributes existing AMIs to additional regions and accounts without rebuilding the image.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_distribute_image/} for full documentation.
}
\keyword{internal}
