% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_component}
\alias{imagebuilder_get_component}
\title{Retrieves a component object}
\usage{
imagebuilder_get_component(componentBuildVersionArn)
}
\arguments{
\item{componentBuildVersionArn}{[required] The Amazon Resource Name (ARN) of the component that you want to get. Regex requires the suffix \verb{/\\\\d+$}.}
}
\description{
Retrieves a component object.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_component/} for full documentation.
}
\keyword{internal}
