% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_image}
\alias{imagebuilder_get_image}
\title{Gets an image}
\usage{
imagebuilder_get_image(imageBuildVersionArn)
}
\arguments{
\item{imageBuildVersionArn}{[required] The Amazon Resource Name (ARN) of the image that you want to get.}
}
\description{
Gets an image.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_image/} for full documentation.
}
\keyword{internal}
