% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_image_pipeline}
\alias{imagebuilder_get_image_pipeline}
\title{Gets an image pipeline}
\usage{
imagebuilder_get_image_pipeline(imagePipelineArn)
}
\arguments{
\item{imagePipelineArn}{[required] The Amazon Resource Name (ARN) of the image pipeline that you want to
retrieve.}
}
\description{
Gets an image pipeline.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_image_pipeline/} for full documentation.
}
\keyword{internal}
