% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_image_recipe_policy}
\alias{imagebuilder_get_image_recipe_policy}
\title{Gets an image recipe policy}
\usage{
imagebuilder_get_image_recipe_policy(imageRecipeArn)
}
\arguments{
\item{imageRecipeArn}{[required] The Amazon Resource Name (ARN) of the image recipe whose policy you want
to retrieve.}
}
\description{
Gets an image recipe policy.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_image_recipe_policy/} for full documentation.
}
\keyword{internal}
