% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_infrastructure_configuration}
\alias{imagebuilder_get_infrastructure_configuration}
\title{Gets an infrastructure configuration}
\usage{
imagebuilder_get_infrastructure_configuration(infrastructureConfigurationArn)
}
\arguments{
\item{infrastructureConfigurationArn}{[required] The Amazon Resource Name (ARN) of the infrastructure configuration that
you want to retrieve.}
}
\description{
Gets an infrastructure configuration.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_infrastructure_configuration/} for full documentation.
}
\keyword{internal}
