% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_lifecycle_execution}
\alias{imagebuilder_get_lifecycle_execution}
\title{Retrieves the runtime information for a specific runtime instance of the
lifecycle policy}
\usage{
imagebuilder_get_lifecycle_execution(lifecycleExecutionId)
}
\arguments{
\item{lifecycleExecutionId}{[required] The unique identifier for a runtime instance of the lifecycle policy.}
}
\description{
Retrieves the runtime information for a specific runtime instance of the lifecycle policy.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_lifecycle_execution/} for full documentation.
}
\keyword{internal}
