% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_lifecycle_policy}
\alias{imagebuilder_get_lifecycle_policy}
\title{Get details for the specified image lifecycle policy}
\usage{
imagebuilder_get_lifecycle_policy(lifecyclePolicyArn)
}
\arguments{
\item{lifecyclePolicyArn}{[required] Specifies the Amazon Resource Name (ARN) of the image lifecycle policy
resource to get.}
}
\description{
Get details for the specified image lifecycle policy.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_lifecycle_policy/} for full documentation.
}
\keyword{internal}
