% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_marketplace_resource}
\alias{imagebuilder_get_marketplace_resource}
\title{Verifies the subscription and performs resource dependency checks on the
requested Amazon Web Services Marketplace resource}
\usage{
imagebuilder_get_marketplace_resource(
  resourceType,
  resourceArn,
  resourceLocation = NULL
)
}
\arguments{
\item{resourceType}{[required] Specifies which type of Amazon Web Services Marketplace resource Image Builder retrieves.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies an Amazon Web Services Marketplace resource.}

\item{resourceLocation}{The bucket path that you can specify to download the resource from Amazon S3.}
}
\description{
Verifies the subscription and performs resource dependency checks on the requested Amazon Web Services Marketplace resource. For Amazon Web Services Marketplace components, the response contains fields to download the components and their artifacts.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_marketplace_resource/} for full documentation.
}
\keyword{internal}
