% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_workflow}
\alias{imagebuilder_get_workflow}
\title{Retrieves a workflow resource object}
\usage{
imagebuilder_get_workflow(workflowBuildVersionArn)
}
\arguments{
\item{workflowBuildVersionArn}{[required] The Amazon Resource Name (ARN) of the workflow resource that you want to get.}
}
\description{
Retrieves a workflow resource object.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_workflow/} for full documentation.
}
\keyword{internal}
