% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_workflow_execution}
\alias{imagebuilder_get_workflow_execution}
\title{Get the runtime information that was logged for a specific runtime
instance of the workflow}
\usage{
imagebuilder_get_workflow_execution(workflowExecutionId)
}
\arguments{
\item{workflowExecutionId}{[required] Use the unique identifier for a runtime instance of the workflow to get
runtime details.}
}
\description{
Get the runtime information that was logged for a specific runtime instance of the workflow.

See \url{https://paws-r.github.io/docs/imagebuilder/get_workflow_execution.html} for full documentation.
}
\keyword{internal}
