% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_get_workflow_step_execution}
\alias{imagebuilder_get_workflow_step_execution}
\title{Get the runtime information that was logged for a specific runtime
instance of the workflow step}
\usage{
imagebuilder_get_workflow_step_execution(stepExecutionId)
}
\arguments{
\item{stepExecutionId}{[required] Use the unique identifier for a specific runtime instance of the
workflow step to get runtime details for that step.}
}
\description{
Get the runtime information that was logged for a specific runtime instance of the workflow step.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_get_workflow_step_execution/} for full documentation.
}
\keyword{internal}
