% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_component_build_versions}
\alias{imagebuilder_list_component_build_versions}
\title{Returns the list of component build versions for the specified component
version Amazon Resource Name (ARN)}
\usage{
imagebuilder_list_component_build_versions(
  componentVersionArn = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{componentVersionArn}{The component version Amazon Resource Name (ARN) whose versions you want to list.}

\item{maxResults}{The maximum number of items to return in a single request.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}
}
\description{
Returns the list of component build versions for the specified component version Amazon Resource Name (ARN).

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_component_build_versions/} for full documentation.
}
\keyword{internal}
