% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_container_recipes}
\alias{imagebuilder_list_container_recipes}
\title{Returns a list of container recipes}
\usage{
imagebuilder_list_container_recipes(
  owner = NULL,
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{owner}{Returns container recipes belonging to the specified owner, that have been shared with you. You can omit this field to return container recipes belonging to your account.}

\item{filters}{Use the following filters to streamline results:
\itemize{
\item \code{containerType}
\item \code{name}
\item \code{parentImage}
\item \code{platform}
}}

\item{maxResults}{The maximum number of items to return in a single request.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}
}
\description{
Returns a list of container recipes.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_container_recipes/} for full documentation.
}
\keyword{internal}
