% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_container_recipes}
\alias{imagebuilder_list_container_recipes}
\title{Returns a list of container recipes}
\usage{
imagebuilder_list_container_recipes(
  owner = NULL,
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{owner}{Returns container recipes belonging to the specified owner, that have
been shared with you. You can omit this field to return container
recipes belonging to your account.}

\item{filters}{Use the following filters to streamline results:
\itemize{
\item \code{containerType}
\item \code{name}
\item \code{parentImage}
\item \code{platform}
}}

\item{maxResults}{The maximum number of results to return in the list.}

\item{nextToken}{Provides a token for pagination, which determines where to begin the
next set of results when the current set reaches the maximum for one
request.}
}
\description{
Returns a list of container recipes.

See \url{https://paws-r.github.io/docs/imagebuilder/list_container_recipes.html} for full documentation.
}
\keyword{internal}
