% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_build_versions}
\alias{imagebuilder_list_image_build_versions}
\title{Returns a list of image build versions}
\usage{
imagebuilder_list_image_build_versions(
  imageVersionArn = NULL,
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{imageVersionArn}{The Amazon Resource Name (ARN) of the image whose build versions you
want to retrieve.}

\item{filters}{Use the following filters to streamline results:
\itemize{
\item \code{name}
\item \code{osVersion}
\item \code{platform}
\item \code{type}
\item \code{version}
}}

\item{maxResults}{Specify the maximum number of items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}
}
\description{
Returns a list of image build versions.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_image_build_versions/} for full documentation.
}
\keyword{internal}
