% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_packages}
\alias{imagebuilder_list_image_packages}
\title{Lists the packages that are associated with an image build version, as
determined by Amazon Web Services Systems Manager Inventory at build
time}
\usage{
imagebuilder_list_image_packages(
  imageBuildVersionArn,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{imageBuildVersionArn}{[required] Filter results for the ListImagePackages request by the Image Build Version ARN}

\item{maxResults}{The maximum number of items to return in a single request.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}
}
\description{
Lists the packages that are associated with an image build version, as determined by Amazon Web Services Systems Manager Inventory at build time.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_image_packages/} for full documentation.
}
\keyword{internal}
