% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_packages}
\alias{imagebuilder_list_image_packages}
\title{List the Packages that are associated with an Image Build Version, as
determined by Amazon Web Services Systems Manager Inventory at build
time}
\usage{
imagebuilder_list_image_packages(
  imageBuildVersionArn,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{imageBuildVersionArn}{[required] Filter results for the ListImagePackages request by the Image Build
Version ARN}

\item{maxResults}{Specify the maximum number of items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}
}
\description{
List the Packages that are associated with an Image Build Version, as determined by Amazon Web Services Systems Manager Inventory at build time.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_image_packages/} for full documentation.
}
\keyword{internal}
