% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_pipelines}
\alias{imagebuilder_list_image_pipelines}
\title{Returns a list of image pipelines}
\usage{
imagebuilder_list_image_pipelines(
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filters}{Use the following filters to streamline results:
\itemize{
\item \code{description}
\item \code{distributionConfigurationArn}
\item \code{imageRecipeArn}
\item \code{infrastructureConfigurationArn}
\item \code{name}
\item \code{status}
}}

\item{maxResults}{Specify the maximum number of items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}
}
\description{
Returns a list of image pipelines.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_image_pipelines/} for full documentation.
}
\keyword{internal}
