% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_recipes}
\alias{imagebuilder_list_image_recipes}
\title{Returns a list of image recipes}
\usage{
imagebuilder_list_image_recipes(
  owner = NULL,
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{owner}{The owner defines which image recipes you want to list. By default, this
request will only show image recipes owned by your account. You can use
this field to specify if you want to view image recipes owned by
yourself, by Amazon, or those image recipes that have been shared with
you by other customers.}

\item{filters}{Use the following filters to streamline results:
\itemize{
\item \code{name}
\item \code{parentImage}
\item \code{platform}
}}

\item{maxResults}{The maximum items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}
}
\description{
Returns a list of image recipes.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_image_recipes/} for full documentation.
}
\keyword{internal}
