% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_recipes}
\alias{imagebuilder_list_image_recipes}
\title{Returns a list of image recipes}
\usage{
imagebuilder_list_image_recipes(
  owner = NULL,
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{owner}{You can specify the recipe owner to filter results by that owner. By
default, this request will only show image recipes owned by your
account. To filter by a different owner, specify one of the
\verb{Valid Values} that are listed for this parameter.}

\item{filters}{Use the following filters to streamline results:
\itemize{
\item \code{name}
\item \code{parentImage}
\item \code{platform}
}}

\item{maxResults}{Specify the maximum number of items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}
}
\description{
Returns a list of image recipes.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_image_recipes/} for full documentation.
}
\keyword{internal}
