% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_scan_finding_aggregations}
\alias{imagebuilder_list_image_scan_finding_aggregations}
\title{Returns a list of image scan aggregations for your account}
\usage{
imagebuilder_list_image_scan_finding_aggregations(
  filter = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filter}{A filter name and value pair that is used to return a more specific list of results from a list operation. Filters can be used to match a set of resources by specific criteria, such as tags, attributes, or IDs.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}
}
\description{
Returns a list of image scan aggregations for your account. You can filter by the type of key that Image Builder uses to group results. For example, if you want to get a list of findings by severity level for one of your pipelines, you might specify your pipeline with the \code{imagePipelineArn} filter. If you don't specify a filter, Image Builder returns an aggregation for your account.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_image_scan_finding_aggregations/} for full documentation.
}
\keyword{internal}
