% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_image_scan_findings}
\alias{imagebuilder_list_image_scan_findings}
\title{Returns a list of image scan findings for your account}
\usage{
imagebuilder_list_image_scan_findings(
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filters}{An array of name value pairs that you can use to filter your results. You can use the following filters to streamline results:
\itemize{
\item \code{imageBuildVersionArn}
\item \code{imagePipelineArn}
\item \code{vulnerabilityId}
\item \code{severity}
}

If you don't request a filter, then all findings in your account are listed.}

\item{maxResults}{The maximum number of items to return in a single request.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}
}
\description{
Returns a list of image scan findings for your account.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_image_scan_findings/} for full documentation.
}
\keyword{internal}
