% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_lifecycle_executions}
\alias{imagebuilder_list_lifecycle_executions}
\title{Retrieves the lifecycle runtime history for the specified resource}
\usage{
imagebuilder_list_lifecycle_executions(
  maxResults = NULL,
  nextToken = NULL,
  resourceArn
)
}
\arguments{
\item{maxResults}{The maximum number of items to return in a single request.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which to get a list of lifecycle runtime instances.}
}
\description{
Retrieves the lifecycle runtime history for the specified resource.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_lifecycle_executions/} for full documentation.
}
\keyword{internal}
