% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_lifecycle_policies}
\alias{imagebuilder_list_lifecycle_policies}
\title{Retrieves a list of lifecycle policies in your Amazon Web Services
account}
\usage{
imagebuilder_list_lifecycle_policies(
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filters}{Streamline results based on one of the following values: \code{Name}, \code{Status}.}

\item{maxResults}{The maximum number of items to return in a single request.}

\item{nextToken}{A token to specify where to start paginating. Use the \code{nextToken} value from a previously truncated response.}
}
\description{
Retrieves a list of lifecycle policies in your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_lifecycle_policies/} for full documentation.
}
\keyword{internal}
