% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_list_waiting_workflow_steps}
\alias{imagebuilder_list_waiting_workflow_steps}
\title{Get a list of workflow steps that are waiting for action for workflows
in your Amazon Web Services account}
\usage{
imagebuilder_list_waiting_workflow_steps(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{Specify the maximum number of items to return in a request.}

\item{nextToken}{A token to specify where to start paginating. This is the nextToken from
a previously truncated response.}
}
\description{
Get a list of workflow steps that are waiting for action for workflows in your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_list_waiting_workflow_steps/} for full documentation.
}
\keyword{internal}
