% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_put_image_policy}
\alias{imagebuilder_put_image_policy}
\title{Applies a policy to an image}
\usage{
imagebuilder_put_image_policy(imageArn, policy)
}
\arguments{
\item{imageArn}{[required] The Amazon Resource Name (ARN) of the image that this policy should be applied to.}

\item{policy}{[required] The policy to apply.}
}
\description{
Applies a policy to an image. To share resources, call the RAM API \href{https://docs.aws.amazon.com/ram/latest/APIReference/API_CreateResourceShare.html}{CreateResourceShare}. If you call this API, you must also call the RAM API \href{https://docs.aws.amazon.com/ram/latest/APIReference/API_PromoteResourceShareCreatedFromPolicy.html}{PromoteResourceShareCreatedFromPolicy} so that the resource is visible to all principals with whom the resource is shared.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_put_image_policy/} for full documentation.
}
\keyword{internal}
