% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_put_image_recipe_policy}
\alias{imagebuilder_put_image_recipe_policy}
\title{Applies a policy to an image recipe}
\usage{
imagebuilder_put_image_recipe_policy(imageRecipeArn, policy)
}
\arguments{
\item{imageRecipeArn}{[required] The Amazon Resource Name (ARN) of the image recipe that this policy
should be applied to.}

\item{policy}{[required] The policy to apply.}
}
\description{
Applies a policy to an image recipe. We recommend that you call the RAM API \href{https://docs.aws.amazon.com/ram/latest/APIReference/API_CreateResourceShare.html}{CreateResourceShare} to share resources. If you call the Image Builder API \code{\link[=imagebuilder_put_image_recipe_policy]{put_image_recipe_policy}}, you must also call the RAM API \href{https://docs.aws.amazon.com/ram/latest/APIReference/API_PromoteResourceShareCreatedFromPolicy.html}{PromoteResourceShareCreatedFromPolicy} in order for the resource to be visible to all principals with whom the resource is shared.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_put_image_recipe_policy/} for full documentation.
}
\keyword{internal}
