% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_update_distribution_configuration}
\alias{imagebuilder_update_distribution_configuration}
\title{Updates a distribution configuration}
\usage{
imagebuilder_update_distribution_configuration(
  distributionConfigurationArn,
  description = NULL,
  distributions,
  clientToken
)
}
\arguments{
\item{distributionConfigurationArn}{[required] The Amazon Resource Name (ARN) of the distribution configuration that you want to update.}

\item{description}{The description of the distribution configuration.}

\item{distributions}{[required] The distributions of the distribution configuration.}

\item{clientToken}{[required] A unique, case-sensitive identifier you provide to ensure that the operation completes no more than one time. If this token matches a previous request, the service ignores the request, but does not return an error. For more information, see \href{https://docs.aws.amazon.com/ec2/latest/devguide/ec2-api-idempotency.html}{Ensuring idempotency} in the \emph{Amazon EC2 API Reference}.}
}
\description{
Updates a distribution configuration. Distribution configurations define and configure the outputs of your pipeline.

See \url{https://www.paws-r-sdk.com/docs/imagebuilder_update_distribution_configuration/} for full documentation.
}
\keyword{internal}
