% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagebuilder_operations.R
\name{imagebuilder_update_image_pipeline}
\alias{imagebuilder_update_image_pipeline}
\title{Updates an image pipeline}
\usage{
imagebuilder_update_image_pipeline(
  imagePipelineArn,
  description = NULL,
  imageRecipeArn = NULL,
  containerRecipeArn = NULL,
  infrastructureConfigurationArn,
  distributionConfigurationArn = NULL,
  imageTestsConfiguration = NULL,
  enhancedImageMetadataEnabled = NULL,
  schedule = NULL,
  status = NULL,
  clientToken
)
}
\arguments{
\item{imagePipelineArn}{[required] The Amazon Resource Name (ARN) of the image pipeline that you want to
update.}

\item{description}{The description of the image pipeline.}

\item{imageRecipeArn}{The Amazon Resource Name (ARN) of the image recipe that will be used to
configure images updated by this image pipeline.}

\item{containerRecipeArn}{The Amazon Resource Name (ARN) of the container pipeline to update.}

\item{infrastructureConfigurationArn}{[required] The Amazon Resource Name (ARN) of the infrastructure configuration that
will be used to build images updated by this image pipeline.}

\item{distributionConfigurationArn}{The Amazon Resource Name (ARN) of the distribution configuration that
will be used to configure and distribute images updated by this image
pipeline.}

\item{imageTestsConfiguration}{The image test configuration of the image pipeline.}

\item{enhancedImageMetadataEnabled}{Collects additional information about the image being created, including
the operating system (OS) version and package list. This information is
used to enhance the overall experience of using EC2 Image Builder.
Enabled by default.}

\item{schedule}{The schedule of the image pipeline.}

\item{status}{The status of the image pipeline.}

\item{clientToken}{[required] The idempotency token used to make this request idempotent.}
}
\description{
Updates an image pipeline. Image pipelines enable you to automate the creation and distribution of images.

See \url{https://paws-r.github.io/docs/imagebuilder/update_image_pipeline.html} for full documentation.
}
\keyword{internal}
