% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_delete_layer_version}
\alias{lambda_delete_layer_version}
\title{Deletes a version of an Lambda layer}
\usage{
lambda_delete_layer_version(LayerName, VersionNumber)
}
\arguments{
\item{LayerName}{[required] The name or Amazon Resource Name (ARN) of the layer.}

\item{VersionNumber}{[required] The version number.}
}
\description{
Deletes a version of an \href{https://docs.aws.amazon.com/lambda/latest/dg/chapter-layers.html}{Lambda layer}. Deleted versions can no longer be viewed or added to functions. To avoid breaking functions, a copy of the version remains in Lambda until no functions refer to it.

See \url{https://www.paws-r-sdk.com/docs/lambda_delete_layer_version/} for full documentation.
}
\keyword{internal}
