% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_event_source_mapping}
\alias{lambda_get_event_source_mapping}
\title{Returns details about an event source mapping}
\usage{
lambda_get_event_source_mapping(UUID)
}
\arguments{
\item{UUID}{[required] The identifier of the event source mapping.}
}
\description{
Returns details about an event source mapping. You can get the
identifier of a mapping from the output of ListEventSourceMappings.
}
\section{Request syntax}{
\preformatted{svc$get_event_source_mapping(
  UUID = "string"
)
}
}

\examples{
\dontrun{
# The following example returns details about an event source mapping. To
# get a mapping's UUID, use ListEventSourceMappings.
svc$get_event_source_mapping(
  UUID = "14e0db71-xmpl-4eb5-b481-8945cf9d10c2"
)
}

}
\keyword{internal}
