% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_event_source_mapping}
\alias{lambda_get_event_source_mapping}
\title{Returns details about an event source mapping}
\usage{
lambda_get_event_source_mapping(UUID)
}
\arguments{
\item{UUID}{[required] The identifier of the event source mapping.}
}
\description{
Returns details about an event source mapping. You can get the identifier of a mapping from the output of \code{\link[=lambda_list_event_source_mappings]{list_event_source_mappings}}.

See \url{https://www.paws-r-sdk.com/docs/lambda_get_event_source_mapping/} for full documentation.
}
\keyword{internal}
