% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_function}
\alias{lambda_get_function}
\title{Returns information about the function or function version, with a link
to download the deployment package that's valid for 10 minutes}
\usage{
lambda_get_function(FunctionName, Qualifier = NULL)
}
\arguments{
\item{FunctionName}{[required] The name or ARN of the Lambda function, version, or alias.

\strong{Name formats}
\itemize{
\item \strong{Function name} – \verb{my-function} (name-only), \verb{my-function:v1}
(with alias).
\item \strong{Function ARN} –
\verb{arn:aws:lambda:us-west-2:123456789012:function:my-function}.
\item \strong{Partial ARN} – \verb{123456789012:function:my-function}.
}

You can append a version number or alias to any of the formats. The
length constraint applies only to the full ARN. If you specify only the
function name, it is limited to 64 characters in length.}

\item{Qualifier}{Specify a version or alias to get details about a published version of
the function.}
}
\description{
Returns information about the function or function version, with a link to download the deployment package that's valid for 10 minutes. If you specify a function version, only details that are specific to that version are returned.

See \url{https://www.paws-r-sdk.com/docs/lambda_get_function/} for full documentation.
}
\keyword{internal}
