% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_layer_version_policy}
\alias{lambda_get_layer_version_policy}
\title{Returns the permission policy for a version of an Lambda layer}
\usage{
lambda_get_layer_version_policy(LayerName, VersionNumber)
}
\arguments{
\item{LayerName}{[required] The name or Amazon Resource Name (ARN) of the layer.}

\item{VersionNumber}{[required] The version number.}
}
\description{
Returns the permission policy for a version of an \href{https://docs.aws.amazon.com/lambda/latest/dg/chapter-layers.html}{Lambda layer}. For more information, see \code{\link[=lambda_add_layer_version_permission]{add_layer_version_permission}}.

See \url{https://www.paws-r-sdk.com/docs/lambda_get_layer_version_policy/} for full documentation.
}
\keyword{internal}
