% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_list_code_signing_configs}
\alias{lambda_list_code_signing_configs}
\title{Returns a list of code signing configurations}
\usage{
lambda_list_code_signing_configs(Marker = NULL, MaxItems = NULL)
}
\arguments{
\item{Marker}{Specify the pagination token that's returned by a previous request to
retrieve the next page of results.}

\item{MaxItems}{Maximum number of items to return.}
}
\description{
Returns a list of \href{https://docs.aws.amazon.com/lambda/latest/dg/configuration-codesigning.html}{code signing configurations}. A request returns up to 10,000 configurations per call. You can use the \code{MaxItems} parameter to return fewer configurations per call.

See \url{https://www.paws-r-sdk.com/docs/lambda_list_code_signing_configs/} for full documentation.
}
\keyword{internal}
