% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_list_function_event_invoke_configs}
\alias{lambda_list_function_event_invoke_configs}
\title{Retrieves a list of configurations for asynchronous invocation for a
function}
\usage{
lambda_list_function_event_invoke_configs(
  FunctionName,
  Marker = NULL,
  MaxItems = NULL
)
}
\arguments{
\item{FunctionName}{[required] The name or ARN of the Lambda function.

\strong{Name formats}
\itemize{
\item \strong{Function name} - \verb{my-function}.
\item \strong{Function ARN} -
\verb{arn:aws:lambda:us-west-2:123456789012:function:my-function}.
\item \strong{Partial ARN} - \verb{123456789012:function:my-function}.
}

The length constraint applies only to the full ARN. If you specify only
the function name, it is limited to 64 characters in length.}

\item{Marker}{Specify the pagination token that's returned by a previous request to
retrieve the next page of results.}

\item{MaxItems}{The maximum number of configurations to return.}
}
\description{
Retrieves a list of configurations for asynchronous invocation for a function.

See \url{https://www.paws-r-sdk.com/docs/lambda_list_function_event_invoke_configs/} for full documentation.
}
\keyword{internal}
