% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_list_layers}
\alias{lambda_list_layers}
\title{Lists Lambda layers and shows information about the latest version of
each}
\usage{
lambda_list_layers(
  CompatibleRuntime = NULL,
  Marker = NULL,
  MaxItems = NULL,
  CompatibleArchitecture = NULL
)
}
\arguments{
\item{CompatibleRuntime}{A runtime identifier. For example, \code{java21}.

The following list includes deprecated runtimes. For more information,
see \href{https://docs.aws.amazon.com/lambda/latest/dg/lambda-runtimes.html#runtime-support-policy}{Runtime deprecation policy}.}

\item{Marker}{A pagination token returned by a previous call.}

\item{MaxItems}{The maximum number of layers to return.}

\item{CompatibleArchitecture}{The compatible \href{https://docs.aws.amazon.com/lambda/latest/dg/foundation-arch.html}{instruction set architecture}.}
}
\description{
Lists \href{https://docs.aws.amazon.com/lambda/latest/dg/chapter-layers.html}{Lambda layers} and shows information about the latest version of each. Specify a \href{https://docs.aws.amazon.com/lambda/latest/dg/lambda-runtimes.html}{runtime identifier} to list only layers that indicate that they're compatible with that runtime. Specify a compatible architecture to include only layers that are compatible with that \href{https://docs.aws.amazon.com/lambda/latest/dg/foundation-arch.html}{instruction set architecture}.

See \url{https://www.paws-r-sdk.com/docs/lambda_list_layers/} for full documentation.
}
\keyword{internal}
