% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_update_capacity_provider}
\alias{lambda_update_capacity_provider}
\title{Updates the configuration of an existing capacity provider}
\usage{
lambda_update_capacity_provider(
  CapacityProviderName,
  CapacityProviderScalingConfig = NULL
)
}
\arguments{
\item{CapacityProviderName}{[required] The name of the capacity provider to update.}

\item{CapacityProviderScalingConfig}{The updated scaling configuration for the capacity provider.}
}
\description{
Updates the configuration of an existing capacity provider.

See \url{https://www.paws-r-sdk.com/docs/lambda_update_capacity_provider/} for full documentation.
}
\keyword{internal}
