% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_attach_load_balancer_tls_certificate}
\alias{lightsail_attach_load_balancer_tls_certificate}
\title{Attaches a Transport Layer Security (TLS) certificate to your load
balancer}
\usage{
lightsail_attach_load_balancer_tls_certificate(
  loadBalancerName,
  certificateName
)
}
\arguments{
\item{loadBalancerName}{[required] The name of the load balancer to which you want to associate the SSL/TLS
certificate.}

\item{certificateName}{[required] The name of your SSL/TLS certificate.}
}
\description{
Attaches a Transport Layer Security (TLS) certificate to your load balancer. TLS is just an updated, more secure version of Secure Socket Layer (SSL).

See \url{https://www.paws-r-sdk.com/docs/lightsail_attach_load_balancer_tls_certificate/} for full documentation.
}
\keyword{internal}
