% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_create_container_service_deployment}
\alias{lightsail_create_container_service_deployment}
\title{Creates a deployment for your Amazon Lightsail container service}
\usage{
lightsail_create_container_service_deployment(
  serviceName,
  containers = NULL,
  publicEndpoint = NULL
)
}
\arguments{
\item{serviceName}{[required] The name of the container service for which to create the deployment.}

\item{containers}{An object that describes the settings of the containers that will be
launched on the container service.}

\item{publicEndpoint}{An object that describes the settings of the public endpoint for the
container service.}
}
\description{
Creates a deployment for your Amazon Lightsail container service.

See \url{https://www.paws-r-sdk.com/docs/lightsail_create_container_service_deployment/} for full documentation.
}
\keyword{internal}
