% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_create_container_service_registry_login}
\alias{lightsail_create_container_service_registry_login}
\title{Creates a temporary set of log in credentials that you can use to log in
to the Docker process on your local machine}
\usage{
lightsail_create_container_service_registry_login()
}
\description{
Creates a temporary set of log in credentials that you can use to log in to the Docker process on your local machine. After you're logged in, you can use the native Docker commands to push your local container images to the container image registry of your Amazon Lightsail account so that you can use them with your Lightsail container service. The log in credentials expire 12 hours after they are created, at which point you will need to create a new set of log in credentials.

See \url{https://www.paws-r-sdk.com/docs/lightsail_create_container_service_registry_login/} for full documentation.
}
\keyword{internal}
