% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_create_domain}
\alias{lightsail_create_domain}
\title{Creates a domain resource for the specified domain (e}
\usage{
lightsail_create_domain(domainName, tags = NULL)
}
\arguments{
\item{domainName}{[required] The domain name to manage (e.g., \code{example.com}).}

\item{tags}{The tag keys and optional values to add to the resource during create.

Use the \code{\link[=lightsail_tag_resource]{tag_resource}} action to tag a
resource after it's created.}
}
\description{
Creates a domain resource for the specified domain (e.g., example.com).

See \url{https://paws-r.github.io/docs/lightsail/create_domain.html} for full documentation.
}
\keyword{internal}
