% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_create_domain}
\alias{lightsail_create_domain}
\title{Creates a domain resource for the specified domain (example}
\usage{
lightsail_create_domain(domainName, tags = NULL)
}
\arguments{
\item{domainName}{[required] The domain name to manage (\code{example.com}).}

\item{tags}{The tag keys and optional values to add to the resource during create.

Use the \code{\link[=lightsail_tag_resource]{tag_resource}} action to tag a
resource after it's created.}
}
\description{
Creates a domain resource for the specified domain (example.com).

See \url{https://www.paws-r-sdk.com/docs/lightsail_create_domain/} for full documentation.
}
\keyword{internal}
