% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_create_domain_entry}
\alias{lightsail_create_domain_entry}
\title{Creates one of the following domain name system (DNS) records in a
domain DNS zone: Address (A), canonical name (CNAME), mail exchanger
(MX), name server (NS), start of authority (SOA), service locator (SRV),
or text (TXT)}
\usage{
lightsail_create_domain_entry(domainName, domainEntry)
}
\arguments{
\item{domainName}{[required] The domain name (\code{example.com}) for which you want to create the domain
entry.}

\item{domainEntry}{[required] An array of key-value pairs containing information about the domain
entry request.}
}
\description{
Creates one of the following domain name system (DNS) records in a domain DNS zone: Address (A), canonical name (CNAME), mail exchanger (MX), name server (NS), start of authority (SOA), service locator (SRV), or text (TXT).

See \url{https://www.paws-r-sdk.com/docs/lightsail_create_domain_entry/} for full documentation.
}
\keyword{internal}
