% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_create_key_pair}
\alias{lightsail_create_key_pair}
\title{Creates a custom SSH key pair that you can use with an Amazon Lightsail
instance}
\usage{
lightsail_create_key_pair(keyPairName, tags = NULL)
}
\arguments{
\item{keyPairName}{[required] The name for your new key pair.}

\item{tags}{The tag keys and optional values to add to the resource during create.

Use the \code{\link[=lightsail_tag_resource]{tag_resource}} action to tag a
resource after it's created.}
}
\description{
Creates a custom SSH key pair that you can use with an Amazon Lightsail instance.

See \url{https://www.paws-r-sdk.com/docs/lightsail_create_key_pair/} for full documentation.
}
\keyword{internal}
