% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_alarm}
\alias{lightsail_delete_alarm}
\title{Deletes an alarm}
\usage{
lightsail_delete_alarm(alarmName)
}
\arguments{
\item{alarmName}{[required] The name of the alarm to delete.}
}
\description{
Deletes an alarm.

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_alarm/} for full documentation.
}
\keyword{internal}
