% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_disk_snapshot}
\alias{lightsail_delete_disk_snapshot}
\title{Deletes the specified disk snapshot}
\usage{
lightsail_delete_disk_snapshot(diskSnapshotName)
}
\arguments{
\item{diskSnapshotName}{[required] The name of the disk snapshot you want to delete (\code{my-disk-snapshot}).}
}
\description{
Deletes the specified disk snapshot.

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_disk_snapshot/} for full documentation.
}
\keyword{internal}
