% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_domain}
\alias{lightsail_delete_domain}
\title{Deletes the specified domain recordset and all of its domain records}
\usage{
lightsail_delete_domain(domainName)
}
\arguments{
\item{domainName}{[required] The specific domain name to delete.}
}
\description{
Deletes the specified domain recordset and all of its domain records.

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_domain/} for full documentation.
}
\keyword{internal}
