% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_domain_entry}
\alias{lightsail_delete_domain_entry}
\title{Deletes a specific domain entry}
\usage{
lightsail_delete_domain_entry(domainName, domainEntry)
}
\arguments{
\item{domainName}{[required] The name of the domain entry to delete.}

\item{domainEntry}{[required] An array of key-value pairs containing information about your domain
entries.}
}
\description{
Deletes a specific domain entry.

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_domain_entry/} for full documentation.
}
\keyword{internal}
