% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_instance}
\alias{lightsail_delete_instance}
\title{Deletes an Amazon Lightsail instance}
\usage{
lightsail_delete_instance(instanceName, forceDeleteAddOns = NULL)
}
\arguments{
\item{instanceName}{[required] The name of the instance to delete.}

\item{forceDeleteAddOns}{A Boolean value to indicate whether to delete all add-ons for the
instance.}
}
\description{
Deletes an Amazon Lightsail instance.

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_instance/} for full documentation.
}
\keyword{internal}
