% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_key_pair}
\alias{lightsail_delete_key_pair}
\title{Deletes the specified key pair by removing the public key from Amazon
Lightsail}
\usage{
lightsail_delete_key_pair(keyPairName, expectedFingerprint = NULL)
}
\arguments{
\item{keyPairName}{[required] The name of the key pair to delete.}

\item{expectedFingerprint}{The RSA fingerprint of the Lightsail default key pair to delete.

The \code{expectedFingerprint} parameter is required only when specifying to
delete a Lightsail default key pair.}
}
\description{
Deletes the specified key pair by removing the public key from Amazon Lightsail.

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_key_pair/} for full documentation.
}
\keyword{internal}
