% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_relational_database_snapshot}
\alias{lightsail_delete_relational_database_snapshot}
\title{Deletes a database snapshot in Amazon Lightsail}
\usage{
lightsail_delete_relational_database_snapshot(relationalDatabaseSnapshotName)
}
\arguments{
\item{relationalDatabaseSnapshotName}{[required] The name of the database snapshot that you are deleting.}
}
\description{
Deletes a database snapshot in Amazon Lightsail.

See \url{https://www.paws-r-sdk.com/docs/lightsail_delete_relational_database_snapshot/} for full documentation.
}
\keyword{internal}
