% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_detach_disk}
\alias{lightsail_detach_disk}
\title{Detaches a stopped block storage disk from a Lightsail instance}
\usage{
lightsail_detach_disk(diskName)
}
\arguments{
\item{diskName}{[required] The unique name of the disk you want to detach from your instance
(\code{my-disk}).}
}
\description{
Detaches a stopped block storage disk from a Lightsail instance. Make sure to unmount any file systems on the device within your operating system before stopping the instance and detaching the disk.

See \url{https://www.paws-r-sdk.com/docs/lightsail_detach_disk/} for full documentation.
}
\keyword{internal}
