% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_detach_instances_from_load_balancer}
\alias{lightsail_detach_instances_from_load_balancer}
\title{Detaches the specified instances from a Lightsail load balancer}
\usage{
lightsail_detach_instances_from_load_balancer(loadBalancerName, instanceNames)
}
\arguments{
\item{loadBalancerName}{[required] The name of the Lightsail load balancer.}

\item{instanceNames}{[required] An array of strings containing the names of the instances you want to
detach from the load balancer.}
}
\description{
Detaches the specified instances from a Lightsail load balancer.

See \url{https://www.paws-r-sdk.com/docs/lightsail_detach_instances_from_load_balancer/} for full documentation.
}
\keyword{internal}
