% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_detach_static_ip}
\alias{lightsail_detach_static_ip}
\title{Detaches a static IP from the Amazon Lightsail instance to which it is
attached}
\usage{
lightsail_detach_static_ip(staticIpName)
}
\arguments{
\item{staticIpName}{[required] The name of the static IP to detach from the instance.}
}
\description{
Detaches a static IP from the Amazon Lightsail instance to which it is attached.

See \url{https://www.paws-r-sdk.com/docs/lightsail_detach_static_ip/} for full documentation.
}
\keyword{internal}
