% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_disable_add_on}
\alias{lightsail_disable_add_on}
\title{Disables an add-on for an Amazon Lightsail resource}
\usage{
lightsail_disable_add_on(addOnType, resourceName)
}
\arguments{
\item{addOnType}{[required] The add-on type to disable.}

\item{resourceName}{[required] The name of the source resource for which to disable the add-on.}
}
\description{
Disables an add-on for an Amazon Lightsail resource. For more information, see the \href{https://docs.aws.amazon.com/en_us/lightsail/latest/userguide/amazon-lightsail-configuring-automatic-snapshots.html}{Amazon Lightsail Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/lightsail_disable_add_on/} for full documentation.
}
\keyword{internal}
