% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_download_default_key_pair}
\alias{lightsail_download_default_key_pair}
\title{Downloads the default SSH key pair from the user's account}
\usage{
lightsail_download_default_key_pair()
}
\value{
A list with the following syntax:\preformatted{list(
  publicKeyBase64 = "string",
  privateKeyBase64 = "string"
)
}
}
\description{
Downloads the default SSH key pair from the user's account.
}
\section{Request syntax}{
\preformatted{svc$download_default_key_pair()
}
}

\keyword{internal}
