% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_enable_add_on}
\alias{lightsail_enable_add_on}
\title{Enables or modifies an add-on for an Amazon Lightsail resource}
\usage{
lightsail_enable_add_on(resourceName, addOnRequest)
}
\arguments{
\item{resourceName}{[required] The name of the source resource for which to enable or modify the
add-on.}

\item{addOnRequest}{[required] An array of strings representing the add-on to enable or modify.}
}
\description{
Enables or modifies an add-on for an Amazon Lightsail resource. For more information, see the \href{https://docs.aws.amazon.com/en_us/lightsail/latest/userguide/amazon-lightsail-configuring-automatic-snapshots.html}{Amazon Lightsail Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/lightsail_enable_add_on/} for full documentation.
}
\keyword{internal}
