% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_export_snapshot}
\alias{lightsail_export_snapshot}
\title{Exports an Amazon Lightsail instance or block storage disk snapshot to
Amazon Elastic Compute Cloud (Amazon EC2)}
\usage{
lightsail_export_snapshot(sourceSnapshotName)
}
\arguments{
\item{sourceSnapshotName}{[required] The name of the instance or disk snapshot to be exported to Amazon EC2.}
}
\description{
Exports an Amazon Lightsail instance or block storage disk snapshot to Amazon Elastic Compute Cloud (Amazon EC2). This operation results in an export snapshot record that can be used with the \verb{create cloud formation stack} operation to create new Amazon EC2 instances.

See \url{https://www.paws-r-sdk.com/docs/lightsail_export_snapshot/} for full documentation.
}
\keyword{internal}
