% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_active_names}
\alias{lightsail_get_active_names}
\title{Returns the names of all active (not deleted) resources}
\usage{
lightsail_get_active_names(pageToken)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_active_names]{get_active_names}} request. If your
results are paginated, the response will return a next page token that
you can specify as the page token in a subsequent request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  activeNames = list(
    "string"
  ),
  nextPageToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns the names of all active (not deleted) resources.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_active_names(
  pageToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
