% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_auto_snapshots}
\alias{lightsail_get_auto_snapshots}
\title{Returns the available automatic snapshots for an instance or disk}
\usage{
lightsail_get_auto_snapshots(resourceName)
}
\arguments{
\item{resourceName}{[required] The name of the source instance or disk from which to get automatic
snapshot information.}
}
\description{
Returns the available automatic snapshots for an instance or disk. For more information, see the \href{https://docs.aws.amazon.com/en_us/lightsail/latest/userguide/amazon-lightsail-configuring-automatic-snapshots.html}{Amazon Lightsail Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_auto_snapshots/} for full documentation.
}
\keyword{internal}
