% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_cloud_formation_stack_records}
\alias{lightsail_get_cloud_formation_stack_records}
\title{Returns the CloudFormation stack record created as a result of the
create cloud formation stack operation}
\usage{
lightsail_get_cloud_formation_stack_records(pageToken = NULL)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial \code{GetClouFormationStackRecords}
request. If your results are paginated, the response will return a next
page token that you can specify as the page token in a subsequent
request.}
}
\description{
Returns the CloudFormation stack record created as a result of the \verb{create cloud formation stack} operation.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_cloud_formation_stack_records/} for full documentation.
}
\keyword{internal}
