% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_container_api_metadata}
\alias{lightsail_get_container_api_metadata}
\title{Returns information about Amazon Lightsail containers, such as the
current version of the Lightsail Control (lightsailctl) plugin}
\usage{
lightsail_get_container_api_metadata()
}
\value{
A list with the following syntax:\preformatted{list(
  metadata = list(
    list(
      "string"
    )
  )
)
}
}
\description{
Returns information about Amazon Lightsail containers, such as the
current version of the Lightsail Control (lightsailctl) plugin.
}
\section{Request syntax}{
\preformatted{svc$get_container_api_metadata()
}
}

\keyword{internal}
