% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_container_api_metadata}
\alias{lightsail_get_container_api_metadata}
\title{Returns information about Amazon Lightsail containers, such as the
current version of the Lightsail Control (lightsailctl) plugin}
\usage{
lightsail_get_container_api_metadata()
}
\description{
Returns information about Amazon Lightsail containers, such as the current version of the Lightsail Control (lightsailctl) plugin.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_container_api_metadata/} for full documentation.
}
\keyword{internal}
