% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_container_services}
\alias{lightsail_get_container_services}
\title{Returns information about one or more of your Amazon Lightsail container
services}
\usage{
lightsail_get_container_services(serviceName = NULL)
}
\arguments{
\item{serviceName}{The name of the container service for which to return information.

When omitted, the response includes all of your container services in
the Amazon Web Services Region where the request is made.}
}
\description{
Returns information about one or more of your Amazon Lightsail container services.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_container_services/} for full documentation.
}
\keyword{internal}
