% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_disk_snapshot}
\alias{lightsail_get_disk_snapshot}
\title{Returns information about a specific block storage disk snapshot}
\usage{
lightsail_get_disk_snapshot(diskSnapshotName)
}
\arguments{
\item{diskSnapshotName}{[required] The name of the disk snapshot (\code{my-disk-snapshot}).}
}
\description{
Returns information about a specific block storage disk snapshot.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_disk_snapshot/} for full documentation.
}
\keyword{internal}
