% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_disk_snapshots}
\alias{lightsail_get_disk_snapshots}
\title{Returns information about all block storage disk snapshots in your AWS
account and region}
\usage{
lightsail_get_disk_snapshots(pageToken = NULL)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_disk_snapshots]{get_disk_snapshots}} request. If your
results are paginated, the response will return a next page token that
you can specify as the page token in a subsequent request.}
}
\description{
Returns information about all block storage disk snapshots in your AWS account and region.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_disk_snapshots/} for full documentation.
}
\keyword{internal}
