% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_distribution_bundles}
\alias{lightsail_get_distribution_bundles}
\title{Returns the bundles that can be applied to your Amazon Lightsail content
delivery network (CDN) distributions}
\usage{
lightsail_get_distribution_bundles()
}
\description{
Returns the bundles that can be applied to your Amazon Lightsail content delivery network (CDN) distributions.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_distribution_bundles/} for full documentation.
}
\keyword{internal}
