% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_distributions}
\alias{lightsail_get_distributions}
\title{Returns information about one or more of your Amazon Lightsail content
delivery network (CDN) distributions}
\usage{
lightsail_get_distributions(distributionName = NULL, pageToken = NULL)
}
\arguments{
\item{distributionName}{The name of the distribution for which to return information.

When omitted, the response includes all of your distributions in the
Amazon Web Services Region where the request is made.}

\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_distributions]{get_distributions}} request. If your
results are paginated, the response will return a next page token that
you can specify as the page token in a subsequent request.}
}
\description{
Returns information about one or more of your Amazon Lightsail content delivery network (CDN) distributions.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_distributions/} for full documentation.
}
\keyword{internal}
