% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_domains}
\alias{lightsail_get_domains}
\title{Returns a list of all domains in the user's account}
\usage{
lightsail_get_domains(pageToken = NULL)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_domains]{get_domains}} request. If your results are
paginated, the response will return a next page token that you can
specify as the page token in a subsequent request.}
}
\description{
Returns a list of all domains in the user's account.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_domains/} for full documentation.
}
\keyword{internal}
