% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_export_snapshot_records}
\alias{lightsail_get_export_snapshot_records}
\title{Returns all export snapshot records created as a result of the export
snapshot operation}
\usage{
lightsail_get_export_snapshot_records(pageToken = NULL)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_export_snapshot_records]{get_export_snapshot_records}}
request. If your results are paginated, the response will return a next
page token that you can specify as the page token in a subsequent
request.}
}
\description{
Returns all export snapshot records created as a result of the \verb{export snapshot} operation.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_export_snapshot_records/} for full documentation.
}
\keyword{internal}
