% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_instance}
\alias{lightsail_get_instance}
\title{Returns information about a specific Amazon Lightsail instance, which is
a virtual private server}
\usage{
lightsail_get_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the instance.}
}
\description{
Returns information about a specific Amazon Lightsail instance, which is a virtual private server.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_instance/} for full documentation.
}
\keyword{internal}
