% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_instance_access_details}
\alias{lightsail_get_instance_access_details}
\title{Returns temporary SSH keys you can use to connect to a specific virtual
private server, or instance}
\usage{
lightsail_get_instance_access_details(instanceName, protocol = NULL)
}
\arguments{
\item{instanceName}{[required] The name of the instance to access.}

\item{protocol}{The protocol to use to connect to your instance. Defaults to \code{ssh}.}
}
\description{
Returns temporary SSH keys you can use to connect to a specific virtual private server, or \emph{instance}.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_instance_access_details/} for full documentation.
}
\keyword{internal}
