% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_instance_snapshot}
\alias{lightsail_get_instance_snapshot}
\title{Returns information about a specific instance snapshot}
\usage{
lightsail_get_instance_snapshot(instanceSnapshotName)
}
\arguments{
\item{instanceSnapshotName}{[required] The name of the snapshot for which you are requesting information.}
}
\description{
Returns information about a specific instance snapshot.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_instance_snapshot/} for full documentation.
}
\keyword{internal}
