% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_instance_state}
\alias{lightsail_get_instance_state}
\title{Returns the state of a specific instance}
\usage{
lightsail_get_instance_state(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the instance to get state information about.}
}
\description{
Returns the state of a specific instance. Works on one instance at a time.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_instance_state/} for full documentation.
}
\keyword{internal}
