% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_instances}
\alias{lightsail_get_instances}
\title{Returns information about all Amazon Lightsail virtual private servers,
or instances}
\usage{
lightsail_get_instances(pageToken = NULL)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_instances]{get_instances}} request. If your results are
paginated, the response will return a next page token that you can
specify as the page token in a subsequent request.}
}
\description{
Returns information about all Amazon Lightsail virtual private servers, or \emph{instances}.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_instances/} for full documentation.
}
\keyword{internal}
