% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_key_pair}
\alias{lightsail_get_key_pair}
\title{Returns information about a specific key pair}
\usage{
lightsail_get_key_pair(keyPairName)
}
\arguments{
\item{keyPairName}{[required] The name of the key pair for which you are requesting information.}
}
\description{
Returns information about a specific key pair.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_key_pair/} for full documentation.
}
\keyword{internal}
