% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_load_balancer_tls_certificates}
\alias{lightsail_get_load_balancer_tls_certificates}
\title{Returns information about the TLS certificates that are associated with
the specified Lightsail load balancer}
\usage{
lightsail_get_load_balancer_tls_certificates(loadBalancerName)
}
\arguments{
\item{loadBalancerName}{[required] The name of the load balancer you associated with your SSL/TLS
certificate.}
}
\description{
Returns information about the TLS certificates that are associated with the specified Lightsail load balancer.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_load_balancer_tls_certificates/} for full documentation.
}
\keyword{internal}
