% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_load_balancers}
\alias{lightsail_get_load_balancers}
\title{Returns information about all load balancers in an account}
\usage{
lightsail_get_load_balancers(pageToken = NULL)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_load_balancers]{get_load_balancers}} request. If your
results are paginated, the response will return a next page token that
you can specify as the page token in a subsequent request.}
}
\description{
Returns information about all load balancers in an account.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_load_balancers/} for full documentation.
}
\keyword{internal}
