% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_operations_for_resource}
\alias{lightsail_get_operations_for_resource}
\title{Gets operations for a specific resource (an instance or a static IP)}
\usage{
lightsail_get_operations_for_resource(resourceName, pageToken = NULL)
}
\arguments{
\item{resourceName}{[required] The name of the resource for which you are requesting information.}

\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_operations_for_resource]{get_operations_for_resource}}
request. If your results are paginated, the response will return a next
page token that you can specify as the page token in a subsequent
request.}
}
\description{
Gets operations for a specific resource (an instance or a static IP).

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_operations_for_resource/} for full documentation.
}
\keyword{internal}
