% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_database}
\alias{lightsail_get_relational_database}
\title{Returns information about a specific database in Amazon Lightsail}
\usage{
lightsail_get_relational_database(relationalDatabaseName)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of the database that you are looking up.}
}
\description{
Returns information about a specific database in Amazon Lightsail.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_relational_database/} for full documentation.
}
\keyword{internal}
