% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_database_blueprints}
\alias{lightsail_get_relational_database_blueprints}
\title{Returns a list of available database blueprints in Amazon Lightsail}
\usage{
lightsail_get_relational_database_blueprints(pageToken)
}
\arguments{
\item{pageToken}{A token used for advancing to a specific page of results for your
\verb{get relational database blueprints} request.}
}
\description{
Returns a list of available database blueprints in Amazon Lightsail. A
blueprint describes the major engine version of a database.
}
\details{
You can use a blueprint ID to create a new database that runs a specific
database engine.
}
\section{Request syntax}{
\preformatted{svc$get_relational_database_blueprints(
  pageToken = "string"
)
}
}

\keyword{internal}
