% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_database_log_streams}
\alias{lightsail_get_relational_database_log_streams}
\title{Returns a list of available log streams for a specific database in
Amazon Lightsail}
\usage{
lightsail_get_relational_database_log_streams(relationalDatabaseName)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database for which to get log streams.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  logStreams = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of available log streams for a specific database in
Amazon Lightsail.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_relational_database_log_streams(
  relationalDatabaseName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
