% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_database_log_streams}
\alias{lightsail_get_relational_database_log_streams}
\title{Returns a list of available log streams for a specific database in
Amazon Lightsail}
\usage{
lightsail_get_relational_database_log_streams(relationalDatabaseName)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database for which to get log streams.}
}
\description{
Returns a list of available log streams for a specific database in Amazon Lightsail.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_relational_database_log_streams/} for full documentation.
}
\keyword{internal}
