% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_database_parameters}
\alias{lightsail_get_relational_database_parameters}
\title{Returns all of the runtime parameters offered by the underlying database
software, or engine, for a specific database in Amazon Lightsail}
\usage{
lightsail_get_relational_database_parameters(
  relationalDatabaseName,
  pageToken = NULL
)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database for which to get parameters.}

\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial
\code{\link[=lightsail_get_relational_database_parameters]{get_relational_database_parameters}}
request. If your results are paginated, the response will return a next
page token that you can specify as the page token in a subsequent
request.}
}
\description{
Returns all of the runtime parameters offered by the underlying database software, or engine, for a specific database in Amazon Lightsail.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_relational_database_parameters/} for full documentation.
}
\keyword{internal}
