% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_database_snapshot}
\alias{lightsail_get_relational_database_snapshot}
\title{Returns information about a specific database snapshot in Amazon
Lightsail}
\usage{
lightsail_get_relational_database_snapshot(relationalDatabaseSnapshotName)
}
\arguments{
\item{relationalDatabaseSnapshotName}{[required] The name of the database snapshot for which to get information.}
}
\description{
Returns information about a specific database snapshot in Amazon Lightsail.

See \url{https://www.paws-r-sdk.com/docs/lightsail_get_relational_database_snapshot/} for full documentation.
}
\keyword{internal}
