% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_relational_databases}
\alias{lightsail_get_relational_databases}
\title{Returns information about all of your databases in Amazon Lightsail}
\usage{
lightsail_get_relational_databases(pageToken)
}
\arguments{
\item{pageToken}{The token to advance to the next page of results from your request.

To get a page token, perform an initial \code{GetRelationalDatabases}
request. If your results are paginated, the response will return a next
page token that you can specify as the page token in a subsequent
request.}
}
\description{
Returns information about all of your databases in Amazon Lightsail.
}
\section{Request syntax}{
\preformatted{svc$get_relational_databases(
  pageToken = "string"
)
}
}

\keyword{internal}
