% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_import_key_pair}
\alias{lightsail_import_key_pair}
\title{Imports a public SSH key from a specific key pair}
\usage{
lightsail_import_key_pair(keyPairName, publicKeyBase64)
}
\arguments{
\item{keyPairName}{[required] The name of the key pair for which you want to import the public key.}

\item{publicKeyBase64}{[required] A base64-encoded public key of the \code{ssh-rsa} type.}
}
\description{
Imports a public SSH key from a specific key pair.

See \url{https://www.paws-r-sdk.com/docs/lightsail_import_key_pair/} for full documentation.
}
\keyword{internal}
