% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_reboot_instance}
\alias{lightsail_reboot_instance}
\title{Restarts a specific instance}
\usage{
lightsail_reboot_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the instance to reboot.}
}
\description{
Restarts a specific instance.

See \url{https://www.paws-r-sdk.com/docs/lightsail_reboot_instance/} for full documentation.
}
\keyword{internal}
