% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_reset_distribution_cache}
\alias{lightsail_reset_distribution_cache}
\title{Deletes currently cached content from your Amazon Lightsail content
delivery network (CDN) distribution}
\usage{
lightsail_reset_distribution_cache(distributionName = NULL)
}
\arguments{
\item{distributionName}{The name of the distribution for which to reset cache.

Use the \code{\link[=lightsail_get_distributions]{get_distributions}} action to get
a list of distribution names that you can specify.}
}
\description{
Deletes currently cached content from your Amazon Lightsail content delivery network (CDN) distribution.

See \url{https://www.paws-r-sdk.com/docs/lightsail_reset_distribution_cache/} for full documentation.
}
\keyword{internal}
