% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_send_contact_method_verification}
\alias{lightsail_send_contact_method_verification}
\title{Sends a verification request to an email contact method to ensure it's
owned by the requester}
\usage{
lightsail_send_contact_method_verification(protocol)
}
\arguments{
\item{protocol}{[required] The protocol to verify, such as \code{Email} or \code{SMS} (text messaging).}
}
\description{
Sends a verification request to an email contact method to ensure it's owned by the requester. SMS contact methods don't need to be verified.

See \url{https://www.paws-r-sdk.com/docs/lightsail_send_contact_method_verification/} for full documentation.
}
\keyword{internal}
