% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_start_relational_database}
\alias{lightsail_start_relational_database}
\title{Starts a specific database from a stopped state in Amazon Lightsail}
\usage{
lightsail_start_relational_database(relationalDatabaseName)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database to start.}
}
\description{
Starts a specific database from a stopped state in Amazon Lightsail. To restart a database, use the \verb{reboot relational database} operation.

See \url{https://www.paws-r-sdk.com/docs/lightsail_start_relational_database/} for full documentation.
}
\keyword{internal}
