% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_stop_instance}
\alias{lightsail_stop_instance}
\title{Stops a specific Amazon Lightsail instance that is currently running}
\usage{
lightsail_stop_instance(instanceName, force = NULL)
}
\arguments{
\item{instanceName}{[required] The name of the instance (a virtual private server) to stop.}

\item{force}{When set to \code{True}, forces a Lightsail instance that is stuck in a
\code{stopping} state to stop.

Only use the \code{force} parameter if your instance is stuck in the
\code{stopping} state. In any other state, your instance should stop normally
without adding this parameter to your API request.}
}
\description{
Stops a specific Amazon Lightsail instance that is currently running.

See \url{https://www.paws-r-sdk.com/docs/lightsail_stop_instance/} for full documentation.
}
\keyword{internal}
