% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_untag_resource}
\alias{lightsail_untag_resource}
\title{Deletes the specified set of tag keys and their values from the
specified Amazon Lightsail resource}
\usage{
lightsail_untag_resource(resourceName, resourceArn = NULL, tagKeys)
}
\arguments{
\item{resourceName}{[required] The name of the resource from which you are removing a tag.}

\item{resourceArn}{The Amazon Resource Name (ARN) of the resource from which you want to
remove a tag.}

\item{tagKeys}{[required] The tag keys to delete from the specified resource.}
}
\description{
Deletes the specified set of tag keys and their values from the specified Amazon Lightsail resource.

See \url{https://www.paws-r-sdk.com/docs/lightsail_untag_resource/} for full documentation.
}
\keyword{internal}
