% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_update_domain_entry}
\alias{lightsail_update_domain_entry}
\title{Updates a domain recordset after it is created}
\usage{
lightsail_update_domain_entry(domainName, domainEntry)
}
\arguments{
\item{domainName}{[required] The name of the domain recordset to update.}

\item{domainEntry}{[required] An array of key-value pairs containing information about the domain
entry.}
}
\description{
Updates a domain recordset after it is created.

See \url{https://www.paws-r-sdk.com/docs/lightsail_update_domain_entry/} for full documentation.
}
\keyword{internal}
