% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_update_relational_database_parameters}
\alias{lightsail_update_relational_database_parameters}
\title{Allows the update of one or more parameters of a database in Amazon
Lightsail}
\usage{
lightsail_update_relational_database_parameters(
  relationalDatabaseName,
  parameters
)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database for which to update parameters.}

\item{parameters}{[required] The database parameters to update.}
}
\description{
Allows the update of one or more parameters of a database in Amazon Lightsail.

See \url{https://www.paws-r-sdk.com/docs/lightsail_update_relational_database_parameters/} for full documentation.
}
\keyword{internal}
