% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_delete_service}
\alias{proton_delete_service}
\title{Delete a service, with its instances and pipeline}
\usage{
proton_delete_service(name)
}
\arguments{
\item{name}{[required] The name of the service to delete.}
}
\description{
Delete a service, with its instances and pipeline.

See \url{https://www.paws-r-sdk.com/docs/proton_delete_service/} for full documentation.
}
\keyword{internal}
