% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_account_settings}
\alias{proton_get_account_settings}
\title{Get detail data for Proton account-wide settings}
\usage{
proton_get_account_settings()
}
\description{
Get detail data for Proton account-wide settings.

See \url{https://paws-r.github.io/docs/proton/get_account_settings.html} for full documentation.
}
\keyword{internal}
