% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_environment}
\alias{proton_get_environment}
\title{Get detailed data for an environment}
\usage{
proton_get_environment(name)
}
\arguments{
\item{name}{[required] The name of the environment that you want to get the detailed data for.}
}
\description{
Get detailed data for an environment.

See \url{https://www.paws-r-sdk.com/docs/proton_get_environment/} for full documentation.
}
\keyword{internal}
