% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_get_repository}
\alias{proton_get_repository}
\title{Get detail data for a linked repository}
\usage{
proton_get_repository(name, provider)
}
\arguments{
\item{name}{[required] The repository name, for example \code{myrepos/myrepo}.}

\item{provider}{[required] The repository provider.}
}
\description{
Get detail data for a linked repository.

See \url{https://www.paws-r-sdk.com/docs/proton_get_repository/} for full documentation.
}
\keyword{internal}
