% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_environment_outputs}
\alias{proton_list_environment_outputs}
\title{List the infrastructure as code outputs for your environment}
\usage{
proton_list_environment_outputs(
  deploymentId = NULL,
  environmentName,
  nextToken = NULL
)
}
\arguments{
\item{deploymentId}{The ID of the deployment whose outputs you want.}

\item{environmentName}{[required] The environment name.}

\item{nextToken}{A token that indicates the location of the next environment output in
the array of environment outputs, after the list of environment outputs
that was previously requested.}
}
\description{
List the infrastructure as code outputs for your environment.

See \url{https://www.paws-r-sdk.com/docs/proton_list_environment_outputs/} for full documentation.
}
\keyword{internal}
