% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_list_repositories}
\alias{proton_list_repositories}
\title{List linked repositories with detail data}
\usage{
proton_list_repositories(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of repositories to list.}

\item{nextToken}{A token that indicates the location of the next repository in the array
of repositories, after the list of repositories previously requested.}
}
\description{
List linked repositories with detail data.

See \url{https://www.paws-r-sdk.com/docs/proton_list_repositories/} for full documentation.
}
\keyword{internal}
