% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton_operations.R
\name{proton_update_service_sync_blocker}
\alias{proton_update_service_sync_blocker}
\title{Update the service sync blocker by resolving it}
\usage{
proton_update_service_sync_blocker(id, resolvedReason)
}
\arguments{
\item{id}{[required] The ID of the service sync blocker.}

\item{resolvedReason}{[required] The reason the service sync blocker was resolved.}
}
\description{
Update the service sync blocker by resolving it.

See \url{https://www.paws-r-sdk.com/docs/proton_update_service_sync_blocker/} for full documentation.
}
\keyword{internal}
