% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_create_application_version}
\alias{serverlessapplicationrepository_create_application_version}
\title{Creates an application version}
\usage{
serverlessapplicationrepository_create_application_version(
  ApplicationId,
  SemanticVersion,
  SourceCodeArchiveUrl = NULL,
  SourceCodeUrl = NULL,
  TemplateBody = NULL,
  TemplateUrl = NULL
)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}

\item{SemanticVersion}{[required] The semantic version of the new version.}

\item{SourceCodeArchiveUrl}{A link to the S3 object that contains the ZIP archive of the source code
for this version of your application.

Maximum size 50 MB}

\item{SourceCodeUrl}{A link to a public repository for the source code of your application,
for example the URL of a specific GitHub commit.}

\item{TemplateBody}{The raw packaged AWS SAM template of your application.}

\item{TemplateUrl}{A link to the packaged AWS SAM template of your application.}
}
\description{
Creates an application version.

See \url{https://www.paws-r-sdk.com/docs/serverlessapplicationrepository_create_application_version/} for full documentation.
}
\keyword{internal}
