% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_delete_application}
\alias{serverlessapplicationrepository_delete_application}
\title{Deletes the specified application}
\usage{
serverlessapplicationrepository_delete_application(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}
}
\description{
Deletes the specified application.

See \url{https://www.paws-r-sdk.com/docs/serverlessapplicationrepository_delete_application/} for full documentation.
}
\keyword{internal}
