% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_get_application}
\alias{serverlessapplicationrepository_get_application}
\title{Gets the specified application}
\usage{
serverlessapplicationrepository_get_application(
  ApplicationId,
  SemanticVersion = NULL
)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}

\item{SemanticVersion}{The semantic version of the application to get.}
}
\description{
Gets the specified application.

See \url{https://www.paws-r-sdk.com/docs/serverlessapplicationrepository_get_application/} for full documentation.
}
\keyword{internal}
