% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_get_application_policy}
\alias{serverlessapplicationrepository_get_application_policy}
\title{Retrieves the policy for the application}
\usage{
serverlessapplicationrepository_get_application_policy(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}
}
\value{
A list with the following syntax:\preformatted{list(
  Statements = list(
    list(
      Actions = list(
        "string"
      ),
      PrincipalOrgIDs = list(
        "string"
      ),
      Principals = list(
        "string"
      ),
      StatementId = "string"
    )
  )
)
}
}
\description{
Retrieves the policy for the application.
}
\section{Request syntax}{
\preformatted{svc$get_application_policy(
  ApplicationId = "string"
)
}
}

\keyword{internal}
