% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_get_cloud_formation_template}
\alias{serverlessapplicationrepository_get_cloud_formation_template}
\title{Gets the specified AWS CloudFormation template}
\usage{
serverlessapplicationrepository_get_cloud_formation_template(
  ApplicationId,
  TemplateId
)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}

\item{TemplateId}{[required] The UUID returned by CreateCloudFormationTemplate.

Pattern:
[0-9a-fA-F]\{8\}\\-[0-9a-fA-F]\{4\}\\-[0-9a-fA-F]\{4\}\\-[0-9a-fA-F]\{4\}\\-[0-9a-fA-F]\{12\}}
}
\description{
Gets the specified AWS CloudFormation template.

See \url{https://www.paws-r-sdk.com/docs/serverlessapplicationrepository_get_cloud_formation_template/} for full documentation.
}
\keyword{internal}
