% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_list_application_versions}
\alias{serverlessapplicationrepository_list_application_versions}
\title{Lists versions for the specified application}
\usage{
serverlessapplicationrepository_list_application_versions(
  ApplicationId,
  MaxItems = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}

\item{MaxItems}{The total number of items to return.}

\item{NextToken}{A token to specify where to start paginating.}
}
\description{
Lists versions for the specified application.

See \url{https://www.paws-r-sdk.com/docs/serverlessapplicationrepository_list_application_versions/} for full documentation.
}
\keyword{internal}
