% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_list_applications}
\alias{serverlessapplicationrepository_list_applications}
\title{Lists applications owned by the requester}
\usage{
serverlessapplicationrepository_list_applications(
  MaxItems = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxItems}{The total number of items to return.}

\item{NextToken}{A token to specify where to start paginating.}
}
\description{
Lists applications owned by the requester.

See \url{https://www.paws-r-sdk.com/docs/serverlessapplicationrepository_list_applications/} for full documentation.
}
\keyword{internal}
